% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-aggregate.R
\docType{methods}
\name{aggregate_replicates}
\alias{aggregate_replicates}
\alias{.aggregate_features}
\alias{.aggregate_replicates}
\alias{aggregate_replicates,bcbioSCDataSet-method}
\title{Aggregation Methods}
\usage{
aggregate_replicates(object, ...)

.aggregate_features(sparse, featureids)

.aggregate_replicates(sparse, cellids)

\S4method{aggregate_replicates}{bcbioSCDataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{sparse}{Sparse counts matrix.}

\item{featureids}{New feature IDs for the collapsed features.}

\item{cellids}{New cellular ids of the collapsed cells, one for each cell.}
}
\value{
\link{bcbioSCDataSet}.

Aggregated sparse matrix.

Aggregated sparse matrix.
}
\description{
Collapse features with the same feature id by summing them.

Pool cellular technical replicate counts.
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\keyword{internal}
