% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateCols-methods.R
\docType{methods}
\name{aggregateCols}
\alias{aggregateCols}
\alias{aggregateCols,SingleCellExperiment-method}
\title{Aggregate Columns}
\usage{
\S4method{aggregateCols}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Aggregate Columns
}
\examples{
object <- indrops_small
sampleNames(object)

# Define groupings factor as`aggregate` column in `colData()`.
glimpse(object$aggregate)

x <- aggregateCols(object)
show(x)
sampleNames(x)
}
\seealso{
Other Data Functions: \code{\link{cell2sample}},
  \code{\link{metricsPerSample}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Data Functions}
