% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-aggregateFeatures.R
\docType{methods}
\name{aggregateFeatures}
\alias{aggregateFeatures}
\alias{aggregateFeatures,bcbioSingleCellANY-method}
\title{Aggregate Features}
\usage{
aggregateFeatures(object, ...)

\S4method{aggregateFeatures}{bcbioSingleCellANY}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{bcbioSingleCell}.
}
\description{
Aggregate Features
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateReplicates}},
  \code{\link{filterCells}}, \code{\link{readMarkers}},
  \code{\link{subsetPerSample}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{Data Management Utilities}
\keyword{internal}
