% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-aggregateReplicates.R
\docType{methods}
\name{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates,bcbioSingleCell-method}
\alias{aggregateReplicates,dgCMatrix-method}
\title{Aggregate Replicates}
\usage{
\S4method{aggregateReplicates}{bcbioSingleCell}(object, cellids)

\S4method{aggregateReplicates}{dgCMatrix}(object, cellids)
}
\arguments{
\item{object}{Sparse counts matrix (e.g. \code{dgCMatrix}).}

\item{cellids}{Cellular barcode identifiers.}
}
\value{
\code{dgCMatrix}.
}
\description{
Aggregate Replicates
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateFeatures}},
  \code{\link{filterCells}}, \code{\link{readMarkers}},
  \code{\link{subsetPerSample}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{Data Management Utilities}
\keyword{internal}
