% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{all_generics}
\alias{all_generics}
\title{S4 Generics}
\arguments{
\item{object}{Object.}

\item{x}{Primary object.}

\item{y}{Secondary object.}

\item{value}{Object to assign.}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
S4 Generics
}
\keyword{internal}
