% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{all_markers_small}
\alias{all_markers_small}
\title{Minimal Seurat Markers Example}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 946 rows and 18 columns.}
\usage{
all_markers_small
}
\description{
Minimal Seurat Markers Example
}
\author{
Michael Steinbaugh
}
\keyword{internal}
