% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-annotable.R
\docType{methods}
\name{annotable}
\alias{annotable}
\alias{annotable,bcbioSingleCell-method}
\alias{annotable,seurat-method}
\title{Ensembl Annotations}
\usage{
\S4method{annotable}{bcbioSingleCell}(object)

\S4method{annotable}{seurat}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{data.frame}
}
\description{
Ensembl Annotations
}
\examples{
load(system.file(
    file.path("inst", "extdata", "bcb.rda"),
    package = "bcbioSingleCell"))
load(system.file(
    file.path("inst", "extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
annotable(bcb) \%>\% glimpse()

# seurat
annotable(seurat) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
