% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_metrics.R
\name{barcode_metrics}
\alias{barcode_metrics}
\title{Generate barcode metrics summary}
\usage{
barcode_metrics(run, sparsecounts)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run}

\item{sparsecounts}{Sparse counts matrix}
}
\value{
Data frame
}
\description{
Generate barcode metrics summary
}
\author{
Rory Kirchner

Michael Steinbaugh
}
