% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{barcode_plot}
\alias{barcode_plot}
\title{Plot a barcode histogram for a given sample}
\usage{
barcode_plot(filename, sample = NULL)
}
\arguments{
\item{filename}{path to a barcode histogram file}

\item{sample}{title for plot}
}
\description{
Plot a barcode histogram for a given sample
}
\author{
Rory Kirchner
}
