% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_plots.R
\name{barcode_plots}
\alias{barcode_plots}
\title{Plot all scRNA-Seq sample barcodes}
\usage{
barcode_plots(project = NULL)
}
\arguments{
\item{\code{bcbio-nextgen}}{project list}
}
\description{
Plot all scRNA-Seq sample barcodes
}
\author{
Michael Steinbaugh
}
