% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-barcodeRanks.R
\docType{methods}
\name{barcodeRanks}
\alias{barcodeRanks}
\alias{barcodeRanks}
\alias{barcodeRanks,SingleCellExperiment-method}
\title{Barcode Ranks}
\usage{
barcodeRanks(object, ...)

\S4method{barcodeRanks}{SingleCellExperiment}(object, lower = 100L,
  fitBounds = NULL, df = 20L)
}
\value{
\code{list}.
}
\description{
Return statistics to construct a barcode-rank plot. Also calculates the knee
and inflection points for further use.
}
\examples{
# SingleCellExperiment ====
x <- barcodeRanks(bcb_small)
names(x)
x[["knee"]]
x[["inflection"]]
}
\seealso{
Modified version of \code{DropletUtils::barcodeRanks()}.
}
\author{
Aaron Lun
}
