% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-barcodeRanksPerSample.R
\docType{methods}
\name{barcodeRanksPerSample}
\alias{barcodeRanksPerSample}
\alias{barcodeRanksPerSample}
\alias{barcodeRanksPerSample,SingleCellExperiment-method}
\title{Barcode Ranks per Sample}
\usage{
barcodeRanksPerSample(object, ...)

\S4method{barcodeRanksPerSample}{SingleCellExperiment}(object)
}
\description{
Barcode Ranks per Sample
}
\examples{
# bcbioSingleCell ====
x <- barcodeRanksPerSample(bcb_small)
names(x)
}
\author{
Michael Steinbaugh
}
