% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cellular_barcodes.R
\name{.read_cellular_barcode_file}
\alias{.read_cellular_barcode_file}
\title{Read sample cellular barcode summary file}
\usage{
.read_cellular_barcode_file(file_name)
}
\arguments{
\item{file_name}{File name.}

\item{sample_dirs}{Sample directories.}
}
\description{
Read sample cellular barcode summary file
}
\author{
Michael Steinbaugh
}
\keyword{internal}
