% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump.R
\name{kables}
\alias{kables}
\alias{set_names_dot}
\alias{set_names_snake}
\title{Utility functions from basejump package}
\usage{
kables(list, captions = NULL)

set_names_dot(data)

set_names_snake(data)
}
\arguments{
\item{data}{Data type that supports name assignments}
}
\description{
Utility functions from basejump package

Set names in dot notation

Set names in snake_case
}
\author{
Michael Steinbaugh
}
\keyword{internal}
