% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcb}
\alias{bcb}
\title{bcbio single-cell RNA-seq example data set}
\format{An object of class \code{bcbioSingleCell} with 50 rows and 100 columns.}
\usage{
bcb
}
\description{
Harvard inDrops v3 example data
}
\note{
Updated 2019-08-12.
}
\examples{
data(bcb)
bcb
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
