% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-bcbio.R
\docType{methods}
\name{bcbio,bcbioSCDataSet-method}
\alias{bcbio,bcbioSCDataSet-method}
\alias{bcbio<-,bcbioSCDataSet-method}
\title{\link{bcbioSCDataSet} object accessors}
\usage{
\S4method{bcbio}{bcbioSCDataSet}(object, type = "counts")

\S4method{bcbio}{bcbioSCDataSet}(object, type = "counts") <- value
}
\arguments{
\item{object}{\link{bcbioSCDataSet}.}

\item{type}{Type of count data to retrieve.}

\item{value}{An integer matrix or other object.}
}
\value{
Matrix/Object containing count data.
}
\description{
\link{bcbioSCDataSet} object accessors
}
