% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioSingleCell-package.R
\name{bcbioRNASeq-package}
\alias{bcbioRNASeq-package}
\title{bcbioSingleCell}
\description{
Import and analyze \href{http://bcbio-nextgen.readthedocs.io}{bcbio} single-cell
RNA-seq data.
}
\keyword{internal}
