% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioSCDataSet-class}
\alias{bcbioSCDataSet-class}
\alias{bcbioSCDataSet}
\title{bcbioSCDataSet}
\description{
This class will be deprecated in favor of \link{bcbioSingleCell} in a future
release.
}
\section{Slots}{

\describe{
\item{\code{callers}}{\link{SimpleList} containing additional bcbio run data with dimensions
that don't match the count matrix. This is currently used to store all
unfiltered cellular barcodes for quality control analysis.}
}}

\author{
Michael Steinbaugh
}
\keyword{internal}
