% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioSCFiltered-class}
\alias{bcbioSCFiltered-class}
\alias{bcbioSCFiltered}
\title{bcbioSCFiltered}
\description{
This class will be deprecated in favor of \link{bcbioSingleCell} in a future
release.
}
\keyword{internal}
