% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioSCSubset-class}
\alias{bcbioSCSubset-class}
\alias{bcbioSCSubset}
\title{bcbioSCSubset}
\description{
\link{bcbioSCSubset} is a subclass of \link{SummarizedExperiment} designed to store a
subset of single-cell RNA-seq counts filtered by quality control analysis.
}
\author{
Michael Steinbaugh
}
