% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{bcbioSCSubset-class}
\alias{bcbioSCSubset-class}
\alias{bcbioSCSubset}
\title{bcbioSCSubset}
\description{
\code{bcbioSCSubset} is a subclass of \link{ExpressionSet}.
}
\author{
Michael Steinbaugh
}
