% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioSingleCell-class}
\alias{bcbioSingleCell-class}
\title{bcbio single-cell RNA-seq data set}
\description{
\code{bcbioSingleCell} is an S4 class that extends \code{SingleCellExperiment}, and is
designed to store a bcbio single-cell RNA-seq analysis. This class contains
read counts saved as a sparse matrix (\code{sparseMatrix}), sample metadata, and
cell quality control metrics.
}
\seealso{
\code{bcbioSingleCell}.
}
\author{
Michael Steinbaugh
}
\concept{S4 classes}
