% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bcbioSingleCell-package}
\title{bcbioSingleCell}
\description{
Import and analyze \href{https://bcbio-nextgen.readthedocs.io/}{bcbio} single-cell
RNA-seq data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/bcbiosinglecell/}
  \item \url{https://github.com/hbc/bcbioSingleCell/}
  \item Report bugs at \url{https://github.com/hbc/bcbioSingleCell/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Authors:
\itemize{
  \item Rory Kirchner \email{roryk@alum.mit.edu} (\href{https://orcid.org/0000-0003-4814-5885}{ORCID})
  \item Mary Piper \email{mary.piper@gmail.com} (\href{https://orcid.org/0000-0003-2699-3840}{ORCID})
  \item Victor Barrera \email{barrera.vic@gmail.com} (\href{https://orcid.org/0000-0003-0590-4634}{ORCID})
  \item Shannan Ho Sui \email{shosui@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-6191-4709}{ORCID})
}

Other contributors:
\itemize{
  \item Harvard Chan Bioinformatics Core \email{bioinformatics@hsph.harvard.edu} [copyright holder, funder]
}

}
\keyword{internal}
