% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioSingleCell.R
\name{bcbioSingleCell}
\alias{bcbioSingleCell}
\title{bcbio single-cell RNA-seq data set}
\usage{
bcbioSingleCell(uploadDir, sampleMetadataFile = NULL, organism = NULL,
  ensemblRelease = NULL, genomeBuild = NULL, gffFile = NULL,
  transgeneNames = NULL, spikeNames = NULL,
  interestingGroups = "sampleName",
  BPPARAM = BiocParallel::SerialParam(), ...)
}
\arguments{
\item{uploadDir}{\code{character(1)}.
Final upload directory path.}

\item{sampleMetadataFile}{\code{character(1)}.
Sample metadata file path. CSV or TSV is preferred, but Excel worksheets
are also supported. Check the documentation for conventions and required
columns.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{ensemblRelease}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
Note: don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{gffFile}{\code{character(1)}.
GFF/GTF (General Feature Format) file.
Generally, we recommend using GTF (GFFv2) instead of GFFv3.}

\item{transgeneNames}{\code{character}.
Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).}

\item{spikeNames}{\code{character}.
Vector indicating which assay rows denote spike-in sequences (e.g. ERCCs).}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{BPPARAM}{
    An optional \code{\link{BiocParallelParam}} instance
    determining the parallel back-end to be used during evaluation, or a
    \code{list} of \code{BiocParallelParam} instances, to be applied in
    sequence for nested calls to \pkg{BiocParallel} functions.
  }

\item{...}{Additional arguments.}
}
\value{
\code{bcbioSingleCell}.
}
\description{
\code{bcbioSingleCell} is an S4 class that extends \code{SingleCellExperiment}, and is
designed to store a bcbio single-cell RNA-seq analysis. This class contains
read counts saved as a sparse matrix (\code{sparseMatrix}), sample metadata, and
cell quality control metrics.
}
\note{
Updated 2019-08-12.
}
\section{Remote data}{


When working in RStudio, we recommend connecting to the bcbio-nextgen run
directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}

\examples{
uploadDir <- system.file("extdata/indrops", package = "bcbioSingleCell")

x <- bcbioSingleCell(uploadDir)
print(x)

x <- bcbioSingleCell(
    uploadDir = uploadDir,
    sampleMetadataFile = file.path(uploadDir, "metadata.csv")
)
print(x)
}
\seealso{
\itemize{
\item \code{SingleCellExperiment::SingleCellExperiment()}.
\item \code{.S4methods(class = "bcbioSingleCell")}.
}
}
\author{
Michael Steinbaugh
}
