% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellularBarcodes-internal.R
\name{.bindCB}
\alias{.bindCB}
\title{Bind Cellular Barcodes}
\usage{
.bindCB(list)
}
\arguments{
\item{list}{List of cellular barcodes.}
}
\value{
\link{tibble}.
}
\description{
Bind Cellular Barcodes
}
\author{
Michael Steinbaugh
}
\keyword{internal}
