% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-calculate_metrics.R
\name{.calculate_metrics}
\alias{.calculate_metrics}
\title{Calculate cellular barcode metrics summary}
\usage{
.calculate_metrics(sparse_counts, annotable)
}
\arguments{
\item{sparse_counts}{Sparse counts matrix.}

\item{annotable}{Annotable.}
}
\value{
\link{matrix}.
}
\description{
This should only be performed during the initial run loading.
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\keyword{internal}
