% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMetrics.R
\name{calculateMetrics}
\alias{calculateMetrics}
\title{Calculate Cellular Barcode Quality Metrics}
\usage{
calculateMetrics(counts, rowRanges = NULL, prefilter = FALSE)
}
\arguments{
\item{counts}{\code{matrix} or \code{sparseMatrix}. Raw counts. Do not input
log-normalized or library scaled counts.}

\item{rowRanges}{\code{GRanges}.
Genomic ranges (e.g. genome annotations). Metadata describing the assay
rows.}

\item{prefilter}{\code{logical(1)}.
Apply prefiltering to remove zero count genes.}
}
\value{
\code{DataFrame}.
}
\description{
Calculate Cellular Barcode Quality Metrics
}
\examples{
data(indrops)
counts <- counts(indrops)
rowRanges <- rowRanges(indrops)

# Metrics using genome annotations (recommended).
x <- calculateMetrics(counts, rowRanges = rowRanges)
head(x)

# Minimal metrics (supported, but not recommended).
x <- calculateMetrics(counts, rowRanges = NULL)
head(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
