% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cbList.R
\name{.cbList}
\alias{.cbList}
\title{Cellular Barcodes List}
\usage{
.cbList(sampleDirs)
}
\arguments{
\item{sampleDirs}{Sample directories.}
}
\value{
\link{list}.
}
\description{
Cellular Barcodes List
}
\seealso{
Other Cellular Barcode Utilities: \code{\link{.bindCB}},
  \code{\link{.proportionalCB}}, \code{\link{.readCBFile}}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
