% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-cell2sample.R
\docType{methods}
\name{cell2sample}
\alias{cell2sample}
\alias{cell2sample}
\alias{cell2sample,character-method}
\alias{cell2sample,bcbioSingleCell-method}
\alias{cell2sample,seurat-method}
\title{Cell to Sample Mappings}
\usage{
cell2sample(object, ...)

\S4method{cell2sample}{character}(object, samples)

\S4method{cell2sample}{bcbioSingleCell}(object)

\S4method{cell2sample}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{samples}{Sample identifiers.}
}
\description{
Cell to Sample Mappings
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioSingleCell"))
load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
cell2sample(bcb) \%>\% glimpse()

# seurat
cell2sample(seurat) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
