% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_cycle_markers}
\alias{cell_cycle_markers}
\alias{cellCycleMarkers}
\title{Cell-Cycle Markers}
\format{An object of class \code{list} of length 3.}
\usage{
cell_cycle_markers
}
\description{
Cell-Cycle Markers
}
\examples{
names(cell_cycle_markers)
}
\seealso{
Other Marker Data: \code{\link{cell_type_markers}}
}
\author{
Michael Steinbaugh
}
\concept{Marker Data}
\keyword{datasets}
