% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_type_markers}
\alias{cell_type_markers}
\alias{cellTypeMarkers}
\title{Cell-Type Markers}
\format{An object of class \code{list} of length 3.}
\usage{
cell_type_markers
}
\description{
Cell-Type Markers
}
\examples{
names(cell_type_markers)
}
\seealso{
Other Marker Data: \code{\link{cell_cycle_markers}}
}
\author{
Michael Steinbaugh
}
\concept{Marker Data}
\keyword{datasets}
