% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellCountsPerCluster-methods.R
\docType{methods}
\name{cellCountsPerCluster}
\alias{cellCountsPerCluster}
\alias{cellCountsPerCluster,SingleCellExperiment-method}
\alias{cellCountsPerCluster,seurat-method}
\title{Cell Counts per Cluster}
\usage{
cellCountsPerCluster(object, ...)

\S4method{cellCountsPerCluster}{SingleCellExperiment}(object,
  interestingGroups)

\S4method{cellCountsPerCluster}{seurat}(object, interestingGroups)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}
}
\value{
\code{tibble} grouped by "\code{ident}" column, arranged by abundance.
}
\description{
Cell Counts per Cluster
}
\examples{
# seurat ====
cellCountsPerCluster(seurat_small)
}
