% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cellCycleMarkers}
\alias{cellCycleMarkers}
\title{Cell-Cycle Markers}
\format{An object of class \code{list} of length 2.}
\usage{
cellCycleMarkers
}
\description{
Cell-Cycle Markers
}
\author{
Michael Steinbaugh
}
\keyword{internal}
