% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cellCycleMarkersHsapiens}
\alias{cellCycleMarkersHsapiens}
\title{Cell Cycle Markers (Human)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 97 rows and 3 columns.}
\usage{
cellCycleMarkersHsapiens
}
\description{
Cell Cycle Markers (Human)
}
\author{
Michael Steinbaugh
}
\keyword{internal}
