% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellRanger-class}
\alias{CellRanger-class}
\title{10X Genomics Cell Ranger Data Set}
\description{
Extends \code{SingleCellExperiment}, with additional validity checks on the
\code{metadata()} slot.
}
\seealso{
\code{\link[=CellRanger]{CellRanger()}}.

Other S4 classes: \code{\link{bcbioSingleCell-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 classes}
