% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellranger-data}
\alias{cellranger-data}
\alias{cellranger}
\title{10X Genomics Cell Ranger Example Data Set}
\format{An object of class \code{CellRanger} with 50 rows and 100 columns.}
\usage{
cellranger
}
\description{
10X Genomics Cell Ranger Example Data Set
}
\examples{
data(cellranger)
cellranger
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
