% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cellranger_small}
\alias{cellranger_small}
\title{CellRanger Example}
\format{An object of class \code{SingleCellExperiment} with 500 rows and 500 columns.}
\usage{
cellranger_small
}
\description{
CellRanger Example
}
\seealso{
Other Minimal Example Data: \code{\link{all_markers_small}},
  \code{\link{bcb_small}},
  \code{\link{known_markers_small}},
  \code{\link{seurat_small}}
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
