% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-cellsPerCluster.R
\docType{methods}
\name{cellsPerCluster}
\alias{cellsPerCluster}
\alias{cellsPerCluster}
\alias{cellsPerCluster,seurat-method}
\title{Cell Counts per Cluster}
\usage{
cellsPerCluster(object, ...)

\S4method{cellsPerCluster}{seurat}(object, interestingGroups,
  return = c("list", "markdown"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in camel case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{return}{Return as \code{list} or \code{markdown}.}
}
\description{
Cell Counts per Cluster
}
\examples{
# seurat ====
x <- cellsPerCluster(seurat_small, return = "list")
names(x)

cellsPerCluster(seurat_small, return = "markdown")
}
