% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-cellTypesPerCluster.R
\docType{methods}
\name{cellTypesPerCluster}
\alias{cellTypesPerCluster}
\alias{cellTypesPerCluster,grouped_df-method}
\title{Cell Types per Cluster}
\usage{
cellTypesPerCluster(object, ...)

\S4method{cellTypesPerCluster}{grouped_df}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{tibble} grouped by cluster, containing the count (\code{n}) of
significant known makers per cell type.
}
\description{
Cell Types per Cluster
}
\examples{
\dontrun{
knownMarkers <- knownMarkersDetected(
    all = seuratAllMarkers,
    known = cellTypeMarkers)
cellTypesPerCluster(knownMarkers)
}
}
\author{
Michael Steinbaugh
}
