% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cellular_barcodes.R
\name{.cellular_barcodes}
\alias{.cellular_barcodes}
\title{Cellular barcode distributions}
\usage{
.cellular_barcodes(sample_dirs)
}
\description{
Cellular barcode distributions
}
\author{
Michael Steinbaugh
}
\keyword{internal}
