% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-colData.R
\docType{methods}
\name{colData}
\alias{colData}
\alias{colData,bcbioSingleCell-method}
\alias{colData,seurat-method}
\title{Column Data}
\usage{
\S4method{colData}{bcbioSingleCell}(x, return = c("data.frame", "DataFrame"))

\S4method{colData}{seurat}(x, return = c("data.frame", "DataFrame"))
}
\arguments{
\item{x}{Primary object.}

\item{return}{Return data as \code{data.frame} or \code{DataFrame}.}
}
\value{
Data describing the columns (cells).
}
\description{
Column Data
}
\examples{
# seurat ====
colData(pbmc_small)
}
\author{
Michael Steinbaugh
}
