% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_technical_replicates.R
\name{combine_technical_replicates}
\alias{combine_technical_replicates}
\title{Combine cellular technical replicates in a sparse matrix}
\usage{
combine_technical_replicates(sparse, cellids)
}
\arguments{
\item{sparse}{a sparse matrix of counts}

\item{cellids}{new cellular ids of the collapsed cells, one for each cell}
}
\value{
sparse matrix with technical replicates combined
}
\description{
Combine cellular technical replicates in a sparse matrix
}
\author{
Rory Kirchner
}
