% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-counts.R
\docType{methods}
\name{.counts}
\alias{.counts}
\alias{counts,bcbioSCDataSet-method}
\alias{counts,SummarizedExperiment-method}
\title{Counts accessor}
\usage{
\S4method{counts}{bcbioSCDataSet}(object, format = "dgCMatrix")

\S4method{counts}{SummarizedExperiment}(object, format = "dgCMatrix")
}
\arguments{
\item{object}{\link{bcbioSCDataSet}.}

\item{format}{Matrix format.}
}
\value{
Counts matrix.
}
\description{
Counts accessor
}
\author{
Michael Steinbaugh
}
