% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bcbio_skeleton.R
\name{create_bcbio_skeleton}
\alias{create_bcbio_skeleton}
\title{Create a skeleton bcbio project}
\usage{
create_bcbio_skeleton(base_dir = getwd(), organism)
}
\arguments{
\item{base_dir}{Base directory in which to create skeleton bcbio structure}

\item{organism}{Organism to use}
}
\value{
bcbio run skeleton
}
\description{
Create a skeleton bcbio project
}
\author{
Rory Kirchner

Michael Steinbaugh
}
