% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_orchestra.R
\name{detect_orchestra}
\alias{detect_orchestra}
\title{Detect if R is running on Orchestra HPC cluster}
\usage{
detect_orchestra()
}
\description{
Utility function to enable CPU intensive tasks
}
\author{
Michael Steinbaugh
}
