% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-detect.R
\name{.detect_pipeline}
\alias{.detect_pipeline}
\title{Detect scRNA-seq analysis pipeline}
\usage{
.detect_pipeline(sample_dirs)
}
\arguments{
\item{sample_dirs}{Sample directories.}
}
\value{
Pipeline string. Stops on detection failure.
}
\description{
Currently supports bcbio-nextgen and 10X Genomics Cell Ranger.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
