% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-detectOrganism.R
\docType{methods}
\name{detectOrganism}
\alias{detectOrganism}
\alias{detectOrganism,bcbioSingleCell-method}
\alias{detectOrganism,seurat-method}
\title{Detect Organism}
\usage{
\S4method{detectOrganism}{bcbioSingleCell}(object)

\S4method{detectOrganism}{seurat}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Detect Organism
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioSingleCell"))
load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
detectOrganism(bcb)

# seurat
detectOrganism(seurat)
}
\author{
Michael Steinbaugh
}
