% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectPipeline-internal.R
\name{.detectPipeline}
\alias{.detectPipeline}
\title{Detect scRNA-Seq Analysis Pipeline}
\usage{
.detectPipeline(uploadDir)
}
\arguments{
\item{uploadDir}{Final upload directory.}
}
\value{
Pipeline string. Stops on detection failure.
}
\description{
Currently supports bcbio-nextgen and 10X Genomics Cell Ranger.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
