% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{comp}
\alias{comp}
\alias{revcomp}
\title{DNA sequence utilites}
\usage{
comp(dna)

revcomp(dna)
}
\arguments{
\item{dna}{DNA sequence (ATGC nucleotides)}
}
\value{
Modified DNA sequence
}
\description{
Complement

Reverse complement
}
\examples{
comp("ATGCATGC")
revcomp("ATGCATGC")
}
\author{
Michael Steinbaugh
}
