% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-aggregateFeatures.R
\name{.aggregateFeatures}
\alias{.aggregateFeatures}
\title{Aggregate Features Constructor}
\usage{
.aggregateFeatures(sparse, featureids)
}
\arguments{
\item{sparse}{Sparse counts matrix (e.g. \link{dgCMatrix}).}

\item{featureids}{Feature identifiers (e.g. gene or transcript IDs).}
}
\description{
Aggregate Features Constructor
}
\author{
Rory Kirchner
}
