% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotTSNE.R
\name{.plotTSNE}
\alias{.plotTSNE}
\title{Plot t-SNE Constructor Function}
\usage{
.plotTSNE(object, interestingGroup = "ident", label = TRUE)
}
\arguments{
\item{object}{\link{data.frame} returned from \code{\link[=fetchTSNEExpressionData]{fetchTSNEExpressionData()}}.}
}
\description{
Plot t-SNE Constructor Function
}
