% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-quantileHeatmap.R
\name{.quantileBreaks}
\alias{.quantileBreaks}
\title{Create Breaks Based on Quantiles of the Data}
\usage{
.quantileBreaks(xs, n = 10L)
}
\arguments{
\item{xs}{Numeric vector.}

\item{n}{The number of breaks to create.}
}
\value{
A vector of \code{n} quantile breaks.
}
\description{
Create Breaks Based on Quantiles of the Data
}
