% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-sparseCounts.R
\name{.sparseCountsTx2Gene}
\alias{.sparseCountsTx2Gene}
\title{Transcript To Gene-Level Sparse Counts}
\usage{
.sparseCountsTx2Gene(txlevel, tx2gene)
}
\arguments{
\item{txlevel}{Transcript-level sparse counts matrix (\code{dgCMatrix}).}

\item{tx2gene}{Transcript to gene identifier mappings.}
}
\value{
\code{dgCMatrix}.
}
\description{
Transcript To Gene-Level Sparse Counts
}
\author{
Michael Steinbaugh
}
\keyword{internal}
