% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-download.R
\docType{methods}
\name{download}
\alias{download}
\alias{download}
\alias{download,missing-method}
\alias{download,character-method}
\title{File Downloads}
\usage{
download(object)

\S4method{download}{missing}()

\S4method{download}{character}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
No return.
}
\description{
If file isn't present, download latest version from the
\href{http://bioinformatics.sph.harvard.edu/}{HBC website}.
}
\details{
File download utility functions for RMarkdown knit reports.
}
\examples{
download()
download("setup.R")
}
\seealso{
Other `bcbio` Utilities: \code{\link{bcbio}},
  \code{\link{loadRun}}
}
\author{
Michael Steinbaugh
}
