% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/downloads-methods.R
\docType{methods}
\name{downloads}
\alias{downloads}
\alias{downloads}
\alias{downloads,missing-method}
\alias{downloads,character-method}
\title{File Downloads}
\usage{
downloads(object)

\S4method{downloads}{missing}()

\S4method{downloads}{character}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
No return.
}
\description{
If file isn't present, download latest version from the
\href{http://bioinformatics.sph.harvard.edu/}{HBC website}.
}
\details{
File download utility functions for RMarkdown knit reports.
}
\examples{
downloads()
downloads("setup.R")
}
\seealso{
Other `bcbio` Utilities: \code{\link{bcbio}},
  \code{\link{loadSinglecell}}
}
\author{
Michael Steinbaugh
}
