% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{examples}
\alias{examples}
\title{Example Data}
\format{An object of class \code{list} of length 1.}
\usage{
examples
}
\description{
Example Data
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
