% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-fetchGeneData.R
\docType{methods}
\name{fetchGeneData}
\alias{fetchGeneData}
\alias{fetchGeneData}
\alias{fetchGeneData,seurat-method}
\title{Fetch Gene Expression Data}
\usage{
fetchGeneData(object, ...)

\S4method{fetchGeneData}{seurat}(object, genes)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{Gene identifiers (matrix rownames).}
}
\value{
\code{matrix}.
}
\description{
Fetch Gene Expression Data
}
\examples{
# seurat ====
genes <- counts(seurat_small) \%>\% rownames() \%>\% head()
fetchGeneData(seurat_small, genes = genes) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
