% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-fetchTSNEExpressionData.R
\docType{methods}
\name{fetchTSNEExpressionData}
\alias{fetchTSNEExpressionData}
\alias{fetchTSNEExpressionData}
\alias{fetchTSNEExpressionData,seurat-method}
\title{Fetch t-SNE Locations, Cellular Metadata and Expression of Genes}
\usage{
fetchTSNEExpressionData(object, ...)

\S4method{fetchTSNEExpressionData}{seurat}(object, genes)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{Gene identifiers. Must match the rownames of the object.}
}
\value{
\code{data.frame} containing tSNE coordinates, sample metadata, and
aggregate marker expression values (\code{mean}, \code{median}, and \code{sum}).
}
\description{
This gets t-SNE locations, cellular metadata, expression of genes and
the geometric mean of the gene expression from a \link{seurat} object.
}
\examples{
# seurat ====
genes <- head(rownames(pbmc_small))
fetchTSNEExpressionData(pbmc_small, genes = genes) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{colData}},
  \code{\link{counts}}, \code{\link{fetchGeneData}},
  \code{\link{fetchPCAData}}, \code{\link{fetchTSNEData}},
  \code{\link{filterCells}}, \code{\link{gene2symbol}},
  \code{\link{interestingGroups}},
  \code{\link{readCellTypeMarkersFile}},
  \code{\link{rowData}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
