% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-filter.R
\docType{methods}
\name{filter,bcbioSCDataSet-method}
\alias{filter,bcbioSCDataSet-method}
\title{Filter cellular barcodes}
\usage{
\S4method{filter}{bcbioSCDataSet}(object, reads = 1000, min_genes = 500,
  max_genes = 5000, mito_ratio = 0.2, novelty = 0.8, show = TRUE)
}
\arguments{
\item{object}{\link{bcbioSCDataSet}.}

\item{reads}{Minimum number of total read counts per cell.}

\item{min_genes}{Minimum number of genes detected.}

\item{max_genes}{Maximum number of genes detected.}

\item{mito_ratio}{Maximum relative mitochondrial abundance (\code{0-1} scale).}

\item{novelty}{Minimum novelty score.}

\item{show}{Show summary statistics and plots.}
}
\value{
New \link{bcbioSCDataSet}.
}
\description{
Apply gene detection, mitochondrial abundance, and novelty score cutoffs to
cellular barcodes.
}
\author{
Michael Steinbaugh
}
