% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_barcodes.R
\name{filter_barcodes}
\alias{filter_barcodes}
\title{Filter cellular barcodes}
\usage{
filter_barcodes(metrics, min_genes = 500, max_genes = 5000,
  percent_mito = 0.2, novelty = 0, plot = TRUE)
}
\arguments{
\item{metrics}{Barcode metrics data frame}

\item{min_genes}{Minimum number of genes detected}

\item{max_genes}{Maximum number of genes detected}

\item{percent_mito}{Maximum mitochondrial pct abundance (\code{0-1} scale)}

\item{novelty}{Minimum novelty score}

\item{plot}{Print relevant plots}
}
\value{
Filtered metrics data frame
}
\description{
Apply gene detection, mitochondrial abundance, and novelty score cutoffs to
cellular barcodes.
}
\author{
Michael Steinbaugh
}
