% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cellular_barcodes.R
\name{filter_cellular_barcodes}
\alias{filter_cellular_barcodes}
\title{Filter cellular barcodes}
\usage{
filter_cellular_barcodes(run, total_counts = 1000, coding_counts = 1000,
  min_genes = 500, max_genes = 5000, mito_ratio = 0.2, novelty = 0.8,
  show = TRUE)
}
\arguments{
\item{run}{\link{bcbioSCDataSet}.}

\item{min_genes}{Minimum number of genes detected.}

\item{max_genes}{Maximum number of genes detected.}

\item{mito_ratio}{Maximum relative mitochondrial abundance (\code{0-1} scale).}

\item{novelty}{Minimum novelty score.}

\item{show}{Show summary statistics and plots.}

\item{reads}{Minimum number of read counts per cell.}
}
\value{
Filtered \link{bcbioSCDataSet}.
}
\description{
Apply gene detection, mitochondrial abundance, and novelty score cutoffs to
cellular barcodes.
}
\author{
Michael Steinbaugh
}
