% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCells-methods.R
\docType{methods}
\name{filterCells}
\alias{filterCells}
\alias{filterCells,SingleCellExperiment-method}
\title{Filter cells}
\usage{
\S4method{filterCells}{SingleCellExperiment}(object, nCells = Inf,
  minUMIs = 1L, maxUMIs = Inf, minGenes = 1L, maxGenes = Inf,
  minNovelty = 0L, maxMitoRatio = 1L, minCellsPerGene = 1L)
}
\arguments{
\item{object}{Object.}

\item{nCells}{\code{integer(1)}.
Expected number of cells per sample.}

\item{minUMIs}{\code{integer(1)}.
Minimum number of UMI disambiguated counts per cell.}

\item{maxUMIs}{\code{integer(1)}.
Maximum number of UMI disambiguated counts per cell.}

\item{minGenes}{\code{integer(1)}.
Minimum number of genes detected.}

\item{maxGenes}{\code{integer(1)}.
Maximum number of genes detected.}

\item{minNovelty}{\code{integer(1)} (\code{0}-\code{1}).
Minimum novelty score (log10 genes per UMI).}

\item{maxMitoRatio}{\code{integer(1)} (\code{0}-\code{1}).
Maximum relative mitochondrial abundance.}

\item{minCellsPerGene}{\code{integer(1)}.
Include genes with non-zero expression in at least this many cells.}
}
\value{
\code{SingleCellExperiment}.
Filtering information gets slotted into \code{metadata} as \code{filterCells} and
\code{filterParams}.
}
\description{
Filter cells
}
\details{
Apply gene detection, novelty score, and mitochondrial abundance cutoffs to
cellular barcodes. By default we recommend applying the same filtering cutoff
to all samples. The filtering parameters now support per-sample cutoffs,
defined using a named \code{numeric} vector. When matching per sample, be sure to
use the \code{sampleNames} return values (i.e. the \code{sampleName} column in
\code{sampleData}).
}
\examples{
data(indrops)

object <- indrops
show(object)

x <- filterCells(object)
show(x)
metadata(x)$filterParams

## Per sample cutoffs.
sampleNames(object)
x <- filterCells(
    object = object,
    minUMIs = c(rep_1 = 100)
)
show(x)
metadata(x)$filterParams
}
\author{
Michael Steinbaugh
}
