% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-flatFiles.R
\docType{methods}
\name{flatFiles}
\alias{flatFiles}
\alias{flatFiles,bcbioSingleCell-method}
\title{Get Flat Files from S4 Object}
\usage{
\S4method{flatFiles}{bcbioSingleCell}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{list}.

bcb <- examples[["bcb"]]
flatFiles(bcb) %>% names()

[["bcb"]: R:[%22bcb%22
}
\description{
Prepare an unstructured list of information saved in the S4 object,
for improved archival data storage.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
