% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-gene2symbol.R
\name{.gene2symbol}
\alias{.gene2symbol}
\title{Get gene symbols from Ensembl identifier rownames}
\usage{
.gene2symbol(object)
}
\arguments{
\item{object}{Object containing SummarizedExperiment with rownames set as
Ensembl identifiers.}
}
\value{
Ensembl gene symbol (\code{external_gene_name}) character vector.
}
\description{
Get gene symbols from Ensembl identifier rownames
}
\author{
Michael Steinbaugh
}
\keyword{internal}
