% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/reexports.R
\name{aggregate_replicates}
\alias{aggregate_replicates}
\alias{bcbio}
\alias{bcbio<-}
\alias{filter_barcodes}
\alias{metadata_table}
\alias{interesting_groups}
\alias{metrics}
\alias{sample_metadata}
\alias{select_samples}
\alias{top_barcodes}
\alias{counts}
\alias{design}
\title{S4 generics}
\usage{
aggregate_replicates(object, ...)

bcbio(object, ...)

bcbio(object, ...) <- value

filter_barcodes(object, ...)

metadata_table(object, ...)

interesting_groups(object)

metrics(object, ...)

sample_metadata(object, ...)

select_samples(object, ...)

top_barcodes(object, ...)

counts(object, ...)

design(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters.}

\item{value}{Value to assign.}
}
\description{
S4 generics
}
