% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_detected_plot.R
\name{genes_detected_plot}
\alias{genes_detected_plot}
\title{Genes detected plot}
\usage{
genes_detected_plot(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\description{
Genes detected plot
}
\author{
Rory Kirchner

Michael Steinbaugh
}
\keyword{plot}
\keyword{qc}
