% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-indrop_metadata.R
\name{.indrop_metadata}
\alias{.indrop_metadata}
\title{Prepare inDrop metadata}
\usage{
.indrop_metadata(metadata)
}
\arguments{
\item{metadata}{Metadata data frame.}
}
\value{
Metadata data frame.
}
\description{
Appends reverse complement sequences by matching against the \code{sequence}
column.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
