% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indrops}
\alias{indrops}
\title{inDrops Example Data Set}
\format{An object of class \code{bcbioSingleCell} with 50 rows and 100 columns.}
\usage{
indrops
}
\description{
inDrops Example Data Set
}
\examples{
data(indrops)
indrops
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
