% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indrops_small}
\alias{indrops_small}
\title{inDrops Example}
\format{An object of class \code{bcbioSingleCell} with 50 rows and 100 columns.}
\usage{
indrops_small
}
\description{
inDrops Example
}
\examples{
data(indrops_small)
indrops_small
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
