% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-inflectionPoint.R
\docType{methods}
\name{inflectionPoint}
\alias{inflectionPoint}
\alias{inflectionPoint}
\alias{inflectionPoint,SingleCellExperiment-method}
\alias{inflectionPoint,seurat-method}
\title{Calculate the Cellular Barcode Read Depth Inflection Point}
\usage{
inflectionPoint(object, ...)

\S4method{inflectionPoint}{SingleCellExperiment}(object)

\S4method{inflectionPoint}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Calculate the Cellular Barcode Read Depth Inflection Point
}
\examples{
# bcbioSingleCell ====
inflectionPoint(bcb_small)
inflectionPoint(cellranger_small)

# seurat ====
inflectionPoint(seurat_small)
inflectionPoint(Seurat::pbmc_small)
}
\seealso{
Modified version of \code{dropbead::estimateCellNumber()}.

Other Quality Control Functions: \code{\link{barcodeRanksPerSample}},
  \code{\link{barcodeRanks}}, \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
