% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_run}
\alias{check_run}
\alias{check_sparse}
\title{Check object integrity}
\usage{
check_run(run)

check_sparse(sparse)
}
\arguments{
\item{run}{\link{bcbioSCDataSet}.}

\item{sparse}{Sparse counts matrix.}
}
\description{
Check object integrity
}
\keyword{internal}
