% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-readCBFile.R
\name{.readCBFile}
\alias{.readCBFile}
\title{Read Cellular Barcode File}
\usage{
.readCBFile(file)
}
\arguments{
\item{file}{Cellular barcode TSV file.}
}
\value{
\link{tibble}.
}
\description{
Read Cellular Barcode File
}
\keyword{internal}
