% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-sparseCounts.R
\name{.stripTranscriptVersions}
\alias{.stripTranscriptVersions}
\title{Strip Transcript Versions}
\usage{
.stripTranscriptVersions(sparseCounts)
}
\arguments{
\item{sparseCounts}{Sparse counts matrix (\code{dgCMatrix}).}
}
\value{
\code{dgCMatrix}.
}
\description{
Strip Transcript Versions
}
\keyword{internal}
