% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_markers.R
\name{known_markers}
\alias{known_markers}
\title{Known markers data frame}
\usage{
known_markers(markers_file, genome_build)
}
\arguments{
\item{markers_file}{Gene markers file (CSV or Excel).}

\item{genome_build}{Genome build.}
}
\description{
Known markers data frame
}
