% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-markers.R
\docType{methods}
\name{known_markers_detected}
\alias{known_markers_detected}
\alias{known_markers_detected,data.frame,tbl_df-method}
\title{Known markers detected}
\usage{
\S4method{known_markers_detected}{data.frame,tbl_df}(x, y, show = TRUE)
}
\arguments{
\item{x}{All markers \link{data.frame} generated by
\code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}}.}

\item{y}{Known markers imported by \code{\link[=read_known_markers]{read_known_markers()}}.}

\item{show}{Show \link{kable}.}
}
\value{
\link{tibble}.
}
\description{
Known markers detected
}
