% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{known_markers_small}
\alias{known_markers_small}
\title{Minimal Known Markers Example}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 19 columns.}
\usage{
known_markers_small
}
\description{
Minimal Known Markers Example
}
\author{
Michael Steinbaugh
}
\keyword{internal}
