% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/knownMarkersDetected-methods.R
\docType{methods}
\name{knownMarkersDetected}
\alias{knownMarkersDetected}
\alias{knownMarkersDetected}
\alias{knownMarkersDetected,data.frame,tbl_df-method}
\title{Known Markers Detected}
\usage{
knownMarkersDetected(x, y, ...)

\S4method{knownMarkersDetected}{data.frame,tbl_df}(x, y, show = TRUE)
}
\arguments{
\item{x}{All markers \link{data.frame} generated by
\code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}}.}

\item{y}{Known markers imported by \code{\link[=readKnownMarkers]{readKnownMarkers()}}.}

\item{...}{Additional arguments.}

\item{show}{Show \link{kable}.}
}
\value{
\link{tibble}.
}
\description{
Known Markers Detected
}
\seealso{
Other Clustering Utilities: \code{\link{plotClusters}},
  \code{\link{plotKnownMarkers}},
  \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
