% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cellranger.R
\name{load_cellranger}
\alias{load_cellranger}
\title{Load Cell Ranger output}
\usage{
load_cellranger(upload_dir, metadata_file, interesting_groups)
}
\arguments{
\item{upload_dir}{Path to final upload directory.}

\item{metadata_file}{(\strong{Required}). Sample barcode metadata file.}

\item{interesting_groups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}
}
\value{
\link{bcbioSCDataSet}.
}
\description{
Read \href{https://www.10xgenomics.com/software/}{10x Genomics Chromium} cell
counts from \code{barcodes.tsv}, \code{genes.tsv}, and \code{matrix.mtx} files.
}
\author{
Michael Steinbaugh
}
