% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_csv_with_rownames.R
\name{load_csv_with_rownames}
\alias{load_csv_with_rownames}
\title{Load a CSV file with readr, setting a given column as the rownames}
\usage{
load_csv_with_rownames(filename, column)
}
\arguments{
\item{filename}{CSV to read}

\item{column}{column to make into the rownames}
}
\value{
dataframe
}
\description{
Load a CSV file with readr, setting a given column as the rownames
}
\author{
Rory Kirchner
}
\keyword{internal}
