% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{load_sparsecounts}
\alias{load_sparsecounts}
\title{Load a gzipped MatrixMart file, setting the row and column names.}
\usage{
load_sparsecounts(gzfile)
}
\arguments{
\item{gzfile}{gzipped MatrixMart file to load}
}
\value{
a sparse matrix of counts
}
\description{
Load a gzipped MatrixMart file, setting the row and column names.
}
\author{
Rory Kirchner
}
