% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCellRanger.R
\name{loadCellRanger}
\alias{loadCellRanger}
\title{Load 10X Genomics CellRanger Data}
\usage{
loadCellRanger(uploadDir, refDataDir, sampleMetadataFile,
  interestingGroups = "sampleName", prefilter = TRUE, ...)
}
\arguments{
\item{uploadDir}{Path to CellRanger output directory. This directory path
must contain \code{filtered_gene_bc_matrices/} as a child.}

\item{refDataDir}{Directory path to cellranger reference annotation data.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
Soft deprecation to new functions.
}
\description{
Read \href{https://www.10xgenomics.com/software/}{10x Genomics Chromium} cell
counts from \code{barcodes.tsv}, \code{genes.tsv}, and \code{matrix.mtx} files.
}
\details{
This function is a simplified version of \code{\link[=loadSingleCellRun]{loadSingleCellRun()}}
optimized for handling CellRanger output.
}
\author{
Michael Steinbaugh
}
