% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-loadRun.R
\docType{methods}
\name{loadRun}
\alias{loadRun}
\alias{loadRun}
\alias{loadRun,character-method}
\title{Load \code{bcbio} Single-Cell RNA-Seq Run}
\usage{
loadRun(object, ...)

\S4method{loadRun}{character}(object = "final", sampleMetadataFile,
  wellMetadataFile = NULL, interestingGroups = "sampleName",
  tx2gene = NULL, ...)
}
\arguments{
\item{object}{Path to final upload directory. This path is set when running
\code{bcbio_nextgen -w template}.}

\item{...}{Additional arguments, passed as metadata.}

\item{sampleMetadataFile}{\strong{Required}. Sample barcode metadata file.}

\item{wellMetadataFile}{\emph{Optional}. Well identifier metadata file.}

\item{interestingGroups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{tx2gene}{tx2gene \link{data.frame}, required when using a custom
transcriptome FASTA file. This argument will be removed in a future update.}
}
\value{
\link{bcbioSCDataSet}.
}
\description{
Load \code{bcbio} Single-Cell RNA-Seq Run
}
\details{
\itemize{
\item \strong{cellranger}: Read \href{https://www.10xgenomics.com/software/}{10x GenomicsChromium} cell counts from
\code{barcodes.tsv}, \code{genes.tsv}, and \code{matrix.mtx} files.
}
}
\note{
When working in RStudio, we recommend connecting to the bcbio-nextgen
run directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\seealso{
Other `bcbio` Utilities: \code{\link{bcbio}},
  \code{\link{download}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
