% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSingleCell.R
\name{loadSingleCell}
\alias{loadSingleCell}
\title{Load bcbio Single-Cell RNA-Seq Data}
\usage{
loadSingleCell(uploadDir, sampleMetadataFile = NULL,
  interestingGroups = "sampleName", organism, genomeBuild = NULL,
  ensemblRelease = NULL, gffFile = NULL, isSpike = NULL,
  prefilter = TRUE, ...)
}
\arguments{
\item{uploadDir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{sampleMetadataFile}{Sample barcode metadata file. Optional for runs
with demultiplixed index barcodes (e.g. SureCell), but otherwise required
for runs with multipliexed FASTQs containing multiple index barcodes (e.g.
inDrop). Consult the GitHub repo for examples and additional information.}

\item{interestingGroups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{organism}{Organism name. Use the full latin name (e.g. "Homo sapiens").
By default, this will be passed to AnnotationHub and ensembldb to obtain
pre-built annotations, unless \code{gffFile} is set.}

\item{genomeBuild}{\emph{Optional.} Ensembl genome build name (e.g. "GRCh38").
Included in database query to AnnotationHub, unless \code{gffFile} is set.}

\item{ensemblRelease}{\emph{Optional.} Ensembl release version. If \code{NULL},
defaults to current release available through AnnotationHub/ensembldb.
Only applies unless \code{gffFile} is set.}

\item{gffFile}{\emph{Optional, not recommended.} By default, we recommend leaving
this \code{NULL} for genomes that are supported on Ensembl. In this case, the
row annotations (\code{\link[=rowRanges]{rowRanges()}}) will be obtained automatically from Ensembl
using AnnotationHub and ensembldb. Internally, they are stored as genomic
ranges (\code{GRanges}). Additionally, these annotations are accessible as a
\code{data.frame} via the \code{\link[=rowData]{rowData()}} function. For a genome that is not
supported on Ensembl or AnnotationHub, a GFF/GTF (General Feature Format)
file is required. Generally, we recommend using a GTF (GFFv2) file here
over a GFF3 file, although both formats are supported. The function will
internally generate a \code{TxDb} containing transcript-to-gene mappings and
construct a \code{GRanges} object containing the genomic ranges (\code{\link[=rowRanges]{rowRanges()}}).}

\item{isSpike}{\emph{Optional.} Gene names corresponding to FASTA spike-in
sequences (e.g. ERCCs, EGFP, TDTOMATO).}

\item{prefilter}{Prefilter counts prior to quality control analysis.}

\item{...}{Additional arguments, to be stashed in the \code{\link[=metadata]{metadata()}} slot.}
}
\value{
\code{bcbioSingleCell}.
}
\description{
Load bcbio Single-Cell RNA-Seq Data
}
\note{
When working in RStudio, we recommend connecting to the bcbio-nextgen
run directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\examples{
extdataDir <- system.file("extdata", package = "bcbioSingleCell")
uploadDir <- file.path(extdataDir, "harvard_indrop_v3")
sampleMetadataFile <- file.path(extdataDir, "harvard_indrop_v3.xlsx")
loadSingleCell(
    uploadDir = uploadDir,
    sampleMetadataFile = sampleMetadataFile
)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
