% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrity_checks.R
\name{make_bcbio_skeleton}
\alias{make_bcbio_skeleton}
\title{create a skeleton bcbio project}
\usage{
make_bcbio_skeleton(basedir = getwd(), organism = "hsapiens")
}
\arguments{
\item{basedir}{base directory in which to create skeleton bcbio structure}
}
\value{
a bcbio run object
}
\description{
create a skeleton bcbio project
}
\author{
Rory Kirchner

Michael Steinbaugh
}
