% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metadata.R
\docType{methods}
\name{metadata}
\alias{metadata}
\alias{metadata,seurat-method}
\alias{metadata<-,seurat-method}
\title{Metadata}
\usage{
\S4method{metadata}{seurat}(x)

\S4method{metadata}{seurat}(x) <- value
}
\description{
Metadata
}
\examples{
# seurat ====
# Only works for objects created with bcbioSingleCell
metadata(pbmc_small)
names(metadata(seurat_small))

metadata(seurat_small)[["stash"]] <- "XXX"
metadata(seurat_small)[["stash"]]
}
\seealso{
\code{getMethod("metadata<-", "Annotated")}
}
