% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_table.R
\name{metadata_table}
\alias{metadata_table}
\title{Metadata table for knit report}
\usage{
metadata_table(run)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run}
}
\value{
kable
}
\description{
Metadata table for knit report
}
\author{
Michael Steinbaugh
}
