% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-metadataTable.R
\docType{methods}
\name{metadataTable}
\alias{metadataTable}
\alias{metadataTable}
\alias{metadataTable,bcbioSCDataSet-method}
\alias{metadataTable,bcbioSCSubset-method}
\title{Metadata Table}
\usage{
metadataTable(object, ...)

\S4method{metadataTable}{bcbioSCDataSet}(object, ...)

\S4method{metadataTable}{bcbioSCSubset}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments, passed to \code{\link[knitr:kable]{knitr::kable()}}.}
}
\value{
\link{kable}.
}
\description{
Table output for a knit report.
}
\author{
Michael Steinbaugh
}
