% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,bcbioSingleCell-method}
\alias{metrics,CellRanger-method}
\title{Metrics}
\usage{
\S4method{metrics}{bcbioSingleCell}(object, recalculate = FALSE)

\S4method{metrics}{CellRanger}(object, recalculate = FALSE)
}
\arguments{
\item{object}{Object.}

\item{recalculate}{\code{boolean}. Force recalculation, using primary
\code{\link[BiocGenerics:counts]{BiocGenerics::counts()}} matrix.}
}
\value{
\code{DataFrame}.

\itemize{
\item \code{"tibble"}: \code{grouped_df}. Grouped by \code{sampleID} column.
\item \code{"DataFrame"}: \code{DataFrame}. Rownames are identical to \code{\link[base:colnames]{base::colnames()}},
of the object, like \code{\link[SummarizedExperiment:colData]{SummarizedExperiment::colData()}}.
}
}
\description{
This function takes data stored in \code{SummarizedExperiment::colData()} and
consistently returns a tibble grouped by sample by default (\code{sampleID}).
}
\details{
\code{\link[=metrics]{metrics()}} always returns \code{sampleName} and \code{interestingGroups} columns, even
when these columns are not defined in \code{colData}. This is designed to
integrate with plotting functions that use ggplot2 internally.
}
\section{Methods (by class)}{

\itemize{
\item \code{SummarizedExperiment}: Metrics are sample level. \code{sampleID} column corresponds
to \code{\link[base:colnames]{base::colnames()}}.

\item \code{SingleCellExperiment}: Metrics are cell level. \code{cellID} column corresponds to
\code{\link[base:colnames]{base::colnames()}}. Tibble is returned grouped by sample
(\code{sampleID} column).
}
}

\examples{
data(indrops_small)
x <- metrics(indrops_small, recalculate = TRUE)
print(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
