% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-metricsPerSample.R
\docType{methods}
\name{metricsPerSample}
\alias{metricsPerSample}
\alias{metricsPerSample}
\alias{metricsPerSample,bcbioSingleCell-method}
\alias{metricsPerSample,seurat-method}
\title{Metrics per Sample}
\usage{
metricsPerSample(object, ...)

\S4method{metricsPerSample}{bcbioSingleCell}(object)

\S4method{metricsPerSample}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{tbl_df} containing summary statistics.
}
\description{
Metrics per Sample
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat_small.rda", package = "bcbioSingleCell"))

# bcbioSingleCell ====
metricsPerSample(bcb_small)

# seurat ====
metricsPerSample(seurat_small)
}
\seealso{
Other Quality Control Functions: \code{\link{calculateMetrics}},
  \code{\link{metrics}}, \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotQuantileHeatmap}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
