% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{new_seurat}
\alias{new_seurat}
\title{Create a new Seurat object}
\usage{
new_seurat(object, ...)
}
\arguments{
\item{object}{\link{SummarizedExperiment} containing filtered counts.}
}
\value{
\link{seurat}.
}
\description{
Create a new Seurat object
}
\author{
Michael Steinbaugh
}
