% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{object}{Object.}

\item{value}{Object to assign.}

\item{x}{Primary object.}

\item{y}{Secondary object.}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
Parameters
}
\keyword{internal}
