% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metrics.R
\name{.plot_cell_counts_barplot}
\alias{.plot_cell_counts_barplot}
\alias{plot_cell_counts}
\alias{plot_umis_per_cell}
\alias{plot_genes_detected}
\alias{plot_umis_vs_genes}
\alias{plot_mito_ratio}
\alias{plot_novelty}
\title{Cell metrics quality control plots}
\usage{
plot_cell_counts(object)

plot_umis_per_cell(object, min = 1000L)

plot_genes_detected(object, min = 500L)

plot_umis_vs_genes(object)

plot_mito_ratio(object, max = 0.1)

plot_novelty(object, min = 0.8)
}
\arguments{
\item{object}{Primary object.}

\item{min}{Recommended minimum value cutoff.}

\item{max}{Recommended maximum value cutoff.}
}
\value{
\link{ggplot}.
}
\description{
Novelty score means log10 genes detected per count.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
