% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode.R
\name{plot_barcode}
\alias{plot_barcode}
\alias{plot_barcodes}
\title{Plot sample barcodes}
\usage{
plot_barcode(file_name, sample = NULL)

plot_barcodes(run)
}
\arguments{
\item{file_name}{path to a barcode histogram file}

\item{sample}{title for plot}

\item{run}{\code{bcbio-nextgen} run}
}
\description{
Plot a barcode histogram for a given sample

Plot all single cell sample barcodes
}
\author{
Rory Kirchner

Michael Steinbaugh
}
\keyword{internal}
