% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcodes.R
\name{plot_barcodes}
\alias{plot_barcodes}
\title{Plot all scRNA-Seq sample barcodes}
\usage{
plot_barcodes(bcbio)
}
\arguments{
\item{bcbio}{bcbio run object}
}
\description{
Plot all scRNA-Seq sample barcodes
}
\author{
Michael Steinbaugh
}
