% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_cell_counts.R
\docType{methods}
\name{plot_cell_counts}
\alias{plot_cell_counts}
\alias{.plot_cell_counts}
\alias{plot_cell_counts,bcbioSCDataSet-method}
\alias{plot_cell_counts,SCSubset-method}
\title{Plot cell counts}
\usage{
\S4method{plot_cell_counts}{bcbioSCDataSet}(object)

\S4method{plot_cell_counts}{SCSubset}(object)
}
\arguments{
\item{object}{Primary object.}
}
\value{
\link{ggplot}.
}
\description{
Plot cell counts
}
\author{
Michael Steinbaugh, Rory Kirchner
}
