% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R,
%   R/methods-plot_cellular_barcodes.R
\docType{methods}
\name{plot_cellular_barcodes}
\alias{plot_cellular_barcodes}
\alias{.plot_cb_df}
\alias{.plot_cb_violin}
\alias{.plot_cb_histogram}
\alias{.plot_cb_histogram2}
\alias{plot_cellular_barcodes,bcbioSCDataSet-method}
\title{Plot cellular barcode distributions per sample}
\usage{
\S4method{plot_cellular_barcodes}{bcbioSCDataSet}(object)
}
\arguments{
\item{object}{\link{bcbioSCDataSet}.}
}
\value{
\link{ggplot}.
}
\description{
Plot cellular barcode distributions per sample
}
\details{
A violin plot is a comact display of a continuous distribution. It
is a blend of \code{\link[=geom_boxplot]{geom_boxplot()}} and \code{\link[=geom_density]{geom_density()}}: a violin plot is a
mirrored density plot displayed in the same way as a boxplot.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
