% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot clusters}
\usage{
plot_clusters(seurat, genes)
}
\arguments{
\item{seurat}{\link{seurat}.}

\item{genes}{Genes.}
}
\value{
\link{ggplot}.
}
\description{
Plot clusters
}
