% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes_detected.R
\name{plot_genes_detected_boxplot}
\alias{plot_genes_detected_boxplot}
\alias{plot_genes_detected_histogram}
\alias{plot_genes_detected}
\title{Genes detected plots}
\usage{
plot_genes_detected_boxplot(metrics)

plot_genes_detected_histogram(metrics)

plot_genes_detected(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\value{
ggplot2 object
}
\description{
Boxplot

Histogram

Show both plots
}
\author{
Rory Kirchner

Michael Steinbaugh
}
