% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes_detected.R
\name{plot_genes_detected_boxplot}
\alias{plot_genes_detected_boxplot}
\title{Genes detected boxplot}
\usage{
plot_genes_detected_boxplot(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\value{
ggplot2 object
}
\description{
Genes detected boxplot
}
\author{
Rory Kirchner

Michael Steinbaugh
}
