% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes_detected.R
\name{plot_genes_detected_histogram}
\alias{plot_genes_detected_histogram}
\title{Genes detected histogram}
\usage{
plot_genes_detected_histogram(metrics)
}
\arguments{
\item{metrics}{barcode metrics data frame}
}
\value{
ggplot2 object
}
\description{
Genes detected histogram
}
\author{
Rory Kirchner

Michael Steinbaugh
}
