% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_genes_per_cell.R
\docType{methods}
\name{plot_genes_per_cell}
\alias{plot_genes_per_cell}
\alias{.plot_genes_per_cell_boxplot}
\alias{.plot_genes_per_cell_histogram}
\alias{.plot_genes_per_cell}
\alias{plot_genes_per_cell,bcbioSCDataSet-method}
\alias{plot_genes_per_cell,SCSubset-method}
\title{Plot genes per cell}
\usage{
\S4method{plot_genes_per_cell}{bcbioSCDataSet}(object, min = 500L,
  max = NULL)

\S4method{plot_genes_per_cell}{SCSubset}(object)
}
\arguments{
\item{object}{Primary object.}

\item{min}{Recommended minimum value cutoff.}

\item{max}{Recommended maximum value cutoff.}
}
\value{
\link{ggplot}.
}
\description{
Plot genes per cell
}
\author{
Michael Steinbaugh, Rory Kirchner
}
