% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mito_counts.R
\name{plot_mito_counts}
\alias{plot_mito_counts}
\title{Total counts vs. genes detected plot}
\usage{
plot_mito_counts(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\description{
Total counts vs. genes detected plot
}
\author{
Rory Kirchner

Michael Steinbaugh
}
