% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mito_counts.R
\name{plot_mito_counts_boxplot}
\alias{plot_mito_counts_boxplot}
\title{Mitochondrial count scatterplot}
\usage{
plot_mito_counts_boxplot(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\value{
ggplot2 object
}
\description{
Mitochondrial count scatterplot
}
\author{
Rory Kirchner

Michael Steinbaugh
}
