% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mito_counts.R
\name{plot_mito_counts_histogram}
\alias{plot_mito_counts_histogram}
\title{Mitochondrial count histogram}
\usage{
plot_mito_counts_histogram(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\value{
ggplot2 object
}
\description{
Mitochondrial count histogram
}
\author{
Rory Kirchner

Michael Steinbaugh
}
