% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_mito_ratio.R
\docType{methods}
\name{plot_mito_ratio}
\alias{plot_mito_ratio}
\alias{.plot_mito_ratio_boxplot}
\alias{.plot_mito_ratio_histogram}
\alias{.plot_mito_ratio_scatterplot}
\alias{.plot_mito_ratio}
\alias{plot_mito_ratio,bcbioSCDataSet-method}
\alias{plot_mito_ratio,SCSubset-method}
\title{Plot mitochondrial transcript abundance}
\usage{
\S4method{plot_mito_ratio}{bcbioSCDataSet}(object, max = 0.1)

\S4method{plot_mito_ratio}{SCSubset}(object)
}
\arguments{
\item{object}{Primary object.}

\item{max}{Recommended maximum value cutoff.}
}
\value{
\link{ggplot}.
}
\description{
Plot mitochondrial transcript abundance
}
\author{
Michael Steinbaugh, Rory Kirchner
}
