% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_novelty.R
\name{plot_novelty}
\alias{plot_novelty}
\title{Novelty plot}
\usage{
plot_novelty(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\description{
log detected per count
}
\author{
Rory Kirchner

Michael Steinbaugh
}
