% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_novelty.R
\name{plot_novelty_histogram}
\alias{plot_novelty_histogram}
\title{Novelty histogram}
\usage{
plot_novelty_histogram(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\value{
ggplot2 object
}
\description{
log detected per count
}
\author{
Rory Kirchner

Michael Steinbaugh
}
