% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_total_cells.R
\name{plot_total_cells}
\alias{plot_total_cells}
\title{Total cells plot}
\usage{
plot_total_cells(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\value{
ggplot2 object
}
\description{
Total cells plot
}
\author{
Rory Kirchner

Michael Steinbaugh
}
