% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_total_counts.R
\name{plot_total_counts}
\alias{plot_total_counts}
\title{Total counts plots}
\usage{
plot_total_counts(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\description{
Total counts plots
}
\author{
Rory Kirchner

Michael Steinbaugh
}
