% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_total_vs_detected.R
\name{plot_total_vs_detected}
\alias{plot_total_vs_detected}
\title{Total counts vs. genes detected plot}
\usage{
plot_total_vs_detected(metrics, colorby = "sample")
}
\arguments{
\item{metrics}{Barcode metrics data frame}

\item{colorby}{column to color the points by}
}
\value{
ggplot2 object
}
\description{
Total counts vs. genes detected plot
}
\author{
Rory Kirchner

Michael Steinbaugh
}
