% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_umis_per_cell.R
\docType{methods}
\name{plot_umis_per_cell}
\alias{plot_umis_per_cell}
\alias{.plot_umis_per_cell_boxplot}
\alias{.plot_umis_per_cell_histogram}
\alias{.plot_umis_per_cell}
\alias{plot_umis_per_cell,bcbioSCDataSet-method}
\alias{plot_umis_per_cell,SCSubset-method}
\title{Plot UMIs per cell}
\usage{
\S4method{plot_umis_per_cell}{bcbioSCDataSet}(object, min = 1000L)

\S4method{plot_umis_per_cell}{SCSubset}(object)
}
\arguments{
\item{object}{Primary object.}

\item{min}{Recommended minimum value cutoff.}
}
\value{
\link{ggplot}.
}
\description{
Plot the universal molecular identifiers (UMIs) per cell.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
