% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_umis_vs_genes.R
\docType{methods}
\name{plot_umis_vs_genes}
\alias{plot_umis_vs_genes}
\alias{.plot_umis_vs_genes}
\alias{plot_umis_vs_genes,bcbioSCDataSet-method}
\alias{plot_umis_vs_genes,bcbioSCSubset-method}
\title{Plot UMI and Gene Correlation}
\usage{
plot_umis_vs_genes(object, ...)

\S4method{plot_umis_vs_genes}{bcbioSCDataSet}(object)

\S4method{plot_umis_vs_genes}{bcbioSCSubset}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{ggplot}.
}
\description{
Plot UMI and Gene Correlation
}
\author{
Michael Steinbaugh, Rory Kirchner
}
