% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metrics.R
\name{plot_zeros_vs_depth}
\alias{plot_zeros_vs_depth}
\title{plot percentage of zeros vs library depth}
\usage{
plot_zeros_vs_depth(object)
}
\description{
plot percentage of zeros vs library depth
}
\author{
Rory Kirchner
}
