% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotBarcodeRanks.R
\docType{methods}
\name{plotBarcodeRanks}
\alias{plotBarcodeRanks}
\alias{plotBarcodeRanks}
\alias{plotBarcodeRanks,SingleCellExperiment-method}
\alias{plotBarcodeRanks,seurat-method}
\title{Plot Barcode Ranks}
\usage{
plotBarcodeRanks(object, ...)

\S4method{plotBarcodeRanks}{SingleCellExperiment}(object, ...)

\S4method{plotBarcodeRanks}{seurat}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[DropletUtils:barcodeRanks]{DropletUtils::barcodeRanks()}}.}
}
\description{
Plot Barcode Ranks
}
\examples{
# bcbioSingleCell ====
plotBarcodeRanks(indrops_small)

# SingleCellExperiment ====
plotBarcodeRanks(cellranger_small)

# seurat ====
plotBarcodeRanks(seurat_small)
}
\seealso{
\code{\link[DropletUtils:barcodeRanks]{DropletUtils::barcodeRanks()}}.
}
\author{
Michael Steinbaugh
}
