% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCellCounts-methods.R
\docType{methods}
\name{plotCellCounts}
\alias{plotCellCounts}
\alias{plotCellCounts,SingleCellExperiment-method}
\title{Plot Cell Counts}
\usage{
plotCellCounts(object, ...)

\S4method{plotCellCounts}{SingleCellExperiment}(object,
  interestingGroups = NULL, fill = getOption("basejump.discrete.fill",
  NULL), title = "cell counts")
}
\value{
\code{ggplot}.
}
\description{
Plot Cell Counts
}
\examples{
data(indrops_small)
plotCellCounts(indrops_small)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
