% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCellTypesPerCluster.R
\name{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\title{Plot Cell Types per Cluster}
\usage{
plotCellTypesPerCluster(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{min}{Minimum number of marker genes per cell type.}

\item{max}{Maximum number of marker genes per cell type.}
}
\value{
\link{ggplot}.
}
\description{
Plot the geometric mean of the significant marker genes for every known cell
type (per unbiased cluster). Cell types with too few (\code{min} cutoff) or too
many (\code{max} cutoff) marker genes will be skipped.
}
\examples{
\dontrun{
cells <- cellTypesPerCluster(knownMakersDetected)
plotCellTypesPerCluster(cells)
}
}
\author{
Michael Steinbaugh
}
