% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCellularBarcodes-methods.R
\docType{methods}
\name{plotCellularBarcodes}
\alias{plotCellularBarcodes}
\alias{plotCellularBarcodes}
\alias{plotCellularBarcodes,bcbioSCDataSet-method}
\title{Plot Cellular Barcode Distributions per Sample}
\usage{
plotCellularBarcodes(object, ...)

\S4method{plotCellularBarcodes}{bcbioSCDataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{ggplot}.
}
\description{
Plot Cellular Barcode Distributions per Sample
}
\details{
A violin plot is a comact display of a continuous distribution. It
is a blend of \code{\link[=geom_boxplot]{geom_boxplot()}} and \code{\link[=geom_density]{geom_density()}}: a violin plot is a
mirrored density plot displayed in the same way as a boxplot.
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
