% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCumulativeReadsPerCell.R
\docType{methods}
\name{plotCumulativeReadsPerCell}
\alias{plotCumulativeReadsPerCell}
\alias{plotCumulativeReadsPerCell}
\alias{plotCumulativeReadsPerCell,bcbioSingleCell-method}
\alias{plotCumulativeReadsPerCell,seurat-method}
\alias{plotKnee}
\title{Plot Cumulative Frequency of Reads Per Cell}
\usage{
plotCumulativeReadsPerCell(object, ...)

\S4method{plotCumulativeReadsPerCell}{bcbioSingleCell}(object,
  maxCells = 100000L)

\S4method{plotCumulativeReadsPerCell}{seurat}(object, maxCells = 100000L)

plotKnee(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{maxCells}{The number of cells to consider as an upper bound and to
compute the inflection point.}
}
\value{
\code{ggplot}.
}
\description{
Cumulative frequency plot that helps visualize the inflection point that is
present in single-cell barcode distributions.
}
\examples{
# bcbioSingleCell ====
plotCumulativeReadsPerCell(bcb_small)
plotCumulativeReadsPerCell(cellranger_small)

# seurat ====
plotCumulativeReadsPerCell(seurat_small)
plotCumulativeReadsPerCell(pbmc_small)
}
\seealso{
Other Quality Control Functions: \code{\link{inflectionPoint}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotQuantileHeatmap}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
