% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCumulativeUMIsPerCell.R
\docType{methods}
\name{plotCumulativeUMIsPerCell}
\alias{plotCumulativeUMIsPerCell}
\alias{plotCumulativeUMIsPerCell}
\alias{plotCumulativeUMIsPerCell,bcbioSingleCell-method}
\alias{plotCumulativeUMIsPerCell,seurat-method}
\alias{plotKnee}
\title{Plot Cumulative Frequency of UMIs Per Cell}
\usage{
plotCumulativeUMIsPerCell(object, ...)

\S4method{plotCumulativeUMIsPerCell}{bcbioSingleCell}(object,
  trans = c("identity", "log10", "log2", "sqrt"))

\S4method{plotCumulativeUMIsPerCell}{seurat}(object, trans = c("identity",
  "log10", "log2", "sqrt"))

plotKnee(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{trans}{Name of the axis scale transformation to apply. See
\code{help("scale_x_continuous", "ggplot2")} for more information.}
}
\value{
\code{ggplot}.
}
\description{
Cumulative frequency plot that helps visualize the inflection point that is
present in single-cell barcode distributions.
}
\details{
These calculates apply to all cellular barcodes in the dataset and are not
grouped by sample.
}
\examples{
# bcbioSingleCell ====
plotCumulativeUMIsPerCell(bcb_small)
plotCumulativeUMIsPerCell(cellranger_small)

# seurat ====
plotCumulativeUMIsPerCell(seurat_small)
plotCumulativeUMIsPerCell(pbmc_small)
}
\seealso{
Other Quality Control Functions: \code{\link{inflectionPoint}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotQuantileHeatmap}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
