% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotFeature-methods.R
\docType{methods}
\name{plotFeatureTSNE}
\alias{plotFeatureTSNE}
\alias{plotFeatureUMAP}
\alias{plotFeature}
\alias{plotFeatureTSNE,SingleCellExperiment-method}
\alias{plotFeatureTSNE,seurat-method}
\alias{plotFeatureUMAP,SingleCellExperiment-method}
\alias{plotFeatureUMAP,seurat-method}
\title{Plot Feature}
\usage{
plotFeatureTSNE(object, ...)

plotFeatureUMAP(object, ...)

\S4method{plotFeatureTSNE}{SingleCellExperiment}(object, features,
  color = NULL, pointSize = getOption("pointSize", 0.75),
  pointAlpha = getOption("pointAlpha", 0.75), label = TRUE,
  labelSize = getOption("labelSize", 6L), dark = FALSE, grid = FALSE,
  legend = FALSE, aspectRatio = 1L)

\S4method{plotFeatureTSNE}{seurat}(object, features, color = NULL,
  pointSize = getOption("pointSize", 0.75),
  pointAlpha = getOption("pointAlpha", 0.75), label = TRUE,
  labelSize = getOption("labelSize", 6L), dark = FALSE, grid = FALSE,
  legend = FALSE, aspectRatio = 1L)

\S4method{plotFeatureUMAP}{SingleCellExperiment}(object, features,
  color = NULL, pointSize = getOption("pointSize", 0.75),
  pointAlpha = getOption("pointAlpha", 0.75), label = TRUE,
  labelSize = getOption("labelSize", 6L), dark = FALSE, grid = FALSE,
  legend = FALSE, aspectRatio = 1L)

\S4method{plotFeatureUMAP}{seurat}(object, features, color = NULL,
  pointSize = getOption("pointSize", 0.75),
  pointAlpha = getOption("pointAlpha", 0.75), label = TRUE,
  labelSize = getOption("labelSize", 6L), dark = FALSE, grid = FALSE,
  legend = FALSE, aspectRatio = 1L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{features}{\code{character}. Features to plot (e.g. gene expression, PC
scores, number of genes detected).}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot color scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{pointSize}{\code{scalar numeric}. Cell point size.}

\item{pointAlpha}{\code{scalar numeric} (\code{0}-\code{1}). Alpha transparency level.
Useful when there many cells in the dataset, and some cells can be masked.}

\item{label}{\code{boolean}. Overlay a cluster identitiy label on the plot.}

\item{labelSize}{\code{scalar integer}. Size of the text label.}

\item{dark}{\code{boolean}. Plot against a dark background using
\code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{grid}{\code{boolean}. Show major grid lines but hide axis lines.}

\item{legend}{\code{boolean}. Include plot legend.}

\item{aspectRatio}{\code{scalar integer}. Aspect ratio.}
}
\value{
\code{ggplot} or \code{list}.
}
\description{
Plot Feature
}
\examples{
# seurat ====
object <- seurat_small
features <- c("nUMI", "nGene", "PC1", "PC2")

# t-SNE
plotFeatureTSNE(object, features)

# UMAP
plotFeatureUMAP(object, features)
}
\seealso{
\code{\link[Seurat:FeaturePlot]{Seurat::FeaturePlot()}}.

Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotTSNE}},
  \code{\link{sanitizeMarkers}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Functions}
