% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotFeatures.R
\docType{methods}
\name{plotFeatures}
\alias{plotFeatures}
\alias{plotFeatures,seurat-method}
\title{Plot Features}
\usage{
plotFeatures(object, ...)

\S4method{plotFeatures}{seurat}(object, features)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{features}{Character vector of features (e.g. gene expression, PC
scores, number of genes detected).}
}
\value{
No return, only graphical output.
}
\description{
Plot Features
}
\author{
Michael Steinbaugh
}
