% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotFeatureTSNE.R
\docType{methods}
\name{plotFeatureTSNE}
\alias{plotFeatureTSNE}
\alias{plotFeatureTSNE,seurat-method}
\title{Plot Feature t-SNE}
\usage{
plotFeatureTSNE(object, ...)

\S4method{plotFeatureTSNE}{seurat}(object, feature,
  color = scale_color_gradient(low = "lightgray", high = "purple"),
  dark = FALSE, label = TRUE, legend = FALSE, returnAsList = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{feature}{Character vector of features (e.g. gene expression, PC
scores, number of genes detected).}

\item{returnAsList}{Return plots as a list.}
}
\value{
No return, only graphical output.
}
\description{
Plot Feature t-SNE
}
\seealso{
\code{\link[Seurat:FeaturePlot]{Seurat::FeaturePlot()}}.
}
\author{
Michael Steinbaugh
}
