% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene-methods.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,SingleCellExperiment-method}
\alias{plotGene,seurat-method}
\title{Plot Gene}
\usage{
\S4method{plotGene}{SingleCellExperiment}(object, genes, geom = c("dot",
  "violin"), legend = getOption("bcbio.legend", TRUE))

\S4method{plotGene}{seurat}(object, genes, geom = c("dot", "violin"),
  legend = getOption("bcbio.legend", TRUE))
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}. Gene identifiers. Must match the rownames of the
object.}

\item{geom}{\code{string}. Plot type. Uses \code{\link[=match.arg]{match.arg()}} to pick the type.
Currently supports "\code{dot}" and "\code{violin}".}

\item{legend}{\code{boolean}. Include plot legend.}
}
\value{
\code{ggplot}.
}
\description{
Visualize genes on a dot or violin plot.
}
\examples{
object <- indrops_small
genes <- head(rownames(object))
glimpse(genes)

# Dot
plotGene(object, genes = genes, geom = "dot")

# Violin
plotGene(object, genes = genes, geom = "violin")
}
\seealso{
\itemize{
\item \code{\link[=plotDot]{plotDot()}}, \code{\link[Seurat:DotPlot]{Seurat::DotPlot()}}.
\item \code{\link[=plotViolin]{plotViolin()}}, \code{\link[Seurat:VlnPlot]{Seurat::VlnPlot()}}.
\item \code{\link[Seurat:RidgePlot]{Seurat::RidgePlot()}}.
}

Other Gene Expression Functions: \code{\link{plotDot}},
  \code{\link{plotViolin}}
}
\author{
Michael Steinbaugh
}
\concept{Gene Expression Functions}
