% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGenesPerCell-methods.R
\docType{methods}
\name{plotGenesPerCell}
\alias{plotGenesPerCell}
\alias{plotGenesPerCell,SingleCellExperiment-method}
\title{Plot Genes per Cell}
\usage{
plotGenesPerCell(object, ...)

\S4method{plotGenesPerCell}{SingleCellExperiment}(object,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  interestingGroups = NULL, min = 0L, max = Inf, trans = "log2",
  fill = getOption("basejump.discrete.fill", NULL),
  title = "genes per cell")
}
\value{
\code{ggplot}.
}
\description{
Plot Genes per Cell
}
\examples{
data(indrops_small)
plotGenesPerCell(indrops_small)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
