% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotKnownMarkers.R
\docType{methods}
\name{plotKnownMarkers}
\alias{plotKnownMarkers}
\alias{plotKnownMarkers}
\alias{plotKnownMarkers,seurat,grouped_df-method}
\title{Plot Known Markers}
\usage{
plotKnownMarkers(object, knownMarkers, ...)

\S4method{plotKnownMarkers}{seurat,grouped_df}(object, knownMarkers,
  color = scale_color_viridis(option = "inferno"), dark = TRUE,
  headerLevel = NULL)
}
\arguments{
\item{object}{Object.}

\item{knownMarkers}{\code{\link[=knownMarkersDetected]{knownMarkersDetected()}} \link{tibble} grouped by cluster.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{color}{Color palette to use for points. Defaults to the viridis color
palette. Use \code{\link[ggplot2:scale_color_gradient]{ggplot2::scale_color_gradient()}} to easily define your own
low/high gradient.}

\item{dark}{Enable dark mode.}

\item{headerLevel}{Include a Markdown header for each gene.}
}
\value{
No value, only graphical output.
}
\description{
Plot Known Markers
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotMarkers}}, \code{\link{plotPCElbow}},
  \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
