% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-seurat-plotKnownMarkersDetected.R
\docType{methods}
\name{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected,seurat-method}
\title{Plot Known Markers Detected}
\usage{
plotKnownMarkersDetected(object, ...)

\S4method{plotKnownMarkersDetected}{seurat}(object, markers, headerLevel = 2L,
  ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{markers}{\code{grouped_df} of known marker genes.}

\item{headerLevel}{R Markdown header level.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Plot Known Markers Detected
}
\examples{
# seurat ====
# Let's plot the first known marker, as a quick example
plotKnownMarkersDetected(
    object = seurat_small,
    markers = known_markers_small[1L, , drop = FALSE]
)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotMarker}},
  \code{\link{plotPCA}}, \code{\link{plotPCElbow}},
  \code{\link{plotTSNE}}, \code{\link{plotTopMarkers}},
  \code{\link{plotUMAP}}, \code{\link{sanitizeMarkers}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
