% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMarkers.R
\docType{methods}
\name{plotMarkers}
\alias{plotMarkers}
\alias{plotMarkers,seurat-method}
\title{Plot Gene Markers}
\usage{
plotMarkers(object, ...)

\S4method{plotMarkers}{seurat}(object, symbols, headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{symbols}{Character vector of gene marker symbols.}

\item{headerLevel}{Include a Markdown header for each gene.}
}
\value{
No value, only graphical output.
}
\description{
Plot Gene Markers
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{pcCutoff}}, \code{\link{plotKnownMarkers}},
  \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Utilities}
