% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMitoRatio-methods.R
\docType{methods}
\name{plotMitoRatio}
\alias{plotMitoRatio}
\alias{plotMitoRatio,bcbioSingleCell-method}
\title{Plot mitochondrial transcript abundance}
\usage{
plotMitoRatio(object, ...)

\S4method{plotMitoRatio}{bcbioSingleCell}(object, geom = c("histogram",
  "ecdf", "violin", "ridgeline", "boxplot"), interestingGroups = NULL,
  max = 1L, fill = getOption(x = "acid.fill.discrete", default =
  acidplots::scale_fill_synesthesia_d()), trans = "sqrt",
  title = "Mito ratio")
}
\arguments{
\item{object}{Object.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{max}{\code{numeric(1)}.
Recommended maximum value cutoff.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{title}{\code{character(1)}.
Plot title.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot mitochondrial transcript abundance
}
\note{
Updated 2019-07-27.
}
\examples{
data(indrops)
plotMitoRatio(indrops)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
