% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMitoVsCoding-methods.R
\docType{methods}
\name{plotMitoVsCoding}
\alias{plotMitoVsCoding}
\alias{plotMitoVsCoding,SingleCellExperiment-method}
\title{Plot Mitochondrial vs. Coding Counts}
\usage{
plotMitoVsCoding(object, ...)

\S4method{plotMitoVsCoding}{SingleCellExperiment}(object,
  interestingGroups = NULL, trendline = FALSE,
  color = getOption("basejump.discrete.color", NULL), trans = "log2",
  title = "mito vs. coding")
}
\value{
\code{ggplot}.
}
\description{
Plot Mitochondrial vs. Coding Counts
}
\examples{
data(indrops_small)
plotMitoVsCoding(indrops_small)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
