% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotNovelty-methods.R
\docType{methods}
\name{plotNovelty}
\alias{plotNovelty}
\alias{plotNovelty,SingleCellExperiment-method}
\title{Plot Novelty Score}
\usage{
plotNovelty(object, ...)

\S4method{plotNovelty}{SingleCellExperiment}(object,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  interestingGroups = NULL, min = 0L,
  fill = getOption("basejump.discrete.fill", NULL), trans = "identity",
  title = "novelty : genes per UMI")
}
\value{
\code{ggplot}.
}
\description{
"Novelty" refers to log10 genes detected per count.
}
\examples{
data(indrops_small)
plotNovelty(indrops_small)
}
\author{
Michael Steinbaugh
}
