% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotPCA.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,seurat-method}
\title{Plot Principal Component Analysis (PCA)}
\usage{
\S4method{plotPCA}{seurat}(object, interestingGroups = "ident",
  color = scale_color_hue(), pointSize = 1, label = TRUE, labelSize = 6,
  dark = TRUE)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Interesting group to use for plot colors.}

\item{pointSize}{Cell point size.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{dark}{Enable dark mode.}
}
\value{
\link{ggplot}.
}
\description{
Generate a t-distributed stochastic neighbor embedding (t-SNE) plot.
}
\examples{
\dontrun{
plotTSNE(seurat)
}
}
\seealso{
\code{\link[=plotTSNE]{plotTSNE()}}.

Other PCA Utilities: \code{\link{fetchPCAData}}
}
\author{
Michael Steinbaugh
}
\concept{PCA Utilities}
