% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC-methods.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\alias{plotQC,SingleCellExperiment-method}
\title{Plot Quality Control}
\usage{
\S4method{plotQC}{SingleCellExperiment}(object, interestingGroups = NULL,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  headerLevel = 2L, legend = getOption("basejump.legend", FALSE),
  return = c("grid", "list", "markdown"))
}
\arguments{
\item{object}{Object.}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). R Markdown header level.}

\item{legend}{\code{boolean}. Include plot legend.}
}
\value{
\itemize{
\item \code{grid}: Graphical output.
\item \code{list}: \code{list} containing \code{ggplot} objects.
\item \code{markdown}: R Markdown report, with reports separated by headers.
}
}
\description{
Plot Quality Control
}
\examples{
data(indrops_small)
plotQC(indrops_small)
}
\author{
Michael Steinbaugh
}
