% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotQuantileHeatmap.R
\docType{methods}
\name{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap,bcbioSingleCell-method}
\alias{plotQuantileHeatmap,seurat-method}
\title{Plot Heatmap with Quantile Breaks}
\usage{
\S4method{plotQuantileHeatmap}{bcbioSingleCell}(object, n = 5L,
  annotationCol = NA, clusterCols = FALSE, clusterRows = FALSE,
  color = viridis, legendColor = viridis, title = NULL)

\S4method{plotQuantileHeatmap}{seurat}(object, n = 5L, annotationCol = NA,
  clusterCols = FALSE, clusterRows = FALSE, color = viridis,
  legendColor = viridis, title = NULL)
}
\arguments{
\item{object}{Object.}

\item{n}{The number of breaks to create.}

\item{annotationCol}{\emph{Optional.} \code{data.frame} that defines annotation
mappings for the columns.}

\item{clusterCols}{Logical determining if columns should be arranged with
hierarchical clustering. Alternatively, can define an \code{hclust} object.}

\item{clusterRows}{Logical determining if rows should be arranged with
hierarchical clustering. Alternatively, can define an \code{hclust} object.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{legendColor}{Colors to use for legend labels.}

\item{title}{Plot title.}
}
\description{
Plot Heatmap with Quantile Breaks
}
\details{
This is helpful for more usefully visualizing single cell data.
Ideas and code from: http://slowkow.com/notes/heatmap-tutorial/
}
\examples{
# bcbioSingleCell ====
plotQuantileHeatmap(bcb_small)
}
\author{
Rory Kirchner, Michael Steinbaugh
}
