% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotStressGenes.R
\docType{methods}
\name{plotStressGenes}
\alias{plotStressGenes}
\alias{plotStressGenes}
\alias{plotStressGenes,ANY-method}
\title{Plot Stress Genes}
\usage{
plotStressGenes(object, ...)

\S4method{plotStressGenes}{ANY}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{ggplot}.
}
\description{
Plot Stress Genes
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotKnownMarkers}},
  \code{\link{plotMarkers}}, \code{\link{plotPCElbow}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
