% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotTSNE.R
\docType{methods}
\name{plotTSNE}
\alias{plotTSNE}
\alias{plotTSNE}
\alias{plotTSNE,seurat-method}
\title{Plot t-SNE}
\usage{
plotTSNE(object, ...)

\S4method{plotTSNE}{seurat}(object, interestingGroups = "ident",
  pointsAsNumbers = FALSE, pointSize = 0.5, pointAlpha = 0.8,
  label = TRUE, labelSize = 6L, color = scale_color_hue(), dark = TRUE,
  title = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{Interesting group to use for plot colors.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{pointSize}{Cell point size.}

\item{pointAlpha}{Alpha transparency level. Useful when there many cells in
the dataset, and some cells can be masked.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{color}{Color palette to use for points.}

\item{dark}{Enable dark mode.}

\item{title}{Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Generate a t-distributed stochastic neighbor embedding (t-SNE) plot.
}
\examples{
# seurat ====
plotTSNE(pbmc_small)
}
\seealso{
Other t-SNE Utilities: \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{plotMarkerTSNE}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
