% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat-tsne.R
\name{PlotTsneExpressionData}
\alias{PlotTsneExpressionData}
\title{Plot fetched t-SNE expression data}
\usage{
PlotTsneExpressionData(df, colorpoints = "geomean")
}
\arguments{
\item{df}{a dataframe from FetchTsneExpressionData}

\item{colorpoints}{color points by geometric mean or expression of
individual gene}
}
\value{
ggplot2 t-SNE plot
}
\description{
Plot fetched t-SNE expression data
}
\examples{
\dontrun{
  dat = FetchTsneExpressionData(seurat, topGenes)
  # to make t-SNE plot colored by geometric mean of topGenes
  PlotTsneExpressionData(dat, colorpoints="geometricmean")
  # to make faceted t-SNE plot of each gene (looks good at up to 6 genes)
  PlotTsneExpression(dat, colorpoints="expression") +
    facet_wrap(~gene)
} @author Rory Kirchner
}
