% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotUMIsPerCell-methods.R
\docType{methods}
\name{plotUMIsPerCell}
\alias{plotUMIsPerCell}
\alias{plotUMIsPerCell,SingleCellExperiment-method}
\title{Plot UMIs per Cell}
\usage{
plotUMIsPerCell(object, ...)

\S4method{plotUMIsPerCell}{SingleCellExperiment}(object,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  interestingGroups = NULL, min = 0L, max = Inf, point = c("none",
  "inflection", "knee"), trans = "log10",
  color = getOption("basejump.discrete.color", NULL),
  fill = getOption("basejump.discrete.fill", NULL),
  title = "UMIs per cell")
}
\arguments{
\item{point}{\code{string}. Label either the "\code{knee}" or "\code{inflection}" points per
sample. To disable, use "\code{none}". Requires \code{geom = "ecdf"}.}
}
\value{
\code{ggplot}.
}
\description{
Plot the universal molecular identifiers (UMIs) per cell.
}
\examples{
data(indrops_small)
plotUMIsPerCell(indrops_small, geom = "violin")
plotUMIsPerCell(indrops_small, geom = "ridgeline")
plotUMIsPerCell(indrops_small, geom = "ecdf")
plotUMIsPerCell(indrops_small, geom = "histogram")
plotUMIsPerCell(indrops_small, geom = "boxplot")
}
\author{
Michael Steinbaugh, Rory Kirchner
}
