% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotUMIsVsGenes-methods.R
\docType{methods}
\name{plotUMIsVsGenes}
\alias{plotUMIsVsGenes}
\alias{plotUMIsVsGenes,SingleCellExperiment-method}
\title{Plot UMI and Gene Correlation}
\usage{
plotUMIsVsGenes(object, ...)

\S4method{plotUMIsVsGenes}{SingleCellExperiment}(object,
  interestingGroups = NULL, trendline = FALSE,
  color = getOption("basejump.discrete.color", NULL), trans = "log2",
  title = "UMIs vs. genes")
}
\value{
\code{ggplot}.
}
\description{
Plot UMI and Gene Correlation
}
\examples{
data(indrops_small)
plotUMIsVsGenes(indrops_small)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
