% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotViolin.R
\docType{methods}
\name{plotViolin}
\alias{plotViolin}
\alias{plotViolin,seurat-method}
\title{Plot Violin}
\usage{
\S4method{plotViolin}{seurat}(object, genes, format = "symbol",
  fill = viridis::scale_fill_viridis(discrete = TRUE), return = "grid",
  headerLevel = 2)
}
\arguments{
\item{object}{Object.}

\item{genes}{Gene identifiers to plot.}

\item{format}{Gene identifier format. Supports \code{ensgene} or \code{symbol}.}

\item{fill}{Fill color palette. Defaults to viridis.}

\item{headerLevel}{Markdown header level. Only applicable when
\code{return = "markdown"}.}
}
\value{
\itemize{
\item \code{grid}: \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}.
\item \code{list}: \link{list}.
\item \code{markdown}: Markdown output, with headers for each plot.
}
}
\description{
Plot Violin
}
\examples{

genes <- slot(seurat, "raw.data") \%>\% rownames() \%>\% .[1:2]
print(genes)

# grid
plotViolin(seurat, genes = genes, return = "grid")

# list
list <- plotViolin(seurat, genes = genes, return = "list")
glimpse(list)

# markdown
plotViolin(seurat, genes = genes, return = "markdown")
}
\author{
Michael Steinbaugh
}
