% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSingleCellTemplate.R
\name{prepareSingleCellTemplate}
\alias{prepareSingleCellTemplate}
\title{Prepare Single-Cell RNA-Seq RMarkdown Template}
\usage{
prepareSingleCellTemplate()
}
\value{
No value.
}
\description{
Prepare Single-Cell RNA-Seq RMarkdown Template
}
\examples{
files <- c(
    "_footer.Rmd",
    "_header.Rmd",
    "_output.yaml",
    "_setup.R",
    "bibliography.bib"
)
prepareSingleCellTemplate()
all(file.exists(files))
unlink(files)
}
\author{
Michael Steinbaugh
}
