% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectDir-internal.R
\name{.programs}
\alias{.programs}
\title{Read Program Versions}
\usage{
.programs(projectDir)
}
\arguments{
\item{projectDir}{Project directory.}
}
\value{
\link{tibble}.
}
\description{
Read Program Versions
}
\author{
Michael Steinbaugh
}
\keyword{internal}
