% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotCellularBarcodes.R
\name{.propTblFromDataSet}
\alias{.propTblFromDataSet}
\title{Proportional Cellular Barcodes}
\usage{
.propTblFromDataSet(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{tibble}.
}
\description{
Proportional Cellular Barcodes
}
\details{
Modified version of Klein Lab MATLAB code.
}
\seealso{
Other Cellular Barcode Utilities: \code{\link{.bindCB}},
  \code{\link{.cbList}}, \code{\link{.readCBFile}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\keyword{internal}
