% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{plot_total_cells}
\alias{plot_total_cells}
\alias{plot_total_counts_histogram}
\alias{plot_total_counts_boxplot}
\alias{plot_genes_detected_boxplot}
\alias{plot_genes_detected_histogram}
\alias{plot_total_vs_detected}
\alias{plot_mito_counts_histogram}
\alias{plot_mito_counts_boxplot}
\alias{plot_mito_counts_scatterplot}
\alias{plot_novelty_histogram}
\alias{plot_novelty_boxplot}
\alias{plot_total_counts}
\alias{plot_genes_detected}
\alias{plot_mito_counts}
\alias{plot_novelty}
\title{Quality control plots}
\usage{
plot_total_cells(metrics)

plot_total_counts_histogram(metrics)

plot_total_counts_boxplot(metrics)

plot_genes_detected_boxplot(metrics, min_genes = get("min_genes", envir =
  parent.frame()), max_genes = get("max_genes", envir = parent.frame()))

plot_genes_detected_histogram(metrics, min_genes = get("min_genes", envir =
  parent.frame()), max_genes = get("max_genes", envir = parent.frame()))

plot_total_vs_detected(metrics, colorby = "sample_name")

plot_mito_counts_histogram(metrics, percent_mito = get("percent_mito", envir =
  parent.frame()))

plot_mito_counts_boxplot(metrics, percent_mito = get("percent_mito", envir =
  parent.frame()))

plot_mito_counts_scatterplot(metrics, percent_mito = NULL)

plot_novelty_histogram(metrics, novelty = get("novelty", envir =
  parent.frame()))

plot_novelty_boxplot(metrics, novelty = get("novelty", envir =
  parent.frame()))

plot_total_counts(...)

plot_genes_detected(...)

plot_mito_counts(...)

plot_novelty(...)
}
\arguments{
\item{metrics}{Barcode metrics data frame}

\item{min_genes}{Recommended minimum gene count cutoff}

\item{max_genes}{Recommended maximum gene count cutoff}

\item{colorby}{column to color the points by}

\item{percent_mito}{Recommended mitochondrial percentage cutoff value}

\item{novelty}{Recommended novelty cutoff value}

\item{...}{Passthrough parameters}
}
\value{
ggplot2 object
}
\description{
Total cells barplot

Total counts histogram

Total counts boxplot

Genes detected boxplot

Genes detected histogram

Total counts vs. genes detected plot

Mitochondrial abundance histogram

Mitochondrial abundance boxplot

Mitochondrial abundance scatterplot

Novelty histogram (log10 genes detected per count)

Novelty boxplot (log10 genes detected per count)

Plot total counts (RMarkdown chunk wrapper)

Plot genes detected (RMarkdown chunk wrapper)

Plot mitochondrial counts (RMarkdown chunk wrapper)

Plot novelty (RMarkdown chunk wrapper)
}
\author{
Michael Steinbaugh

Rory Kirchner
}
