% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots_barcodes.R
\name{plot_barcode}
\alias{plot_barcode}
\alias{plot_barcodes}
\title{Sample barcode distribution quality control plots}
\usage{
plot_barcode(file_name, sample_name = NULL)

plot_barcodes(run)
}
\arguments{
\item{file_name}{Barcode histogram file}

\item{sample_name}{Sample name (title for plot)}

\item{run}{bcbio-nextgen run}
}
\description{
Plot an individual barcode histogram per sample file

Plot all single cell sample barcodes (RMarkdown chunk wrapper)
}
\author{
Michael Steinbaugh

Rory Kirchner
}
