% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metrics.R
\name{.plot_cell_counts_barplot}
\alias{.plot_cell_counts_barplot}
\alias{plot_cell_counts}
\alias{plot_read_counts}
\alias{plot_genes_detected}
\alias{plot_reads_vs_genes}
\alias{plot_mito_ratio}
\alias{plot_novelty}
\title{Cell metrics quality control plots}
\usage{
.plot_cell_counts_barplot(bcb)

plot_cell_counts(bcb)

plot_read_counts(bcb, min = 1000L)

plot_genes_detected(bcb, min = 500, max = NULL)

plot_reads_vs_genes(bcb, intgroup = NULL)

plot_mito_ratio(bcb, max = 0.2)

plot_novelty(bcb, min = 0.8)
}
\arguments{
\item{bcb}{\link{bcbioSCDataSet}.}

\item{min}{Recommended minimum value cutoff.}

\item{max}{Recommended maximum value cutoff.}

\item{intgroup}{Interesting group used to define colors.}
}
\value{
\link{ggplot}.
}
\description{
Novelty score means log10 genes detected per count.
}
\author{
Michael Steinbaugh

Rory Kirchner
}
