% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-heatmap.R
\name{quantileHeatmap}
\alias{quantileHeatmap}
\title{heatmap with quantile breaks}
\usage{
quantileHeatmap(mat, annotation = NA, clusterRows = TRUE,
  clusterCols = TRUE)
}
\arguments{
\item{mat}{matrix of data}

\item{annotation}{column annotation}

\item{clusterRows}{perform row clustering}

\item{clusterCols}{perform column clustering}
}
\value{
pheatmap object
}
\description{
heatmap with quantile breaks
}
\details{
This is helpful for more usefully visualizing single cell data.
ideas and code from: http://slowkow.com/notes/heatmap-tutorial/
}
\author{
Rory Kirchner
}
