% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{read_barcode_file}
\alias{read_barcode_file}
\title{Read sample cellular barcode summary file}
\usage{
read_barcode_file(file_name)
}
\arguments{
\item{file_name}{File name.}
}
\description{
Read sample cellular barcode summary file
}
\author{
Rory Kirchner
}
\keyword{internal}
