% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_counts.R
\name{read_counts}
\alias{read_counts}
\title{Read a MatrixMart file, setting the row and column names}
\usage{
read_counts(matrix_file)
}
\arguments{
\item{matrix_file}{MatrixMart file to read}

\item{strip_tx_version}{Strip Ensembl transcript version numbers}
}
\value{
a sparse matrix of counts
}
\description{
Note that for a bcbio run, this function will return transcript-level counts
}
\author{
Rory Kirchner

Michael Steinbaugh
}
\keyword{internal}
