% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-read_file.R
\name{.read_file}
\alias{.read_file}
\title{Read file}
\usage{
.read_file(file, ...)
}
\arguments{
\item{file}{File path.}

\item{...}{Additional parameters.}
}
\value{
Generally as a \link{tibble}, or a sparse matrix for \code{.mtx} files.
}
\description{
Supports automatic loading of standard \code{.csv}, \code{.mtx}, \code{.tsv}, and \code{.xlsx}
files. Also supports bcbio-nextgen pipeline-specific \code{.counts}, \code{.colnames},
and \code{.rownames} files.
}
\seealso{
\itemize{
\item \href{http://readr.tidyverse.org}{readr}.
\item \href{http://readxl.tidyverse.org}{readxl}.
\item \code{\link[=readMM]{readMM()}}: Read a MatrixMarket file.
}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
