% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_known_markers}
\alias{read_known_markers}
\title{Read Known Markers}
\usage{
read_known_markers(markers_file, genome_build, show = TRUE)
}
\arguments{
\item{markers_file}{Gene markers file (CSV or Excel).}

\item{genome_build}{Genome build.}

\item{show}{Show \link{kable}.}
}
\value{
\link{tibble}.
}
\description{
Read Known Markers
}
\author{
Michael Steinbaugh
}
