% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata}
\usage{
read_metadata(run, pattern = NULL, pattern_col = "sample_name")
}
\arguments{
\item{run}{bcbio-nextgen run.}

\item{pattern}{Apply grep pattern matching to samples.}

\item{pattern_col}{Column in data frame used for pattern subsetting.}
}
\value{
Metadata data frame.
}
\description{
Read metadata
}
\author{
Michael Steinbaugh
}
\keyword{internal}
