% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sample_barcodes_metadata.R
\name{read_sample_barcodes_metadata}
\alias{read_sample_barcodes_metadata}
\title{Read sample barcodes metadata}
\usage{
read_sample_barcodes_metadata(file)
}
\arguments{
\item{file}{Sample barcodes metadata file}
}
\value{
Data frame
}
\description{
Read sample barcodes metadata
}
\author{
Michael Steinbaugh
}
