% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sparsecounts.R
\name{read_sparsecounts}
\alias{read_sparsecounts}
\title{Read a MatrixMart file, setting the row and column names}
\usage{
read_sparsecounts(matfile)
}
\arguments{
\item{matfile}{MatrixMart file to read}
}
\value{
a sparse matrix of counts
}
\description{
Read a MatrixMart file, setting the row and column names
}
\author{
Rory Kirchner

Michael Steinbaugh
}
