% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sparsecounts_gz.R
\name{read_sparsecounts_gz}
\alias{read_sparsecounts_gz}
\title{Read a gzipped MatrixMart file, setting the row and column names}
\usage{
read_sparsecounts_gz(gzfile)
}
\arguments{
\item{gzfile}{gzipped MatrixMart file to read}
}
\value{
a sparse matrix of counts
}
\description{
Read a gzipped MatrixMart file, setting the row and column names
}
\author{
Rory Kirchner
}
\keyword{internal}
