% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCBFile-internal.R
\name{.readCBFile}
\alias{.readCBFile}
\title{Read Cellular Barcode File}
\usage{
.readCBFile(file)
}
\arguments{
\item{file}{Cellular barcode TSV file.}
}
\value{
\link{tibble}.
}
\description{
Read Cellular Barcode File
}
\seealso{
Other Cellular Barcode Utilities: \code{\link{.bindCB}},
  \code{\link{.cbList}}, \code{\link{.proportionalCB}}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
