% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCellRanger.R
\name{readCellRanger}
\alias{readCellRanger}
\title{Read 10X Genomics Cell Ranger Data}
\usage{
readCellRanger(uploadDir, format = c("mtx", "hdf5"), filtered = TRUE,
  organism = NULL, sampleMetadataFile = NULL,
  interestingGroups = "sampleName", refdataDir = NULL,
  gffFile = NULL, transgeneNames = NULL, spikeNames = NULL, ...)
}
\arguments{
\item{uploadDir}{\code{string}. Path to Cell Ranger output directory. This
directory path must contain \code{filtered_gene_bc_matrices*} as a child
directory.}

\item{format}{\code{string}. Output format, either MatrixMarket ("\code{mtx}") or HDF5
("\code{hdf5}").}

\item{filtered}{\code{boolean}. Use filtered (recommended) or raw counts. Note
that raw counts still contain only whitelisted cellular barcodes.}

\item{organism}{\code{string} or \code{NULL}. Organism name. Use the full Latin name
(e.g. "Homo sapiens"), since this will be input downstream to AnnotationHub
and ensembldb, unless \code{gffFile} is set. If left \code{NULL} (\emph{not recommended}),
the function call will skip loading gene-level annotations into
\code{\link[=rowRanges]{rowRanges()}}. This can be useful for poorly annotation genomes or
experiments involving multiple genomes.}

\item{sampleMetadataFile}{\code{string} or \code{NULL}. Sample barcode metadata file.
Optional for runs with demultiplixed index barcodes (e.g. SureCell) and can
be left \code{NULL}, but otherwise required for runs with multipliexed FASTQs
containing multiple index barcodes (e.g. inDrop).}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}

\item{refdataDir}{\code{string} or \code{NULL}. Directory path to Cell Ranger reference
annotation data.}

\item{gffFile}{\code{string} or \code{NULL}. \emph{Advanced use only; not recommended.} By
default, we recommend leaving this \code{NULL} for genomes that are supported on
Ensembl. In this case, the row annotations (\code{\link[=rowRanges]{rowRanges()}}) will be obtained
automatically from Ensembl by passing the \code{organism}, \code{genomeBuild}, and
\code{ensemblRelease} arguments to AnnotationHub and ensembldb. For a genome
that is not supported on Ensembl and/or AnnotationHub, a GFF/GTF (General
Feature Format) file is required. Generally, we recommend using a GTF
(GFFv2) file here over a GFF3 file if possible, although all GFF formats
are supported. The function will internally generate transcript-to-gene
mappings and construct a \code{GRanges} object containing the genomic ranges
(\code{\link[=rowRanges]{rowRanges()}}).}

\item{transgeneNames}{\code{character} or \code{NULL}. Vector indicating which
\code{\link[=assay]{assay()}} rows denote transgenes (e.g. EGFP, TDTOMATO).}

\item{spikeNames}{\code{character} or \code{NULL}. Vector indicating which \code{\link[=assay]{assay()}}
rows denote spike-in sequences (e.g. ERCCs).}

\item{...}{Additional arguments, to be stashed in the \code{\link[=metadata]{metadata()}} slot.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Read \href{https://www.10xgenomics.com/software/}{10x Genomics Chromium} cell
counts from \code{barcodes.tsv}, \code{genes.tsv}, and \code{matrix.mtx} files.
}
\section{Directory structure}{

Cell Ranger can vary in its output directory structure, but we're requiring a
single, consistent directory structure for all datasets, even those that only
contain a single sample:

\preformatted{
file.path(
    "<uploadDir>",
    "<sampleName>",
    "outs",
    "filtered_gene_bc_matrices*",
    "outs",
    "<genomeBuild>",
    "matrix.mtx"
)
}
}

\section{Sample metadata}{

A user-supplied sample metadata file defined by \code{sampleMetadataFile} is
required for multiplexed datasets. Otherwise this can be left \code{NULL}, and
minimal sample data will be used, based on the directory names.
}

\section{Reference data}{

We strongly recommend supplying the corresponding reference data required for
Cell Ranger with the \code{refdataDir} argument. It will convert the gene
annotations defined in the GTF file into a \code{GRanges} object, which get
slotted in \code{\link[=rowRanges]{rowRanges()}}. Otherwise, the function will attempt to use the
most current annotations available from Ensembl, and some gene IDs may not
match, due to deprecation in the current Ensembl release.
}

\examples{
uploadDir <- system.file("extdata/cellranger", package = "bcbioSingleCell")
x <- readCellRanger(uploadDir)
print(x)
}
\author{
Michael Steinbaugh
}
\concept{Read Functions}
