% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCellTypeMarkers.R
\name{readCellTypeMarkers}
\alias{readCellTypeMarkers}
\title{Read Cell Type Markers File}
\usage{
readCellTypeMarkers(file, gene2symbol)
}
\arguments{
\item{file}{\code{string}. Gene markers file path (CSV or Excel).}

\item{gene2symbol}{\code{data.frame}. Gene-to-symbol mappings. Columns must
contain \code{geneID} and \code{geneName}.}
}
\value{
\code{grouped_df}, grouped by "\code{cellType}" column.
}
\description{
Read Cell Type Markers File
}
\examples{
# Homo sapiens
file <- system.file(
    file.path("extdata", "cell_type_markers.csv"),
    package = "bcbioSingleCell"
)
gene2symbol <- makeGene2symbolFromEnsembl("Homo sapiens")
readCellTypeMarkers(file, gene2symbol = gene2symbol)
}
\seealso{
Other Read Functions: \code{\link{readCellRanger}}
}
\author{
Michael Steinbaugh
}
\concept{Read Functions}
