% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-readCellTypeMarkersFile.R
\docType{methods}
\name{readCellTypeMarkersFile}
\alias{readCellTypeMarkersFile}
\alias{readCellTypeMarkersFile}
\alias{readCellTypeMarkersFile,character-method}
\title{Read Cell Type Markers File}
\usage{
readCellTypeMarkersFile(object, ...)

\S4method{readCellTypeMarkersFile}{character}(object, gene2symbol)
}
\arguments{
\item{object}{Gene markers file (CSV or Excel).}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{gene2symbol}{Gene-to-symbol annotation \link{data.frame}.}
}
\value{
\link{tibble}, gropued by \code{cell} column.
}
\description{
Read Cell Type Markers File
}
\examples{
cellTypeMarkersFile <- system.file(
    file.path("extdata", "cellTypeMarkers.csv"),
    package = "bcbioSingleCell")
gene2symbol <- annotable("Mus musculus", format = "gene2symbol")
readCellTypeMarkersFile(cellTypeMarkersFile, gene2symbol = gene2symbol)
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateReplicates}},
  \code{\link{filterCells}}, \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
