% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/readKnownMarkers-methods.R
\docType{methods}
\name{readKnownMarkers}
\alias{readKnownMarkers}
\alias{readKnownMarkers,character-method}
\title{Read Known Markers}
\usage{
readKnownMarkers(object, ...)

\S4method{readKnownMarkers}{character}(object, genomeBuild, show = TRUE)
}
\arguments{
\item{object}{Gene markers file (CSV or Excel).}

\item{...}{Additional arguments.}

\item{genomeBuild}{Genome build.}

\item{show}{Show \link{kable}.}
}
\value{
\link{tibble}.
}
\description{
Read Known Markers
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}},
  \code{\link{filterCells}},
  \code{\link{interestingGroups}},
  \code{\link{sampleMetadata}}, \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh
}
