% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readMarkersFile.R
\docType{methods}
\name{readMarkersFile}
\alias{readMarkersFile}
\alias{readMarkersFile}
\alias{readMarkersFile,character-method}
\title{Read Known Markers File}
\usage{
readMarkersFile(object, ...)

\S4method{readMarkersFile}{character}(object, gene2symbol)
}
\arguments{
\item{object}{Gene markers file (CSV or Excel).}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{gene2symbol}{Gene-to-symbol annotation \link{data.frame}.}
}
\value{
\link{tibble}.
}
\description{
Read Known Markers File
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateReplicates}},
  \code{\link{filterCells}}, \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
