% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSampleMetadataFile-internal.R
\name{.readSampleMetadataFile}
\alias{.readSampleMetadataFile}
\title{Read Sample Metadata File}
\usage{
.readSampleMetadataFile(file, sampleDirs, pipeline)
}
\arguments{
\item{file}{Sample metadata file.}

\item{sampleDirs}{Sample directory paths.}

\item{pipeline}{Pipeline used to process the samples.}

\item{uniqueNames}{Ensure sample names are unique.}
}
\value{
\link{tibble}.
}
\description{
Appends reverse complement sequences by matching against the \code{sequence}
column.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
