% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseCounts-internal.R
\name{.readSparseCounts}
\alias{.readSparseCounts}
\title{Read Sparse Counts}
\usage{
.readSparseCounts(sampleDir, pipeline = "bcbio")
}
\arguments{
\item{sampleDir}{Named character vector of sample directory containing the
MatrixMart file.}

\item{pipeline}{Pipeline used to generate the MatrixMarket file. Defaults to
bcbio-nextgen (\code{bcbio}). Also supports 10X Chromium CellRanger
(\code{cellranger}).}
}
\value{
\code{dgCMatrix}.
}
\description{
Read single-cell RNA-seq counts saved as a
\href{https://people.sc.fsu.edu/~jburkardt/data/mm/mm.html}{MatrixMarket} file
into a sparse counts matrix (\code{dgCMatrix}). Transcript (\link{rownames}) and
molecular barcodes (\link{colnames}) identifiers are set from the corresponding
dependency files automatically.
}
\details{
This function supports automatic handling of compressed matrix files.
}
\note{
bcbio-nextgen outputs counts at transcript level. 10X Chromium
CellRanger outputs counts at gene level.
}
\seealso{
\code{help("dgCMatrix-class")}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\keyword{internal}
