% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/metrics-methods.R,
%   R/plotQC-methods.R, R/show-methods.R, R/updateObject-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plotGene}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump.generics}{\code{\link[basejump.generics]{metrics}}, \code{\link[basejump.generics]{plotQC}}}

  \item{basejump.plots}{\code{\link[basejump.plots]{plotGene}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{updateObject}}}

  \item{methods}{\code{\link[methods]{show}}}
}}

