% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-rowData.R
\docType{methods}
\name{rowData}
\alias{rowData}
\alias{rowData,bcbioSingleCell-method}
\alias{rowData,seurat-method}
\title{Row Data}
\usage{
\S4method{rowData}{bcbioSingleCell}(x, return = c("data.frame", "DataFrame"))

\S4method{rowData}{seurat}(x, return = c("data.frame", "DataFrame"))
}
\arguments{
\item{x}{Primary object.}
}
\value{
Return data as \code{data.frame}, or \code{DataFrame}.
}
\description{
Row Data
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat.rda", package = "bcbioSingleCell"))

# bcbioSingleCell ====
rowData(bcb) \%>\% glimpse()

# seurat ====
rowData(pbmc_small) \%>\% glimpse()
rowData(seurat) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
