% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-sample_dirs.R
\name{.sample_dirs}
\alias{.sample_dirs}
\title{Detect sample directories}
\usage{
.sample_dirs(upload_dir, pipeline = "bcbio")
}
\arguments{
\item{upload_dir}{Upload directory.}

\item{pipeline}{Pipeline used to generate the samples.}
}
\value{
Named character vector containing sample directory paths. Function
will \link{stop} if no complete sample directories match.
}
\description{
Detect sample directories
}
\author{
Michael Steinbaugh
}
\keyword{internal}
