% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-sample_metadata.R
\docType{methods}
\name{sample_metadata}
\alias{sample_metadata}
\alias{.sample_metadata}
\alias{sample_metadata,bcbioSCDataSet-method}
\alias{sample_metadata,bcbioSCSubset-method}
\title{Sample Metadata}
\usage{
sample_metadata(object, ...)

\S4method{sample_metadata}{bcbioSCDataSet}(object)

\S4method{sample_metadata}{bcbioSCSubset}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{tibble}.
}
\description{
Sample Metadata
}
\author{
Michael Steinbaugh
}
