% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-sample_metadata_file.R
\name{.sample_metadata_file}
\alias{.sample_metadata_file}
\title{Prepare sample metadata from file}
\usage{
.sample_metadata_file(file, sample_dirs, pipeline, unique_names = FALSE)
}
\arguments{
\item{file}{Sample metadata file.}

\item{sample_dirs}{Sample directory paths.}

\item{pipeline}{Pipeline used to process the samples.}

\item{unique_names}{Ensure sample names are unique.}
}
\value{
\link{tibble}.
}
\description{
Appends reverse complement sequences by matching against the \code{sequence}
column.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
