% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleData.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData,bcbioSingleCell-method}
\alias{sampleData,seurat-method}
\title{Sample Data}
\usage{
\S4method{sampleData}{bcbioSingleCell}(object, interestingGroups)

\S4method{sampleData}{seurat}(object, interestingGroups)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in \code{\link[=camel]{camel()}} case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}
}
\value{
\code{data.frame}.
}
\description{
Metadata in columns describing the samples, which are defined in the
rownames. Similar to \code{\link[=colData]{colData()}}, which for \code{bcbioSingleCell} and
\code{SingleCellExperiment} objects describes cells in the columns, rather than
the samples.
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioSingleCell"))

# bcbioSingleCell ====
sampleData(bcb_small) \%>\% glimpse()

# seurat ====
sampleData(pbmc_small) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
