% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDirs-internal.R
\name{.sampleDirs}
\alias{.sampleDirs}
\title{Detect Sample Directories}
\usage{
.sampleDirs(uploadDir, pipeline = "bcbio")
}
\arguments{
\item{uploadDir}{Upload directory.}

\item{pipeline}{Pipeline used to generate the samples.}
}
\value{
Named character vector containing sample directory paths. Function
will \link{stop} if no complete sample directories match.
}
\description{
Detect Sample Directories
}
\author{
Michael Steinbaugh
}
\keyword{internal}
