% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{SCSubset-class}
\alias{SCSubset-class}
\alias{SCSubset}
\title{SCSubset}
\description{
\link{SCSubset} is a subclass of
\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment} designed to store a subset
of single-cell RNA-seq counts filtered by quality control analysis.
}
\author{
Michael Steinbaugh
}
