% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-select_samples.R
\docType{methods}
\name{select_samples}
\alias{select_samples}
\alias{.select_samples}
\alias{select_samples,bcbioSCDataSet-method}
\alias{select_samples,SCSubset-method}
\title{Select samples}
\usage{
\S4method{select_samples}{bcbioSCDataSet}(object, ...)

\S4method{select_samples}{SCSubset}(object, ...)
}
\arguments{
\item{object}{Primary object.}

\item{...}{Columns to use for grep pattern matching. Supply a named character
vector containing the column name and the grep pattern.}
}
\value{
\link{SCSubset}.
}
\description{
Select samples
}
\details{
Internally, pattern matching against sample and file names is
applied using \code{\link[=str_detect]{str_detect()}}.
}
\note{
Supports an object containing filtered counts generated by
\code{\link[=filter_barcodes]{filter_barcodes()}} on a \link{bcbioSCDataSet}.
}
