% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rownames.R
\name{set_rownames}
\alias{set_rownames}
\title{Set row names}
\usage{
set_rownames(df, column)
}
\arguments{
\item{df}{Data frame}

\item{column}{Column to use for \code{rownames}}
}
\description{
Wrapper function for \code{rownames} that works in a chain operation
}
\examples{
\dontrun{
set_rownames(df, "description")
}
}
\author{
Michael Steinbaugh
}
