% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-seurat.R
\docType{methods}
\name{seurat}
\alias{seurat}
\alias{assay,seurat,ANY-method}
\alias{colnames,seurat-method}
\alias{gene2symbol,seurat-method}
\alias{metadata,seurat-method}
\alias{metadata<-,seurat-method}
\alias{rownames,seurat-method}
\title{Extend S4 Methods for \code{seurat} Objects}
\usage{
\S4method{assay}{seurat,ANY}(x)

\S4method{colnames}{seurat}(x)

\S4method{gene2symbol}{seurat}(object)

\S4method{metadata}{seurat}(x)

\S4method{metadata}{seurat}(x) <- value

\S4method{rownames}{seurat}(x)
}
\arguments{
\item{x}{Primary object.}

\item{object}{Object.}

\item{value}{Object to assign.}
}
\value{
Match \code{SummarizedExperiment} method return.
}
\description{
Provide \code{SummarizedExperiment} method support.
}
\examples{
# assay ====
assay(pbmc_small) \%>\% glimpse()
# colnames ====
colnames(pbmc_small) \%>\% head()
# gene2symbol ====
gene2symbol(seurat_small) \%>\% glimpse()
# metadata ====
names(metadata(seurat_small))
metadata(seurat_small)[["stash"]] <- "XXX"
metadata(seurat_small)[["stash"]]
# rownames ====
rownames(pbmc_small) \%>\% head()
}
\seealso{
\code{getMethod("metadata<-", "Annotated")}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
