% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_small}
\alias{seurat_small}
\title{Seurat Example}
\format{An object of class \code{seurat} of length 1.}
\usage{
seurat_small
}
\description{
Seurat Example
}
\examples{
show(seurat_small)
}
\seealso{
Other Minimal Example Data: \code{\link{all_markers_small}},
  \code{\link{cellranger_small}},
  \code{\link{indrops_small}},
  \code{\link{known_markers_small}}
}
\author{
Michael Steinbaugh
}
\concept{Minimal Example Data}
\keyword{datasets}
