% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/show-methods.R
\name{show}
\alias{show}
\alias{show,bcbioSingleCell-method}
\title{Show}
\usage{
show(object)

\S4method{show}{bcbioSingleCell}(object)
}
\description{
Show
}
\note{
Updated 2020-12-22.
}
\examples{
## bcbioSingleCell ====
data(bcb)
show(bcb)
}
\author{
Michael Steinbaugh
}
