% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{.aggregate_features}
\alias{.aggregate_features}
\alias{.aggregate_replicates}
\title{Sparse matrix aggregation operations}
\usage{
.aggregate_features(sparse, featureids)

.aggregate_replicates(sparse, cellids)
}
\arguments{
\item{sparse}{Sparse counts matrix.}

\item{featureids}{New feature IDs for the collapsed features.}

\item{cellids}{New cellular ids of the collapsed cells, one for each cell.}
}
\value{
\link{"dgCMatrix-class"}.
}
\description{
Collapse features with the same feature id by summing them.

Pool cellular technical replicate counts.
}
\author{
Rory Kirchner, Michael Steinbaugh
}
