% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-sparse_counts.R
\name{.sparse_counts}
\alias{.sparse_counts}
\title{Read MatrixMarket file}
\usage{
.sparse_counts(matrix_file, pipeline = "bcbio")
}
\arguments{
\item{matrix_file}{MatrixMart file to read.}

\item{pipeline}{Pipeline used to generate the MatrixMarket file. Defaults to
bcbio-nextgen (\code{bcbio}). Also supports 10X Chromium CellRanger
(\code{cellranger}).}
}
\value{
Sparse counts matrix.
}
\description{
Load single-cell RNA-seq counts saved as a \href{https://people.sc.fsu.edu/~jburkardt/data/mm/mm.html}{MatrixMarketfile} into a sparse
counts matrix (\code{dgCMatrix}). Transcript (\link{rownames}) and molecular barcodes
(\link{colnames}) identifiers are set from the corresponding dependency files
automatically.
}
\details{
This function supports automatic handling of compressed matrix files.
}
\note{
bcbio-nextgen outputs counts at transcript level. 10X Chromium
CellRanger outputs counts at gene level.
}
\seealso{
\code{help("dgCMatrix-class")}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\keyword{internal}
