% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseCounts-internal.R
\name{.sparseCountsTx2Gene}
\alias{.sparseCountsTx2Gene}
\title{Transcript To Gene-Level Sparse Counts}
\usage{
.sparseCountsTx2Gene(sparseCounts, tx2gene)
}
\arguments{
\item{sparseCounts}{Sparse counts matrix (\code{dgCMatrix}).}

\item{tx2gene}{Transcript to gene identifier mappings.}
}
\value{
\code{dgCMatrix}.
}
\description{
Transcript To Gene-Level Sparse Counts
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\keyword{internal}
