% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stripTranscriptVersions.R
\docType{methods}
\name{stripTranscriptVersions}
\alias{stripTranscriptVersions}
\title{Strip Transcript Versions}
\usage{
\S4method{stripTranscriptVersions}{dgCMatrix}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Object with the transcript versions in the rownames removed.
}
\description{
Strip Transcript Versions
}
\author{
Michael Steinbaugh
}
