% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-summarized_experiment.R
\name{.summarized_experiment}
\alias{.summarized_experiment}
\title{Package scRNA-Seq counts into \link{SummarizedExperiment}}
\usage{
.summarized_experiment(sparse_counts, col_data, row_data, metadata = NULL)
}
\arguments{
\item{sparse_counts}{Gene-level sparse counts matrix.}

\item{col_data}{Cellular barcode metrics.}

\item{row_data}{Annotable.}

\item{metadata}{Custom metadata.}
}
\value{
\link{SummarizedExperiment}.
}
\description{
Package scRNA-Seq counts into \link{SummarizedExperiment}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
