% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{synonyms}
\alias{synonyms}
\title{Gene Symbol Synonyms}
\format{An object of class \code{list} of length 4.}
\usage{
synonyms
}
\description{
Gene Symbol Synonyms
}
\author{
Michael Steinbaugh
}
\keyword{internal}
