% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-top_barcodes.R
\docType{methods}
\name{top_barcodes}
\alias{top_barcodes}
\alias{.top_barcodes}
\alias{top_barcodes,bcbioSCDataSet-method}
\alias{top_barcodes,bcbioSCSubset-method}
\title{Top Barcodes}
\usage{
top_barcodes(object, ...)

\S4method{top_barcodes}{bcbioSCDataSet}(object, n = 2L)

\S4method{top_barcodes}{bcbioSCSubset}(object, n = 2L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{n}{Number of barcodes to return per sample.}
}
\value{
\link{tibble}.
}
\description{
Top Barcodes
}
\author{
Michael Steinbaugh
}
