% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-markers.R
\docType{methods}
\name{top_markers}
\alias{top_markers}
\alias{top_markers,data.frame-method}
\title{Top Markers}
\usage{
top_markers(object, ...)

\S4method{top_markers}{data.frame}(object, n = 4L, show = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{n}{Number of genes per cluster.}

\item{show}{Show \link{kable}.}
}
\value{
\link{tibble}.
}
\description{
Top Markers
}
