% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-topMarkers.R
\docType{methods}
\name{topMarkers}
\alias{topMarkers}
\alias{topMarkers}
\alias{topMarkers,grouped_df-method}
\title{Top Markers}
\usage{
topMarkers(object, ...)

\S4method{topMarkers}{grouped_df}(object, n = 10L, direction = "positive",
  coding = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{n}{Number of genes per cluster.}

\item{direction}{Whether to include only \code{positive}, \code{negative}, or \code{both}
directions of association per cluster. Defaults to \code{positive}.}

\item{coding}{Only include protein coding genes.}
}
\value{
\code{tibble}.
}
\description{
Top Markers
}
\examples{
# grouped_df ====
topMarkers(all_markers_small, n = 1L)
}
\seealso{
\itemize{
\item \code{\link[dplyr:slice]{dplyr::slice()}}
\item \code{\link[dplyr:top_n]{dplyr::top_n()}}.
}

Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkers}}, \code{\link{plotPCElbow}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
