% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_cells_plot.R
\name{total_cells_plot}
\alias{total_cells_plot}
\title{Total cells plot}
\usage{
total_cells_plot(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\description{
Total cells plot
}
\author{
Rory Kirchner

Michael Steinbaugh
}
