% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_counts_plot.R
\name{total_counts_plot}
\alias{total_counts_plot}
\title{Total counts plots}
\usage{
total_counts_plot(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\description{
Total counts plots
}
\author{
Rory Kirchner

Michael Steinbaugh
}
\keyword{plot}
\keyword{qc}
