% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_vs_detected_plot.R
\name{total_vs_detected_plot}
\alias{total_vs_detected_plot}
\title{Total counts vs. genes detected plot}
\usage{
total_vs_detected_plot(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\description{
Total counts vs. genes detected plot
}
\author{
Rory Kirchner

Michael Steinbaugh
}
\keyword{plot}
\keyword{qc}
