% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-tx2symbol.R
\name{.tx2symbol}
\alias{.tx2symbol}
\title{Convert a transcript matrix to a gene symbol matrix}
\usage{
.tx2symbol(matrix, strip = TRUE)
}
\arguments{
\item{matrix}{Counts matrix.}

\item{strip}{Strip transcript versions from the matrix before aggregating.}
}
\value{
Matrix of counts aggregated by summing the "to" ids.
}
\description{
Convert a transcript matrix to a gene symbol matrix
}
\author{
Rory Kirchner
}
\keyword{internal}
