% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-unlist_barcodes.R
\name{.unlist_barcodes}
\alias{.unlist_barcodes}
\title{Unlist cellular barcodes}
\usage{
.unlist_barcodes(bcb)
}
\arguments{
\item{bcb}{\link{bcbioSCDataSet}.}
}
\description{
Convert named list of cellular barcodes per sample to a data frame.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
