% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/updateObject-methods.R
\name{updateObject}
\alias{updateObject}
\alias{updateObject,bcbioSingleCell-method}
\title{Update object}
\usage{
updateObject(object, ..., verbose = FALSE)

\S4method{updateObject}{bcbioSingleCell}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{verbose}{\code{logical(1)}.
Verbose output.}
}
\value{
Modified object.
}
\description{
Update object
}
\note{
Updated 2021-02-22.
}
\examples{
data(bcb)
updateObject(bcb)

## Example that depends on remote file.
## > x <- import(
## >     file = file.path(
## >         bcbioSingleCellTestsURL,
## >         "bcbioSingleCell_0.1.0.rds"
## >     )
## > )
## > x <- updateObject(x)
## > x
}
\author{
Michael Steinbaugh
}
