% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_counts.R
\name{write_counts}
\alias{write_counts}
\title{Write MatrixMart .mtx, .colnames, and .rownames files}
\usage{
write_counts(counts, gzip = FALSE)
}
\arguments{
\item{counts}{Sparse counts matrix}

\item{gzip}{Compress matrix file with gzip}
}
\value{
Path to matrix file
}
\description{
Write MatrixMart .mtx, .colnames, and .rownames files
}
\author{
Rory Kirchner

Michael Steinbaugh
}
