% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{write_sparsecounts}
\alias{write_sparsecounts}
\title{Write a gzipped MatrixMart file, storing the row and column names.}
\usage{
write_sparsecounts(sparse, gzfile)
}
\arguments{
\item{sparse}{sparse matrix of counts}

\item{gzfile}{outfile with gzip extension}
}
\value{
gzfile
}
\description{
Write a gzipped MatrixMart file, storing the row and column names.
}
\author{
Rory Kirchner
}
