% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-aggregate.R
\docType{methods}
\name{aggregate_replicates,bcbioSCDataSet-method}
\alias{aggregate_replicates,bcbioSCDataSet-method}
\title{Aggregate technical replicates}
\usage{
\S4method{aggregate_replicates}{bcbioSCDataSet}(object)
}
\arguments{
\item{object}{\link{bcbioSCDataSet}.}
}
\value{
\link{bcbioSCDataSet}.
}
\description{
Aggregate technical replicates
}
\author{
Michael Steinbaugh
}
