% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-aggregate.R
\docType{methods}
\name{aggregate_replicates}
\alias{aggregate_replicates}
\alias{aggregate_replicates,bcbioSCDataSet-method}
\alias{.aggregate_features}
\alias{.aggregate_replicates}
\title{Aggregation methods}
\usage{
\S4method{aggregate_replicates}{bcbioSCDataSet}(object)
}
\arguments{
\item{object}{Primary object.}

\item{sparse}{Sparse counts matrix.}

\item{featureids}{New feature IDs for the collapsed features.}

\item{cellids}{New cellular ids of the collapsed cells, one for each cell.}
}
\value{
\link{bcbioSCDataSet}.

Aggregated sparse matrix.

Aggregated sparse matrix.
}
\description{
Aggregation methods

Collapse features with the same feature id by summing them.

Pool cellular technical replicate counts.
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\keyword{internal}
