% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-aggregate.R
\docType{methods}
\name{aggregate_replicates}
\alias{aggregate_replicates}
\alias{aggregate_replicates,bcbioSCDataSet-method}
\title{Aggregate technical replicates}
\usage{
aggregate_replicates(object, ...)

\S4method{aggregate_replicates}{bcbioSCDataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters.}
}
\value{
\link{bcbioSCDataSet}.
}
\description{
Aggregate technical replicates
}
\author{
Michael Steinbaugh
}
