% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-aggregate.R
\docType{methods}
\name{aggregate_replicates}
\alias{aggregate_replicates}
\alias{aggregate_replicates,bcbioSCDataSet-method}
\title{Aggregate replicates}
\usage{
aggregate_replicates(object, ...)

\S4method{aggregate_replicates}{bcbioSCDataSet}(object)
}
\value{
New \link{bcbioSCDataSet}.
}
\description{
Aggregate replicates
}
