% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-aggregateFeatures.R
\docType{methods}
\name{aggregateFeatures}
\alias{aggregateFeatures}
\alias{aggregateFeatures,bcbioSingleCell-method}
\alias{aggregateFeatures,dgCMatrix-method}
\alias{aggregateFeatures,dgTMatrix-method}
\title{Aggregate Features}
\usage{
aggregateFeatures(object, ...)

\S4method{aggregateFeatures}{bcbioSingleCell}(object, featureids)

\S4method{aggregateFeatures}{dgCMatrix}(object, featureids)

\S4method{aggregateFeatures}{dgTMatrix}(object, featureids)
}
\arguments{
\item{object}{Sparse counts matrix (e.g. \link{dgCMatrix}).}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{featureids}{Feature identifiers (e.g. gene or transcript IDs).}
}
\value{
\link{dgCMatrix}.

\link{bcbioSingleCell}.
}
\description{
Aggregate Features
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateReplicates}},
  \code{\link{filterCells}}, \code{\link{readMarkers}},
  \code{\link{subsetPerSample}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{Data Management Utilities}
\keyword{internal}
