% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-aggregateFeatures.R
\docType{methods}
\name{aggregateFeatures}
\alias{aggregateFeatures}
\alias{aggregateFeatures}
\alias{aggregateFeatures,bcbioSCDataSet-method}
\title{Aggregate Features}
\usage{
aggregateFeatures(object, ...)

\S4method{aggregateFeatures}{bcbioSCDataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{bcbioSCDataSet}.
}
\description{
Aggregate Features
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateReplicates}},
  \code{\link{filterCells}},
  \code{\link{interestingGroups}},
  \code{\link{readMarkers}}, \code{\link{sampleMetadata}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
