% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-aggregateFeatures.R
\docType{methods}
\name{aggregateFeatures}
\alias{aggregateFeatures}
\alias{aggregateFeatures,bcbioSingleCell-method}
\title{Aggregate Features}
\usage{
\S4method{aggregateFeatures}{bcbioSingleCell}(object, features)
}
\arguments{
\item{object}{Object.}

\item{features}{Feature identifiers (e.g. gene or transcript IDs). These are
the rownames of the counts matrix.}
}
\value{
Object of same class with aggregated features (rows).
}
\description{
Aggregate Features
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateReplicates}},
  \code{\link{filterCells}}, \code{\link{readMarkers}},
  \code{\link{subsetPerSample}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
