% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-aggregateReplicates.R
\docType{methods}
\name{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates,bcbioSingleCell-method}
\title{Aggregate Replicates}
\usage{
\S4method{aggregateReplicates}{bcbioSingleCell}(object)
}
\arguments{
\item{object}{Object.}

\item{groupings}{Factor that defines the aggregation groupings. The new
aggregation names are defined as the factor levels, and the original
replicates are defined as the names of the factor.}
}
\description{
Aggregate Replicates
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
print(bcb)
aggregateReplicates(bcb)
}
\seealso{
Other Data Management Utilities: \code{\link{filterCells}},
  \code{\link{readCellTypeMarkersFile}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
