% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-aggregateReplicates.R
\docType{methods}
\name{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates,bcbioSingleCell-method}
\title{Aggregate Replicates}
\usage{
\S4method{aggregateReplicates}{bcbioSingleCell}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Aggregate Replicates
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioSingleCell"))

# bcbioSingleCell ====
x <- aggregateReplicates(bcb_small)
sampleData(x)
}
\seealso{
Other Data Management Utilities: \code{\link{filterCells}},
  \code{\link{readCellTypeMarkersFile}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
