% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateReplicates-methods.R
\docType{methods}
\name{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates,SingleCellExperiment-method}
\title{Aggregate Replicates}
\usage{
\S4method{aggregateReplicates}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Aggregate Replicates
}
\examples{
object <- indrops_small
sampleNames(object)

# Define groupings factor as`aggregate` column in `colData()`
glimpse(object$aggregate)

x <- aggregateReplicates(object)
show(x)
sampleNames(x)
}
\seealso{
Other Data Functions: \code{\link{cell2sample}},
  \code{\link{metricsPerSample}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Data Functions}
