% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aggregateReplicates-methods.R
\docType{methods}
\name{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates,bcbioSCDataSet-method}
\title{Aggregate Replicates}
\usage{
aggregateReplicates(object, ...)

\S4method{aggregateReplicates}{bcbioSCDataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{bcbioSCDataSet}.
}
\description{
Aggregate Replicates
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateFeatures}},
  \code{\link{filterCells}},
  \code{\link{interestingGroups}},
  \code{\link{readKnownMarkers}},
  \code{\link{sampleMetadata}}, \code{\link{selectSamples}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
