% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-aggregateReplicates.R
\docType{methods}
\name{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates,bcbioSingleCell-method}
\title{Aggregate Replicates}
\usage{
\S4method{aggregateReplicates}{bcbioSingleCell}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Aggregate Replicates
}
\examples{
# bcbioSingleCell ====
x <- aggregateReplicates(bcb_small)
sampleData(x)
}
\seealso{
Other Data Functions: \code{\link{calculateMetrics}},
  \code{\link{cell2sample}}, \code{\link{colData}},
  \code{\link{counts}}, \code{\link{fetchGeneData}},
  \code{\link{fetchPCAData}}, \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}},
  \code{\link{interestingGroups}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{rowData}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
