% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-aggregateReplicates.R
\docType{methods}
\name{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates,bcbioSCDataSet-method}
\title{Aggregate Replicates}
\usage{
\S4method{aggregateReplicates}{bcbioSCDataSet}(object)
}
\value{
\link{bcbioSCDataSet}.
}
\description{
Aggregate Replicates
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateFeatures}},
  \code{\link{filterCells}}, \code{\link{readMarkers}},
  \code{\link{subsetPerSample}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{Data Management Utilities}
