% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-aggregateReplicates.R
\docType{methods}
\name{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates,bcbioSingleCell-method}
\title{Aggregate Replicates}
\usage{
\S4method{aggregateReplicates}{bcbioSingleCell}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Object of same class with aggregated counts per pooled sample
(columns).
}
\description{
Aggregate Replicates
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
aggregateReplicates(bcb)
}
\seealso{
Other Data Management Utilities: \code{\link{filterCells}},
  \code{\link{readMarkersFile}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
