% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/basejump.R
\name{all_generics}
\alias{all_generics}
\alias{assay}
\alias{assays}
\alias{colData}
\alias{metadata}
\alias{rowData}
\title{S4 Generics}
\usage{
assay(x, i, ...)

assays(x, ..., withDimnames = TRUE)

colData(x, ...)

metadata(x, ...)

rowData(x, ...)
}
\arguments{
\item{x}{Primary object.}

\item{...}{Additional arguments.}

\item{object}{Object.}

\item{y}{Secondary object.}
}
\description{
S4 Generics
}
\keyword{internal}
