% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{all_markers_small}
\alias{all_markers_small}
\title{Sanitized Markers Example}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1380 rows and 18 columns.}
\usage{
all_markers_small
}
\description{
Sanitized Markers Example
}
\examples{
glimpse(all_markers_small)
}
\seealso{
Other Minimal Example Data: \code{\link{cellranger_small}},
  \code{\link{indrops_small}},
  \code{\link{known_markers_small}},
  \code{\link{seurat_small}}
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
