% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_metrics.R
\name{barcode_metrics}
\alias{barcode_metrics}
\title{Generate barcode metrics summary}
\usage{
barcode_metrics(counts, annotations, metadata, tx2gene = TRUE)
}
\arguments{
\item{counts}{`bcbio-nextgen` scRNA-seq counts sparse matrix}

\item{annotations}{Ensembl annotations data frame}

\item{metadata}{Sample metadata data frame}

\item{tx2gene}{Convert transcript-level annotations to gene-level}
}
\value{
Data frame
}
\description{
Generate barcode metrics summary
}
\author{
Rory Kirchner

Michael Steinbaugh
}
