% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{barcode_metrics}
\alias{barcode_metrics}
\title{Generate barcode metrics summary}
\usage{
barcode_metrics(run)
}
\arguments{
\item{run}{\link{bcbioSCDataSet}.}
}
\value{
Tibble grouped by sample name.
}
\description{
Generate barcode metrics summary
}
\author{
Rory Kirchner

Michael Steinbaugh
}
