% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-barcodeRanks.R
\docType{methods}
\name{barcodeRanks}
\alias{barcodeRanks}
\alias{barcodeRanks}
\alias{barcodeRanks,matrix-method}
\alias{barcodeRanks,dgCMatrix-method}
\alias{barcodeRanks,dgTMatrix-method}
\alias{barcodeRanks,SingleCellExperiment-method}
\title{Barcode Ranks}
\usage{
barcodeRanks(object, ...)

\S4method{barcodeRanks}{matrix}(object, lower = 100L, fitBounds = NULL,
  df = 20L)

\S4method{barcodeRanks}{dgCMatrix}(object, lower = 100L, fitBounds = NULL,
  df = 20L)

\S4method{barcodeRanks}{dgTMatrix}(object, lower = 100L, fitBounds = NULL,
  df = 20L)

\S4method{barcodeRanks}{SingleCellExperiment}(object, ...)
}
\value{
\code{list}.
}
\description{
Return statistics to construct a barcode-rank plot. Also calculates the knee
and inflection points for further use.
}
\examples{
# SingleCellExperiment ====
x <- barcodeRanks(bcb_small)
names(x)
x[["knee"]]
x[["inflection"]]
}
\seealso{
Modified version of \code{DropletUtils::barcodeRanks()} that works on
a \code{SingleCellExperiment} object instead of a matrix.
}
\author{
Aaron Lun, Michael Steinbaugh
}
