% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-barcodeRanks.R
\docType{methods}
\name{barcodeRanks}
\alias{barcodeRanks}
\alias{barcodeRanks}
\alias{barcodeRanks,matrix-method}
\alias{barcodeRanks,dgCMatrix-method}
\alias{barcodeRanks,dgTMatrix-method}
\alias{barcodeRanks,SingleCellExperiment-method}
\title{Barcode Ranks}
\usage{
barcodeRanks(object, ...)

\S4method{barcodeRanks}{matrix}(object, lower = 100L, fitBounds = NULL,
  df = 20L)

\S4method{barcodeRanks}{dgCMatrix}(object, lower = 100L, fitBounds = NULL,
  df = 20L)

\S4method{barcodeRanks}{dgTMatrix}(object, lower = 100L, fitBounds = NULL,
  df = 20L)

\S4method{barcodeRanks}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{lower}{\code{integer} scalar specifying the lower bound on the total UMI
count, at or below which all barcodes are assumed to correspond to empty
droplets.}

\item{fitBounds}{\code{numeric} vector of length 2, specifying the lower and upper
bounds on the total UMI count for spline fitting.}

\item{df}{\code{integer} scalar specifying the number of degress of freedom to
use for spline fitting with \code{\link[stats:smooth.spline]{stats::smooth.spline()}}.}
}
\value{
\code{list}.
}
\description{
Return statistics to construct a barcode-rank plot. Also calculates the knee
and inflection points for further use.
}
\examples{
# SingleCellExperiment ====
x <- barcodeRanks(indrops_small)
names(x)
x[["knee"]]
x[["inflection"]]
}
\seealso{
Modified version of \code{DropletUtils::barcodeRanks()} that works on
a \code{SingleCellExperiment} object instead of a matrix.

Other Quality Control Functions: \code{\link{barcodeRanksPerSample}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Aaron Lun, Michael Steinbaugh
}
