% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-barcodeRanksPerSample.R
\docType{methods}
\name{barcodeRanksPerSample}
\alias{barcodeRanksPerSample}
\alias{barcodeRanksPerSample}
\alias{barcodeRanksPerSample,SingleCellExperiment-method}
\alias{barcodeRanksPerSample,seurat-method}
\title{Barcode Ranks per Sample}
\usage{
barcodeRanksPerSample(object, ...)

\S4method{barcodeRanksPerSample}{SingleCellExperiment}(object, ...)

\S4method{barcodeRanksPerSample}{seurat}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[=barcodeRanks]{barcodeRanks()}}.}
}
\value{
\code{list}.
}
\description{
Barcode Ranks per Sample
}
\examples{
# bcbioSingleCell ====
x <- barcodeRanksPerSample(indrops_small)
names(x)

# seurat ====
x <- barcodeRanksPerSample(seurat_small)
names(x)
}
\seealso{
Other Quality Control Functions: \code{\link{barcodeRanks}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
