% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-barcodes.R
\name{.read_barcode_file}
\alias{.read_barcode_file}
\alias{.barcodes}
\title{Read sample cellular barcode summary file}
\usage{
.read_barcode_file(file_name)

.barcodes(sample_dirs)
}
\arguments{
\item{file_name}{File name.}

\item{sample_dirs}{Sample directories.}
}
\description{
Read sample cellular barcode summary file
}
\author{
Michael Steinbaugh
}
\keyword{internal}
