% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump.R
\name{set_names_snake}
\alias{set_names_snake}
\alias{set_names_dot}
\title{Utility functions from basejump package}
\usage{
set_names_snake(data)

set_names_dot(data)
}
\arguments{
\item{data}{Data type that supports name assignments}
}
\description{
Set names in snake_case

Set names in dot notation
}
\author{
Michael Steinbaugh
}
