% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bcb_small}
\alias{bcb_small}
\title{Minimal \code{bcbioSingleCell} Example}
\format{An object of class \code{bcbioSingleCell} with 500 rows and 500 columns.}
\usage{
bcb_small
}
\description{
Minimal \code{bcbioSingleCell} Example
}
\author{
Michael Steinbaugh
}
\keyword{internal}
