% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-bcbio.R
\docType{methods}
\name{bcbio}
\alias{bcbio}
\alias{bcbio<-}
\alias{bcbio,bcbioSCDataSet-method}
\alias{bcbio<-,bcbioSCDataSet-method}
\title{\code{bcbioSCDataSet} Object Accessors}
\usage{
bcbio(object, ...)

bcbio(object, ...) <- value

\S4method{bcbio}{bcbioSCDataSet}(object, type = "counts")

\S4method{bcbio}{bcbioSCDataSet}(object, type = "counts") <- value
}
\arguments{
\item{object}{Primary object.}

\item{...}{Additional arguments.}

\item{value}{Object to assign.}

\item{type}{Type of count data to retrieve.}
}
\value{
Matrix or object containing count data.
}
\description{
\code{bcbioSCDataSet} Object Accessors
}
