% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-bcbio.R
\docType{methods}
\name{bcbio}
\alias{bcbio}
\alias{bcbio<-}
\alias{bcbio,bcbioSCDataSet-method}
\alias{bcbio<-,bcbioSCDataSet-method}
\title{bcbioSCDataSet object accessors}
\usage{
\S4method{bcbio}{bcbioSCDataSet}(object, type = "counts")

\S4method{bcbio}{bcbioSCDataSet}(object, type = "counts") <- value
}
\arguments{
\item{object}{Primary object.}

\item{value}{An integer matrix or other object.}

\item{type}{Type of count data to retrieve.}
}
\value{
Matrix/Object containing count data.
}
\description{
bcbioSCDataSet object accessors
}
