% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/all_methods.R
\docType{methods}
\name{bcbio}
\alias{bcbio}
\alias{bcbio<-}
\alias{bcbio}
\alias{bcbio,bcbioSCDataSet-method}
\alias{bcbio<-,bcbioSCDataSet,matrix-method}
\alias{bcbio.bcbioSCDataSet}
\alias{bcbio,bcbioSCDataSet-method}
\alias{bcbio<-,bcbioSCDataSet-method}
\title{Accessors for the sparse count matrix of a \link{bcbioSCDataSet} object}
\usage{
bcbio(object, ...)

bcbio(object, ...) <- value

bcbio.bcbioSCDataSet(object, type = "counts")

\S4method{bcbio}{bcbioSCDataSet}(object, type = "counts")

\S4method{bcbio}{bcbioSCDataSet}(object, type = "counts") <- value
}
\arguments{
\item{object}{\link{bcbioSCDataSet} object.}

\item{...}{Matrix count data.}

\item{value}{An integer matrix or other object.}

\item{type}{type of count data to retrieve}
}
\value{
Matrix/Object containing count data.
}
\description{
Accessors for the sparse count matrix of a \link{bcbioSCDataSet} object
}
