% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-bcbio.R
\docType{methods}
\name{bcbio}
\alias{bcbio}
\alias{bcbio,seurat-method}
\alias{bcbio<-,seurat-method}
\title{Additional bcbio Run Data Accessor}
\usage{
\S4method{bcbio}{seurat}(object, type)

\S4method{bcbio}{seurat}(object, type) <- value
}
\arguments{
\item{object}{Object.}

\item{type}{Type of count data to retrieve.}

\item{value}{Object to assign.}
}
\value{
\link{bcbioSingleCell}.
}
\description{
Additional bcbio Run Data Accessor
}
\examples{
# seurat ====
names(bcbio(seurat_small))

# Assignment method support
bcbio(seurat_small, "stash") <- "testing"
bcbio(seurat_small, "stash")
}
\author{
Michael Steinbaugh
}
