% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-bcbio.R
\docType{methods}
\name{bcbio}
\alias{bcbio}
\alias{bcbio,bcbioSingleCell-method}
\alias{bcbio,seurat-method}
\alias{bcbio<-,bcbioSingleCell-method}
\alias{bcbio,bcbioSCDataSet-method}
\alias{bcbio<-,bcbioSCDataSet-method}
\title{\code{bcbioSingleCell} Additional bcbio Run Data Accessor}
\usage{
\S4method{bcbio}{bcbioSingleCell}(object, type)

\S4method{bcbio}{seurat}(object, type)

\S4method{bcbio}{bcbioSingleCell}(object, type) <- value

\S4method{bcbio}{bcbioSCDataSet}(object, type)

\S4method{bcbio}{bcbioSCDataSet}(object, type) <- value
}
\arguments{
\item{object}{Object.}

\item{type}{Type of count data to retrieve.}

\item{value}{Object to assign.}
}
\value{
\link{bcbioSingleCell}.
}
\description{
\code{bcbioSingleCell} Additional bcbio Run Data Accessor
}
\author{
Michael Steinbaugh
}
