% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-bcbio.R
\docType{methods}
\name{bcbio}
\alias{bcbio}
\alias{bcbio<-}
\alias{bcbio}
\alias{bcbio,bcbioSCDataSet-method}
\alias{bcbio<-,bcbioSCDataSet-method}
\title{\code{bcbioSCDataSet} Object Accessors}
\usage{
bcbio(object, ...)

bcbio(object, ...) <- value

\S4method{bcbio}{bcbioSCDataSet}(object, type)

\S4method{bcbio}{bcbioSCDataSet}(object, type) <- value
}
\arguments{
\item{object}{Primary object.}

\item{...}{Additional arguments.}

\item{value}{Object to assign.}

\item{type}{Type of count data to retrieve.}
}
\value{
\link{bcbioSCDataSet}.
}
\description{
\code{bcbioSCDataSet} Object Accessors
}
\seealso{
Other `bcbio` Utilities: \code{\link{download}},
  \code{\link{loadRun}}
}
