% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-bcbio.R
\docType{methods}
\name{bcbio}
\alias{bcbio}
\alias{bcbio,seurat-method}
\alias{bcbio<-,seurat-method}
\title{Additional bcbio Run Data Accessor}
\usage{
\S4method{bcbio}{seurat}(object, type)

\S4method{bcbio}{seurat}(object, type) <- value
}
\arguments{
\item{object}{Object.}

\item{type}{Type of count data to retrieve.}

\item{value}{Object to assign.}
}
\value{
\link{bcbioSingleCell}.
}
\description{
Additional bcbio Run Data Accessor
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat.rda", package = "bcbioSingleCell"))

# bcbioSingleCell
bcbio(bcb) \%>\% names()

# Assignment method support
bcbio(bcb, "stash") <- "testing"
bcbio(bcb, "stash")

# seurat
bcbio(seurat) \%>\% names()

# Assignment method support
bcbio(seurat, "stash") <- "testing"
bcbio(seurat, "stash")
}
\author{
Michael Steinbaugh
}
