% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioSCDataSet-class}
\alias{bcbioSCDataSet-class}
\alias{bcbioSCDataSet}
\title{bcbioSCDataSet}
\description{
\link{bcbioSCDataSet} is an extension of \link{SummarizedExperiment} designed to store
a single-cell RNA-seq analysis. This class contains read counts save as a
sparse matrix (\code{dgCMatrix}), sample barcodes, run metadata, and barcode
summary statistics for each sample analyzed.
}
\section{Slots}{

\describe{
\item{\code{bcbio}}{\link{SimpleList} containing additional bcbio run data with dimensions
that don't match the count matrix. This is currently used to store all
unfiltered cellular barcodes for quality control analysis.}
}}

\author{
Michael Steinbaugh
}
