% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{bcbioSCDataSet-class}
\alias{bcbioSCDataSet-class}
\alias{bcbioSCDataSet}
\title{\link{bcbioSCDataSet}}
\description{
\link{bcbioSCDataSet} is a subclass of
\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment} designed to store a single-cell
RNA-seq analysis. This class contains read counts save as a sparse matrix
(\code{dgCMatrix}), sample barcodes, run metadata, and barcode summary statistics
for each sample analyzed.
}
\author{
Michael Steinbaugh
}
