% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioSCDataSet-class}
\alias{bcbioSCDataSet-class}
\alias{bcbioSCDataSet}
\title{bcbioSCDataSet}
\description{
This class will be deprecated in favor of \link{bcbioSingleCell} in a future
release.
}
\keyword{internal}
