% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{bcbioSCDataSet-class}
\alias{bcbioSCDataSet-class}
\alias{bcbioSCDataSet}
\title{bcbioSCDataSet}
\description{
\code{bcbioSCDataSet} is a subclass of \link{SummarizedExperiment} designed to store a
single-cell RNA-seq analysis. This class contains experiment metadata, raw
counts, and summary statistics for each sample analyzed.
}
\details{
Methods for this objects ...

\code{metadata} contains ...
}
\keyword{internal}
