% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioSCFiltered-class}
\alias{bcbioSCFiltered-class}
\alias{bcbioSCFiltered}
\title{bcbioSCFiltered}
\description{
\link{bcbioSCFiltered} is a subclass of \link{SummarizedExperiment} designed to store
single-cell RNA-seq counts of cells filtered by quality control analysis.
}
\author{
Michael Steinbaugh
}
