% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioSingleCell-class}
\alias{bcbioSingleCell-class}
\title{\code{bcbioSingleCell} Class}
\description{
Extends \code{SingleCellExperiment}, with additional validity checks on the
\code{metadata()} slot.
}
