% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioSingleCell-class}
\alias{bcbioSingleCell-class}
\alias{bcbioSingleCell}
\title{bcbioSingleCell Class}
\description{
\code{bcbioSingleCell} extends \code{SingleCellExperiment} and designed to store a
bcbio single-cell RNA-seq analysis. This class contains read counts saved as
a sparse matrix (\code{dgCMatrix}), sample metadata, and cell quality control
metrics.
}
\section{Slots}{

\describe{
\item{\code{bcbio}}{\code{SimpleList} containing additional bcbio run data with dimensions
that don't match the count matrix. This is currently used to store all
unfiltered cellular barcodes for quality control analysis.}
}}

\note{
\code{bcbioSingleCell} extended \code{SummarizedExperiment} prior to v0.1.0,
where we migrated to \code{SingleCellExperiment}.
}
\seealso{
\itemize{
\item \code{\link[=loadSingleCell]{loadSingleCell()}}, \code{\link[=loadCellRanger]{loadCellRanger()}}.
\item \code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment::SingleCellExperiment()}}.
\item \code{.S4methods(class = "bcbioSingleCell")}.
}
}
\author{
Michael Steinbaugh
}
