% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioSingleCell-class}
\alias{bcbioSingleCell-class}
\title{\code{bcbioSingleCell} Class}
\description{
\code{bcbioSingleCell} is an S4 class that extends \code{SingleCellExperiment}, and is
designed to store a bcbio single-cell RNA-seq analysis. This class contains
read counts saved as a sparse matrix (\code{dgCMatrix}), sample metadata, and cell
quality control metrics.
}
\seealso{
Other S4 Classes: \code{\link{CellRanger-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
