% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioSingleCell-class}
\alias{bcbioSingleCell-class}
\alias{bcbioSingleCell}
\title{bcbioSingleCell Object Class}
\description{
\code{bcbioSingleCell} is an extension of \code{SummarizedExperiment} designed to store
bcbio single-cell RNA-seq counts. This class contains read counts save as a
sparse matrix (\code{dgCMatrix}), sample barcodes, run metadata, and barcode
summary statistics for each sample analyzed.
}
\section{Slots}{

\describe{
\item{\code{bcbio}}{\code{SimpleList} containing additional bcbio run data with dimensions
that don't match the count matrix. This is currently used to store all
unfiltered cellular barcodes for quality control analysis.}
}}

\note{
\code{bcbioSingleCell} extended \code{SummarizedExperiment} until v0.0.31 of
this package, when we migrated over to using \code{SingleCellExperiment}.
}
\seealso{
\itemize{
\item \code{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment()}}.
\item \code{.S4methods(class = "bcbioSingleCell")}.
}
}
\author{
Michael Steinbaugh
}
