% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioSinglecell.R
\docType{package}
\name{bcbioSinglecell}
\alias{bcbioSinglecell}
\alias{bcbioSinglecell-package}
\title{bcbioSinglecell.}
\description{
Utility functions for analysis of bcbio-nextgen single-cell RNA-seq data.
}
\details{
Consult the
\href{http://bioinformatics.sph.harvard.edu/bcbioSinglecell}{package website}
for more information.
}
\keyword{internal}
