% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioSinglecell.R
\docType{package}
\name{bcbioSinglecell}
\alias{bcbioSinglecell}
\alias{bcbioSinglecell-package}
\title{\code{bcbioSinglecell} package}
\description{
Utility functions for analysis of \code{bcbio-nextgen} single-cell RNA-seq
data.
}
\details{
See the README on \href{https://github.com/hbc/bcbioSinglecell}{GitHub}.
}
