% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{bcbioSingleCell}
\alias{bcbioSingleCell}
\title{Read bcbio Single-Cell RNA-Seq Data}
\usage{
bcbioSingleCell(uploadDir, sampleMetadataFile = NULL, organism = NULL,
  ensemblRelease = NULL, genomeBuild = NULL, gffFile = NULL,
  transgeneNames = NULL, spikeNames = NULL,
  interestingGroups = "sampleName", ...)
}
\arguments{
\item{uploadDir}{\code{string}. Path to final upload directory. This path is set
when running "\code{bcbio_nextgen -w template}".}

\item{sampleMetadataFile}{\code{string} or \code{NULL}. Sample barcode metadata file.
Optional for runs with demultiplixed index barcodes (e.g. SureCell) and can
be left \code{NULL}, but otherwise suggested for runs with multipliexed FASTQs
containing multiple index barcodes (e.g. inDrops).}

\item{organism}{\code{string} or \code{NULL}. Organism name. Use the full Latin name
(e.g. "Homo sapiens"), since this will be input downstream to AnnotationHub
and ensembldb, unless \code{gffFile} is set. If left \code{NULL} (\emph{not recommended}),
the function call will skip loading gene-level annotations into
\code{\link[=rowRanges]{rowRanges()}}. This can be useful for poorly annotation genomes or
experiments involving multiple genomes.}

\item{ensemblRelease}{\code{scalar integer} or \code{NULL}. Ensembl release version. If
\code{NULL}, defaults to current release, and does not typically need to be
user-defined. Passed to AnnotationHub for \code{EnsDb} annotation matching,
unless \code{gffFile} is set.}

\item{genomeBuild}{\code{string} or \code{NULL}. Ensembl genome build name (e.g.
"GRCh38"). This will be passed to AnnotationHub for \code{EnsDb} annotation
matching, unless \code{gffFile} is set.}

\item{gffFile}{\code{string} or \code{NULL}. \emph{Advanced use only; not recommended.} By
default, we recommend leaving this \code{NULL} for genomes that are supported on
Ensembl. In this case, the row annotations (\code{\link[=rowRanges]{rowRanges()}}) will be obtained
automatically from Ensembl by passing the \code{organism}, \code{genomeBuild}, and
\code{ensemblRelease} arguments to AnnotationHub and ensembldb. For a genome
that is not supported on Ensembl and/or AnnotationHub, a GFF/GTF (General
Feature Format) file is required. Generally, we recommend using a GTF
(GFFv2) file here over a GFF3 file if possible, although all GFF formats
are supported. The function will internally generate transcript-to-gene
mappings and construct a \code{GRanges} object containing the genomic ranges
(\code{\link[=rowRanges]{rowRanges()}}).}

\item{transgeneNames}{\code{character} or \code{NULL}. Vector indicating which
\code{\link[=assay]{assay()}} rows denote transgenes (e.g. EGFP, TDTOMATO).}

\item{spikeNames}{\code{character} or \code{NULL}. Vector indicating which \code{\link[=assay]{assay()}}
rows denote spike-in sequences (e.g. ERCCs).}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}

\item{...}{Additional arguments, to be stashed in the \code{\link[=metadata]{metadata()}} slot.}
}
\value{
\code{bcbioSingleCell}.
}
\description{
Read bcbio Single-Cell RNA-Seq Data
}
\section{Remote Data}{


When working in RStudio, we recommend connecting to the bcbio-nextgen run
directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}

\examples{
uploadDir <- system.file("extdata/indrops", package = "bcbioSingleCell")

x <- bcbioSingleCell(uploadDir)
print(x)

x <- bcbioSingleCell(
    uploadDir = uploadDir,
    sampleMetadataFile = file.path(uploadDir, "metadata.csv")
)
print(x)
}
\seealso{
\itemize{
\item \code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment::SingleCellExperiment()}}.
\item \code{.S4methods(class = "bcbioSingleCell")}.
}

Other S4 Generators: \code{\link{CellRanger}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{S4 Generators}
