% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{bcbioSingleCellTestsURL}
\alias{bcbioSingleCellTestsURL}
\title{Cache URL}
\format{An object of class \code{character} of length 1.}
\usage{
bcbioSingleCellTestsURL
}
\description{
Cache URL
}
\examples{
bcbioSingleCellTestsURL
}
\keyword{internal}
