% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-bindCB.R
\name{.bindCB}
\alias{.bindCB}
\title{Bind Cellular Barcodes}
\usage{
.bindCB(list)
}
\arguments{
\item{list}{List of cellular barcodes.}
}
\value{
\link{tibble}.
}
\description{
Bind Cellular Barcodes
}
\seealso{
Other Cellular Barcode Utilities: \code{\link{.cbList}},
  \code{\link{.propTblFromDataSet}},
  \code{\link{.readCBFile}}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
