% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-calculateMetrics.R
\docType{methods}
\name{calculateMetrics}
\alias{calculateMetrics}
\alias{calculateMetrics}
\alias{calculateMetrics,dgCMatrix-method}
\alias{calculateMetrics,bcbioSingleCell-method}
\title{Calculate Cellular Barcode Metrics Summary}
\usage{
calculateMetrics(object, ...)

\S4method{calculateMetrics}{dgCMatrix}(object, annotable, prefilter = TRUE)

\S4method{calculateMetrics}{bcbioSingleCell}(object, prefilter = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{annotable}{Annotable.}

\item{prefilter}{Whether to apply pre-filtering to the cellular barcodes.}
}
\value{
\link{data.frame}.
}
\description{
Calculate Cellular Barcode Metrics Summary
}
\note{
This should only be performed during the initial run loading.
}
\author{
Rory Kirchner, Michael Steinbaugh
}
