% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-calculateMetrics.R
\docType{methods}
\name{calculateMetrics}
\alias{calculateMetrics}
\alias{calculateMetrics}
\alias{calculateMetrics,dgCMatrix-method}
\alias{calculateMetrics,bcbioSingleCell-method}
\title{Calculate Cellular Barcode Metrics Summary}
\usage{
calculateMetrics(object, ...)

\S4method{calculateMetrics}{dgCMatrix}(object, annotable = TRUE,
  prefilter = TRUE)

\S4method{calculateMetrics}{bcbioSingleCell}(object, prefilter = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{annotable}{Annotable.}

\item{prefilter}{Whether to apply pre-filtering to the cellular barcodes.}
}
\value{
\link{data.frame}.
}
\description{
Calculate Cellular Barcode Metrics Summary
}
\note{
This should only be performed during the initial run loading.
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
calculateMetrics(bcb) \%>\% glimpse()

# dgCMatrix
counts <- counts(bcb)
annotable <- annotable(bcb)
calculateMetrics(counts, annotable = annotable) \%>\% glimpse()
}
\author{
Rory Kirchner, Michael Steinbaugh
}
