% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-calculateMetrics.R
\docType{methods}
\name{calculateMetrics}
\alias{calculateMetrics}
\alias{calculateMetrics}
\alias{calculateMetrics,dgCMatrix-method}
\alias{calculateMetrics,bcbioSingleCell-method}
\title{Calculate Cellular Barcode Metrics Summary}
\usage{
calculateMetrics(object, ...)

\S4method{calculateMetrics}{dgCMatrix}(object, rowData, prefilter = TRUE)

\S4method{calculateMetrics}{bcbioSingleCell}(object, prefilter = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{rowData}{Data describing the rows of the object.}

\item{prefilter}{Whether to apply pre-filtering to the cellular barcodes.}
}
\value{
\code{data.frame}.
}
\description{
Calculate Cellular Barcode Metrics Summary
}
\note{
This should only be performed during the initial run loading.
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioSingleCell"))

# bcbioSingleCell ====
calculateMetrics(bcb_small) \%>\% glimpse()

# dgCMatrix ====
counts <- counts(bcb_small)
rowData <- rowData(bcb_small)
calculateMetrics(counts, rowData = rowData) \%>\% glimpse()
}
\seealso{
Other Quality Control Functions: \code{\link{metricsPerSample}},
  \code{\link{metrics}}, \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotQuantileHeatmap}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
