% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-calculateMetrics.R
\docType{methods}
\name{calculateMetrics}
\alias{calculateMetrics}
\alias{calculateMetrics}
\alias{calculateMetrics,dgCMatrix-method}
\alias{calculateMetrics,bcbioSingleCell-method}
\title{Calculate Cellular Barcode Metrics Summary}
\usage{
calculateMetrics(object, ...)

\S4method{calculateMetrics}{dgCMatrix}(object, rowData, prefilter = TRUE)

\S4method{calculateMetrics}{bcbioSingleCell}(object, prefilter = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{rowData}{Data describing the rows of the object.}

\item{prefilter}{Whether to apply pre-filtering to the cellular barcodes.}
}
\value{
\code{data.frame}.
}
\description{
Calculate Cellular Barcode Metrics Summary
}
\note{
This step is performed automatically during the initial load.
}
\examples{
# bcbioSingleCell ====
calculateMetrics(bcb_small) \%>\% glimpse()

# dgCMatrix ====
counts <- counts(bcb_small)
rowData <- rowData(bcb_small)
calculateMetrics(counts, rowData = rowData) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{colData}},
  \code{\link{counts}}, \code{\link{fetchGeneData}},
  \code{\link{fetchPCAData}}, \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}}, \code{\link{gene2symbol}},
  \code{\link{interestingGroups}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{rowData}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
