% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-cell2sample.R
\docType{methods}
\name{cell2sample}
\alias{cell2sample}
\alias{cell2sample}
\alias{cell2sample,character-method}
\alias{cell2sample,bcbioSingleCell-method}
\title{Cell to Sample Mappings}
\usage{
cell2sample(object, ...)

\S4method{cell2sample}{character}(object, samples)

\S4method{cell2sample}{bcbioSingleCell}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{samples}{Sample identifiers.}
}
\description{
Cell to Sample Mappings
}
\author{
Michael Steinbaugh
}
