% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cell2sample-methods.R
\docType{methods}
\name{cell2sample}
\alias{cell2sample}
\alias{cell2sample}
\alias{cell2sample,SingleCellExperiment-method}
\title{Cell to Sample Mappings}
\usage{
cell2sample(object, ...)

\S4method{cell2sample}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Named \code{factor} containing sample IDs as the levels and cellular
barcode IDs as the names.
}
\description{
Cell to Sample Mappings
}
\examples{
# SingleCellExperiment ====
x <- cell2sample(cellranger_small)
table(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{fetchGeneData}},
  \code{\link{metricsPerSample}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
