% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-cell2sample.R
\docType{methods}
\name{cell2sample}
\alias{cell2sample}
\alias{cell2sample}
\alias{cell2sample,bcbioSingleCell-method}
\alias{cell2sample,seurat-method}
\title{Cell to Sample Mappings}
\usage{
cell2sample(object, ...)

\S4method{cell2sample}{bcbioSingleCell}(object)

\S4method{cell2sample}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Cell to Sample Mappings
}
\examples{
# bcbioSingleCell
cell2sample(bcb) \%>\% table()

# seurat
cell2sample(pbmc_small) \%>\% table()
}
