% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-cell2sample.R
\docType{methods}
\name{cell2sample}
\alias{cell2sample}
\alias{cell2sample}
\alias{cell2sample,bcbioSingleCell-method}
\alias{cell2sample,seurat-method}
\title{Cell to Sample Mappings}
\usage{
cell2sample(object, ...)

\S4method{cell2sample}{bcbioSingleCell}(object)

\S4method{cell2sample}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Cell to Sample Mappings
}
\examples{
# bcbioSingleCell ====
cell2sample(bcb_small) \%>\% table()

# seurat ====
cell2sample(pbmc_small) \%>\% table()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{calculateMetrics}}, \code{\link{colData}},
  \code{\link{counts}}, \code{\link{fetchGeneData}},
  \code{\link{fetchPCAData}}, \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}}, \code{\link{gene2symbol}},
  \code{\link{interestingGroups}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{rowData}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
