% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cellCycleMarkersMmusculus}
\alias{cellCycleMarkersMmusculus}
\title{Cell Cycle Markers (Mouse)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 97 rows and 3 columns.}
\usage{
cellCycleMarkersMmusculus
}
\description{
Cell Cycle Markers (Mouse)
}
\author{
Michael Steinbaugh
}
\keyword{internal}
