% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellRanger-class}
\alias{CellRanger-class}
\title{\code{CellRanger} Class}
\description{
Extends \code{SingleCellExperiment}, with additional validity checks on the
\code{metadata()} slot.
}
