% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellranger_small}
\alias{cellranger_small}
\title{Cell Ranger Example}
\format{An object of class \code{CellRanger} with 200 rows and 2000 columns.}
\usage{
cellranger_small
}
\description{
Cell Ranger Example
}
\examples{
bcbioSingleCell::cellranger_small
}
\seealso{
Other Minimal Example Data: \code{\link{indrops_small}}
}
\author{
Michael Steinbaugh
}
\concept{Minimal Example Data}
\keyword{datasets}
