% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cellTypeMarkers}
\alias{cellTypeMarkers}
\title{Cell-Type Markers}
\format{An object of class \code{list} of length 3.}
\usage{
cellTypeMarkers
}
\description{
Cell-Type Markers
}
\seealso{
Other Marker Data: \code{\link{cellCycleMarkers}}
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
