% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-cellTypesPerCluster.R
\docType{methods}
\name{cellTypesPerCluster}
\alias{cellTypesPerCluster}
\alias{cellTypesPerCluster}
\alias{cellTypesPerCluster,grouped_df-method}
\title{Cell Types per Cluster}
\usage{
cellTypesPerCluster(object, ...)

\S4method{cellTypesPerCluster}{grouped_df}(object, min = 1L, max = Inf)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{min}{Minimum number of marker genes per cluster.}

\item{max}{Maximum number of marker genes per cluster.}
}
\value{
\code{grouped_df} grouped by "\code{cluster}", containing the count (\code{n}) of
significant known makers per cell type.
}
\description{
Cell Types per Cluster
}
\examples{
# grouped_df ====
cellTypesPerCluster(known_markers_small) \%>\% glimpse()
}
\seealso{
Other Clustering Functions: \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotMarker}},
  \code{\link{plotPCElbow}}, \code{\link{plotTSNE}},
  \code{\link{plotTopMarkers}},
  \code{\link{sanitizeMarkers}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
