% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-cellTypesPerCluster.R
\docType{methods}
\name{cellTypesPerCluster}
\alias{cellTypesPerCluster}
\alias{cellTypesPerCluster}
\alias{cellTypesPerCluster,grouped_df-method}
\title{Cell Types per Cluster}
\usage{
cellTypesPerCluster(object, ...)

\S4method{cellTypesPerCluster}{grouped_df}(object, min = 1L, max = Inf)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{min}{Minimum number of marker genes per cluster.}

\item{max}{Maximum number of marker genes per cluster.}
}
\value{
\link{tibble} grouped by cluster, containing the count (\code{n}) of
significant known makers per cell type.
}
\description{
Cell Types per Cluster
}
\note{
This function only returns the positive markers per cluster.
}
\examples{
load(system.file(
    file.path("extdata", "knownMarkersDetected.rda"),
    package = "bcbioSingleCell"))

cellTypesPerCluster(knownMarkersDetected) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
