% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellTypesPerCluster-methods.R
\docType{methods}
\name{cellTypesPerCluster}
\alias{cellTypesPerCluster}
\alias{cellTypesPerCluster,grouped_df-method}
\title{Cell Types per Cluster}
\usage{
cellTypesPerCluster(object, ...)

\S4method{cellTypesPerCluster}{grouped_df}(object, min = 1L, max = Inf)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{min}{\code{scalar integer}. Minimum number of marker genes per cluster.}

\item{max}{\code{scalar integer}. Maximum number of marker genes per cluster.}
}
\value{
\code{grouped_df} grouped by "\code{cluster}", containing the count (\code{n}) of
significant known makers per cell type.
}
\description{
Cell Types per Cluster
}
\examples{
# grouped_df ====
x <- cellTypesPerCluster(known_markers_small)
glimpse(x)
}
\seealso{
Other Clustering Functions: \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeature}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotReducedDim}},
  \code{\link{sanitizeSeuratMarkers}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Functions}
