% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cellular_barcodes.R
\name{.proportional_cb}
\alias{.proportional_cb}
\alias{.read_cellular_barcode_file}
\alias{.cellular_barcodes}
\alias{.bind_cellular_barcodes}
\title{Cellular barcode operations}
\usage{
.proportional_cb(object)

.read_cellular_barcode_file(file)

.cellular_barcodes(sample_dirs)

.bind_cellular_barcodes(list)
}
\arguments{
\item{file}{Cellular barcode TSV file.}

\item{sample_dirs}{Sample directories.}

\item{list}{Cellular barcodes list.}
}
\value{
\link{tibble}.

\link{tibble}.

\link{list}.

\link{tibble}.
}
\description{
Cellular barcode operations
}
\details{
Modified version of Klein Lab MATLAB code.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\keyword{internal}
