% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cellular_barcodes.R
\name{.read_cellular_barcode_file}
\alias{.read_cellular_barcode_file}
\alias{.cellular_barcodes}
\alias{.bind_cellular_barcodes}
\title{Cellular barcode operations}
\usage{
.read_cellular_barcode_file(file)

.cellular_barcodes(sample_dirs)

.bind_cellular_barcodes(list)
}
\arguments{
\item{file}{Cellular barcode TSV file.}

\item{sample_dirs}{Sample directories.}

\item{list}{Cellular barcodes list}
}
\description{
Cellular barcode operations
}
\author{
Michael Steinbaugh
}
\keyword{internal}
