% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{coerce}
\alias{coerce}
\title{Coerce \code{bcbio} Data to Other Single-Cell Classes}
\arguments{
\item{from}{Object to coerce.}
}
\description{
Coerce \code{bcbio} Data to Other Single-Cell Classes
}
\author{
Michael Steinbaugh
}
