% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coerce.R
\name{coerce}
\alias{coerce}
\alias{as}
\alias{coerce-SingleCellExperiment-seurat}
\title{Methods for Coercing an Object to a Class}
\value{
Object of new class.
}
\description{
Methods for Coercing an Object to a Class
}
\section{SingleCellExperiment to seurat}{

Interally \code{\link[Seurat:CreateSeuratObject]{Seurat::CreateSeuratObject()}} is called without applying any
additional filtering cutoffs, since we have already defined them during our
quality control analysis. Here we are passing the raw gene-level counts of
the filtered cells into a new \code{seurat} class object.
}

\examples{
# SingleCellExperiment to seurat ====
x <- as(bcb_small, "seurat")
class(x)
print(x)
}
\seealso{
\itemize{
\item \code{\link[methods:as]{methods::as()}}.
\item \code{\link[methods:coerce]{methods::coerce()}}.
}
}
\author{
Michael Steinbaugh
}
