% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coerce.R
\name{coerce}
\alias{coerce}
\alias{coerce-bcbioSingleCell-seurat}
\title{Coerce Object}
\arguments{
\item{from}{Class for which the coerce method will perform coercion.}
}
\description{
Coerce Object
}
\section{bcbioSingleCell to seurat}{

Interally \code{\link[Seurat:CreateSeuratObject]{Seurat::CreateSeuratObject()}} is called without applying any
additional filtering cutoffs, since we have already defined them during our
quality control analysis. Here we are passing the raw gene-level counts of
the filtered cells into a new \code{seurat} class object, using \code{\link[=as]{as()}} object
coercion.
}

\examples{
# Coerce bcbioSingleCell to seurat
as(bcb_small, "seurat")
}
\seealso{
\code{help(topic = "coerce", package = "methods")}.
}
\author{
Michael Steinbaugh
}
