% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAs.R
\name{coerce}
\alias{coerce}
\alias{upgrade-bcbioSingleCell}
\alias{coerce-bcbioSingleCell-seurat}
\alias{coerce-bcbioSingleCell-SummarizedExperiment}
\title{Coerce Object}
\arguments{
\item{from}{Class for which the coerce method will perform coercion.}
}
\description{
Coerce Object
}
\section{Upgrade bcbioSingleCell to current version}{

This method adds support for upgrading \code{bcbioSCDataSet} objects to the latest
\link{bcbioSingleCell} class version. This should be backwards compatible to
\link{bcbioSingleCell} version 0.0.17. Previous objects saved using
\code{bcbioSinglecell} (note case) will likely fail to load with newer versions of
the package.
}

\section{bcbioSingleCell to seurat}{

Interally \code{\link[Seurat:CreateSeuratObject]{Seurat::CreateSeuratObject()}} is called without applying any
additional filtering cutoffs, since we have already defined them during our
quality control analysis. Here we are passing the raw gene-level counts of
the filtered cells into a new \link{seurat} class object, using \code{\link[=as]{as()}} object
coercion.
}

\section{bcbioSingleCell to SummarizedExperiment}{

Since \link{bcbioSingleCell} is an extension of \link{SummarizedExperiment}, this
coercion method is very simple. Here we're simply dropping our \code{@bcbio} slot,
which contains raw cellular barcodes and other bcbio-specific metadata.
}

\examples{
load(system.file(
    file.path("inst", "extdata", "filtered.rda"),
    package = "bcbioSingleCell"))

# Coerce bcbioSingleCell to seurat
as(filtered, "seurat")
}
\seealso{
\code{help(topic = "coerce", package = "methods")}.
}
\author{
Michael Steinbaugh
}
