% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAs.R
\name{coerce}
\alias{coerce}
\alias{coerce-bcbioSCFiltered-seurat}
\title{Coerce Object}
\description{
Coerce Object
}
\section{\link{bcbioSCFiltered} to \link{seurat}}{

Interally, this begins by calling \code{\link[Seurat:CreateSeuratObject]{Seurat::CreateSeuratObject()}} without any
additional filtering cutoffs, since we already applied them during our
quality control analysis. Here we are passing the raw gene-level counts of
the filtered cells into a new \link{seurat} class object, using \code{\link[=as]{as()}} object
coercion. Next, global-scaling normalization is applied to the raw counts
with \code{\link[Seurat:NormalizeData]{Seurat::NormalizeData()}}, which (1) normalizes the gene expression
measurements for each cell by the total expression, (2) multiplies this by a
scale factor (10,000 by default), and (3) log-transforms the result.
\code{\link[Seurat:FindVariableGenes]{Seurat::FindVariableGenes()}} is then called, which calculates the average
expression and dispersion for each gene, places these genes into bins, and
then calculates a z-score for dispersion within each bin. This helps control
for the relationship between variability and average expression. Finally, the
genes are scaled and centered using the \code{\link[Seurat:ScaleData]{Seurat::ScaleData()}} function.
}

\examples{
\dontrun{
seurat <- as(bcb, "seurat")
}
}
\seealso{
\code{help("coerce")}
}
\author{
Michael Steinbaugh
}
