% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-methods.R
\name{coerce}
\alias{coerce}
\alias{as}
\alias{coerce-SingleCellExperiment-seurat}
\alias{coerce-seurat-SingleCellExperiment}
\title{Methods for Coercing an Object to a Class}
\value{
Object of new class.
}
\description{
Methods for Coercing an Object to a Class
}
\note{
Use \code{\link[=convertGenesToSymbols]{convertGenesToSymbols()}} to convert gene IDs to names (symbols).
This step is no longer automatic, as of v0.1.18.
}
\section{\code{SingleCellExperiment} to \code{seurat}}{

Interally \code{\link[Seurat:CreateSeuratObject]{Seurat::CreateSeuratObject()}} is called without applying any
additional filtering cutoffs, since we have already defined them during our
quality control analysis. Here we are passing the raw gene-level counts of
the filtered cells into a new \code{seurat} class object.
}

\section{\code{seurat} to \code{SingleCellExperiment}}{

Super basic S4 coercion support for taking the raw counts matrix from
a \code{seurat} class object and coercing to a \code{SingleCellExperiment}.
}

\examples{
# SingleCellExperiment to seurat ====
x <- as(indrops_small, "seurat")
class(x)
print(x)
# seurat to SingleCellExperiment ====
x <- as(seurat_small, "SingleCellExperiment")
class(x)
print(x)
}
\seealso{
\itemize{
\item \code{\link[Seurat:CreateSeuratObject]{Seurat::CreateSeuratObject()}}.
\item \code{\link[Seurat:Convert]{Seurat::Convert()}}.
\item \code{\link[methods:as]{methods::as()}}.
\item \code{\link[methods:coerce]{methods::coerce()}}.
}

Other S4 Object: \code{\link{bcbioSingleCell}},
  \code{\link{extract}}, \code{\link{show}},
  \code{\link{updateObject}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Object}
