% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-colData.R
\docType{methods}
\name{colData}
\alias{colData}
\alias{colData,seurat-method}
\title{Column Data}
\usage{
\S4method{colData}{seurat}(x)
}
\arguments{
\item{x}{Primary object.}
}
\value{
No value.
}
\description{
Column Data
}
\author{
Michael Steinbaugh
}
