% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-colData.R
\docType{methods}
\name{colData}
\alias{colData}
\alias{colData,SingleCellExperiment-method}
\title{Column Data}
\usage{
\S4method{colData}{SingleCellExperiment}(x)
}
\arguments{
\item{x}{Primary object.}
}
\description{
Column Data
}
\examples{
# bcbioSingleCell ====
colData(bcb_small) \%>\% glimpse()

# seurat ====
colData(seurat_small) \%>\% glimpse()
colData(pbmc_small) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
\keyword{internal}
