% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-colData.R
\docType{methods}
\name{colData}
\alias{colData}
\alias{colData,bcbioSingleCell-method}
\alias{colData,seurat-method}
\title{Column Data}
\usage{
\S4method{colData}{bcbioSingleCell}(x, return = c("data.frame", "DataFrame"))

\S4method{colData}{seurat}(x, return = c("data.frame", "DataFrame"))
}
\arguments{
\item{x}{Primary object.}

\item{return}{Return data as \code{data.frame} or \code{DataFrame}.}
}
\value{
Data describing the columns (cells).
}
\description{
Column Data
}
\examples{
# seurat ====
colData(pbmc_small)
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{fetchGeneData}}, \code{\link{fetchPCAData}},
  \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}}, \code{\link{gene2symbol}},
  \code{\link{interestingGroups}},
  \code{\link{readCellTypeMarkers}}, \code{\link{rowData}},
  \code{\link{sampleData}}, \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
