% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-method.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,SingleCellExperiment,SingleCellExperiment-method}
\title{Combine Multiple \code{SingleCellExperiment} Objects}
\usage{
\S4method{combine}{SingleCellExperiment,SingleCellExperiment}(x, y, ...)
}
\arguments{
\item{x}{One of the values.}

\item{y}{A second value.}

\item{...}{Any other objects of the same class as \code{x} and \code{y}.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Combine Multiple \code{SingleCellExperiment} Objects
}
\examples{
# Rows (genes) must be identical
# Organism and genome build must be identical
# For this simple working example, let's duplicate our minimal dataset,
# modify the cell barcodes, and combine

x <- indrops_small
validObject(x)
sampleData(x)
sampleNames(x)

# Here we're faking a distinct replicate, just as an example
y <- indrops_small
# Modify the cellular barcodes prefix
colnames(y) <- gsub(
    pattern = "^multiplexed_AAAAAAAA",
    replacement = "multiplexed_CCCCCCCC",
    x = colnames(y)
)
# Modify the column data
colData(y)[["description"]] <- factor("multiplexed-CCCCCCCC")
colData(y)[["sampleID"]] <- factor("multiplexed_CCCCCCCC")
colData(y)[["sampleName"]] <- factor("rep_2")
colData(y)[["sequence"]] <- factor("GGGGGGGG")
colData(y)[["revcomp"]] <- factor("CCCCCCCC")
# Modify the sample data
sampleData(y)[["sampleName"]] <- factor("rep_2")
sampleData(y)[["description"]] <- factor("multiplexed-CCCCCCCC")
sampleData(y)[["sequence"]] <- factor("GGGGGGGG")
sampleData(y)[["revcomp"]] <- factor("CCCCCCCC")
rownames(sampleData(y)) <- "multiplexed_CCCCCCCC"
validObject(y)
sampleData(y)
sampleNames(y)

# Combine two SingleCellExperiment objects
combine(x, y)
}
\seealso{
\code{help("merge.Matrix", "Matrix.utils")}.
}
\author{
Michael Steinbaugh
}
