% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_features.R
\name{combine_features}
\alias{combine_features}
\title{Combine features of a sparse matrix}
\usage{
combine_features(sparse, featureids)
}
\arguments{
\item{sparse}{A sparse matrix of counts}

\item{featureids}{New feature ids for the collapsed features}
}
\value{
Sparse matrix
}
\description{
This collapsed features with the same feature id by summing them
}
\author{
Rory Kirchner
}
