% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_technical_replicates.R
\name{combine_technical_replicates}
\alias{combine_technical_replicates}
\title{combine technical replicates}
\usage{
combine_technical_replicates(counts, cellids)
}
\arguments{
\item{counts}{a sparse matrix of counts}

\item{cellids}{new cellular ids of the collapsed cells}
}
\value{
sparse matrix with technical replicates combined
}
\description{
combine technical replicates
}
\author{
Rory Kirchner
}
