% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump.R, R/methods-counts.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{.counts}
\alias{counts,bcbioSCDataSet-method}
\alias{counts,bcbioSCSubset-method}
\title{Counts Accessor}
\usage{
counts(object, ...)

\S4method{counts}{bcbioSCDataSet}(object, gene2symbol = FALSE,
  format = "dgCMatrix")

\S4method{counts}{bcbioSCSubset}(object, gene2symbol = FALSE,
  format = "dgCMatrix")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{gene2symbol}{Convert Ensembl gene identifiers (rownames) to gene
symbols. Recommended for passing counts to Seurat.}

\item{format}{Return counts as sparse matrix (\strong{recommended}; \code{dgCMatrix},
\code{dgTMatrix}) or dense matrix (\code{matrix}).}
}
\value{
\link{matrix}.
}
\description{
Counts Accessor
}
\author{
Michael Steinbaugh
}
