% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-counts.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,bcbioSingleCell-method}
\alias{counts,seurat-method}
\title{Counts Accessor}
\usage{
\S4method{counts}{bcbioSingleCell}(object)

\S4method{counts}{seurat}(object, normalized = FALSE)
}
\arguments{
\item{object}{Object.}

\item{normalized}{Normalized (\code{TRUE}) or raw (\code{FALSE}) counts.}
}
\value{
Counts matrix (typically a \code{dgCMatrix}).
}
\description{
Counts Accessor
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioSingleCell"))

# bcbioSingleCell ====
counts(bcb_small) \%>\% glimpse()

# seurat ====
counts(pbmc_small, normalized = FALSE) \%>\% glimpse()
counts(pbmc_small, normalized = TRUE) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
