% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-bcbio_counts.R
\name{.bcbio_counts}
\alias{.bcbio_counts}
\title{Import transcript-level count data from a bcbio run into a sparse matrix}
\usage{
.bcbio_counts(project_dir, strip_version = TRUE)
}
\arguments{
\item{strip_version}{Strip transcript version from identifier.}
}
\value{
Sparse counts matrix.
}
\description{
Import transcript-level count data from a bcbio run into a sparse matrix
}
\author{
Michael Steinbaugh
}
