% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-counts.R
\docType{methods}
\name{.counts}
\alias{.counts}
\alias{counts,bcbioSCDataSet-method}
\alias{counts,SummarizedExperiment-method}
\title{Counts accessor}
\usage{
\S4method{counts}{bcbioSCDataSet}(object, gene2symbol = FALSE,
  format = "dgCMatrix")

\S4method{counts}{SummarizedExperiment}(object, gene2symbol = FALSE,
  format = "dgCMatrix")
}
\arguments{
\item{object}{\link{bcbioSCDataSet}.}

\item{gene2symbol}{Convert Ensembl gene identifiers (rownames) to gene
symbols. Recommended for passing counts to Seurat.}

\item{format}{Return counts as sparse matrix (\strong{recommended}; \code{dgCMatrix},
\code{dgTMatrix}) or dense matrix (\code{matrix}).}
}
\value{
Counts matrix.
}
\description{
Counts accessor
}
\author{
Michael Steinbaugh
}
