% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-counts.R, R/reexports.R
\docType{methods}
\name{counts,bcbioSCDataSet-method}
\alias{counts,bcbioSCDataSet-method}
\alias{counts}
\title{Counts accessor}
\usage{
\S4method{counts}{bcbioSCDataSet}(object)

counts(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters.}
}
\value{
Sparse counts (\code{dgCMatrix}).
}
\description{
Counts accessor
}
\author{
Michael Steinbaugh
}
