% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-counts.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,bcbioSCDataSet-method}
\alias{counts,bcbioSCFiltered-method}
\title{Counts Accessor}
\usage{
\S4method{counts}{bcbioSCDataSet}(object, gene2symbol = FALSE,
  as = "dgCMatrix")

\S4method{counts}{bcbioSCFiltered}(object, gene2symbol = FALSE,
  as = "dgCMatrix")
}
\arguments{
\item{object}{Object.}

\item{gene2symbol}{Convert Ensembl gene identifiers (rownames) to gene
symbols. Recommended for passing counts to Seurat.}

\item{as}{Return class (\strong{recommended}; \code{dgCMatrix},
\code{dgTMatrix}) or dense matrix (\code{matrix}).}
}
\value{
Matrix class object.
}
\description{
Counts Accessor
}
\author{
Michael Steinbaugh
}
