% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-counts.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,bcbioSCDataSet-method}
\title{Counts accessor}
\usage{
counts(object, ...)

\S4method{counts}{bcbioSCDataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters.}
}
\value{
Sparse counts (\code{dgCMatrix}).
}
\description{
Counts accessor
}
\author{
Michael Steinbaugh
}
