% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-counts.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,bcbioSingleCell-method}
\alias{counts,seurat-method}
\title{Counts Accessor}
\usage{
\S4method{counts}{bcbioSingleCell}(object, gene2symbol = FALSE)

\S4method{counts}{seurat}(object, normalized = FALSE)
}
\arguments{
\item{object}{Object.}

\item{gene2symbol}{Convert Ensembl gene identifiers (rownames) to gene
symbols. Recommended for passing counts to Seurat.}

\item{normalized}{Normalized (\code{TRUE}) or raw (\code{FALSE}) counts.}
}
\value{
Counts matrix.
}
\description{
Counts Accessor
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat.rda", package = "bcbioSingleCell"))

# bcbioSingleCell
counts(bcb, gene2symbol = FALSE) \%>\% glimpse()
counts(bcb, gene2symbol = TRUE) \%>\% glimpse()

# seurat
counts(seurat, normalized = FALSE) \%>\% glimpse()
counts(seurat, normalized = TRUE) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
