% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-seurat-counts.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,seurat-method}
\title{Counts Accessor}
\usage{
\S4method{counts}{seurat}(object, normalized = FALSE)
}
\arguments{
\item{object}{Object.}

\item{normalized}{Normalized (\code{TRUE}) or raw (\code{FALSE}) counts.}
}
\value{
Counts matrix (typically a \code{dgCMatrix}).
}
\description{
Counts Accessor
}
\examples{
# SingleCellExperiment ====
counts(bcb_small) \%>\% glimpse()
counts(cellranger_small) \%>\% glimpse()

# seurat ====
counts(pbmc_small, normalized = FALSE) \%>\% glimpse()
counts(pbmc_small, normalized = TRUE) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{fetchGeneData}},
  \code{\link{fetchPCAData}}, \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
