% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-counts.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,bcbioSingleCell-method}
\alias{counts,seurat-method}
\title{Counts Accessor}
\usage{
\S4method{counts}{bcbioSingleCell}(object)

\S4method{counts}{seurat}(object, normalized = FALSE)
}
\arguments{
\item{object}{Object.}

\item{normalized}{Normalized (\code{TRUE}) or raw (\code{FALSE}) counts.}
}
\value{
Counts matrix (typically a \code{dgCMatrix}).
}
\description{
Counts Accessor
}
\examples{
# bcbioSingleCell ====
counts(bcb_small) \%>\% glimpse()

# seurat ====
counts(pbmc_small, normalized = FALSE) \%>\% glimpse()
counts(pbmc_small, normalized = TRUE) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{colData}},
  \code{\link{fetchGeneData}}, \code{\link{fetchPCAData}},
  \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}}, \code{\link{gene2symbol}},
  \code{\link{interestingGroups}},
  \code{\link{readCellTypeMarkersFile}},
  \code{\link{rowData}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
