% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_create_bcbio_skeleton.R
\name{create_bcbio_skeleton}
\alias{create_bcbio_skeleton}
\title{Create a skeleton bcbio project.}
\usage{
create_bcbio_skeleton(base_dir = getwd(), run_dir = "skeleton",
  config_dir = "config", upload_dir = "final", project_dir = NULL,
  sample_dirs = NULL, organism, intgroup = "sample_name")
}
\arguments{
\item{base_dir}{Base directory in which to create skeleton bcbio structure.}

\item{config_dir}{Configuration directory.}

\item{upload_dir}{Upload directory.}

\item{project_dir}{Project directory (YYYY-MM-DD_template), nested inside
upload directory.}

\item{sample_dirs}{Sample directories, nested inside upload directory.}

\item{organism}{Organism to use.}
}
\value{
bcbio run skeleton.
}
\description{
Create a skeleton bcbio project.
}
\author{
Rory Kirchner

Michael Steinbaugh
}
\keyword{internal}
