% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{plotClusters}
\alias{plotTSNEExpressionData}
\alias{loadSingleCellRun}
\alias{darkTheme}
\alias{pcCutoff}
\alias{quantileHeatmap}
\alias{plotKnownMarkers}
\alias{readMarkers}
\alias{readMarkersFile}
\alias{plotMarker}
\title{Defunct Functions}
\usage{
plotClusters(...)

plotTSNEExpressionData(...)

loadSingleCellRun(...)

darkTheme(...)

pcCutoff(...)

quantileHeatmap(...)

plotKnownMarkers(...)

readMarkers(...)

readMarkersFile(...)

plotMarker(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{\link[=.Defunct]{.Defunct()}}.
}
\description{
Defunct Functions
}
\keyword{internal}
