% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{loadSingleCellRun}
\alias{darkTheme}
\alias{pcCutoff}
\alias{calculateMetrics}
\alias{inflectionPoint}
\alias{plotCumulativeUMIsPerCell}
\alias{cellCountsPerCluster}
\alias{cellTypesPerCluster}
\alias{clusterCellCountsPerSample}
\alias{diffExp}
\alias{fetchGeneData}
\alias{fetchPCAData}
\alias{fetchTSNEData}
\alias{fetchTSNEExpressionData}
\alias{fetchUMAPData}
\alias{fetchUMAPExpressionData}
\alias{knownMarkersDetected}
\alias{plotCellTypesPerCluster}
\alias{plotFeatureTSNE}
\alias{plotFeatureUMAP}
\alias{plotKnownMarkers}
\alias{plotKnownMarkersDetected}
\alias{plotMarker}
\alias{plotMarkerTSNE}
\alias{plotMarkerUMAP}
\alias{plotMarkers}
\alias{plotPCElbow}
\alias{plotTSNE}
\alias{plotTSNEExpressionData}
\alias{plotTopMarkers}
\alias{plotUMAP}
\alias{readCellTypeMarkers}
\alias{sanitizeMarkers}
\alias{topMarkers}
\title{Defunct Functions}
\usage{
loadSingleCellRun(...)

darkTheme(...)

pcCutoff(...)

calculateMetrics(...)

inflectionPoint(...)

plotCumulativeUMIsPerCell(...)

cellCountsPerCluster(...)

cellTypesPerCluster(...)

clusterCellCountsPerSample(...)

diffExp(...)

fetchGeneData(...)

fetchPCAData(...)

fetchTSNEData(...)

fetchTSNEExpressionData(...)

fetchUMAPData(...)

fetchUMAPExpressionData(...)

knownMarkersDetected(...)

plotCellTypesPerCluster(...)

plotFeatureTSNE(...)

plotFeatureUMAP(...)

plotKnownMarkers(...)

plotKnownMarkersDetected(...)

plotMarker(...)

plotMarkerTSNE(...)

plotMarkerUMAP(...)

plotMarkers(...)

plotPCElbow(...)

plotTSNE(...)

plotTSNEExpressionData(...)

plotTopMarkers(...)

plotUMAP(...)

readCellTypeMarkers(...)

sanitizeMarkers(...)

topMarkers(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{\link[=.Defunct]{.Defunct()}}.
}
\description{
Defunct Functions
}
\keyword{internal}
