% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{loadSingleCell}
\title{Deprecated Functions}
\usage{
loadSingleCell(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated Functions
}
\keyword{internal}
