% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{loadCellRanger}
\alias{loadSingleCell}
\alias{readCellRanger}
\title{Deprecated Functions}
\usage{
loadCellRanger(...)

loadSingleCell(...)

readCellRanger(...)
}
\value{
\code{\link[=.Deprecated]{.Deprecated()}}.
}
\description{
Deprecated Functions
}
\keyword{internal}
