% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{calculateMetrics}
\alias{inflectionPoint}
\alias{plotCumulativeUMIsPerCell}
\alias{loadCellRanger}
\alias{loadSingleCell}
\title{Deprecated Functions}
\usage{
calculateMetrics(...)

inflectionPoint(...)

plotCumulativeUMIsPerCell(...)

loadCellRanger(...)

loadSingleCell(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{\link[=.Deprecated]{.Deprecated()}}.
}
\description{
Deprecated Functions
}
\keyword{internal}
