% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{calculateMetrics}
\alias{inflectionPoint}
\alias{plotCumulativeUMIsPerCell}
\alias{loadCellRanger}
\alias{loadSingleCell}
\alias{cellCountsPerCluster}
\alias{cellTypesPerCluster}
\alias{clusterCellCountsPerSample}
\alias{diffExp}
\alias{fetchGeneData}
\alias{fetchPCAData}
\alias{fetchTSNEData}
\alias{fetchTSNEExpressionData}
\alias{fetchUMAPData}
\alias{fetchUMAPExpressionData}
\alias{knownMarkersDetected}
\alias{plotCellTypesPerCluster}
\alias{plotFeatureTSNE}
\alias{plotFeatureUMAP}
\alias{plotGene}
\alias{plotKnownMarkers}
\alias{plotKnownMarkersDetected}
\alias{plotMarker}
\alias{plotMarkerTSNE}
\alias{plotMarkerUMAP}
\alias{plotMarkers}
\alias{plotPCElbow}
\alias{plotTSNE}
\alias{plotTSNEExpressionData}
\alias{plotTopMarkers}
\alias{plotUMAP}
\alias{readCellTypeMarkers}
\alias{sanitizeMarkers}
\alias{topMarkers}
\title{Deprecated Functions}
\usage{
calculateMetrics(...)

inflectionPoint(...)

plotCumulativeUMIsPerCell(...)

loadCellRanger(...)

loadSingleCell(...)

cellCountsPerCluster(...)

cellTypesPerCluster(...)

clusterCellCountsPerSample(...)

diffExp(...)

fetchGeneData(...)

fetchPCAData(...)

fetchTSNEData(...)

fetchTSNEExpressionData(...)

fetchUMAPData(...)

fetchUMAPExpressionData(...)

knownMarkersDetected(...)

plotCellTypesPerCluster(...)

plotFeatureTSNE(...)

plotFeatureUMAP(...)

plotGene(...)

plotKnownMarkers(...)

plotKnownMarkersDetected(...)

plotMarker(...)

plotMarkerTSNE(...)

plotMarkerUMAP(...)

plotMarkers(...)

plotPCElbow(...)

plotTSNE(...)

plotTSNEExpressionData(...)

plotTopMarkers(...)

plotUMAP(...)

readCellTypeMarkers(...)

sanitizeMarkers(...)

topMarkers(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{\link[=.Deprecated]{.Deprecated()}}.
}
\description{
Deprecated Functions
}
\keyword{internal}
