% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{barcode_metrics}
\alias{filter_barcodes}
\alias{plot_genes_detected}
\alias{plot_mito_counts}
\alias{plot_total_cells}
\alias{plot_total_vs_detected}
\title{Deprecated Functions}
\usage{
barcode_metrics()

filter_barcodes()

plot_genes_detected()

plot_mito_counts()

plot_total_cells()

plot_total_vs_detected()
}
\value{
Varies by function.
}
\description{
Deprecated Functions
}
\keyword{internal}
