% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{plotClusters}
\alias{plotTSNEExpressionData}
\title{Deprecated Functions}
\usage{
plotClusters(...)

plotTSNEExpressionData(...)
}
\arguments{
\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
Soft deprecation to new functions.
}
\description{
Deprecated Functions
}
\keyword{internal}
