% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{calculateMetrics}
\alias{readCellTypeMarkersFile}
\alias{inflectionPoint}
\alias{plotCumulativeUMIsPerCell}
\alias{loadCellRanger}
\alias{loadSingleCell}
\alias{sanitizeMarkers}
\alias{fetchPCAData}
\alias{fetchTSNEData}
\alias{fetchTSNEExpressionData}
\alias{fetchUMAPData}
\alias{fetchUMAPExpressionData}
\alias{plotFeatureTSNE}
\alias{plotFeatureUMAP}
\alias{plotMarkerTSNE}
\alias{plotMarkerUMAP}
\title{Deprecated Functions}
\usage{
calculateMetrics(...)

readCellTypeMarkersFile(...)

inflectionPoint(...)

plotCumulativeUMIsPerCell(...)

loadCellRanger(...)

loadSingleCell(...)

sanitizeMarkers(...)

fetchPCAData(object, ...)

fetchTSNEData(object, ...)

fetchTSNEExpressionData(object, ...)

fetchUMAPData(object, ...)

fetchUMAPExpressionData(object, ...)

plotFeatureTSNE(object, ...)

plotFeatureUMAP(object, ...)

plotMarkerTSNE(object, ...)

plotMarkerUMAP(object, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}
}
\value{
\code{\link[=.Deprecated]{.Deprecated()}}.
}
\description{
Deprecated Functions
}
\keyword{internal}
