% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{plotClusters}
\alias{plotTSNEExpressionData}
\alias{loadSingleCellRun}
\alias{plotFeatures}
\alias{pcCutoff}
\title{Deprecated Functions}
\usage{
plotClusters(...)

plotTSNEExpressionData(...)

loadSingleCellRun(...)

plotFeatures(object, features, ...)

pcCutoff(...)
}
\arguments{
\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{object}{Object.}
}
\value{
Soft deprecation to new functions.
}
\description{
Deprecated Functions
}
\keyword{internal}
