% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{barcode_metrics}
\alias{barcode_metrics}
\alias{filter_barcodes}
\alias{plot_mito_counts}
\alias{plot_total_cells}
\alias{plot_total_vs_detected}
\title{Deprecated functions}
\usage{
barcode_metrics(...)

filter_barcodes(...)

plot_mito_counts(...)

plot_total_cells(...)

plot_total_vs_detected(...)
}
\arguments{
\item{...}{Passthrough parameters}
}
\description{
Deprecated functions
}
\author{
Michael Steinbaugh
}
\keyword{internal}
