% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-detect.R
\name{.detect_tenx}
\alias{.detect_tenx}
\title{Detect scRNA-seq analysis pipeline}
\usage{
.detect_tenx(upload_dir)
}
\arguments{
\item{upload_dir}{Final upload directory.}
}
\description{
Currently supports bcbio-nextgen and 10X Chromium.
}
\author{
Michael Steinbaugh
}
