% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-detect.R
\name{.detect_pipeline}
\alias{.detect_pipeline}
\title{Detect scRNA-seq analysis pipeline}
\usage{
.detect_pipeline(upload_dir)
}
\arguments{
\item{upload_dir}{Final upload directory.}
}
\value{
Pipeline string. Stops on detection failure.
}
\description{
Currently supports bcbio-nextgen and 10X Chromium.
}
\author{
Michael Steinbaugh
}
