% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-detectOrganism.R
\docType{methods}
\name{detectOrganism}
\alias{detectOrganism}
\alias{detectOrganism,bcbioSingleCell-method}
\alias{detectOrganism,seurat-method}
\title{Detect Organism}
\usage{
\S4method{detectOrganism}{bcbioSingleCell}(object)

\S4method{detectOrganism}{seurat}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Full latin organism name.
}
\description{
Detect Organism
}
\details{
Currently supported organisms:
\itemize{
\item \emph{Homo sapiens} (human)
\item \emph{Mus musculus} (mouse)
\item \emph{Rattus norvegicus} (rat)
\item \emph{Drosophila melanogaster} (fruitfly)
\item \emph{Caenorhabditis elegans} (roundworm)
\item \emph{Saccharomyces cerevisiae} (yeast)
\item \emph{Gallus gallus} (chicken)
\item \emph{Ovis aries} (sheep)
}
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioSingleCell"))
load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell

# seurat
}
\seealso{
Other Gene Annotation Utilities: \code{\link{annotable}},
  \code{\link{gene2symbolFromGFF}},
  \code{\link{gene2symbol}}, \code{\link{symbol2gene}},
  \code{\link{tx2geneFromGFF}}, \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh
}
