% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-diffExp.R
\docType{methods}
\name{diffExp}
\alias{diffExp}
\alias{diffExp}
\alias{diffExp,SingleCellExperiment-method}
\alias{diffExp,seurat-method}
\title{Differential Expression}
\usage{
diffExp(object, ...)

\S4method{diffExp}{SingleCellExperiment}(object, numerator, denominator,
  zeroWeights = c("zinbwave", "zingeR"), caller = c("edgeR", "DESeq2"),
  minCellsPerGene = 25L, minCountsPerCell = 5L)

\S4method{diffExp}{seurat}(object, numerator, denominator,
  zeroWeights = c("zinbwave", "zingeR"), caller = c("edgeR", "DESeq2"),
  minCellsPerGene = 25L, minCountsPerCell = 5L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{numerator}{Group of cells to use in the numerator of the contrast
(e.g. treatment).}

\item{denominator}{Group of cells to use in the denominator of the contrast
(e.g. control).}

\item{zeroWeights}{Package to use for zero weight calculations. Defaults to
zinbwave but zingeR is also supported.}

\item{caller}{Package to use for differential expression calling. Defaults to
edgeR (faster for large datasets) but DESeq2 is also supported.}

\item{minCellsPerGene}{The minimum number of cells where a gene is expressed,
to pass low expression filtering. Set to \code{0} to disable (not recommended).}

\item{minCountsPerCell}{Minimum number of counts per cell for a gene to pass
low expression filtering. The number of cells is defined by
\code{minCellsPerGene}. Set to \code{0} to disable (not recommended).}
}
\value{
\itemize{
\item \code{caller = "edgeR"}: \code{DEGLRT}.
\item \code{caller = "DESeq2"}: Unshrunken \code{DESeqResults}. Use \code{lfcShrink()} if
shrunken results are desired.
}
}
\description{
We now generally recommend the ZINB-WaVE method over zingeR, since it is
faster, and has been show to be more sensitive for most single-cell RNA-seq
datasets.
}
\section{zinbwave}{

We are currently using an epsilon setting of 1e12, as recommended by the
ZINB-WaVE integration paper.
}

\section{zingeR}{

We first perform a differential expression analysis using zingeR posterior
probabilities and focussing on the count component of the ZINB model. The
weights are estimated with the core function of zingeR, \code{zeroWeightsLS()}. It
is important to be consistent with the normalization procedure, i.e. if you
use TMM normalization for the analysis, you should also use it for estimating
the zingeR posterior probabilities.
}

\section{edgeR}{

After estimation of the dispersions and posterior probabilities, the
\code{glmWeightedF()} function is used for statistical inference. This is an
adapted function from the \code{glmLRT()} function of edgeR. It uses an F-test for
which the denominator degrees of freedom are by default adjusted according to
the downweighting of excess zeros (\code{ZI = TRUE}). Also, independent filtering
can be performed on the obtained p-values (\code{independentFiltering = TRUE}). We
use the independent filtering strategy that was originally implemented in
DESeq2. By default, the average fitted values are used as a filter criterion.
}

\section{DESeq2}{

We're providing preliminary support for DESeq2 as the differential expression
caller. It is currently considerably slower for large datasets than edgeR.
}

\examples{
# SingleCellExperiment ====
object <- cellranger_small
# Use S4 methods to get cells (colnames) of object
numerator <- colnames(object)[which(object[["sampleName"]] == "proximal")]
glimpse(numerator)
denominator <- colnames(object)[which(object[["sampleName"]] == "distal")]
glimpse(denominator)

# zinbwave-edgeR (fast, recommended)
x <- diffExp(
    object = object,
    numerator = numerator,
    denominator = denominator,
    zeroWeights = "zinbwave",
    caller = "edgeR"
)
class(x)
head(x[["table"]])

# zinbwave-DESeq2 (slower)
\dontrun{
x <- diffExp(
    object = object,
    numerator = numerator,
    denominator = denominator,
    zeroWeights = "zinbwave",
    caller = "DESeq2"
)
class(x)
head(x)
}

# seurat ====
# Expression in cluster 3 relative to cluster 2
\dontrun{
object <- Seurat::pbmc_small
numerator <- Seurat::WhichCells(object, ident = 3L)
denominator <- Seurat::WhichCells(object, ident = 2L)
x <- diffExp(
    object = object,
    numerator = numerator,
    denominator = denominator,
    caller = "edgeR",
    minCellsPerGene = 5L,
    minCountsPerCell = 5L
)
class(x)
head(x[["table"]])
}
}
\seealso{
Consult the following sources for more information:
\itemize{
\item \href{https://doi.org/10.1186/s13059-018-1406-4}{zinbwave paper}.
\item \href{https://bit.ly/2wtDdpS}{zinbwave vignette}.
\item \href{https://github.com/mikelove/zinbwave-deseq2}{zinbwave-DESeq2 workflow}.
\item \href{https://goo.gl/4rTK1w}{zingeR vignette}.
}

\itemize{
\item DESeq2: We're trying to follow the conventions used in DESeq2 for
contrasts, defining the name of the factor in the design formula,
numerator, and denominator level for the fold change calculations. See
\code{\link[DESeq2:results]{DESeq2::results()}} for more information.
\item Seurat: Note that Seurat currently uses the convention \code{cells.1} for the
numerator and \code{cells.2} for the denominator. See \code{\link[Seurat:DiffExpTest]{Seurat::DiffExpTest()}}
for additional information.
}
}
\author{
Michael Steinbaugh
}
