% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_annotations.R
\name{ensembl_annotations}
\alias{ensembl_annotations}
\title{Get transcript annotations from Ensembl}
\usage{
ensembl_annotations(organism, gene_name = "external_gene_name")
}
\arguments{
\item{organism}{Organism identifier}

\item{gene_name}{Ensembl gene name identifier}
}
\value{
Data frame
}
\description{
This function also defines broad classes, which are used in quality control
analysis
}
\examples{
\dontrun{
ensembl_annotations("mmusculus")
}
}
\author{
Rory Kirchner

Michael Steinbaugh
}
\keyword{internal}
