% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_annotations.R
\name{ensembl_annotations}
\alias{ensembl_annotations}
\title{Get transcript annotations from Ensembl}
\usage{
ensembl_annotations(run)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run}
}
\value{
Data frame
}
\description{
This function also defines broad classes, which are used in quality control
analysis
}
\author{
Rory Kirchner

Michael Steinbaugh
}
\keyword{internal}
