% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-fetchData.R
\docType{methods}
\name{fetchGeneData}
\alias{fetchGeneData}
\alias{fetchPCAData}
\alias{fetchTSNEData}
\alias{fetchTSNEExpressionData}
\alias{fetchUMAPData}
\alias{fetchUMAPExpressionData}
\alias{fetchData}
\alias{fetchGeneData,SingleCellExperiment-method}
\alias{fetchGeneData,seurat-method}
\alias{fetchPCAData,seurat-method}
\alias{fetchTSNEData,seurat-method}
\alias{fetchTSNEExpressionData,seurat-method}
\alias{fetchUMAPData,seurat-method}
\alias{fetchUMAPExpressionData,seurat-method}
\title{Fetch Data Functions}
\usage{
fetchGeneData(object, ...)

fetchPCAData(object, ...)

fetchTSNEData(object, ...)

fetchTSNEExpressionData(object, ...)

fetchUMAPData(object, ...)

fetchUMAPExpressionData(object, ...)

\S4method{fetchGeneData}{SingleCellExperiment}(object, genes)

\S4method{fetchGeneData}{seurat}(object, genes)

\S4method{fetchPCAData}{seurat}(object)

\S4method{fetchTSNEData}{seurat}(object)

\S4method{fetchTSNEExpressionData}{seurat}(object, genes)

\S4method{fetchUMAPData}{seurat}(object)

\S4method{fetchUMAPExpressionData}{seurat}(object, genes)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{Gene identifiers. Must match the rownames of the object.}
}
\value{
\itemize{
\item \code{fetchGeneData()}: \code{matrix}.
\item \code{fetchTSNEExpressionData()}: \code{data.frame} containing t-SNE coordinates,
sample metadata, and aggregate marker expression values (\code{mean}, \code{median},
and \code{sum}).
\item Other functions: \code{data.frame}, containing metrics.
}
}
\description{
Fetch Data Functions
}
\examples{
# SingleCellExperiment ====
object <- indrops_small
genes <- head(rownames(object))

# fetchGeneData
x <- fetchGeneData(object, genes = genes)
glimpse(x)

# seurat ====
object <- seurat_small
genes <- head(rownames(object))

# t-SNE
x <- fetchTSNEData(object)
glimpse(x)

# PCA
x <- fetchPCAData(object)
glimpse(x)

# UMAP
x <- fetchUMAPData(object)
glimpse(x)

# t-SNE gene expression
x <- fetchTSNEExpressionData(seurat_small, genes = genes)
glimpse(x)

# UMAP gene expession
genes <- head(rownames(seurat_small))
x <- fetchUMAPExpressionData(seurat_small, genes = genes)
glimpse(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
