% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-fetchGeneData.R
\docType{methods}
\name{fetchGeneData}
\alias{fetchGeneData}
\alias{fetchGeneData}
\alias{fetchGeneData,seurat-method}
\title{Fetch Gene Expression Data}
\usage{
fetchGeneData(object, ...)

\S4method{fetchGeneData}{seurat}(object, genes)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{Gene identifiers. Must match the rownames of the object.}
}
\value{
\code{matrix}.
}
\description{
Fetch Gene Expression Data
}
\examples{
load(system.file("extdata/seurat.rda", package = "bcbioSingleCell"))

# seurat
genes <- slot(seurat, "data") \%>\%
    rownames() \%>\%
    head()
fetchGeneData(seurat, genes = genes) \%>\%
    head()
}
\author{
Michael Steinbaugh
}
