% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-fetchGeneData.R
\docType{methods}
\name{fetchGeneData}
\alias{fetchGeneData}
\alias{fetchGeneData}
\alias{fetchGeneData,seurat-method}
\title{Fetch Gene Expression Data}
\usage{
fetchGeneData(object, ...)

\S4method{fetchGeneData}{seurat}(object, genes)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{Gene identifiers. Must match the rownames of the object.}
}
\value{
\code{matrix}.
}
\description{
Fetch Gene Expression Data
}
\examples{
# seurat ====
genes <- head(rownames(pbmc_small))
fetchGeneData(pbmc_small, genes = genes) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{calculateMetrics}},
  \code{\link{cell2sample}}, \code{\link{colData}},
  \code{\link{counts}}, \code{\link{fetchPCAData}},
  \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}}, \code{\link{gene2symbol}},
  \code{\link{interestingGroups}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{rowData}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
