% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-fetchGeneData.R,
%   R/methods-seurat-SingleCellExperiment.R
\docType{methods}
\name{fetchGeneData}
\alias{fetchGeneData}
\alias{fetchGeneData}
\alias{fetchGeneData,SingleCellExperiment-method}
\alias{fetchGeneData,seurat-method}
\title{Fetch Gene Expression Data}
\usage{
fetchGeneData(object, ...)

\S4method{fetchGeneData}{SingleCellExperiment}(object, genes)

\S4method{fetchGeneData}{seurat}(object, genes)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{Gene identifiers. Must match the rownames of the object.}
}
\value{
\code{matrix}.
}
\description{
Fetch Gene Expression Data
}
\examples{
# SingleCellExperiment ====
genes <- rownames(bcb_small) \%>\% head(2L)
fetchGeneData(bcb_small, genes = genes) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{fetchPCAData}}, \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
