% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-fetchPCAData.R
\docType{methods}
\name{fetchPCAData}
\alias{fetchPCAData}
\alias{fetchPCAData}
\alias{fetchPCAData,seurat-method}
\title{Fetch PCA Dimensions and Cellular Metadata}
\usage{
fetchPCAData(object, ...)

\S4method{fetchPCAData}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{data.frame} of PCA points and metadata for each cell.
}
\description{
Fetch PCA Dimensions and Cellular Metadata
}
\examples{
\dontrun{
data(seurat)
pca <- fetchPCAData(seurat)
}
}
\seealso{
\code{\link[Seurat:PCAPlot]{Seurat::PCAPlot()}}.

Other PCA Utilities: \code{\link{plotPCA}}
}
\author{
Michael Steinbaugh
}
