% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-fetchPCAData.R
\docType{methods}
\name{fetchPCAData}
\alias{fetchPCAData}
\alias{fetchPCAData}
\alias{fetchPCAData,seurat-method}
\title{Fetch PCA Dimensions and Cellular Metadata}
\usage{
fetchPCAData(object, ...)

\S4method{fetchPCAData}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{data.frame} of PCA points and metadata for each cell.
}
\description{
Fetch PCA Dimensions and Cellular Metadata
}
\examples{
# seurat ====
fetchPCAData(pbmc_small) \%>\% glimpse()
}
\seealso{
\code{\link[Seurat:PCAPlot]{Seurat::PCAPlot()}}.

Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{fetchGeneData}}, \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
