% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-fetchTSNEData.R
\docType{methods}
\name{fetchTSNEData}
\alias{fetchTSNEData}
\alias{fetchTSNEData}
\alias{fetchTSNEData,seurat-method}
\title{Fetch t-SNE Locations and Cellular Metadata}
\usage{
fetchTSNEData(object, ...)

\S4method{fetchTSNEData}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{data.frame} of t-SNE points and metadata for each cell.
}
\description{
Fetch t-SNE Locations and Cellular Metadata
}
\examples{
# seurat ====
fetchTSNEData(pbmc_small) \%>\% glimpse()
}
\seealso{
\code{\link[Seurat:TSNEPlot]{Seurat::TSNEPlot()}}.

Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{calculateMetrics}},
  \code{\link{cell2sample}}, \code{\link{colData}},
  \code{\link{counts}}, \code{\link{fetchGeneData}},
  \code{\link{fetchPCAData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}},
  \code{\link{interestingGroups}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{rowData}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
