% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-fetchTSNEData.R
\docType{methods}
\name{fetchTSNEData}
\alias{fetchTSNEData}
\alias{fetchTSNEData,seurat-method}
\title{Fetch t-SNE Locations and Cellular Metadata}
\usage{
fetchTSNEData(object, ...)

\S4method{fetchTSNEData}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{data.frame} of t-SNE points and metadata for each cell.
}
\description{
Fetch t-SNE Locations and Cellular Metadata
}
\examples{
\dontrun{
data(seurat)
tsne <- fetchTSNEData(seurat)
}
}
\seealso{
Other t-SNE Utilities: \code{\link{fetchTSNEExpressionData}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotTSNE}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{t-SNE Utilities}
