% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-fetchTSNEExpressionData.R
\docType{methods}
\name{fetchTSNEExpressionData}
\alias{fetchTSNEExpressionData}
\title{Fetch t-SNE locations, cellular metadata and expression of genes}
\usage{
fetchTSNEExpressionData(object, ...)

\S4method{fetchTSNEExpressionData}{seurat}(object, genes)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{genes}{Genes (by symbol name) of which to get expression data.}
}
\value{
tidy \link{data.frame} of t-SNE points, cellular metadata, and gene
expression.
}
\description{
This gets t-SNE locations, cellular metadata, expression of genes and
the geometric mean of the gene expression from a seurat object.
}
\author{
Rory Kirchner, Michael Steinbaugh
}
