% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat-tsne.R
\name{FetchTsneExpressionData}
\alias{FetchTsneExpressionData}
\title{Fetch t-SNE locations, cellular metadata and expression of genes}
\usage{
FetchTsneExpressionData(seurat, genes)
}
\arguments{
\item{seurat}{a seurat object}

\item{genes}{genes of which to get expression data}
}
\value{
tidy dataframe of t-SNE points, cellular metadata and gene expression
}
\description{
This gets t-SNE locations, cellular metadata, expression of genes and
the geometric mean of the gene expression from a seurat object.
}
\author{
Rory Kirchner
}
