% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-fetchTSNEExpressionData.R
\docType{methods}
\name{fetchTSNEExpressionData}
\alias{fetchTSNEExpressionData}
\alias{fetchTSNEExpressionData}
\alias{fetchTSNEExpressionData,seurat-method}
\title{Fetch t-SNE Locations, Cellular Metadata and Expression of Genes}
\usage{
fetchTSNEExpressionData(object, ...)

\S4method{fetchTSNEExpressionData}{seurat}(object, genes, format = "symbol")
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{genes}{Genes (by symbol name) of which to get expression data.}
}
\value{
\link{tibble} grouped by \code{gene}, containing t-SNE points, cellular
metadata, and gene expression.
}
\description{
This gets t-SNE locations, cellular metadata, expression of genes and
the geometric mean of the gene expression from a \link{seurat} object.
}
\examples{
load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

symbol <- counts(seurat) \%>\% rownames() \%>\% head()
print(symbol)

ensgene <- bcbio(seurat, "gene2symbol") \%>\%
    .[which(.[["symbol"]] \%in\% symbol), "ensgene", drop = TRUE]

fetchTSNEExpressionData(
    seurat,
    genes = symbol,
    format = "symbol") \%>\%
    glimpse()
fetchTSNEExpressionData(
    seurat,
    genes = ensgene,
    format = "ensgene") \%>\%
    glimpse()
}
\seealso{
Other t-SNE Utilities: \code{\link{fetchTSNEData}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotTSNE}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
