% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-filter.R
\docType{methods}
\name{filter,bcbioSCDataSet-method}
\alias{filter,bcbioSCDataSet-method}
\title{Filter cellular barcodes}
\usage{
\S4method{filter}{bcbioSCDataSet}(object, min_umis = 1000L,
  min_genes = 500L, max_genes = NULL, max_mito_ratio = 0.1,
  min_novelty = 0.8, show_report = TRUE)
}
\arguments{
\item{object}{Primary object.}

\item{min_umis}{Minimum number of UMI disambiguated counts per cell.}

\item{min_genes}{Minimum number of genes detected.}

\item{max_genes}{Maximum number of genes detected.}

\item{max_mito_ratio}{Maximum relative mitochondrial abundance (\code{0-1} scale).}

\item{min_novelty}{Minimum novelty score.}

\item{show_report}{Show summary statistics report and plots.}
}
\value{
\link{SCSubset}.
}
\description{
Apply gene detection, mitochondrial abundance, and novelty score cutoffs to
cellular barcodes.
}
\author{
Michael Steinbaugh
}
