% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-filter_barcodes.R
\docType{methods}
\name{filter_barcodes}
\alias{filter_barcodes}
\alias{filter_barcodes,bcbioSCDataSet-method}
\title{Filter cellular barcodes}
\usage{
\S4method{filter_barcodes}{bcbioSCDataSet}(object, umis = 1000L,
  genes = 500L, mito_ratio = 0.1, novelty = 0.8, show_report = TRUE)
}
\arguments{
\item{object}{Primary object.}

\item{umis}{Minimum number of UMI disambiguated counts per cell.}

\item{genes}{Minimum number of genes detected.}

\item{mito_ratio}{Maximum relative mitochondrial abundance (\code{0-1} scale).}

\item{novelty}{Minimum novelty score.}

\item{show_report}{Show summary statistics report and plots.}
}
\value{
\link{bcbioSCDataSet}, with low quality cellular barcodes that
don't pass quality control cutoffs removed.
}
\description{
Apply gene detection, mitochondrial abundance, and novelty score cutoffs to
cellular barcodes.
}
\author{
Michael Steinbaugh
}
