% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-filter_barcodes.R
\docType{methods}
\name{filter_barcodes}
\alias{filter_barcodes}
\alias{filter_barcodes,bcbioSCDataSet-method}
\title{Filter cellular barcodes}
\usage{
filter_barcodes(object, ...)

\S4method{filter_barcodes}{bcbioSCDataSet}(object, reads = 1000L,
  min_genes = 500L, max_genes = 10000L, mito_ratio = 0.1, novelty = 0.8,
  show = TRUE)
}
\arguments{
\item{object}{\link{bcbioSCDataSet}.}

\item{...}{Additional parameters.}

\item{reads}{Minimum number of total read counts per cell.}

\item{min_genes}{Minimum number of genes detected.}

\item{max_genes}{Maximum number of genes detected.}

\item{mito_ratio}{Maximum relative mitochondrial abundance (\code{0-1} scale).}

\item{novelty}{Minimum novelty score.}

\item{show}{Show summary statistics and plots.}
}
\value{
Filtered \link{bcbioSCDataSet}, with low quality cellular barcodes that
don't pass quality control cutoffs removed.
}
\description{
Apply gene detection, mitochondrial abundance, and novelty score cutoffs to
cellular barcodes.
}
\author{
Michael Steinbaugh
}
