% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cellular_barcodes.R
\docType{data}
\name{filter_cellular_barcodes}
\alias{filter_cellular_barcodes}
\alias{min_genes}
\alias{max_genes}
\alias{novelty}
\alias{percent_mito}
\title{Filter cellular barcodes}
\format{An object of class \code{numeric} of length 1.}
\usage{
filter_cellular_barcodes(metrics, min_genes = 500, max_genes = 5000,
  percent_mito = 25, novelty = 0.75, plot = TRUE)

min_genes

max_genes

novelty

percent_mito
}
\arguments{
\item{metrics}{Barcode metrics data frame}

\item{min_genes}{Minimum number of genes detected}

\item{max_genes}{Maximum number of genes detected}

\item{percent_mito}{Maximum mitochondrial pct abundance (\code{0-1} scale)}

\item{novelty}{Minimum novelty score}

\item{plot}{Print relevant plots}
}
\value{
Filtered metrics data frame
}
\description{
Apply gene detection, mitochondrial abundance, and novelty score cutoffs to
cellular barcodes.

Default minimum gene count cutoff (500)

Default maximum gene count cutoff (5000)

Default minimum novelty score (0.75)

Default maximum mitochondrial percent abundance (25)
}
\examples{
min_genes
max_genes
novelty
percent_mito
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
