% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cellular_barcodes.R
\docType{data}
\name{filter_cellular_barcodes}
\alias{filter_cellular_barcodes}
\alias{min_genes}
\alias{max_genes}
\alias{mito_ratio}
\alias{novelty}
\title{Filter cellular barcodes.}
\format{An object of class \code{numeric} of length 1.}
\usage{
filter_cellular_barcodes(run, min_genes = get("min_genes", envir =
  parent.frame()), max_genes = get("max_genes", envir = parent.frame()),
  mito_ratio = get("mito_ratio", envir = parent.frame()),
  novelty = get("novelty", envir = parent.frame()), plot = TRUE)

min_genes

max_genes

mito_ratio

novelty
}
\arguments{
\item{run}{bcbio-nextgen run.}

\item{min_genes}{Minimum number of genes detected.}

\item{max_genes}{Maximum number of genes detected.}

\item{mito_ratio}{Maximum relative mitochondrial abundance (\code{0-1}
scale).}

\item{novelty}{Minimum novelty score.}

\item{plot}{Print relevant plots.}
}
\value{
Filtered bcbio run object.
}
\description{
Apply gene detection, mitochondrial abundance, and novelty score cutoffs to
cellular barcodes.

Default minimum gene count cutoff.

Default maximum gene count cutoff.

Default maximum relative mitochondrial abundance.

Default minimum novelty score.
}
\examples{
min_genes
max_genes
mito_ratio
novelty
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
