% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-filterCells.R
\docType{methods}
\name{filterCells}
\alias{filterCells}
\alias{filterCells}
\alias{filterCells,bcbioSingleCell-method}
\alias{filterCells,bcbioSCDataSet-method}
\title{Filter Cells}
\usage{
filterCells(object, ...)

\S4method{filterCells}{bcbioSingleCell}(object, minUMIs = 1000,
  minGenes = 500, maxGenes = Inf, maxMitoRatio = 0.1, minNovelty = 0.8,
  minCellsPerGene = 3, quiet = FALSE)

\S4method{filterCells}{bcbioSCDataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{minUMIs}{Minimum number of UMI disambiguated counts per cell.}

\item{minGenes}{Minimum number of genes detected.}

\item{maxGenes}{Maximum number of genes detected.}

\item{maxMitoRatio}{Maximum relative mitochondrial abundance (\code{0-1} scale).}

\item{minNovelty}{Minimum novelty score.}

\item{minCellsPerGene}{Include genes with non-zero expression in at least
this many cells.}

\item{quiet}{If \code{TRUE}, don't show the filtering parameter summary.}
}
\value{
\link{bcbioSingleCell} object, with filtering information slotted into
\code{\link[=metadata]{metadata()}} as \code{filterCells} and \code{filterParams}.
}
\description{
Apply gene detection, mitochondrial abundance, and novelty score cutoffs to
cellular barcodes.
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
print(bcb)
filterCells(bcb)
}
\seealso{
\code{\link[Seurat:CreateSeuratObject]{Seurat::CreateSeuratObject()}}.

Other Data Management Utilities: \code{\link{aggregateReplicates}},
  \code{\link{readCellTypeMarkersFile}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
