% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-filterCells.R
\docType{methods}
\name{filterCells}
\alias{filterCells}
\alias{filterCells,bcbioSingleCell-method}
\alias{filterCells,bcbioSCDataSet-method}
\title{Filter Cells}
\usage{
filterCells(object, ...)

\S4method{filterCells}{bcbioSingleCell}(object, aggregateReplicates = TRUE,
  minUMIs = 1000, minGenes = 500, maxGenes = Inf, maxMitoRatio = 0.1,
  minNovelty = 0.8, showReport = TRUE, headerLevel = 2)

\S4method{filterCells}{bcbioSCDataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{aggregateReplicates}{Aggregate technical replicates, if specified. This
function uses values assigned in the \code{sampleNameAggregate} column of the
internal sample metadata \link{data.frame}.}

\item{minUMIs}{Minimum number of UMI disambiguated counts per cell.}

\item{minGenes}{Minimum number of genes detected.}

\item{maxGenes}{Maximum number of genes detected.}

\item{maxMitoRatio}{Maximum relative mitochondrial abundance (\code{0-1} scale).}

\item{minNovelty}{Minimum novelty score.}

\item{showReport}{Show summary statistics report and plots.}

\item{headerLevel}{RMarkdown header level, if \code{showReport = TRUE}.}
}
\value{
\link{bcbioSingleCell} object, with filtering information slotted into
\code{\link[=metadata]{metadata()}} as \code{filterCells} and \code{filterParams}.
}
\description{
Apply gene detection, mitochondrial abundance, and novelty score cutoffs to
cellular barcodes.
}
\note{
This operation can be re-run on a \link{bcbioSingleCell} object that has
been previously filtered. This function simply saves \code{filterCells} and
\code{filterParams} vectors into the \code{\link[=metadata]{metadata()}} slot.
}
\examples{
\dontrun{
data(bcb)
bcb <- filterCells(bcb)
}
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}},
  \code{\link{readMarkers}}, \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
\concept{Data Management Utilities}
