% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-filterCells.R
\docType{methods}
\name{filterCells}
\alias{filterCells}
\alias{filterCells}
\alias{filterCells,SingleCellExperiment-method}
\title{Filter Cells}
\usage{
filterCells(object, ...)

\S4method{filterCells}{SingleCellExperiment}(object, minUMIs = 0L,
  maxUMIs = Inf, minGenes = 0L, maxGenes = Inf, minNovelty = 0L,
  maxMitoRatio = 1L, minCellsPerGene = 10L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{minUMIs}{Minimum number of UMI disambiguated counts per cell.}

\item{maxUMIs}{Maximum number of UMI disambiguated counts per cell.}

\item{minGenes}{Minimum number of genes detected.}

\item{maxGenes}{Maximum number of genes detected.}

\item{minNovelty}{Minimum novelty score (log10 genes per UMI).}

\item{maxMitoRatio}{Maximum relative mitochondrial abundance (\code{0-1} scale).}

\item{minCellsPerGene}{Include genes with non-zero expression in at least
this many cells.}
}
\value{
\code{bcbioSingleCell}, with filtering information slotted into
\code{\link[=metadata]{metadata()}} as \code{filterCells} and \code{filterParams}.
}
\description{
Apply gene detection, novelty score, and mitochondrial abundance cutoffs to
cellular barcodes. By default we recommend applying the same filtering cutoff
to all samples. The filtering parameters now support per-sample cutoffs,
defined using a named \code{numeric} vector. When matching per sample, be sure to
use the \code{\link[=sampleNames]{sampleNames()}} return values (i.e. the \code{sampleName} column in
\code{\link[=sampleData]{sampleData()}}).
}
\examples{
# bcbioSingleCell ====
show(indrops_small)
x <- filterCells(indrops_small, minGenes = 0L)
show(x)

# SingleCellExperiment ====
show(cellranger_small)
x <- filterCells(cellranger_small, minNovelty = 0L)
show(x)

# Per sample cutoffs
sampleNames(cellranger_small)
x <- filterCells(
    object = cellranger_small,
    minUMIs = c(
        distal = 100,
        proximal = 200
    )
)
x
}
\seealso{
\code{\link[Seurat:CreateSeuratObject]{Seurat::CreateSeuratObject()}}.

Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{fetchGeneData}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
