% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/filterCells-methods.R
\docType{methods}
\name{filterCells}
\alias{filterCells}
\alias{filterCells}
\alias{filterCells,bcbioSCDataSet-method}
\title{Filter Cells}
\usage{
filterCells(object, ...)

\S4method{filterCells}{bcbioSCDataSet}(object, minUMIs = 1000L,
  minGenes = 500L, maxGenes = NULL, maxMitoRatio = 0.1,
  minNovelty = 0.8, showReport = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{minUMIs}{Minimum number of UMI disambiguated counts per cell.}

\item{minGenes}{Minimum number of genes detected.}

\item{maxGenes}{Maximum number of genes detected.}

\item{maxMitoRatio}{Maximum relative mitochondrial abundance (\code{0-1} scale).}

\item{minNovelty}{Minimum novelty score.}

\item{showReport}{Show summary statistics report and plots.}
}
\value{
\link{bcbioSCSubset}.
}
\description{
Apply gene detection, mitochondrial abundance, and novelty score cutoffs to
cellular barcodes.
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}},
  \code{\link{interestingGroups}},
  \code{\link{readKnownMarkers}},
  \code{\link{sampleMetadata}}, \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh
}
