% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/filterCells-methods.R
\docType{methods}
\name{filterCells}
\alias{filterCells}
\alias{filterCells}
\alias{filterCells,SingleCellExperiment-method}
\title{Filter Cells}
\usage{
filterCells(object, ...)

\S4method{filterCells}{SingleCellExperiment}(object, nCells = Inf,
  minUMIs = 0L, maxUMIs = Inf, minGenes = 0L, maxGenes = Inf,
  minNovelty = 0L, maxMitoRatio = 1L, minCellsPerGene = 10L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{nCells}{Expected number of cells per sample.}

\item{minUMIs}{Minimum number of UMI disambiguated counts per cell.}

\item{maxUMIs}{Maximum number of UMI disambiguated counts per cell.}

\item{minGenes}{Minimum number of genes detected.}

\item{maxGenes}{Maximum number of genes detected.}

\item{minNovelty}{Minimum novelty score (log10 genes per UMI).}

\item{maxMitoRatio}{Maximum relative mitochondrial abundance (\code{0-1} scale).}

\item{minCellsPerGene}{Include genes with non-zero expression in at least
this many cells.}
}
\value{
\code{bcbioSingleCell}, with filtering information slotted into
\code{\link[=metadata]{metadata()}} as \code{filterCells} and \code{filterParams}.
}
\description{
Apply gene detection, novelty score, and mitochondrial abundance cutoffs to
cellular barcodes. By default we recommend applying the same filtering cutoff
to all samples. The filtering parameters now support per-sample cutoffs,
defined using a named \code{numeric} vector. When matching per sample, be sure to
use the \code{\link[=sampleNames]{sampleNames()}} return values (i.e. the \code{sampleName} column in
\code{\link[=sampleData]{sampleData()}}).
}
\examples{
# SingleCellExperiment ====
object <- cellranger_small
show(object)

x <- filterCells(object, minNovelty = 0L)
show(x)
metadata(x)$filterParams

# Per sample cutoffs
sampleNames(object)
x <- filterCells(
    object = object,
    minUMIs = c(pbmc = 100)
)
show(x)
metadata(x)$filterParams
}
\seealso{
\code{\link[Seurat:CreateSeuratObject]{Seurat::CreateSeuratObject()}}.

Other Quality Control Functions: \code{\link{barcodeRanksPerSample}},
  \code{\link{metrics}}, \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
