% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-filterCells.R
\docType{methods}
\name{filterCells}
\alias{filterCells}
\alias{filterCells}
\alias{filterCells,bcbioSingleCell-method}
\title{Filter Cells}
\usage{
filterCells(object, ...)

\S4method{filterCells}{bcbioSingleCell}(object, minUMIs = 1000L,
  maxUMIs = Inf, minGenes = 500L, maxGenes = Inf, maxMitoRatio = 0.1,
  minNovelty = 0.75, minCellsPerGene = 3L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{minUMIs}{Minimum number of UMI disambiguated counts per cell.}

\item{maxUMIs}{Maximum number of UMI disambiguated counts per cell.}

\item{minGenes}{Minimum number of genes detected.}

\item{maxGenes}{Maximum number of genes detected.}

\item{maxMitoRatio}{Maximum relative mitochondrial abundance (\code{0-1} scale).}

\item{minNovelty}{Minimum novelty score.}

\item{minCellsPerGene}{Include genes with non-zero expression in at least
this many cells.}
}
\value{
\code{bcbioSingleCell}, with filtering information slotted into
\code{\link[=metadata]{metadata()}} as \code{filterCells} and \code{filterParams}.
}
\description{
Apply gene detection, mitochondrial abundance, and novelty score cutoffs to
cellular barcodes.
}
\details{
The filtering cutoff values now support a named numeric vector. By
default we recommend applying the same filtering cutoff to all samples.
When matching the samples, be sure to use \code{sampleID} column (i.e. the
rownames of \code{\link[=sampleData]{sampleData()}}).
}
\examples{
# bcbioSingleCell ====
show(bcb_small)
x <- filterCells(bcb_small, minGenes = 0L)
show(x)
}
\seealso{
\code{\link[Seurat:CreateSeuratObject]{Seurat::CreateSeuratObject()}}.

Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{calculateMetrics}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{fetchGeneData}}, \code{\link{fetchPCAData}},
  \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{sampleData}}, \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
