% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-flatFiles.R
\docType{methods}
\name{flatFiles}
\alias{flatFiles}
\alias{flatFiles,bcbioSingleCell-method}
\title{Get Flat Files from S4 Object}
\usage{
\S4method{flatFiles}{bcbioSingleCell}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{list}.
}
\description{
Prepare an unstructured list of information saved in the S4 object,
for improved archival data storage.
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
flatFiles(bcb) \%>\% names()
}
\author{
Michael Steinbaugh, Rory Kirchner
}
