% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,seurat-method}
\title{Gene to Symbol Mappings}
\usage{
\S4method{gene2symbol}{seurat}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Gene to Symbol Mappings
}
\examples{
# bcbioRNASeq ====
gene2symbol(bcb_small) \%>\% glimpse()

# seurat ====
gene2symbol(seurat_small) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
