% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,bcbioSingleCell-method}
\alias{gene2symbol,seurat-method}
\title{Gene to Symbol Mappings}
\usage{
\S4method{gene2symbol}{bcbioSingleCell}(object)

\S4method{gene2symbol}{seurat}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Gene to Symbol Mappings
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat.rda", package = "bcbioSingleCell"))

# bcbioRNASeq
gene2symbol(bcb) \%>\% glimpse()

# seurat
gene2symbol(seurat) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
