% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,seurat-method}
\title{Gene to Symbol Mappings}
\usage{
\S4method{gene2symbol}{seurat}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Gene to Symbol Mappings
}
\examples{
# bcbioRNASeq ====
gene2symbol(bcb_small) \%>\% glimpse()

# seurat ====
gene2symbol(seurat_small) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{colData}},
  \code{\link{counts}}, \code{\link{fetchGeneData}},
  \code{\link{fetchPCAData}}, \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}},
  \code{\link{interestingGroups}},
  \code{\link{readCellTypeMarkersFile}},
  \code{\link{rowData}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
