% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{general}
\alias{general}
\title{General Arguments}
\arguments{
\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.
To set the discrete fill palette globally, use
\code{options(bcbio.discrete.fill = scale_fill_viridis_d())}.}

\item{geom}{\code{string}. Plot type. Uses \code{\link[=match.arg]{match.arg()}} and defaults to the first
argument in the \code{character} vector.}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}

\item{min}{\code{scalar numeric}. Recommended minimum value cutoff.}

\item{max}{\code{scalar numeric}. Recommended maximum value cutoff.}

\item{object}{Object.}

\item{prefilter}{\code{boolean}. Apply prefiltering to remove zero count genes.}

\item{return}{\code{string}. Return type. Uses \code{\link[base:match.arg]{base::match.arg()}} internally and
defaults to the first argument in the \code{character} vector.}

\item{title}{\code{string} or \code{NULL}. Plot title.}

\item{trans}{\code{string}. Name of the axis scale transformation to apply. See
\code{help("scale_x_continuous", "ggplot2")} for more information.}

\item{trendline}{\code{boolean}. Include trendline on plot.}

\item{value}{Object to assign.}

\item{x}{Primary object.}

\item{y}{Secondary object.}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
General Arguments
}
\keyword{internal}
