% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sparsecounts.R
\name{import_sparsecounts}
\alias{import_sparsecounts}
\title{Import scRNA-Seq count data}
\usage{
import_sparsecounts(bcbio)
}
\arguments{
\item{bcbio}{bcbio run object}
}
\value{
Sparse counts matrix
}
\description{
Import transcript-level count data from a bcbio run into a sparse matrix
}
\examples{
\dontrun{
import_sparsecounts(bcbio)
}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
