% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-metadata.R
\name{.indrop_metadata}
\alias{.indrop_metadata}
\title{Prepare inDrop metadata}
\usage{
.indrop_metadata(file, sample_dirs)
}
\arguments{
\item{sample_dirs}{Named character vector of sample directory paths.}

\item{metadata}{Metadata data frame.}
}
\value{
\link{DataFrame}.
}
\description{
Appends reverse complement sequences by matching against the \code{sequence}
column.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
