% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indrops}
\alias{indrops}
\title{inDrops example data set}
\format{An object of class \code{bcbioSingleCell} with 50 rows and 100 columns.}
\usage{
indrops
}
\description{
inDrops example data set
}
\note{
Updated 2019-07-24.
}
\examples{
data(indrops)
indrops
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
