% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{indrops_small}
\alias{indrops_small}
\title{inDrops Example}
\format{An object of class \code{bcbioSingleCell} with 500 rows and 500 columns.}
\usage{
indrops_small
}
\description{
inDrops Example
}
\examples{
show(indrops_small)
}
\seealso{
Other Minimal Example Data: \code{\link{all_markers_small}},
  \code{\link{cellranger_small}},
  \code{\link{known_markers_small}},
  \code{\link{seurat_small}}
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
