% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-inflectionPoint.R
\docType{methods}
\name{inflectionPoint}
\alias{inflectionPoint}
\alias{inflectionPoint}
\alias{inflectionPoint,bcbioSingleCell-method}
\alias{inflectionPoint,seurat-method}
\title{Calculate the Cellular Barcode Read Depth Inflection Point}
\usage{
inflectionPoint(object, ...)

\S4method{inflectionPoint}{bcbioSingleCell}(object, maxCells = 100000L)

\S4method{inflectionPoint}{seurat}(object, maxCells = 100000L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{maxCells}{The number of cells to consider as an upper bound and to
compute the inflection point.}
}
\description{
Calculate the Cellular Barcode Read Depth Inflection Point
}
\examples{
# bcbioSingleCell ====
inflectionPoint(bcb_small)
inflectionPoint(cellranger_small)

# seurat ====
inflectionPoint(seurat_small)
inflectionPoint(pbmc_small)
}
\seealso{
Modified version of \code{dropbead::estimateCellNumber()}.

Other Quality Control Functions: \code{\link{plotCellCounts}},
  \code{\link{plotCumulativeReadsPerCell}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotQuantileHeatmap}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
