% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrity_checks.R
\name{check_run}
\alias{check_run}
\title{Perform object integrity checks}
\usage{
check_run(run)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run}
}
\description{
Check \code{bcbio-nextgen} run
}
\author{
Michael Steinbaugh
}
