% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_integrity_checks.R
\name{check_run}
\alias{check_run}
\alias{check_sparse}
\title{Perform object integrity checks}
\usage{
check_run(run)

check_sparse(sparse)
}
\arguments{
\item{run}{bcbio-nextgen run.}

\item{sparse}{Sparse counts matrix.}
}
\description{
Check bcbio-nextgen run.

Check sparse counts matrix.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
