% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-interesting_groups.R
\docType{methods}
\name{interesting_groups}
\alias{interesting_groups}
\alias{interesting_groups,bcbioSCDataSet-method}
\alias{interesting_groups,bcbioSCSubset-method}
\title{Interesting groups}
\usage{
interesting_groups(object)

\S4method{interesting_groups}{bcbioSCDataSet}(object)

\S4method{interesting_groups}{bcbioSCSubset}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Interesting groups
}
