% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-interesting_groups.R
\docType{methods}
\name{.interesting_groups}
\alias{.interesting_groups}
\alias{interesting_groups,bcbioSCDataSet-method}
\alias{interesting_groups,SummarizedExperiment-method}
\title{Interesting groups}
\usage{
\S4method{interesting_groups}{bcbioSCDataSet}(object)

\S4method{interesting_groups}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{\link{bcbioSCDataSet}.}
}
\value{
Character vector.
}
\description{
Interesting groups
}
