% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-interesting_groups.R
\docType{methods}
\name{interesting_groups}
\alias{interesting_groups}
\alias{.interesting_groups}
\alias{interesting_groups,bcbioSCDataSet-method}
\alias{interesting_groups,SCSubset-method}
\title{Interesting groups}
\usage{
\S4method{interesting_groups}{bcbioSCDataSet}(object)

\S4method{interesting_groups}{SCSubset}(object)
}
\arguments{
\item{object}{Primary object.}
}
\value{
Character vector.
}
\description{
Interesting groups
}
