% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-interestingGroups.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups,bcbioSingleCell-method}
\alias{interestingGroups,seurat-method}
\alias{interestingGroups<-,bcbioSingleCell,character-method}
\alias{interestingGroups<-,seurat,character-method}
\title{Interesting Groups}
\usage{
\S4method{interestingGroups}{bcbioSingleCell}(object)

\S4method{interestingGroups}{seurat}(object)

\S4method{interestingGroups}{bcbioSingleCell,character}(object) <- value

\S4method{interestingGroups}{seurat,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Object to assign.}
}
\value{
Character vector.
}
\description{
Interesting Groups
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat_small.rda", package = "bcbioSingleCell"))

# bcbioSingleCell ====
interestingGroups(bcb_small)

# Assignment method support
interestingGroups(bcb_small) <- "sampleID"
interestingGroups(bcb_small)

# seurat ====
interestingGroups(seurat_small)

# Assignment method support
interestingGroups(seurat_small) <- "sampleID"
interestingGroups(seurat_small)
}
\author{
Michael Steinbaugh
}
