% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-interestingGroups.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups,bcbioSingleCell-method}
\alias{interestingGroups,seurat-method}
\alias{interestingGroups<-,bcbioSingleCell,character-method}
\alias{interestingGroups<-,seurat,character-method}
\title{Interesting Groups}
\usage{
\S4method{interestingGroups}{bcbioSingleCell}(object)

\S4method{interestingGroups}{seurat}(object)

\S4method{interestingGroups}{bcbioSingleCell,character}(object) <- value

\S4method{interestingGroups}{seurat,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Object to assign.}
}
\value{
Character vector.
}
\description{
Interesting Groups
}
\author{
Michael Steinbaugh
}
