% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-interestingGroups.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups,bcbioSCDataSet-method}
\alias{interestingGroups,bcbioSCFiltered-method}
\title{Interesting Groups}
\usage{
\S4method{interestingGroups}{bcbioSCDataSet}(object)

\S4method{interestingGroups}{bcbioSCFiltered}(object)
}
\value{
Character vector.
}
\description{
Interesting Groups
}
\author{
Michael Steinbaugh
}
