% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-interestingGroups.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups,bcbioSCDataSet-method}
\alias{interestingGroups,bcbioSCSubset-method}
\title{Interesting Groups}
\usage{
interestingGroups(object)

\S4method{interestingGroups}{bcbioSCDataSet}(object)

\S4method{interestingGroups}{bcbioSCSubset}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Character vector.
}
\description{
Interesting Groups
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}},
  \code{\link{filterCells}},
  \code{\link{readKnownMarkers}},
  \code{\link{sampleMetadata}}, \code{\link{selectSamples}}
}
