% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-interestingGroups.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups<-,SingleCellExperiment,character-method}
\title{Interesting Groups}
\usage{
\S4method{interestingGroups}{SingleCellExperiment,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Object to assign.}
}
\description{
Interesting Groups
}
\examples{
# bcbioSingleCell ====
interestingGroups(bcb_small)
interestingGroups(bcb_small) <- "sampleName"
interestingGroups(bcb_small)
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{fetchGeneData}}, \code{\link{fetchPCAData}},
  \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
