% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-interestingGroups.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups,bcbioSCDataSet-method}
\alias{interestingGroups,bcbioSCFiltered-method}
\title{Interesting Groups}
\usage{
interestingGroups(object)

\S4method{interestingGroups}{bcbioSCDataSet}(object)

\S4method{interestingGroups}{bcbioSCFiltered}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Character vector.
}
\description{
Interesting Groups
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}},
  \code{\link{filterCells}}, \code{\link{readMarkers}},
  \code{\link{sampleMetadata}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
\concept{Data Management Utilities}
