% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-interestingGroups.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups,bcbioSingleCell-method}
\alias{interestingGroups,seurat-method}
\alias{interestingGroups<-,bcbioSingleCell,character-method}
\alias{interestingGroups<-,seurat,character-method}
\title{Interesting Groups}
\usage{
\S4method{interestingGroups}{bcbioSingleCell}(object)

\S4method{interestingGroups}{seurat}(object)

\S4method{interestingGroups}{bcbioSingleCell,character}(object) <- value

\S4method{interestingGroups}{seurat,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Object to assign.}
}
\value{
Character vector.
}
\description{
Interesting Groups
}
\examples{
# bcbioSingleCell ====
interestingGroups(bcb_small)

# Assignment method support
interestingGroups(bcb_small) <- "sampleID"
interestingGroups(bcb_small)

# seurat ====
interestingGroups(seurat_small)

# Assignment method support
interestingGroups(seurat_small) <- "sampleID"
interestingGroups(seurat_small)
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{colData}},
  \code{\link{counts}}, \code{\link{fetchGeneData}},
  \code{\link{fetchPCAData}}, \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}}, \code{\link{gene2symbol}},
  \code{\link{readCellTypeMarkers}}, \code{\link{rowData}},
  \code{\link{sampleData}}, \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
