% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{known_markers_small}
\alias{known_markers_small}
\title{Known Markers Example}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 28 rows and 14 columns.}
\usage{
known_markers_small
}
\description{
Known Markers Example
}
\examples{
glimpse(known_markers_small)
}
\seealso{
Other Minimal Example Data: \code{\link{all_markers_small}},
  \code{\link{cellranger_small}},
  \code{\link{indrops_small}}, \code{\link{seurat_small}}
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
