% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-knownMarkersDetected.R
\docType{methods}
\name{knownMarkersDetected}
\alias{knownMarkersDetected}
\alias{knownMarkersDetected}
\alias{knownMarkersDetected,grouped_df,tbl_df-method}
\title{Known Markers Detected}
\usage{
knownMarkersDetected(all, known, ...)

\S4method{knownMarkersDetected}{grouped_df,tbl_df}(all, known, alpha = 0.05,
  filterPromiscuous = FALSE, promiscuousCutoff = 5L)
}
\arguments{
\item{all}{All markers, generated by \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} return.}

\item{known}{Known markers \link{data.frame} imported by
\code{\link[=readCellTypeMarkersFile]{readCellTypeMarkersFile()}} or pulled from internal \link{cellCycleMarkers}
data.}

\item{...}{Additional arguments.}

\item{alpha}{Alpha cutoff (adjusted P value; false discovery rate).}

\item{filterPromiscuous}{Remove markers with poor specificity, that are
present in as many clusters as defined by \code{promiscuousCutoff}.}

\item{promiscuousCutoff}{Minimum number of clusters required to consider a
gene marker promiscuous.}
}
\value{
Grouped \link{tibble}.
}
\description{
Known Markers Detected
}
\note{
Both the \code{all} and \code{known} objects must contain Ensembl gene
identifiers in the \code{geneID} column. We must avoid any matching operations
based on the gene names, since these change often and can mismatch
easily.
}
\examples{
load(system.file("extdata/seuratAllMarkers.rda", package = "bcbioSingleCell"))

# Use internal known markers for Homo sapiens
knownMarkers <- cellTypeMarkers[["homoSapiens"]]

# grouped_df
knownMarkersDetected(
    all = seuratAllMarkers,
    known = knownMarkers) \%>\%
    glimpse()
}
\seealso{
Other Clustering Utilities: \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkers}}, \code{\link{plotPCElbow}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
