% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-knownMarkersDetected.R
\docType{methods}
\name{knownMarkersDetected}
\alias{knownMarkersDetected}
\alias{knownMarkersDetected}
\alias{knownMarkersDetected,grouped_df-method}
\title{Known Markers Detected}
\usage{
knownMarkersDetected(object, ...)

\S4method{knownMarkersDetected}{grouped_df}(object, known, alpha = 0.05,
  filterPromiscuous = FALSE, promiscuousCutoff = 5L)
}
\arguments{
\item{object}{All markers data, generated by \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}}
return and sanitized with \code{\link[=sanitizeMarkers]{sanitizeMarkers()}}.}

\item{...}{Additional arguments.}

\item{known}{Known markers \code{data.frame} imported by
\code{\link[=readCellTypeMarkers]{readCellTypeMarkers()}} or pulled from internal \link{cellCycleMarkers} data.}

\item{alpha}{Alpha cutoff (adjusted P value; false discovery rate).}

\item{filterPromiscuous}{Remove genes with poor specificity, that are present
in as many clusters as defined by \code{promiscuousCutoff}.}

\item{promiscuousCutoff}{Minimum number of clusters required to consider a
gene marker promiscuous.}
}
\value{
\code{grouped_df}, grouped by "\code{cellType}" column.
}
\description{
Known Markers Detected
}
\note{
Both the \code{all} and \code{known} objects must contain Ensembl gene
identifiers in the \code{geneID} column. We must avoid any matching operations
based on the gene names, since these change often and can mismatch
easily.
}
\examples{
# grouped_df ====
x <- knownMarkersDetected(
    object = all_markers_small,
    known = cellTypeMarkers[["homoSapiens"]]
)
glimpse(x)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeature}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotTSNE}},
  \code{\link{sanitizeMarkers}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
