% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-knownMarkersDetected.R
\docType{methods}
\name{knownMarkersDetected}
\alias{knownMarkersDetected}
\alias{knownMarkersDetected,data.frame,tbl_df-method}
\title{Known Markers Detected}
\usage{
knownMarkersDetected(object, knownMarkers, ...)

\S4method{knownMarkersDetected}{data.frame,tbl_df}(object, knownMarkers,
  show = FALSE)
}
\arguments{
\item{object}{\link{data.frame}: \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} return.}

\item{knownMarkers}{Known markers \link{data.frame} imported by \code{\link[=readMarkers]{readMarkers()}} or
pulled from internal \link{cellCycleData} object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{show}{Show \link{kable}.}
}
\value{
\link{tibble}.
}
\description{
Known Markers Detected
}
\seealso{
Other Clustering Utilities: \code{\link{pcCutoff}},
  \code{\link{plotKnownMarkers}},
  \code{\link{plotMarkers}}, \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Utilities}
