% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-knownMarkersDetected.R
\docType{methods}
\name{knownMarkersDetected}
\alias{knownMarkersDetected}
\alias{knownMarkersDetected,grouped_df,tbl_df-method}
\title{Known Markers Detected}
\usage{
knownMarkersDetected(all, known, ...)

\S4method{knownMarkersDetected}{grouped_df,tbl_df}(all, known,
  removePromiscuous = FALSE)
}
\arguments{
\item{all}{\code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} return.}

\item{known}{Known markers \link{data.frame} imported by \code{\link[=readMarkers]{readMarkers()}} or
pulled from internal \link{cellCycleMarkers} object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{removePromiscuous}{Remove markers with very poor specificity, which are
present in at least 5 clusters.}
}
\value{
\link{tibble}.
}
\description{
Known Markers Detected
}
\seealso{
Other Clustering Utilities: \code{\link{pcCutoff}},
  \code{\link{plotKnownMarkers}},
  \code{\link{plotMarkers}}, \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Utilities}
