% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-knownMarkersDetected.R
\docType{methods}
\name{knownMarkersDetected}
\alias{knownMarkersDetected}
\alias{knownMarkersDetected}
\alias{knownMarkersDetected,grouped_df,tbl_df-method}
\title{Known Markers Detected}
\usage{
knownMarkersDetected(all, known, ...)

\S4method{knownMarkersDetected}{grouped_df,tbl_df}(all, known, alpha = 0.05,
  filterPromiscuous = FALSE, promiscuousCutoff = 5)
}
\arguments{
\item{all}{All markers, generated by \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} return.}

\item{known}{Known markers \link{data.frame} imported by \code{\link[=readMarkers]{readMarkers()}} or
pulled from internal \link{cellCycleMarkers} data.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{alpha}{Alpha cutoff (adjusted P value; false discovery rate).}

\item{filterPromiscuous}{Remove markers with poor specificity, that are
present in as many clusters as defined by \code{promiscuousCutoff}.}

\item{promiscuousCutoff}{Minimum number of clusters required to consider a
gene marker promiscuous.}
}
\value{
Grouped \link{tibble}.
}
\description{
Known Markers Detected
}
\note{
Both the \code{all} and \code{known} objects must contain Ensembl gene
identifiers in the \code{ensgene} column. We must avoid any matching operations
based on the gene symbols, since these change often and can mismatch
easily.
}
\seealso{
Other Clustering Utilities: \code{\link{pcCutoff}},
  \code{\link{plotKnownMarkers}},
  \code{\link{plotMarkers}}, \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
