% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knownMarkersDetected.R
\name{knownMarkersDetected}
\alias{knownMarkersDetected}
\title{Known Markers Detected}
\usage{
knownMarkersDetected(all, known, alpha = 0.05,
  filterPromiscuous = FALSE, promiscuousCutoff = 5L, ...)
}
\arguments{
\item{all}{\code{grouped_df}, grouped by \code{cluster}. All markers \code{data.frame},
generated by \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} return and sanitized with
\code{\link[=sanitizeMarkers]{sanitizeMarkers()}}. Note that an unmodified Seurat markers \code{data.frame}
will error here by design, since it doesn't consistently contain Ensembl
gene identifiers.}

\item{known}{\code{grouped_df}, grouped by \code{cellType} column. Known markers
\code{data.frame} imported by \code{\link[=readCellTypeMarkers]{readCellTypeMarkers()}} or pulled from internal
\link{cell_cycle_markers} data.}

\item{alpha}{\code{scalar numeric}. Alpha cutoff (adjusted P value; false
discovery rate).}

\item{filterPromiscuous}{\code{boolean}. Remove genes with poor specificity, that
are present in as many clusters as defined by \code{promiscuousCutoff}.}

\item{promiscuousCutoff}{\code{scalar integer}. Minimum number of clusters
required to consider a gene marker promiscuous.}

\item{...}{Additional arguments.}
}
\value{
\code{grouped_df}, grouped by "\code{cellType}" column.
}
\description{
Known Markers Detected
}
\note{
Both the \code{all} and \code{known} objects must contain Ensembl gene
identifiers in the \code{geneID} column. We must avoid any matching operations
based on the gene names, since these change often and can mismatch
easily.
}
\examples{
# grouped_df ====
x <- knownMarkersDetected(
    all = all_markers_small,
    known = cell_type_markers[["homoSapiens"]]
)
head(x)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeature}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotReducedDim}},
  \code{\link{sanitizeSeuratMarkers}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Functions}
