% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_run.R
\name{load_run}
\alias{load_run}
\title{Load \code{bcbio} Run}
\usage{
load_run(upload_dir = "final", sample_metadata_file,
  well_metadata_file = NULL, interesting_groups = "sample_name",
  tx2gene = NULL, ...)
}
\arguments{
\item{upload_dir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{sample_metadata_file}{\strong{Required}. Sample barcode metadata file.}

\item{well_metadata_file}{\emph{Optional}. Well identifier metadata file.}

\item{interesting_groups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{tx2gene}{tx2gene \link{data.frame}, required when using a custom
transcriptome FASTA file. This argument will be removed in a future update.}

\item{...}{Additional arguments, passed as metadata.}
}
\value{
\link{bcbioSCDataSet}.
}
\description{
Load \code{bcbio} Run
}
\details{
\itemize{
\item \strong{cellranger}: Read \href{https://www.10xgenomics.com/software/}{10x GenomicsChromium} cell counts from
\code{barcodes.tsv}, \code{genes.tsv}, and \code{matrix.mtx} files.
}
}
\note{
When working in RStudio, we recommend connecting to the bcbio-nextgen
run directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
