% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_run.R
\name{load_run}
\alias{load_run}
\title{Load bcbio-nextgen run}
\usage{
load_run(upload_dir = "final", intgroup = "sample_name", organism, metadata,
  read_counts = TRUE)
}
\arguments{
\item{upload_dir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{intgroup}{Character vector of interesting groups. First entry is used
for plot colors during quality control (QC) analysis. Entire vector is used
for PCA and heatmap QC functions.}

\item{organism}{Organism name, following Ensembl/Biomart conventions. Must be
lowercase and one word (e.g. hsapiens). This will be detected automatically
for common reference genomes.}

\item{metadata}{Optional custom metadata file to import}

\item{read_counts}{Automatically read in the count data using
\code{read_bcbio_counts()}}
}
\value{
\code{bcbio-nextgen} run object
}
\description{
We recommend loading the \code{bcbio-nextgen} run as a remote connection over
\code{sshfs}.
}
\author{
Michael Steinbaugh
}
