% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_run.R
\name{load_run}
\alias{load_run}
\title{Load bcbio-nextgen run}
\usage{
load_run(upload_dir = "final", metadata_file, organism = NULL,
  interesting_groups = "sample_name")
}
\arguments{
\item{upload_dir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{metadata_file}{(\strong{Required}). Sample barcode metadata file.}

\item{organism}{Organism name, following Ensembl conventions. Must be
lowercase and one word (e.g. hsapiens). This will be detected automatically
for common reference genomes.}

\item{interesting_groups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}
}
\value{
\link{bcbioSCDataSet}.
}
\description{
Load bcbio-nextgen run
}
\note{
When working in RStudio, we recommend connecting to the bcbio-nextgen
run directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\author{
Michael Steinbaugh
}
