% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-loadCellRanger.R
\docType{methods}
\name{loadCellRanger}
\alias{loadCellRanger}
\alias{loadCellRanger}
\alias{loadCellRanger,character-method}
\title{Load 10X Genomics CellRanger Data}
\usage{
loadCellRanger(object, ...)

\S4method{loadCellRanger}{character}(object, ...)
}
\arguments{
\item{object}{Path to CellRanger output directory. This directory path must
contain \code{filtered_gene_bc_matrices/} as a child.}

\item{...}{Additional arguments.}
}
\value{
\link{bcbioSCDataSet}.
}
\description{
Read \href{https://www.10xgenomics.com/software/}{10x Genomics Chromium} cell
counts from \code{barcodes.tsv}, \code{genes.tsv}, and \code{matrix.mtx} files.
}
\details{
This function is a simplified version of \code{\link[=loadSingleCellRun]{loadSingleCellRun()}}
optimized for handling CellRanger output.
}
\author{
Michael Steinbaugh
}
