% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCellRanger.R
\name{loadCellRanger}
\alias{loadCellRanger}
\title{Load 10X Genomics CellRanger Data}
\usage{
loadCellRanger(uploadDir, refDataDir, sampleMetadataFile,
  interestingGroups = "sampleName", prefilter = TRUE, ...)
}
\arguments{
\item{uploadDir}{Path to CellRanger output directory. This directory path
must contain \code{filtered_gene_bc_matrices/} as a child.}

\item{refDataDir}{Directory path to cellranger reference annotation data.}

\item{sampleMetadataFile}{Sample barcode metadata file.}

\item{interestingGroups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{prefilter}{Prefilter counts prior to quality control analysis.}

\item{...}{Additional arguments, to be stashed in the \code{\link[=metadata]{metadata()}} slot.}
}
\value{
\link{bcbioSingleCell}.
}
\description{
Read \href{https://www.10xgenomics.com/software/}{10x Genomics Chromium} cell
counts from \code{barcodes.tsv}, \code{genes.tsv}, and \code{matrix.mtx} files.
}
\details{
This function is a simplified version of \code{\link[=loadSingleCellRun]{loadSingleCellRun()}}
optimized for handling CellRanger output.
}
\author{
Michael Steinbaugh
}
