% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCellRanger.R
\name{loadCellRanger}
\alias{loadCellRanger}
\title{Load 10X Genomics CellRanger Data}
\usage{
loadCellRanger(uploadDir, refDataDir, interestingGroups = "sampleName",
  sampleMetadataFile, prefilter = TRUE, annotable, ...)
}
\arguments{
\item{uploadDir}{Path to CellRanger output directory. This directory path
must contain \code{filtered_gene_bc_matrices/} as a child.}

\item{refDataDir}{Directory path to cellranger reference annotation data.}

\item{interestingGroups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{sampleMetadataFile}{Sample barcode metadata file.}

\item{prefilter}{Prefilter counts prior to quality control analysis.}

\item{annotable}{\emph{Optional}. User-defined gene annotations (a.k.a.
"annotable"), which will be slotted into \code{\link[=rowData]{rowData()}}. Typically this should
be left undefined. By default, the function will automatically generate an
annotable from the annotations available on Ensembl. If set \code{NULL}, then
\code{\link[=rowData]{rowData()}} inside the resulting \link{bcbioSingleCell} object will be left
empty. This is recommended for projects dealing with genes or transcripts
that are poorly annotated.}

\item{...}{Additional arguments, to be stashed in the \code{\link[=metadata]{metadata()}} slot.}
}
\value{
\link{bcbioSingleCell}.
}
\description{
Read \href{https://www.10xgenomics.com/software/}{10x Genomics Chromium} cell
counts from \code{barcodes.tsv}, \code{genes.tsv}, and \code{matrix.mtx} files.
}
\details{
This function is a simplified version of \code{\link[=loadSingleCell]{loadSingleCell()}}
optimized for handling CellRanger output.
}
\examples{
\dontrun{
# Mus musculus
# Run with Ensembl 88 transcriptome FASTA and GTF files
bcb <- loadSingleCell(
    uploadDir = file.path("indrop_rnaseq", "final"),
    interestingGroups = c("genotype", "treatment"),
    sampleMetadataFile = file.path("meta", "sample_metadata.xlsx"),
    gtfFile = file.path(
        "annotations",
        "Mus_musculus.GRCm38.88.chr_patch_hapl_scaff.gtf.gz"),
    ensemblVersion = 88)
}
}
\author{
Michael Steinbaugh
}
