% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCellRanger.R
\name{loadCellRanger}
\alias{loadCellRanger}
\title{Load 10X Genomics Cell Ranger Data}
\usage{
loadCellRanger(uploadDir, refdataDir, sampleMetadataFile,
  interestingGroups = "sampleName", isSpike = NULL, prefilter = TRUE, ...)
}
\arguments{
\item{uploadDir}{Path to Cell Ranger output directory. This directory path
must contain \code{filtered_gene_bc_matrices*} as a child directory.}

\item{refdataDir}{Directory path to Cell Ranger reference annotation data.}

\item{sampleMetadataFile}{Sample barcode metadata file. Optional for runs
with demultiplixed index barcodes (e.g. SureCell), but otherwise required
for runs with multipliexed FASTQs containing multiple index barcodes (e.g.
inDrop). Consult the GitHub repo for examples and additional information.}

\item{interestingGroups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{isSpike}{\emph{Optional.} Gene names corresponding to FASTA spike-in
sequences (e.g. ERCCs, EGFP, TDTOMATO).}

\item{prefilter}{Prefilter counts prior to quality control analysis.}

\item{...}{Additional arguments, to be stashed in the \code{\link[=metadata]{metadata()}} slot.}
}
\value{
\code{bcbioSingleCell}.
}
\description{
Read \href{https://www.10xgenomics.com/software/}{10x Genomics Chromium} cell
counts from \code{barcodes.tsv}, \code{genes.tsv}, and \code{matrix.mtx} files.
}
\details{
This function is a simplified version of \code{\link[=loadSingleCell]{loadSingleCell()}}
optimized for handling Cell Ranger output.
}
\note{
Unlike \code{\link[=loadSingleCell]{loadSingleCell()}}, the \code{organism}, \code{ensemblRelease}, and
\code{genomeBuild} are always detected automatically, based on the 10X
\code{refdataDir} YAML metadata. Therefore, these parameters cannot be set by
the user.
}
\examples{
extdataDir <- system.file("extdata", package = "bcbioSingleCell")
loadCellRanger(
    uploadDir = file.path(extdataDir, "cellranger"),
    refdataDir = file.path(extdataDir, "refdata-cellranger-hg19-1.2.0"),
    sampleMetadataFile = file.path(extdataDir, "cellranger.csv")
)
}
\author{
Michael Steinbaugh
}
