% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSingleCell.R
\name{loadSingleCell}
\alias{loadSingleCell}
\title{Load bcbio Single-Cell RNA-Seq Data}
\usage{
loadSingleCell(uploadDir, interestingGroups = "sampleName",
  sampleMetadataFile = NULL, gtfFile = NULL, prefilter = TRUE,
  ensemblVersion = "current", ...)
}
\arguments{
\item{uploadDir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{interestingGroups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{sampleMetadataFile}{Sample barcode metadata file.}

\item{gtfFile}{\emph{Optional}. GTF (Gene Transfer Format) file, which will be
used for transcript-to-gene (\code{tx2gene}) and gene-to-symbol (\code{gene2symbol})
annotation mappings.}

\item{prefilter}{Prefilter counts prior to quality control analysis.}

\item{ensemblVersion}{Ensembl release version. Defaults to current, and does
not typically need to be user-defined. This parameter can be useful for
matching Ensembl annotations against an outdated bcbio annotation build.}

\item{...}{Additional arguments, to be stashed in the \code{\link[=metadata]{metadata()}} slot.}
}
\value{
\link{bcbioSingleCell}.
}
\description{
Load bcbio Single-Cell RNA-Seq Data
}
\note{
When working in RStudio, we recommend connecting to the bcbio-nextgen
run directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\examples{
\dontrun{
# Mus musculus
# Run with Ensembl 88 transcriptome FASTA and GTF files
bcb <- loadSingleCell(
    uploadDir = file.path("indrop_rnaseq", "final"),
    interestingGroups = c("genotype", "treatment"),
    sampleMetadataFile = file.path("meta", "sample_metadata.xlsx"),
    gtfFile = file.path(
        "annotations",
        "Mus_musculus.GRCm38.88.chr_patch_hapl_scaff.gtf.gz"),
    ensemblVersion = 88)
}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
