% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSingleCell.R
\name{loadSingleCell}
\alias{loadSingleCell}
\title{Load bcbio Single-Cell RNA-Seq Data}
\usage{
loadSingleCell(uploadDir, sampleMetadataFile,
  interestingGroups = "sampleName", prefilter = TRUE, gtfFile = NULL,
  annotable = TRUE, organism = NULL, ensemblVersion = NULL,
  genomeBuild = NULL, ...)
}
\arguments{
\item{uploadDir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{sampleMetadataFile}{Sample barcode metadata file. Optional for runs
with demultiplixed index barcodes (e.g. SureCell), but otherwise required
for runs with multipliexed FASTQs containing multiple index barcodes (e.g.
inDrop). Consult the GitHub repo for examples and additional information.}

\item{interestingGroups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{prefilter}{Prefilter counts prior to quality control analysis.}

\item{gtfFile}{\emph{Optional but recommended}. GTF (Gene Transfer Format) file,
which will be used for gene-to-symbol (\code{gene2symbol}) and
transcript-to-gene (\code{tx2gene}) annotation mappings.}

\item{annotable}{User-defined gene annotations (a.k.a.
"annotable"), which will be slotted into \code{\link[=rowData]{rowData()}}. Typically this should
be left set to \code{TRUE}. By default, the function will automatically generate
an annotable from the annotations available on Ensembl. If set \code{FALSE},
then \code{\link[=rowData]{rowData()}} inside the resulting \link{bcbioSingleCell} object will be left
empty. This is recommended for projects dealing with genes or transcripts
that are poorly annotated. Additionally, a manually constructed annotable
can be passed in as a \link{data.frame}, but this isn't generally recommended.}

\item{organism}{\emph{Optional}. Organism name. Use the full latin name (e.g.
"Homo sapiens"), since this will be input downstream to
AnnotationHub/ensembldb. If set, this genome must be supported on Ensembl.
Normally this can be left \code{NULL}, and the function will attempt to detect
the organism automatically using \code{\link[basejump:detectOrganism]{basejump::detectOrganism()}}.}

\item{ensemblVersion}{\emph{Optional}. Ensembl release version. If \code{NULL},
defaults to current release, and does not typically need to be
user-defined. This parameter can be useful for matching Ensembl annotations
against an outdated bcbio annotation build.}

\item{genomeBuild}{\emph{Optional}. Genome build. Normally this can be left \code{NULL}
and the build will be detected from the bcbio run data. This can be set
manually (e.g. "hg19" for the older \emph{Homo sapiens} reference genome). Note
that this must match the genome build identifier on Ensembl for annotations
to download correctly.}

\item{...}{Additional arguments, to be stashed in the \code{\link[=metadata]{metadata()}} slot.}
}
\value{
\link{bcbioSingleCell}.
}
\description{
Load bcbio Single-Cell RNA-Seq Data
}
\note{
When working in RStudio, we recommend connecting to the bcbio-nextgen
run directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\examples{
# Homo sapiens
# Minimal working example dataset
extdataDir <- system.file("extdata", package = "bcbioSingleCell")
uploadDir <- file.path(extdataDir, "harvard_indrop_v3")
sampleMetadataFile <- file.path(extdataDir, "harvard_indrop_v3.xlsx")
bcb <- loadSingleCell(
    uploadDir = uploadDir,
    sampleMetadataFile = sampleMetadataFile)

# Mus musculus
# Run with Ensembl 88 transcriptome FASTA and GTF files
\dontrun{
bcb <- loadSingleCell(
    uploadDir = file.path("indrop_rnaseq", "final"),
    interestingGroups = c("genotype", "treatment"),
    sampleMetadataFile = file.path("meta", "sample_metadata.xlsx"),
    gtfFile = file.path(
        "annotations",
        "Mus_musculus.GRCm38.88.chr_patch_hapl_scaff.gtf.gz"),
    ensemblVersion = 88)
}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
