% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSingleCell.R
\name{loadSingleCell}
\alias{loadSingleCell}
\title{Load bcbio Single-Cell RNA-Seq Data}
\usage{
loadSingleCell(uploadDir, interestingGroups = "sampleName",
  prefilter = TRUE, sampleMetadataFile, gtfFile, ensemblVersion, annotable,
  ...)
}
\arguments{
\item{uploadDir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{interestingGroups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{prefilter}{Prefilter counts prior to quality control analysis.}

\item{sampleMetadataFile}{Sample barcode metadata file. Optional for runs
with demultiplixed index barcodes (e.g. SureCell), but otherwise required
for runs with multipliexed FASTQs containing multiple index barcodes (e.g.
inDrop). Consult the GitHub repo for examples and additional information.}

\item{gtfFile}{\emph{Optional but recommended}. GTF (Gene Transfer Format) file,
which will be used for gene-to-symbol (\code{gene2symbol}) and
transcript-to-gene (\code{tx2gene}) annotation mappings.}

\item{ensemblVersion}{\emph{Optional}. Ensembl release version. If \code{NULL},
defaults to current release, and does not typically need to be
user-defined. This parameter can be useful for matching Ensembl annotations
against an outdated bcbio annotation build.}

\item{annotable}{\emph{Optional}. User-defined gene annotations (a.k.a.
"annotable"), which will be slotted into \code{\link[=rowData]{rowData()}}. Typically this should
be left undefined. By default, the function will automatically generate an
annotable from the annotations available on Ensembl. If set \code{NULL}, then
\code{\link[=rowData]{rowData()}} inside the resulting \link{bcbioSingleCell} object will be left
empty. This is recommended for projects dealing with genes or transcripts
that are poorly annotated.}

\item{...}{Additional arguments, to be stashed in the \code{\link[=metadata]{metadata()}} slot.}
}
\value{
\link{bcbioSingleCell}.
}
\description{
Load bcbio Single-Cell RNA-Seq Data
}
\note{
When working in RStudio, we recommend connecting to the bcbio-nextgen
run directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\examples{
\dontrun{
# Mus musculus
# Run with Ensembl 88 transcriptome FASTA and GTF files
bcb <- loadSingleCell(
    uploadDir = file.path("indrop_rnaseq", "final"),
    interestingGroups = c("genotype", "treatment"),
    sampleMetadataFile = file.path("meta", "sample_metadata.xlsx"),
    gtfFile = file.path(
        "annotations",
        "Mus_musculus.GRCm38.88.chr_patch_hapl_scaff.gtf.gz"),
    ensemblVersion = 88)
}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
