% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-loadSingleCellRun.R
\docType{methods}
\name{loadSingleCellRun}
\alias{loadSingleCellRun}
\alias{loadSingleCellRun}
\alias{loadSingleCellRun,character-method}
\title{Load bcbio Single-Cell RNA-Seq Run}
\usage{
loadSingleCellRun(object, ...)

\S4method{loadSingleCellRun}{character}(object, sampleMetadataFile,
  interestingGroups = "sampleName", gtfFile = NULL,
  wellMetadataFile = NULL, ...)
}
\arguments{
\item{object}{Path to final upload directory. This path is set when running
\code{bcbio_nextgen -w template}.}

\item{...}{Additional arguments, passed as metadata.}

\item{sampleMetadataFile}{Sample barcode metadata file.}

\item{interestingGroups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{gtfFile}{\emph{Optional}. Gene transfer format (GTF) file, which will be
used for transcript-to-gene (\code{tx2gene}) and gene-to-symbol (\code{gene2symbol})
annotation mappings.}

\item{wellMetadataFile}{\emph{Optional}. Well identifier metadata file.}
}
\value{
\link{bcbioSCDataSet}.
}
\description{
Load bcbio Single-Cell RNA-Seq Run
}
\note{
When working in RStudio, we recommend connecting to the bcbio-nextgen
run directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
