% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCellsToSamples.R
\name{mapCellsToSamples}
\alias{mapCellsToSamples}
\title{Define Cell to Sample Mappings}
\usage{
mapCellsToSamples(cells, samples)
}
\arguments{
\item{cells}{Cell identifiers.}

\item{samples}{Sample identifiers.}
}
\value{
Named factor containing cell IDs as the names, and samples as the
factor levels.
}
\description{
This function extracts \code{sampleID} from the \code{cellID} column using grep
matching.
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioSingleCell"))

cells <- colnames(bcb)
samples <- sampleData(bcb)[["sampleID"]]

map <- mapCellsToSamples(cells, samples)
head(map)
levels(map)
}
\author{
Michael Steinbaugh
}
