% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCellsToSamples.R
\name{mapCellsToSamples}
\alias{mapCellsToSamples}
\title{Define Cell to Sample Mappings}
\usage{
mapCellsToSamples(cells, samples)
}
\arguments{
\item{cells}{Cell identifiers.}

\item{samples}{Sample identifiers.}
}
\value{
Named \code{factor} containing cells as the names, and samples as the
factor levels.
}
\description{
This function extracts \code{sampleID} from the \code{cellID} column using grep
matching.
}
\examples{
# bcbioSingleCell ====
cells <- colnames(bcb_small)
samples <- sampleData(bcb_small)[["sampleID"]]
map <- mapCellsToSamples(cells, samples)
head(map)
levels(map)
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{calculateMetrics}},
  \code{\link{cell2sample}}, \code{\link{colData}},
  \code{\link{counts}}, \code{\link{fetchGeneData}},
  \code{\link{fetchPCAData}}, \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{rowData}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
