% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-metadata_table.R
\docType{methods}
\name{metadata_table}
\alias{metadata_table}
\alias{.metadata_table}
\alias{metadata_table,bcbioSCDataSet-method}
\alias{metadata_table,bcbioSCSubset-method}
\title{Metadata Table}
\usage{
metadata_table(object, ...)

\S4method{metadata_table}{bcbioSCDataSet}(object, ...)

\S4method{metadata_table}{bcbioSCSubset}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments, passed to \code{\link[knitr:kable]{knitr::kable()}}.}
}
\value{
\link{kable}.
}
\description{
Table output for a knit report.
}
\author{
Michael Steinbaugh
}
