% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-metadata_table.R
\docType{methods}
\name{metadata_table}
\alias{metadata_table}
\alias{.metadata_table}
\alias{metadata_table,bcbioSCDataSet-method}
\alias{metadata_table,SCSubset-method}
\title{Metadata table for knit report}
\usage{
\S4method{metadata_table}{bcbioSCDataSet}(object, ...)

\S4method{metadata_table}{SCSubset}(object, ...)
}
\arguments{
\item{object}{Primary object.}

\item{...}{Passthrough parameters to \code{\link[knitr:kable]{knitr::kable()}}.}
}
\value{
\link{kable}.
}
\description{
Metadata table for knit report
}
\author{
Michael Steinbaugh
}
