% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-ensembl.R
\name{metadata_table}
\alias{metadata_table}
\title{Metadata table for knit report.}
\usage{
metadata_table(run)
}
\arguments{
\item{run}{bcbio-nextgen run.}
}
\value{
\code{\link[knitr]{kable}}.
}
\description{
Metadata table for knit report.
}
\author{
Michael Steinbaugh
}
