% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-metadata_table.R
\docType{methods}
\name{metadata_table}
\alias{metadata_table}
\alias{metadata_table,bcbioSCDataSet-method}
\title{Metadata table for knit report.}
\usage{
metadata_table(object, ...)

\S4method{metadata_table}{bcbioSCDataSet}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters.}
}
\value{
\link{kable}.
}
\description{
Metadata table for knit report.
}
\author{
Michael Steinbaugh
}
