% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metadataTable.R
\docType{methods}
\name{metadataTable}
\alias{metadataTable}
\alias{metadataTable,bcbioSingleCellANY-method}
\title{Metadata Table}
\usage{
\S4method{metadataTable}{bcbioSingleCellANY}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments, passed to \code{\link[knitr:kable]{knitr::kable()}}.}
}
\value{
\link{kable}.
}
\description{
Table output for a knit report.
}
\author{
Michael Steinbaugh
}
