% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/internal-metrics.R,
%   R/methods-metrics.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{.metrics}
\alias{metrics,bcbioSCDataSet-method}
\title{Generate barcode metrics summary}
\usage{
metrics(object, ...)

.metrics(sparse_counts, annotable)

\S4method{metrics}{bcbioSCDataSet}(object, unique_names = FALSE)
}
\arguments{
\item{object}{\link{bcbioSCDataSet}.}

\item{sparse_counts}{Sparse counts matrix.}

\item{annotable}{Annotable.}

\item{unique_names}{Unique sample names.}
}
\value{
Tibble grouped by sample name.

\link{data.frame}.
}
\description{
Generate barcode metrics summary

Sample barcode metrics
}
\author{
Rory Kirchner

Michael Steinbaugh
}
\keyword{internal}
