% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metrics.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,matrix-method}
\alias{metrics,dgCMatrix-method}
\alias{metrics,dgTMatrix-method}
\alias{metrics,SingleCellExperiment-method}
\alias{metrics,seurat-method}
\title{Sample Barcode Metrics}
\usage{
\S4method{metrics}{matrix}(object, rowData = NULL, prefilter = TRUE)

\S4method{metrics}{dgCMatrix}(object, rowData = NULL, prefilter = TRUE)

\S4method{metrics}{dgTMatrix}(object, rowData = NULL, prefilter = TRUE)

\S4method{metrics}{SingleCellExperiment}(object, interestingGroups)

\S4method{metrics}{seurat}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{rowData}{Data describing the rows of the object.}

\item{prefilter}{Whether to apply pre-filtering to the cellular barcodes.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in camel case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{...}{Additional arguments.}
}
\value{
\code{data.frame} with cellular barcodes as rows.
}
\description{
Sample Barcode Metrics
}
\examples{
# SingleCellExperiment ====
x <- metrics(cellranger_small)
glimpse(x)

# seurat ====
x <- metrics(Seurat::pbmc_small)
glimpse(x)

# dgCMatrix ====
counts <- counts(cellranger_small)
class(counts)
x <- metrics(counts)
glimpse(x)
}
\seealso{
Other Quality Control Functions: \code{\link{barcodeRanksPerSample}},
  \code{\link{filterCells}}, \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
