% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metrics.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,bcbioSingleCell-method}
\alias{metrics,bcbioSingleCellLegacy-method}
\alias{metrics,seurat-method}
\title{Sample Barcode Metrics}
\usage{
\S4method{metrics}{bcbioSingleCell}(object, filterCells = FALSE,
  aggregateReplicates = FALSE)

\S4method{metrics}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{filterCells}{Show only the cells that have passed filtering cutoffs.}

\item{aggregateReplicates}{Aggregate technical replicates, if present. If
\code{TRUE}, this function uses the values slotted in
\code{sampleMetadata(object)[["sampleNameAggregate"]])}.

[["sampleNameAggregate"]: R:[%22sampleNameAggregate%22}
}
\value{
\link{data.frame} with cellular barcodes as rows.
}
\description{
Sample Barcode Metrics
}
\seealso{
Other Quality Control Metrics: \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Metrics}
