% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metrics.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,bcbioSingleCell-method}
\alias{metrics,seurat-method}
\title{Sample Barcode Metrics}
\usage{
\S4method{metrics}{bcbioSingleCell}(object, interestingGroups)

\S4method{metrics}{seurat}(object, interestingGroups)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in \code{\link[=camel]{camel()}} case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}
}
\value{
\code{data.frame} with cellular barcodes as rows.
}
\description{
Sample Barcode Metrics
}
\examples{
# bcbioSingleCell ====
metrics(bcb_small) \%>\% glimpse()

# seurat ====
metrics(seurat_small) \%>\% glimpse()
metrics(pbmc_small) \%>\% glimpse()
}
\seealso{
\code{\link[=sampleData]{sampleData()}}.

Other Quality Control Functions: \code{\link{calculateMetrics}},
  \code{\link{metricsPerSample}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotQuantileHeatmap}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
