% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metrics.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,matrix-method}
\alias{metrics,dgCMatrix-method}
\alias{metrics,dgTMatrix-method}
\alias{metrics,SingleCellExperiment-method}
\alias{metrics,seurat-method}
\title{Sample Barcode Metrics}
\usage{
\S4method{metrics}{matrix}(object, rowData = NULL, prefilter = TRUE)

\S4method{metrics}{dgCMatrix}(object, rowData = NULL, prefilter = TRUE)

\S4method{metrics}{dgTMatrix}(object, rowData = NULL, prefilter = TRUE)

\S4method{metrics}{SingleCellExperiment}(object, interestingGroups)

\S4method{metrics}{seurat}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{rowData}{Data describing the rows of the object.}

\item{prefilter}{Whether to apply pre-filtering to the cellular barcodes.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in camel case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{...}{Additional arguments.}
}
\value{
\code{data.frame} with cellular barcodes as rows.
}
\description{
Sample Barcode Metrics
}
\examples{
# bcbioSingleCell ====
metrics(indrops_small) \%>\% glimpse()

# SingleCellExperiment ====
metrics(cellranger_small) \%>\% glimpse()

# dgCMatrix ====
counts <- counts(indrops_small)
class(counts)
metrics(counts) \%>\% glimpse()

# seurat ====
metrics(seurat_small) \%>\% glimpse()
metrics(Seurat::pbmc_small) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{fetchGeneData}},
  \code{\link{filterCells}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
