% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,bcbioSingleCell-method}
\alias{metrics,CellRanger-method}
\title{Metrics}
\usage{
\S4method{metrics}{bcbioSingleCell}(object, recalculate = FALSE)

\S4method{metrics}{CellRanger}(object, recalculate = FALSE)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{grouped_df}. Grouped by \code{sampleID} column.
}
\description{
This function takes data stored in \code{\link[=colData]{colData()}} and consistently returns a
tibble grouped by sample, to be used for plotting with \code{ggplot2}.
}
\section{SingleCellExperiment}{



Note that metrics are cell level but grouped by sample.

}

\examples{
x <- metrics(indrops_small, recalculate = TRUE)
print(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{makeSingleCellExperiment}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapCellsToSamples}},
  \code{\link{mapGenesToIDs}}, \code{\link{meltCounts}},
  \code{\link{metricsPerSample}},
  \code{\link{minimalSampleData}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}},
  \code{\link{topCellsPerSample}}, \code{\link{tpm}},
  \code{\link{tx2gene}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
