% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-metrics.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{.metrics}
\alias{metrics,bcbioSCDataSet-method}
\alias{metrics,SCSubset-method}
\alias{metrics,seurat-method}
\title{Sample barcode metrics}
\usage{
\S4method{metrics}{bcbioSCDataSet}(object)

\S4method{metrics}{SCSubset}(object)

\S4method{metrics}{seurat}(object)
}
\arguments{
\item{object}{Primary object.}
}
\value{
\link{data.frame}.
}
\description{
Sample barcode metrics
}
