% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,SingleCellExperiment-method}
\title{Sample Barcode Metrics}
\usage{
\S4method{metrics}{SingleCellExperiment}(object,
  interestingGroups = NULL, recalculate = FALSE)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}

\item{recalculate}{\code{boolean}. Force recalculation, using primary \code{\link[=counts]{counts()}}
matrix.}
}
\value{
\code{data.frame} with cellular barcodes as rows. If \code{recalculate = TRUE},
returns the original object class (e.g. \code{SingleCellExperiment}), with
updated metrics in \code{\link[=colData]{colData()}}.
}
\description{
Sample Barcode Metrics
}
\examples{
x <- metrics(indrops_small)
glimpse(x)

# Recalculate mode
x <- metrics(indrops_small, recalculate = TRUE)
print(x)
}
\seealso{
Other Quality Control Functions: \code{\link{barcodeRanksPerSample}},
  \code{\link{filterCells}}, \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Functions}
