% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,bcbioSingleCell-method}
\alias{metrics,CellRanger-method}
\title{Metrics}
\usage{
metrics(object, ...)

\S4method{metrics}{bcbioSingleCell}(object, recalculate = FALSE)

\S4method{metrics}{CellRanger}(object, recalculate = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{recalculate}{\code{boolean}. Force recalculation, using primary \code{\link[=counts]{counts()}}
matrix.}
}
\value{
\code{grouped_df}. Grouped by \code{sampleID} column.
}
\description{
This function takes data stored in \code{\link[=colData]{colData()}} and consistently returns a
tibble grouped by sample, to be used for plotting with \code{ggplot2}.
}
\section{SingleCellExperiment}{



Note that metrics are cell level but grouped by sample.

}

\examples{
data(indrops_small)
x <- metrics(indrops_small, recalculate = TRUE)
print(x)
}
\seealso{
Other SummarizedExperiment Functions: \code{\link{counts}},
  \code{\link{interestingGroups}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenes}}, \code{\link{meltCounts}},
  \code{\link{sampleData}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}, \code{\link{tpm}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
