% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-metrics.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{.metrics}
\alias{metrics,bcbioSCDataSet-method}
\alias{metrics,SCSubset-method}
\title{Sample barcode metrics}
\usage{
\S4method{metrics}{bcbioSCDataSet}(object, unique_names = FALSE)

\S4method{metrics}{SCSubset}(object, unique_names = FALSE)
}
\arguments{
\item{object}{Primary object.}

\item{unique_names}{Unique sample names.}
}
\value{
\link{data.frame}.
}
\description{
Sample barcode metrics
}
