% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metrics.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,bcbioSingleCell-method}
\alias{metrics,seurat-method}
\title{Sample Barcode Metrics}
\usage{
\S4method{metrics}{bcbioSingleCell}(object, interestingGroups,
  filterCells = TRUE, aggregateReplicates = TRUE)

\S4method{metrics}{seurat}(object, interestingGroups)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Interesting group, to use for colors.}

\item{filterCells}{Show only the cells that have passed filtering cutoffs.}

\item{aggregateReplicates}{Aggregate technical replicates, if specified. This
function uses values assigned in the \code{sampleNameAggregate} column of the
internal sample metadata \link{data.frame}.}
}
\value{
\link{data.frame} with cellular barcodes as rows.
}
\description{
Sample Barcode Metrics
}
\seealso{
\code{\link[=sampleMetadata]{sampleMetadata()}}.

Other Quality Control Metrics: \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Metrics}
