% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-methods.R, R/zzz-seurat-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,matrix-method}
\alias{metrics,dgCMatrix-method}
\alias{metrics,dgTMatrix-method}
\alias{metrics,SingleCellExperiment-method}
\alias{metrics,seurat-method}
\title{Sample Barcode Metrics}
\usage{
\S4method{metrics}{matrix}(object, rowRanges = NULL, prefilter = FALSE)

\S4method{metrics}{dgCMatrix}(object, rowRanges = NULL,
  prefilter = FALSE)

\S4method{metrics}{dgTMatrix}(object, rowRanges = NULL,
  prefilter = FALSE)

\S4method{metrics}{SingleCellExperiment}(object, interestingGroups)

\S4method{metrics}{seurat}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{rowRanges}{\code{GRanges}. Data describing the rows of the object.}

\item{prefilter}{\code{boolean}. Apply pre-filtering to the cellular barcodes
(\emph{recommended}).}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}

\item{...}{Additional arguments.}
}
\value{
\code{data.frame} with cellular barcodes as rows.
}
\description{
Sample Barcode Metrics
}
\examples{
# SingleCellExperiment ====
x <- metrics(cellranger_small)
glimpse(x)

# matrix ====
counts <- counts(cellranger_small)
class(counts)
x <- metrics(counts)
glimpse(x)
}
\seealso{
Other Quality Control Functions: \code{\link{barcodeRanksPerSample}},
  \code{\link{filterCells}}, \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Functions}
