% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metrics.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,dgCMatrix-method}
\alias{metrics,SingleCellExperiment-method}
\alias{metrics,seurat-method}
\title{Sample Barcode Metrics}
\usage{
\S4method{metrics}{dgCMatrix}(object, rowData = NULL, prefilter = TRUE)

\S4method{metrics}{SingleCellExperiment}(object)

\S4method{metrics}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{rowData}{Data describing the rows of the object.}

\item{prefilter}{Whether to apply pre-filtering to the cellular barcodes.}
}
\value{
\code{data.frame} with cellular barcodes as rows.
}
\description{
Sample Barcode Metrics
}
\examples{
# bcbioSingleCell ====
metrics(bcb_small) \%>\% glimpse()

# seurat ====
metrics(seurat_small) \%>\% glimpse()
metrics(pbmc_small) \%>\% glimpse()

# dgCMatrix ====
mat <- counts(bcb_small)
metrics(mat) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{fetchGeneData}}, \code{\link{fetchPCAData}},
  \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
