% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-metricsPerSample.R
\docType{methods}
\name{metricsPerSample}
\alias{metricsPerSample}
\alias{metricsPerSample}
\alias{metricsPerSample,bcbioSingleCell-method}
\alias{metricsPerSample,seurat-method}
\title{Metrics per Sample}
\usage{
metricsPerSample(object, ...)

\S4method{metricsPerSample}{bcbioSingleCell}(object)

\S4method{metricsPerSample}{seurat}(object)
}
\value{
\link{tibble} containing summary statistics.
}
\description{
Metrics per Sample
}
\author{
Michael Steinbaugh
}
