% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-metricsPerSample.R
\docType{methods}
\name{metricsPerSample}
\alias{metricsPerSample}
\alias{metricsPerSample}
\alias{metricsPerSample,bcbioSingleCell-method}
\alias{metricsPerSample,seurat-method}
\title{Metrics per Sample}
\usage{
metricsPerSample(object, ...)

\S4method{metricsPerSample}{bcbioSingleCell}(object)

\S4method{metricsPerSample}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{tibble} containing summary statistics.
}
\description{
Metrics per Sample
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioSingleCell"))
load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
metricsPerSample(bcb)

# seurat
metricsPerSample(seurat)
}
\author{
Michael Steinbaugh
}
