% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/metricsPerSample-methods.R
\docType{methods}
\name{metricsPerSample}
\alias{metricsPerSample}
\alias{metricsPerSample,SingleCellExperiment-method}
\title{Metrics per Sample}
\usage{
metricsPerSample(object, ...)

\S4method{metricsPerSample}{SingleCellExperiment}(object, f = c("mean",
  "median", "sum"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{f}{\code{string}. Mathematical function name to apply. Defaults to "\code{sum}"
but "\code{mean}" and "\code{median}" are also supported.}
}
\value{
\code{data.frame}.
}
\description{
Calculate summary statistics per sample.
}
\examples{
# mean
x <- metricsPerSample(indrops_small, f = "mean")
head(x)

# median
x <- metricsPerSample(indrops_small, f = "median")
head(x)

# sum
x <- metricsPerSample(indrops_small, f = "sum")
head(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{cell2sample}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
