% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-metricsPerSample.R
\docType{methods}
\name{metricsPerSample}
\alias{metricsPerSample}
\alias{metricsPerSample}
\alias{metricsPerSample,bcbioSingleCell-method}
\alias{metricsPerSample,seurat-method}
\title{Metrics per Sample}
\usage{
metricsPerSample(object, ...)

\S4method{metricsPerSample}{bcbioSingleCell}(object)

\S4method{metricsPerSample}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{tbl_df} containing summary statistics.
}
\description{
Metrics per Sample
}
\examples{
# bcbioSingleCell ====
metricsPerSample(bcb_small)

# seurat ====
metricsPerSample(seurat_small)
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{calculateMetrics}},
  \code{\link{cell2sample}}, \code{\link{colData}},
  \code{\link{counts}}, \code{\link{fetchGeneData}},
  \code{\link{fetchPCAData}}, \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}}, \code{\link{gene2symbol}},
  \code{\link{interestingGroups}},
  \code{\link{mapCellsToSamples}}, \code{\link{metrics}},
  \code{\link{rowData}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
