% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-metricsPerSample.R,
%   R/methods-seurat-SingleCellExperiment.R
\docType{methods}
\name{metricsPerSample}
\alias{metricsPerSample}
\alias{metricsPerSample}
\alias{metricsPerSample,SingleCellExperiment-method}
\alias{metricsPerSample,seurat-method}
\title{Metrics per Sample}
\usage{
metricsPerSample(object, ...)

\S4method{metricsPerSample}{SingleCellExperiment}(object, f = c("mean",
  "median", "sum"))

\S4method{metricsPerSample}{seurat}(object, f = c("mean", "median", "sum"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{f}{Mathematical function to apply. Defaults to "\code{sum}".}
}
\value{
\code{data.frame}.
}
\description{
Calculate summary statistics per sample.
}
\examples{
# SingleCellExperiment ====
# mean
x <- metricsPerSample(bcb_small, f = "mean")
head(x)

# median
x <- metricsPerSample(bcb_small, f = "median")
head(x)

# sum
x <- metricsPerSample(bcb_small, f = "sum")
head(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{fetchGeneData}}, \code{\link{fetchPCAData}},
  \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}},
  \code{\link{mapCellsToSamples}}, \code{\link{metrics}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
