% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mito_counts_plot.R
\name{mito_counts_plot}
\alias{mito_counts_plot}
\title{Total counts vs. genes detected plot}
\usage{
mito_counts_plot(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\description{
Total counts vs. genes detected plot
}
\author{
Rory Kirchner

Michael Steinbaugh
}
\keyword{plot}
\keyword{qc}
