% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novelty_plot.R
\name{novelty_plot}
\alias{novelty_plot}
\title{Novelty plot}
\usage{
novelty_plot(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\description{
log detected per count
}
\author{
Rory Kirchner

Michael Steinbaugh
}
\keyword{plot}
\keyword{qc}
