% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode.R
\name{plot_barcode}
\alias{plot_barcode}
\title{Plot a barcode histogram for a given sample}
\usage{
plot_barcode(filename, sample = NULL)
}
\arguments{
\item{filename}{path to a barcode histogram file}

\item{sample}{title for plot}
}
\description{
Plot a barcode histogram for a given sample
}
\author{
Rory Kirchner

Michael Steinbaugh
}
\keyword{internal}
