% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-barcodes.R
\name{plot_barcodes}
\alias{plot_barcodes}
\title{Plot cellular barcode distributions per sample}
\usage{
plot_barcodes(run)
}
\arguments{
\item{run}{\link{bcbioSCDataSet}.}
}
\description{
Plot cellular barcode distributions per sample
}
\details{
A violin plot is a compact display of a continuous distribution. It
is a blend of \code{geom_boxplot()} and \code{geom_density()}: a violin plot is a
mirrored density plot displayed in the same way as a boxplot.
}
\author{
Rory Kirchner

Michael Steinbaugh
}
