% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_cell_counts.R
\docType{methods}
\name{plot_cell_counts}
\alias{plot_cell_counts}
\alias{.plot_cell_counts}
\alias{plot_cell_counts,bcbioSCDataSet-method}
\alias{plot_cell_counts,bcbioSCSubset-method}
\title{Plot Cell Counts}
\usage{
plot_cell_counts(object, ...)

\S4method{plot_cell_counts}{bcbioSCDataSet}(object)

\S4method{plot_cell_counts}{bcbioSCSubset}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{ggplot}.
}
\description{
Plot Cell Counts
}
\author{
Michael Steinbaugh, Rory Kirchner
}
