% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_clusters.R
\docType{methods}
\name{plot_clusters}
\alias{plot_clusters}
\alias{plot_clusters,seurat-method}
\title{Plot Clusters}
\usage{
plot_clusters(object, ...)

\S4method{plot_clusters}{seurat}(object, symbols)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{symbols}{Character vector of gene symbols.}
}
\value{
No value, only graphical output.
}
\description{
Plot Clusters
}
\author{
Michael Steinbaugh
}
