% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-seurat.R
\docType{methods}
\name{plot_clusters}
\alias{plot_clusters}
\alias{plot_clusters,seurat-method}
\title{Plot clusters}
\usage{
\S4method{plot_clusters}{seurat}(object, symbols)
}
\arguments{
\item{object}{Primary object.}

\item{symbols}{Character vector of gene symbols.}
}
\value{
No value, only graphical output.
}
\description{
Plot clusters
}
\author{
Michael Steinbaugh
}
