% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes_detected.R
\name{plot_genes_detected}
\alias{plot_genes_detected}
\title{Genes detected plot}
\usage{
plot_genes_detected(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\description{
Genes detected plot
}
\author{
Rory Kirchner

Michael Steinbaugh
}
