% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes_detected.R
\name{plot_genes_detected_boxplot}
\alias{plot_genes_detected_boxplot}
\alias{plot_genes_detected_histogram}
\alias{plot_genes_detected}
\title{Genes detected plots}
\usage{
plot_genes_detected_boxplot(metrics, min_genes = get("min_genes", envir =
  parent.frame()), max_genes = get("max_genes", envir = parent.frame()))

plot_genes_detected_histogram(metrics, min_genes = get("min_genes", envir =
  parent.frame()), max_genes = get("max_genes", envir = parent.frame()))

plot_genes_detected(...)
}
\arguments{
\item{metrics}{Barcode metrics data frame}

\item{min_genes}{Recommended minimum gene count cutoff}

\item{max_genes}{Recommended maximum gene count cutoff}

\item{...}{Passthrough parameters}
}
\value{
ggplot2 object
}
\description{
Boxplot

Histogram

Show both plots
}
\author{
Rory Kirchner

Michael Steinbaugh
}
