% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-seurat.R
\docType{methods}
\name{plot_known_markers}
\alias{plot_known_markers}
\alias{plot_known_markers,seurat-method}
\title{Plot known markers}
\usage{
\S4method{plot_known_markers}{seurat}(object, known_markers_detected)
}
\arguments{
\item{object}{Primary object.}

\item{known_markers_detected}{Known markers detected \link{data.frame}.}
}
\description{
Plot known markers
}
