% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{plot_known_markers}
\alias{plot_known_markers}
\title{Plot known markers}
\usage{
plot_known_markers(seurat, cell_type)
}
\arguments{
\item{seurat}{\link{seurat}.}

\item{cell_type}{Cell type.}
}
\description{
Plot known markers
}
