% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{plot_known_markers}
\alias{plot_known_markers}
\title{Plot known markers}
\usage{
plot_known_markers(seurat, known_markers_detected)
}
\arguments{
\item{seurat}{\link{seurat}.}

\item{known_markers_detected}{Known markers detected \link{data.frame}.}
}
\description{
Plot known markers
}
