% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-seurat.R
\docType{methods}
\name{plot_known_markers}
\alias{plot_known_markers}
\alias{plot_known_markers,seurat,grouped_df-method}
\title{Plot known markers}
\usage{
\S4method{plot_known_markers}{seurat,grouped_df}(x, y, markdown = TRUE)
}
\arguments{
\item{x}{Primary object.}

\item{y}{\code{\link[=known_markers_detected]{known_markers_detected()}} \link{tibble} grouped by cluster.}

\item{markdown}{Print Markdown headers.}
}
\description{
Plot known markers
}
