% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mito_counts.R
\name{plot_mito_counts_histogram}
\alias{plot_mito_counts_histogram}
\alias{plot_mito_counts_boxplot}
\alias{plot_mito_counts}
\title{Mitochondrial count plots}
\usage{
plot_mito_counts_histogram(metrics)

plot_mito_counts_boxplot(metrics)

plot_mito_counts(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\value{
ggplot2 object
}
\description{
Histogram

Boxplot

Show both plots
}
\author{
Rory Kirchner

Michael Steinbaugh
}
