% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mito_counts.R
\name{plot_mito_counts_histogram}
\alias{plot_mito_counts_histogram}
\alias{plot_mito_counts_boxplot}
\alias{plot_mito_counts_scatterplot}
\alias{plot_mito_counts}
\title{Mitochondrial count plots}
\usage{
plot_mito_counts_histogram(metrics, percent_mito = get("percent_mito", envir =
  parent.frame()))

plot_mito_counts_boxplot(metrics, percent_mito = get("percent_mito", envir =
  parent.frame()))

plot_mito_counts_scatterplot(metrics, percent_mito = NULL)

plot_mito_counts(...)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\value{
ggplot2 object
}
\description{
Histogram

Boxplot

Scatterplot

Show both plots
}
\author{
Rory Kirchner

Michael Steinbaugh
}
