% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_novelty.R
\name{plot_novelty_histogram}
\alias{plot_novelty_histogram}
\alias{plot_novelty_boxplot}
\alias{plot_novelty}
\title{Novelty plots (log detected per count)}
\usage{
plot_novelty_histogram(metrics)

plot_novelty_boxplot(metrics)

plot_novelty(metrics)
}
\arguments{
\item{metrics}{Barcode metrics data frame}
}
\value{
ggplot2 object
}
\description{
Histogram

Boxplot

Show both plots
}
\author{
Rory Kirchner

Michael Steinbaugh
}
