% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_novelty.R
\docType{methods}
\name{plot_novelty}
\alias{plot_novelty}
\alias{.plot_novelty_boxplot}
\alias{.plot_novelty_histogram}
\alias{.plot_novelty}
\alias{plot_novelty,bcbioSCDataSet-method}
\alias{plot_novelty,bcbioSCSubset-method}
\title{Plot Novelty Score}
\usage{
plot_novelty(object, ...)

\S4method{plot_novelty}{bcbioSCDataSet}(object, min = 0.8)

\S4method{plot_novelty}{bcbioSCSubset}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{min}{Recommended minimum value cutoff.}
}
\value{
\link{ggplot} grid.
}
\description{
Plot Novelty Score
}
\details{
"Novelty" refers to log10 genes detected per count.
}
