% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-seurat.R
\docType{methods}
\name{plot_top_markers}
\alias{plot_top_markers}
\alias{plot_top_markers,seurat,grouped_df-method}
\title{Plot top markers}
\usage{
\S4method{plot_top_markers}{seurat,grouped_df}(x, y, markdown = TRUE)
}
\arguments{
\item{x}{Object.}

\item{y}{Top markers grouped \link{tibble} returned by \code{\link[=top_markers]{top_markers()}}.}

\item{markdown}{Print Markdown headers.}
}
\value{
\link{ggplot}.
}
\description{
Plot top markers
}
