% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-seurat.R
\docType{methods}
\name{plot_top_markers}
\alias{plot_top_markers}
\alias{plot_top_markers,seurat-method}
\title{Plot top markers}
\usage{
\S4method{plot_top_markers}{seurat}(object, top_markers)
}
\arguments{
\item{object}{Primary object.}

\item{top_markers}{\link{data.frame} returned by \code{\link[=top_markers]{top_markers()}}.}
}
\value{
\link{ggplot}.
}
\description{
Plot top markers
}
