% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{plot_top_markers}
\alias{plot_top_markers}
\title{Plot top markers}
\usage{
plot_top_markers(seurat, top)
}
\arguments{
\item{seurat}{\link{seurat}.}

\item{top}{Top genes data frame.}
}
\value{
\link{ggplot}.
}
\description{
Plot top markers
}
