% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_total_counts.R
\name{plot_total_counts_histogram}
\alias{plot_total_counts_histogram}
\alias{plot_total_counts_boxplot}
\alias{plot_total_counts}
\title{Total counts plots}
\usage{
plot_total_counts_histogram(metrics)

plot_total_counts_boxplot(metrics)

plot_total_counts(...)
}
\arguments{
\item{metrics}{Barcode metrics data frame}

\item{...}{Passthrough parameters}
}
\value{
ggplot2 object
}
\description{
Histogram

Boxplot

Show both plots
}
\author{
Rory Kirchner

Michael Steinbaugh
}
