% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_umis_vs_genes.R
\docType{methods}
\name{plot_umis_vs_genes}
\alias{plot_umis_vs_genes}
\alias{.plot_umis_vs_genes}
\alias{plot_umis_vs_genes,bcbioSCDataSet-method}
\alias{plot_umis_vs_genes,SCSubset-method}
\title{Plot UMI and gene correlation}
\usage{
\S4method{plot_umis_vs_genes}{bcbioSCDataSet}(object)

\S4method{plot_umis_vs_genes}{SCSubset}(object)
}
\arguments{
\item{object}{Primary object.}
}
\value{
\link{ggplot}.
}
\description{
Plot UMI and gene correlation
}
\author{
Michael Steinbaugh, Rory Kirchner
}
