% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellCounts-methods.R
\docType{methods}
\name{plotCellCounts}
\alias{plotCellCounts}
\alias{plotCellCounts,bcbioSingleCell-method}
\title{Plot cell counts}
\usage{
plotCellCounts(object, ...)

\S4method{plotCellCounts}{bcbioSingleCell}(object,
  interestingGroups = NULL, fill = getOption(x = "acid.fill.discrete",
  default = acidplots::scale_fill_synesthesia_d()),
  title = "Cell counts")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{title}{\code{character(1)}.
Title.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot cell counts
}
\note{
Updated 2019-07-29.
}
\examples{
data(indrops)
plotCellCounts(indrops)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
