% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellCounts-methods.R
\docType{methods}
\name{plotCellCounts}
\alias{plotCellCounts}
\alias{plotCellCounts,SingleCellExperiment-method}
\title{Plot cell counts}
\usage{
\S4method{plotCellCounts}{SingleCellExperiment}(object,
  interestingGroups = NULL, fill = getOption("basejump.fill.discrete",
  NULL), title = "cell counts")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(basejump.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot cell counts
}
\examples{
data(indrops)
plotCellCounts(indrops)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
