% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotCellCounts.R
\docType{methods}
\name{plotCellCounts}
\alias{plotCellCounts}
\alias{plotCellCounts,bcbioSingleCellANY-method}
\title{Plot Cell Counts}
\usage{
plotCellCounts(object, ...)

\S4method{plotCellCounts}{bcbioSingleCellANY}(object, interestingGroup,
  filterCells = FALSE, aggregateReplicates = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroup}{Interesting group, to use for colors.}

\item{filterCells}{Show only the cells that have passed filtering cutoffs.}

\item{aggregateReplicates}{Aggregate technical replicates, if present. If
\code{TRUE}, this function uses the values slotted in
\code{sampleMetadata(object)[["sampleNameAggregate"]])}.

[["sampleNameAggregate"]: R:[%22sampleNameAggregate%22}
}
\value{
\link{ggplot} grid.
}
\description{
Plot Cell Counts
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Metrics}
