% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCellCounts-methods.R
\docType{methods}
\name{plotCellCounts}
\alias{plotCellCounts}
\alias{plotCellCounts}
\alias{plotCellCounts,bcbioSCDataSet-method}
\alias{plotCellCounts,bcbioSCSubset-method}
\title{Plot Cell Counts}
\usage{
plotCellCounts(object, ...)

\S4method{plotCellCounts}{bcbioSCDataSet}(object)

\S4method{plotCellCounts}{bcbioSCSubset}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{ggplot} grid.
}
\description{
Plot Cell Counts
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellularBarcodes}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
