% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotCellCounts.R
\docType{methods}
\name{plotCellCounts}
\alias{plotCellCounts}
\alias{plotCellCounts}
\alias{plotCellCounts,bcbioSingleCell-method}
\alias{plotCellCounts,data.frame-method}
\alias{plotCellCounts,seurat-method}
\title{Plot Cell Counts}
\usage{
plotCellCounts(object, ...)

\S4method{plotCellCounts}{bcbioSingleCell}(object, interestingGroups,
  fill = viridis::scale_fill_viridis(discrete = TRUE))

\S4method{plotCellCounts}{data.frame}(object, metadata, interestingGroups,
  fill = viridis::scale_fill_viridis(discrete = TRUE))

\S4method{plotCellCounts}{seurat}(object, interestingGroups,
  fill = viridis::scale_fill_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{Interesting group, to use for colors.}

\item{fill}{Desired ggplot fill scale. Defaults to
\code{\link[viridis:scale_fill_viridis]{viridis::scale_fill_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{metadata}{Sample metadata \link{data.frame}.}
}
\value{
\link{ggplot}.
}
\description{
Plot Cell Counts
}
\details{
A violin plot is a comact display of a continuous distribution. It
is a blend of \code{\link[=geom_boxplot]{geom_boxplot()}} and \code{\link[=geom_density]{geom_density()}}: a violin plot is a
mirrored density plot displayed in the same way as a boxplot.
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioSingleCell"))
load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
plotCellCounts(bcb)

# seurat
plotCellCounts(seurat)

# data.frame
df <- metrics(bcb)
metadata <- sampleMetadata(bcb)
plotCellCounts(df, metadata = metadata)
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
