% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCellCounts-methods.R
\docType{methods}
\name{plotCellCounts}
\alias{plotCellCounts}
\alias{plotCellCounts,SingleCellExperiment-method}
\title{Plot Cell Counts}
\usage{
plotCellCounts(object, ...)

\S4method{plotCellCounts}{SingleCellExperiment}(object,
  interestingGroups = NULL, fill = getOption("bcbio.discrete.fill",
  NULL), title = "cell counts")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.
To set the discrete fill palette globally, use
\code{options(bcbio.discrete.fill = scale_fill_viridis_d())}.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot Cell Counts
}
\examples{
plotCellCounts(indrops_small)
}
\seealso{
Other Quality Control Functions: \code{\link{barcodeRanksPerSample}},
  \code{\link{filterCells}},
  \code{\link{plotBarcodeRanks}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Functions}
