% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotCellCounts.R
\docType{methods}
\name{plotCellCounts}
\alias{plotCellCounts}
\alias{plotCellCounts,bcbioSingleCell-method}
\alias{plotCellCounts,data.frame-method}
\alias{plotCellCounts,seurat-method}
\title{Plot Cell Counts}
\usage{
plotCellCounts(object, ...)

\S4method{plotCellCounts}{bcbioSingleCell}(object, interestingGroups,
  filterCells = TRUE, aggregateReplicates = TRUE,
  fill = scale_fill_viridis(discrete = TRUE))

\S4method{plotCellCounts}{data.frame}(object, metadata, interestingGroups,
  multiplexed = FALSE, fill = scale_fill_viridis(discrete = TRUE))

\S4method{plotCellCounts}{seurat}(object, interestingGroups,
  fill = scale_fill_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroups}{Interesting group, to use for colors.}

\item{filterCells}{Show only the cells that have passed filtering cutoffs.}

\item{aggregateReplicates}{Aggregate technical replicates, if specified. This
function uses values assigned in the \code{sampleNameAggregate} column of the
internal sample metadata \link{data.frame}.}

\item{fill}{Desired ggplot fill scale. Defaults to
\code{\link[viridis:scale_fill_viridis]{viridis::scale_fill_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{metadata}{Sample metadata \link{data.frame}.}

\item{multiplexed}{Whether the FASTQ files are multiplexed. This is
detected automatically for \link{bcbioSingleCell} objects.}
}
\value{
\link{ggplot}.
}
\description{
Plot Cell Counts
}
\examples{
# bcbioSingleCell
\dontrun{
plotCellCounts(bcb)
}

# seurat
\dontrun{
plotCellCounts(seurat)
}

# metrics data.frame
\dontrun{
metrics <- metrics(bcb)
metadata <- sampleMetadata(bcb)
plotCellCounts(metrics, metadata)
}
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Metrics}
