% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCellTypesPerCluster.R
\docType{methods}
\name{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster,seurat,grouped_df-method}
\title{Plot Cell Types per Cluster}
\usage{
plotCellTypesPerCluster(object, cellTypesPerCluster, ...)

\S4method{plotCellTypesPerCluster}{seurat,grouped_df}(object,
  cellTypesPerCluster)
}
\arguments{
\item{object}{Object.}

\item{cellTypesPerCluster}{Cell types per cluster grouped \link{tibble}. This must
be the return from \code{\link[=cellTypesPerCluster]{cellTypesPerCluster()}}.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{ggplot}.
}
\description{
Plot the geometric mean of the significant marker genes for every known cell
type (per unbiased cluster). Cell types with too few (\code{min} cutoff) or too
many (\code{max} cutoff) marker genes will be skipped.
}
\examples{
\dontrun{
cellTypes <- cellTypesPerCluster(knownMakersDetected)
plotCellTypesPerCluster(cellTypes)
}
}
\author{
Michael Steinbaugh
}
