% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCellTypesPerCluster.R
\docType{methods}
\name{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster,seurat,grouped_df-method}
\title{Plot Cell Types per Cluster}
\usage{
plotCellTypesPerCluster(object, cellTypesPerCluster, ...)

\S4method{plotCellTypesPerCluster}{seurat,grouped_df}(object,
  cellTypesPerCluster, color = scale_color_viridis(), dark = TRUE,
  headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{cellTypesPerCluster}{\code{grouped_df}, grouped by \code{cellType}. This must
be the return from \code{\link[=cellTypesPerCluster]{cellTypesPerCluster()}}.}

\item{...}{Additional arguments.}

\item{color}{Color palette.}

\item{dark}{Plot the TSNE against a dark background.}

\item{headerLevel}{Include a Markdown header for each gene.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} plotlist.
}
\description{
Plot the geometric mean of the significant marker genes for every known cell
type (per unbiased cluster). Cell types with too few (\code{min} cutoff) or too
many (\code{max} cutoff) marker genes will be skipped.
}
\examples{
# seurat ====
cellTypesPerCluster <- cellTypesPerCluster(known_markers_small)
plotCellTypesPerCluster(
    object = seurat_small,
    cellTypesPerCluster = cellTypesPerCluster
)
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotTopMarkers}},
  \code{\link{topBarcodes}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
