% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCellTypesPerCluster.R
\docType{methods}
\name{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster,seurat,grouped_df-method}
\title{Plot Cell Types per Cluster}
\usage{
plotCellTypesPerCluster(object, cellTypesPerCluster, ...)

\S4method{plotCellTypesPerCluster}{seurat,grouped_df}(object,
  cellTypesPerCluster, color = viridis::scale_color_viridis(), dark = TRUE,
  headerLevel = NULL)
}
\arguments{
\item{object}{Object.}

\item{cellTypesPerCluster}{Cell types per cluster grouped \link{tibble}. This must
be the return from \code{\link[=cellTypesPerCluster]{cellTypesPerCluster()}}.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{color}{Color palette to use for points. Defaults to the viridis color
palette. Use \code{\link[ggplot2:scale_color_gradient]{ggplot2::scale_color_gradient()}} to easily define your own
low/high gradient.}

\item{dark}{Enable dark mode.}

\item{headerLevel}{Include a Markdown header for each gene.}
}
\value{
Show graphical output. Invisibly return \link{ggplot} plotlist.
}
\description{
Plot the geometric mean of the significant marker genes for every known cell
type (per unbiased cluster). Cell types with too few (\code{min} cutoff) or too
many (\code{max} cutoff) marker genes will be skipped.
}
\examples{
load(system.file(
    file.path("extdata", "knownMarkersDetected.rda"),
    package = "bcbioSingleCell"))
load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

cellTypesPerCluster <- cellTypesPerCluster(knownMarkersDetected)
glimpse(cellTypesPerCluster)

# seurat
# Let's plot the first row, as an example
cellTypesPerCluster <- cellTypesPerCluster[1, , drop = FALSE]
plotCellTypesPerCluster(
    seurat,
    cellTypesPerCluster = cellTypesPerCluster)
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotTopMarkers}},
  \code{\link{topBarcodes}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
