% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCellTypesPerCluster.R
\docType{methods}
\name{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster,seurat,grouped_df-method}
\title{Plot Cell Types per Cluster}
\usage{
plotCellTypesPerCluster(object, cellTypesPerCluster, ...)

\S4method{plotCellTypesPerCluster}{seurat,grouped_df}(object,
  cellTypesPerCluster, color = scale_color_viridis(), dark = TRUE,
  headerLevel = NULL)
}
\arguments{
\item{object}{Object.}

\item{cellTypesPerCluster}{Cell types per cluster \code{grouped_df}. This must be the
return from \code{\link[=cellTypesPerCluster]{cellTypesPerCluster()}}.}

\item{...}{Additional arguments.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{dark}{Plot against a dark background using \code{\link[=midnightTheme]{midnightTheme()}}.}

\item{headerLevel}{R Markdown header level.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} plotlist.
}
\description{
Plot the geometric mean of the significant marker genes for every known cell
type (per unbiased cluster). Cell types with too few (\code{min} cutoff) or too
many (\code{max} cutoff) marker genes will be skipped.
}
\examples{
load(system.file("extdata/knownMarkersDetected.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat.rda", package = "bcbioSingleCell"))

cellTypesPerCluster <- cellTypesPerCluster(knownMarkersDetected)
glimpse(cellTypesPerCluster)

# seurat
# Let's plot the first row, as an example
cellTypesPerCluster <- cellTypesPerCluster[1L, , drop = FALSE]
plotCellTypesPerCluster(
    seurat,
    cellTypesPerCluster = cellTypesPerCluster
)
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotTopMarkers}},
  \code{\link{topBarcodes}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
