% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCellTypesPerCluster.R
\docType{methods}
\name{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster,seurat-method}
\title{Plot Cell Types per Cluster}
\usage{
plotCellTypesPerCluster(object, ...)

\S4method{plotCellTypesPerCluster}{seurat}(object, cellTypesPerCluster,
  color = "auto", dark = FALSE, grid = TRUE, headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{cellTypesPerCluster}{Cell types per cluster \code{grouped_df}. This must be
the return from \code{\link[=cellTypesPerCluster]{cellTypesPerCluster()}}.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{dark}{Plot against a dark background using
\code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{grid}{Show major grid lines but hide axis lines.}

\item{headerLevel}{R Markdown header level.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Plot the geometric mean of the significant marker genes for every known cell
type (per unbiased cluster). Cell types with too few (\code{min} cutoff) or too
many (\code{max} cutoff) marker genes will be skipped.
}
\examples{
# seurat ====
per_cluster <- cellTypesPerCluster(known_markers_small)
glimpse(per_cluster)

# Let's plot the first row, as an example
plotCellTypesPerCluster(
    object = seurat_small,
    cellTypesPerCluster = per_cluster[1L, , drop = FALSE]
)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotMarker}},
  \code{\link{plotPCElbow}}, \code{\link{plotTSNE}},
  \code{\link{sanitizeMarkers}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
