% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotCellTypesPerCluster-methods.R
\docType{methods}
\name{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster,SingleCellExperiment-method}
\title{Plot Cell Types per Cluster}
\usage{
plotCellTypesPerCluster(object, ...)

\S4method{plotCellTypesPerCluster}{SingleCellExperiment}(object,
  cellTypesPerCluster, reduction = c("TSNE", "UMAP"), headerLevel = 2L,
  ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[=plotMarkerTSNE]{plotMarkerTSNE()}} or \code{\link[=plotMarkerUMAP]{plotMarkerUMAP()}}.}

\item{cellTypesPerCluster}{\code{grouped_df}. Cell types per cluster data. This
must be the return from \code{\link[=cellTypesPerCluster]{cellTypesPerCluster()}}.}

\item{reduction}{\code{string}. Dimensional reduction method to apply. Defaults to
t-SNE ("\code{TSNE}") but UMAP is also supported ("\code{UMAP}").}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). R Markdown header level.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Plot the geometric mean of the significant marker genes for every known cell
type (per unbiased cluster). Cell types with too few (\code{min} cutoff) or too
many (\code{max} cutoff) marker genes will be skipped.
}
\examples{
# SingleCellExperiment ====
per_cluster <- cellTypesPerCluster(known_markers_small)
glimpse(per_cluster)

# Let's plot the first row, as an example
plotCellTypesPerCluster(
    object = cellranger_small,
    cellTypesPerCluster = head(per_cluster, n = 2L)
)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotTSNE}},
  \code{\link{sanitizeSeuratMarkers}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Functions}
