% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotCellTypesPerCluster-methods.R
\docType{methods}
\name{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster,seurat-method}
\title{Plot Cell Types per Cluster}
\usage{
plotCellTypesPerCluster(object, ...)

\S4method{plotCellTypesPerCluster}{seurat}(object, cellTypesPerCluster,
  reduction = c("TSNE", "UMAP"), headerLevel = 2L, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[=plotMarkerTSNE]{plotMarkerTSNE()}} or \code{\link[=plotMarkerUMAP]{plotMarkerUMAP()}}.}

\item{cellTypesPerCluster}{Cell types per cluster \code{grouped_df}. This must be
the return from \code{\link[=cellTypesPerCluster]{cellTypesPerCluster()}}.}

\item{reduction}{Dimensional reduction method to apply. Defaults to t-SNE
("\code{TSNE}") but UMAP is also supported ("\code{UMAP}").}

\item{headerLevel}{R Markdown header level.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Plot the geometric mean of the significant marker genes for every known cell
type (per unbiased cluster). Cell types with too few (\code{min} cutoff) or too
many (\code{max} cutoff) marker genes will be skipped.
}
\examples{
# seurat ====
per_cluster <- cellTypesPerCluster(known_markers_small)
glimpse(per_cluster)

# Let's plot the first row, as an example
plotCellTypesPerCluster(
    object = seurat_small,
    cellTypesPerCluster = head(per_cluster, 1),
)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotTSNE}},
  \code{\link{sanitizeMarkers}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
