% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotClusters.R
\docType{methods}
\name{plotClusters}
\alias{plotClusters}
\alias{plotClusters,seurat-method}
\title{Plot Clusters}
\usage{
plotClusters(object, ...)

\S4method{plotClusters}{seurat}(object, symbols, headerLevel = 4L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{symbols}{Character vector of gene symbols.}

\item{headerLevel}{Include a Markdown header for each gene.}
}
\value{
No value, only graphical output.
}
\description{
Plot Clusters
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{pcCutoff}}, \code{\link{plotKnownMarkers}},
  \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Utilities}
