% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCumulativeReadsPerCell.R
\docType{methods}
\name{plotCumulativeReadsPerCell}
\alias{plotCumulativeReadsPerCell}
\alias{plotCumulativeReadsPerCell}
\alias{plotCumulativeReadsPerCell,bcbioSingleCell-method}
\alias{plotCumulativeReadsPerCell,seurat-method}
\alias{plotKnee}
\title{Plot Cumulative Frequency of Reads Per Cell}
\usage{
plotCumulativeReadsPerCell(object, ...)

\S4method{plotCumulativeReadsPerCell}{bcbioSingleCell}(object,
  trans = c("log10", "log2", "sqrt", "identity"))

\S4method{plotCumulativeReadsPerCell}{seurat}(object, trans = c("log10",
  "log2", "sqrt", "identity"))

plotKnee(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Cumulative frequency plot that helps visualize the inflection point that is
present in single-cell barcode distributions.
}
\details{
These calculates apply to all cellular barcodes in the dataset and are not
grouped by sample.
}
\examples{
# bcbioSingleCell ====
plotCumulativeReadsPerCell(bcb_small)
plotCumulativeReadsPerCell(cellranger_small)

# seurat ====
plotCumulativeReadsPerCell(seurat_small)
plotCumulativeReadsPerCell(pbmc_small)
}
\seealso{
Other Quality Control Functions: \code{\link{inflectionPoint}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotQuantileHeatmap}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
