% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDot.R
\docType{methods}
\name{plotDot}
\alias{plotDot}
\alias{plotDot}
\alias{plotDot,seurat-method}
\title{Plot Dot}
\usage{
plotDot(object, ...)

\S4method{plotDot}{seurat}(object, genes, colors = c(low = "lightgray", high =
  "purple"), colMin = -2.5, colMax = 2.5, dotMin = 0, dotScale = 6)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{genes}{Character vector of gene symbols.}

\item{colors}{Named character vector (\code{low}, \code{high}) for plot colors.}

\item{colMin}{Minimum scaled average expression threshold. Everything
smaller will be set to this.}

\item{colMax}{Maximum scaled average expression threshold. Everything larger
will be set to this.}

\item{dotMin}{The fraction of cells at which to draw the smallest dot. All
cell groups with less than this expressing the given gene will have no dot
drawn.}

\item{dotScale}{Scale the size of the points, similar to \code{cex}.}
}
\description{
Plot Dot
}
\note{
\code{\link[Seurat:DotPlot]{Seurat::DotPlot()}} is still plotting even when \code{do.return = TRUE}.
In the meantime, we've broken out the code into this generic to fix
RMarkdown looping of our marker plots.
}
\examples{
load(system.file(
    file.path("inst", "extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

genes <- slot(seurat, "data") \%>\% rownames() \%>\% .[1:2]
print(genes)

# seurat
plotDot(seurat, genes = genes)
}
\seealso{
Modified version of \code{\link[Seurat:DotPlot]{Seurat::DotPlot()}}.
}
\author{
Michael Steinbaugh
}
