% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-seurat-plotDot.R
\docType{methods}
\name{plotDot}
\alias{plotDot}
\alias{plotDot,seurat-method}
\title{Plot Dot}
\usage{
\S4method{plotDot}{seurat}(object, genes, color = scale_color_gradient(low =
  "lightgray", high = "purple"), dark = FALSE, colMin = -2.5,
  colMax = 2.5, dotMin = 0L, dotScale = 6L)
}
\arguments{
\item{object}{Object.}

\item{genes}{Gene identifiers. Must match the rownames of the object.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{dark}{Plot against a dark background using \code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{colMin}{Minimum scaled average expression threshold. Everything
smaller will be set to this.}

\item{colMax}{Maximum scaled average expression threshold. Everything larger
will be set to this.}

\item{dotMin}{The fraction of cells at which to draw the smallest dot. All
cell groups with less than this expressing the given gene will have no dot
drawn.}

\item{dotScale}{Scale the size of the points, similar to \code{cex}.}
}
\description{
Plot Dot
}
\examples{
# seurat ====
genes <- head(rownames(seurat_small))
plotDot(seurat_small, genes = genes)
}
\seealso{
Modified version of \code{\link[Seurat:DotPlot]{Seurat::DotPlot()}}.

Other Gene Expression Functions: \code{\link{plotViolin}}
}
\author{
Michael Steinbaugh
}
