% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotGene.R
\docType{methods}
\name{plotDot}
\alias{plotDot}
\alias{plotGene,seurat-method}
\title{Plot Gene}
\usage{
\S4method{plotGene}{seurat}(object, genes, geom = c("dot", "violin"),
  legend = FALSE)
}
\arguments{
\item{geom}{Plot type. Uses \code{\link[=match.arg]{match.arg()}} to pick the type. Currently
supports "\code{dot}" and "\code{violin}".}
}
\description{
Plot Gene
}
\examples{
object <- seurat_small
genes <- head(rownames(object))

# Dot
plotGene(object, genes = genes, geom = "dot")

# Violin
plotGene(object, genes = genes, geom = "violin")
}
\seealso{
\itemize{
\item \code{\link[Seurat:DotPlot]{Seurat::DotPlot()}}.
\item \code{\link[Seurat:RidgePlot]{Seurat::RidgePlot()}}.
\item \code{\link[Seurat:ViolinPlot]{Seurat::ViolinPlot()}}.
}
}
\author{
Michael Steinbaugh
}
