% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotFeature-methods.R,
%   R/zzz-seurat-methods.R
\docType{methods}
\name{plotFeature}
\alias{plotFeature}
\alias{plotFeatureTSNE}
\alias{plotFeatureUMAP}
\alias{plotFeatureTSNE,SingleCellExperiment-method}
\alias{plotFeatureUMAP,SingleCellExperiment-method}
\alias{plotFeature,seurat-method}
\alias{plotFeatureTSNE,seurat-method}
\alias{plotFeatureUMAP,seurat-method}
\title{Plot Feature}
\usage{
plotFeature(object, ...)

plotFeatureTSNE(object, ...)

plotFeatureUMAP(object, ...)

\S4method{plotFeatureTSNE}{SingleCellExperiment}(object, ...)

\S4method{plotFeatureUMAP}{SingleCellExperiment}(object, ...)

\S4method{plotFeature}{seurat}(object, features, reducedDim = c("TSNE",
  "UMAP"), color = getOption("bcbio.discrete.color", NULL),
  pointSize = getOption("bcbio.pointSize", 0.75),
  pointAlpha = getOption("bcbio.pointAlpha", 0.75),
  label = getOption("bcbio.label", TRUE),
  labelSize = getOption("bcbio.labelSize", 6L),
  dark = getOption("bcbio.dark", FALSE), grid = getOption("bcbio.grid",
  FALSE), legend = getOption("bcbio.legend", FALSE),
  aspectRatio = getOption("bcbio.aspectRatio", 1L))

\S4method{plotFeatureTSNE}{seurat}(object, ...)

\S4method{plotFeatureUMAP}{seurat}(object, ...)

\S4method{plotFeature}{seurat}(object, features, reducedDim = c("TSNE",
  "UMAP"), color = getOption("bcbio.discrete.color", NULL),
  pointSize = getOption("bcbio.pointSize", 0.75),
  pointAlpha = getOption("bcbio.pointAlpha", 0.75),
  label = getOption("bcbio.label", TRUE),
  labelSize = getOption("bcbio.labelSize", 6L),
  dark = getOption("bcbio.dark", FALSE), grid = getOption("bcbio.grid",
  FALSE), legend = getOption("bcbio.legend", FALSE),
  aspectRatio = getOption("bcbio.aspectRatio", 1L))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{features}{\code{character}. Features to plot (e.g. gene expression, PC
scores, number of genes detected).}

\item{reducedDim}{\code{string}. Name of reduced dimension matrix slotted in
\code{\link[=reducedDims]{reducedDims()}}. Includes TNSE, UMAP, PCA, for example.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{pointSize}{\code{scalar numeric}. Cell point size.}

\item{pointAlpha}{\code{scalar numeric} (\code{0}-\code{1}). Alpha transparency level.
Useful when there many cells in the dataset, and some cells can be masked.}

\item{label}{\code{boolean}. Overlay a cluster identitiy label on the plot.}

\item{labelSize}{\code{scalar integer}. Size of the text label.}

\item{dark}{\code{boolean}. Plot against a dark background using
\code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{grid}{\code{boolean}. Show major grid lines but hide axis lines.}

\item{legend}{\code{boolean}. Include plot legend.}

\item{aspectRatio}{\code{scalar integer}. Aspect ratio.}
}
\value{
\code{ggplot} or \code{list}.
}
\description{
Plot Feature
}
\examples{
# SingleCellExperiment ====
object <- indrops_small
features <- c("nUMI", "nGene", "mitoRatio")

# t-SNE
plotFeatureTSNE(object, features)

# UMAP
plotFeatureUMAP(object, features)
}
\seealso{
\code{\link[Seurat:FeaturePlot]{Seurat::FeaturePlot()}}.

Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotReducedDim}},
  \code{\link{sanitizeSeuratMarkers}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Functions}
