% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotFeature.R
\docType{methods}
\name{plotFeatureTSNE}
\alias{plotFeatureTSNE}
\alias{plotFeatureUMAP}
\alias{plotFeature}
\alias{plotFeatureTSNE,seurat-method}
\alias{plotFeatureUMAP,seurat-method}
\title{Plot Feature}
\usage{
plotFeatureTSNE(object, ...)

plotFeatureUMAP(object, ...)

\S4method{plotFeatureTSNE}{seurat}(object, features, color = "auto",
  pointSize = 0.5, pointAlpha = 0.8, label = TRUE, labelSize = 6L,
  dark = TRUE, grid = TRUE, legend = FALSE, aspectRatio = 1L)

\S4method{plotFeatureUMAP}{seurat}(object, features, color = "auto",
  pointSize = 0.5, pointAlpha = 0.8, label = TRUE, labelSize = 6L,
  dark = TRUE, grid = TRUE, legend = FALSE, aspectRatio = 1L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{features}{Character vector of features (e.g. gene expression, PC
scores, number of genes detected).}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{pointSize}{Cell point size.}

\item{pointAlpha}{Alpha transparency level. Useful when there many cells in
the dataset, and some cells can be masked.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{dark}{Plot against a dark background using
\code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{grid}{Show major grid lines but hide axis lines.}

\item{legend}{Include plot legend.}

\item{aspectRatio}{Aspect ratio.}
}
\value{
\code{ggplot} or \code{list}.
}
\description{
Plot Feature
}
\examples{
# seurat ====
object <- seurat_small
features <- c("nUMI", "nGene", "PC1", "PC2")

# t-SNE grid
plotFeatureTSNE(
    object = object,
    features = features,
    dark = TRUE,
    grid = TRUE
)
plotFeatureTSNE(
    object = object,
    features = features,
    dark = FALSE,
    grid = FALSE
)

# UMAP grid
plotFeatureUMAP(
    object,
    features = features,
    dark = TRUE,
    grid = TRUE
)
plotFeatureUMAP(
    object,
    features = features,
    dark = FALSE,
    grid = FALSE
)
}
\seealso{
\code{\link[Seurat:FeaturePlot]{Seurat::FeaturePlot()}}.

Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotTSNE}},
  \code{\link{sanitizeMarkers}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
