% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotFeatures.R
\docType{methods}
\name{plotFeatures}
\alias{plotFeatures}
\alias{plotFeatures,seurat-method}
\title{Plot Features of a Data Set}
\usage{
plotFeatures(object, ...)

\S4method{plotFeatures}{seurat}(object, features, headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{features}{Character vector of parameters supported by
\code{\link[Seurat:FetchData]{Seurat::FetchData()}} (e.g. nUMI, mitoRatio, PC).}

\item{headerLevel}{Include a Markdown header for each gene.}

\item{returnAsList}{Return the \code{gg} objects as a list instead of plotting
with \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}.}
}
\value{
No value, only graphical output.
}
\description{
Plot Features of a Data Set
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{pcCutoff}}, \code{\link{plotKnownMarkers}},
  \code{\link{plotMarkers}}, \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Utilities}
