% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotFeatureTSNE.R
\docType{methods}
\name{plotFeatureTSNE}
\alias{plotFeatureTSNE}
\alias{plotFeatureTSNE}
\alias{plotFeatureTSNE,seurat-method}
\title{Plot Feature t-SNE}
\usage{
plotFeatureTSNE(object, ...)

\S4method{plotFeatureTSNE}{seurat}(object, features, pointSize = 0.5,
  pointAlpha = 0.8, color = scale_color_gradient(low = "lightgray", high =
  "purple"), dark = FALSE, label = TRUE, labelSize = 6L, legend = FALSE,
  return = c("plot", "list"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{features}{Character vector of features (e.g. gene expression, PC
scores, number of genes detected).}

\item{pointSize}{Cell point size.}

\item{pointAlpha}{Alpha transparency level. Useful when there many cells in
the dataset, and some cells can be masked.}

\item{color}{Color palette to use for points.}

\item{dark}{Enable dark mode.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{legend}{Show legends in paneled plots. Defaults to \code{FALSE} because
typically these look too busy and the legends can get cut off.}

\item{return}{Return as "\code{plot}" or "\code{list}".}
}
\value{
Depends on the return argument.
}
\description{
Plot Feature t-SNE
}
\examples{
# seurat ====
plotFeatureTSNE(pbmc_small, features = "PC1")
}
\seealso{
\code{\link[Seurat:FeaturePlot]{Seurat::FeaturePlot()}}.
}
\author{
Michael Steinbaugh
}
