% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotFeatureTSNE.R
\docType{methods}
\name{plotFeatureTSNE}
\alias{plotFeatureTSNE}
\alias{plotFeatureTSNE,seurat-method}
\title{Plot Feature t-SNE}
\usage{
plotFeatureTSNE(object, ...)

\S4method{plotFeatureTSNE}{seurat}(object, feature, pointSize = 1,
  color = scale_color_gradient(low = "lightgray", high = "purple"),
  dark = FALSE, label = TRUE, labelSize = 6, legend = FALSE,
  returnAsList = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{feature}{Character vector of features (e.g. gene expression, PC
scores, number of genes detected).}

\item{pointSize}{Cell point size.}

\item{color}{Color palette to use for points. Defaults to the viridis color
palette. Use \code{\link[ggplot2:scale_color_gradient]{ggplot2::scale_color_gradient()}} to easily define your own
low/high gradient.}

\item{dark}{Enable dark mode.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{returnAsList}{Return plots as a list.}
}
\value{
No return, only graphical output.
}
\description{
Plot Feature t-SNE
}
\seealso{
\code{\link[Seurat:FeaturePlot]{Seurat::FeaturePlot()}}.
}
\author{
Michael Steinbaugh
}
