% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-seurat-plotFeatureTSNE.R
\docType{methods}
\name{plotFeatureTSNE}
\alias{plotFeatureTSNE}
\alias{plotFeatureTSNE}
\alias{plotFeatureTSNE,seurat-method}
\title{Plot Feature t-SNE}
\usage{
plotFeatureTSNE(object, ...)

\S4method{plotFeatureTSNE}{seurat}(object, features, pointSize = 0.5,
  pointAlpha = 0.8, color = scale_color_gradient(low = "lightgray", high =
  "purple"), dark = FALSE, label = TRUE, labelSize = 6L, legend = FALSE,
  return = c("ggplot", "list"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{features}{Character vector of features (e.g. gene expression, PC
scores, number of genes detected).}

\item{pointSize}{Cell point size.}

\item{pointAlpha}{Alpha transparency level. Useful when there many cells in
the dataset, and some cells can be masked.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{dark}{Plot against a dark background using \code{\link[basejump:midnightTheme]{basejump::midnightTheme()}}.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{legend}{Show legends in paneled plots. Defaults to \code{FALSE} because
typically these look too busy and the legends can get cut off.}

\item{return}{Return type. Uses \code{\link[base:match.arg]{base::match.arg()}} internally and defaults
to the first argument in the vector.}
}
\value{
\code{ggplot} or \code{list}.
}
\description{
Plot Feature t-SNE
}
\examples{
# seurat ====
plotFeatureTSNE(pbmc_small, features = "PC1")
}
\seealso{
\code{\link[Seurat:FeaturePlot]{Seurat::FeaturePlot()}}.

Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotMarkers}},
  \code{\link{plotPCA}}, \code{\link{plotPCElbow}},
  \code{\link{plotTSNE}}, \code{\link{plotTopMarkers}},
  \code{\link{sanitizeMarkers}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
