% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGenesPerCell.R
\docType{methods}
\name{plotGenesPerCell}
\alias{plotGenesPerCell}
\alias{plotGenesPerCell,bcbioSingleCell-method}
\title{Plot Genes per Cell}
\usage{
plotGenesPerCell(object, ...)

\S4method{plotGenesPerCell}{bcbioSingleCell}(object, interestingGroups,
  geom = "violin", min, max, filterCells = TRUE,
  aggregateReplicates = TRUE, samplesOnYAxis = TRUE,
  fill = scale_fill_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroups}{Interesting group, to use for colors.}

\item{geom}{Plot type. Currently support formats: \code{boxplot}, \code{histogram},
\code{ridgeline}, and \code{violin} (default).}

\item{min}{Recommended minimum value cutoff.}

\item{max}{Recommended maximum value cutoff.}

\item{filterCells}{Show only the cells that have passed filtering cutoffs.}

\item{aggregateReplicates}{Aggregate technical replicates, if specified. This
function uses values assigned in the \code{sampleNameAggregate} column of the
internal sample metadata \link{data.frame}.}

\item{samplesOnYAxis}{Plot the samples on the y axis. Doesn't apply to
histogram.}

\item{fill}{Desired ggplot fill scale. Defaults to
\code{\link[viridis:scale_fill_viridis]{viridis::scale_fill_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}
}
\value{
\link{ggplot} grid.
}
\description{
Plot Genes per Cell
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Metrics}
