% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGenesPerCell.R
\docType{methods}
\name{plotGenesPerCell}
\alias{plotGenesPerCell}
\alias{.plotGenesPerCellBoxplot}
\alias{.plotGenesPerCellHistogram}
\alias{.plotGenesPerCell}
\alias{plotGenesPerCell,bcbioSCDataSet-method}
\alias{plotGenesPerCell,bcbioSCSubset-method}
\title{Plot Genes per Cell}
\usage{
plotGenesPerCell(object, ...)

\S4method{plotGenesPerCell}{bcbioSCDataSet}(object, min = 500L, max = NULL)

\S4method{plotGenesPerCell}{bcbioSCSubset}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{min}{Recommended minimum value cutoff.}

\item{max}{Recommended maximum value cutoff.}
}
\value{
\link{ggplot} grid.
}
\description{
Plot Genes per Cell
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotCellularBarcodes}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
