% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGenesPerCell.R
\docType{methods}
\name{plotGenesPerCell}
\alias{plotGenesPerCell}
\alias{plotGenesPerCell}
\alias{plotGenesPerCell,SingleCellExperiment-method}
\alias{plotGenesPerCell,seurat-method}
\title{Plot Genes per Cell}
\usage{
plotGenesPerCell(object, ...)

\S4method{plotGenesPerCell}{SingleCellExperiment}(object, geom = c("violin",
  "boxplot", "histogram", "ridgeline"), interestingGroups, min = 0L,
  max = Inf, fill = scale_fill_hue(), title = "genes per cell")

\S4method{plotGenesPerCell}{seurat}(object, geom = c("violin", "boxplot",
  "histogram", "ridgeline"), interestingGroups, min = 0L, max = Inf,
  fill = scale_fill_hue(), title = "genes per cell")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{geom}{Plot type. Uses \code{\link[=match.arg]{match.arg()}} internally and defaults to the
first argument in the vector.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in camel case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{min}{Recommended minimum value cutoff.}

\item{max}{Recommended maximum value cutoff.}

\item{fill}{Desired ggplot fill scale. Must supply discrete values. When set
to \code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{title}{Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot Genes per Cell
}
\examples{
# bcbioSingleCell ====
plotGenesPerCell(indrops_small)

# SingleCellExperiment ====
plotGenesPerCell(cellranger_small)

# seurat ====
plotGenesPerCell(Seurat::pbmc_small)
}
\seealso{
Other Quality Control Functions: \code{\link{barcodeRanksPerSample}},
  \code{\link{plotCellCounts}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
