% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGenesPerCell.R
\docType{methods}
\name{plotGenesPerCell}
\alias{plotGenesPerCell}
\alias{plotGenesPerCell,bcbioSingleCell-method}
\title{Plot Genes per Cell}
\usage{
plotGenesPerCell(object, ...)

\S4method{plotGenesPerCell}{bcbioSingleCell}(object, interestingGroups, min,
  max, filterCells = TRUE, aggregateReplicates = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroups}{Interesting group, to use for colors.}

\item{min}{Recommended minimum value cutoff.}

\item{max}{Recommended maximum value cutoff.}

\item{filterCells}{Show only the cells that have passed filtering cutoffs.}

\item{aggregateReplicates}{Aggregate technical replicates, if specified. This
function uses values assigned in the \code{sampleNameAggregate} column of the
internal sample metadata \link{data.frame}.}
}
\value{
\link{ggplot} grid.
}
\description{
Plot Genes per Cell
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Metrics}
