% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotKnownMarkers.R
\name{plotKnownMarkers}
\alias{plotKnownMarkers}
\alias{plotKnownMarkers}
\title{Plot Known Markers}
\usage{
plotKnownMarkers(object, knownMarkers, ...)
}
\arguments{
\item{object}{Object.}

\item{knownMarkers}{\code{\link[=knownMarkersDetected]{knownMarkersDetected()}} \link{tibble} grouped by cluster.}

\item{...}{Additional arguments.}

\item{markdown}{Print Markdown headers.}
}
\value{
\code{\link[=writeLines]{writeLines()}}.
}
\description{
Plot Known Markers
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotClusters}},
  \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
