% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotKnownMarkers-methods.R
\docType{methods}
\name{plotKnownMarkers}
\alias{plotKnownMarkers}
\alias{plotKnownMarkers,seurat,grouped_df-method}
\title{Plot Known Markers}
\usage{
plotKnownMarkers(x, y, ...)

\S4method{plotKnownMarkers}{seurat,grouped_df}(x, y, markdown = TRUE)
}
\arguments{
\item{x}{Primary object.}

\item{y}{\code{\link[=knownMarkersDetected]{knownMarkersDetected()}} \link{tibble} grouped by cluster.}

\item{...}{Additional arguments.}

\item{markdown}{Print Markdown headers.}
}
\value{
\code{\link[=writeLines]{writeLines()}}.
}
\description{
Plot Known Markers
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotClusters}},
  \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
