% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotKnownMarkers.R
\docType{methods}
\name{plotKnownMarkers}
\alias{plotKnownMarkers}
\alias{plotKnownMarkers}
\alias{plotKnownMarkers,seurat,grouped_df-method}
\title{Plot Known Markers}
\usage{
plotKnownMarkers(object, knownMarkers, ...)

\S4method{plotKnownMarkers}{seurat,grouped_df}(object, knownMarkers,
  headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{knownMarkers}{\code{\link[=knownMarkersDetected]{knownMarkersDetected()}} \link{tibble} grouped by cluster.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{headerLevel}{Include a Markdown header for each gene.}
}
\value{
No value, only graphical output.
}
\description{
Plot Known Markers
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{pcCutoff}}, \code{\link{plotClusters}},
  \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
