% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotKnownMarkersDetected.R
\docType{methods}
\name{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected,seurat,grouped_df-method}
\title{Plot Known Markers Detected}
\usage{
plotKnownMarkersDetected(object, knownMarkersDetected, ...)

\S4method{plotKnownMarkersDetected}{seurat,grouped_df}(object,
  knownMarkersDetected, tsneColor = scale_color_viridis(discrete = FALSE),
  violinFill = scale_fill_viridis(discrete = TRUE),
  dotColor = scale_color_gradient(low = "lightgray", high = "purple"),
  dark = TRUE, pointsAsNumbers = FALSE, headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{knownMarkersDetected}{\code{grouped_df}.}

\item{...}{Additional arguments.}

\item{tsneColor}{Color palette to use for tSNE plot.}

\item{violinFill}{Color palette to use for violin plot.}

\item{dotColor}{Color palette to use for dot plot.}

\item{dark}{Plot against a dark background using \code{\link[=midnightTheme]{midnightTheme()}}.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{headerLevel}{R Markdown header level.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} plotlist.
}
\description{
Plot Known Markers Detected
}
\examples{
# seurat ====
# Let's plot the first 2 markers, as a quick example
plotKnownMarkersDetected(
    object = seurat_small,
    knownMarkersDetected = known_markers_small[seq_len(2L), ]
)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotMarkers}},
  \code{\link{plotPCA}}, \code{\link{plotPCElbow}},
  \code{\link{plotTSNE}}, \code{\link{plotTopMarkers}},
  \code{\link{sanitizeMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
