% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotKnownMarkersDetected.R
\docType{methods}
\name{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected,seurat,grouped_df-method}
\title{Plot Known Markers Detected}
\usage{
plotKnownMarkersDetected(object, knownMarkersDetected, ...)

\S4method{plotKnownMarkersDetected}{seurat,grouped_df}(object,
  knownMarkersDetected, tsneColor = viridis::scale_color_viridis(),
  violinFill = viridis::scale_fill_viridis(discrete = TRUE),
  dotColor = ggplot2::scale_color_gradient(low = "lightgray", high =
  "purple"), dark = TRUE, pointsAsNumbers = FALSE, headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{knownMarkersDetected}{\code{\link[=knownMarkersDetected]{knownMarkersDetected()}} return \link{tibble} grouped
by cluster.}

\item{...}{Additional arguments.}

\item{tsneColor}{Color palette to use for tSNE plot.}

\item{violinFill}{Color palette to use for violin plot.}

\item{dotColor}{Color palette to use for dot plot.}

\item{dark}{Plot the TSNE against a dark background.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{headerLevel}{Include a Markdown header for each gene.}
}
\value{
No value, only graphical output.
}
\description{
Plot Known Markers Detected
}
\examples{
load(system.file(
    file.path("extdata", "knownMarkersDetected.rda"),
    package = "bcbioSingleCell"))
load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

# seurat
# Let's plot the first 2 markers, as a quick example
plotKnownMarkersDetected(
    seurat,
    knownMarkersDetected = knownMarkersDetected[1:2, ])
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotMarkers}}, \code{\link{plotPCElbow}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
