% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotKnownMarkersDetected.R
\docType{methods}
\name{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected,seurat,grouped_df-method}
\title{Plot Known Markers Detected}
\usage{
plotKnownMarkersDetected(object, knownMarkersDetected, ...)

\S4method{plotKnownMarkersDetected}{seurat,grouped_df}(object,
  knownMarkersDetected, color = viridis::scale_color_viridis(), dark = TRUE,
  headerLevel = NULL)
}
\arguments{
\item{object}{Object.}

\item{knownMarkersDetected}{\code{\link[=knownMarkersDetected]{knownMarkersDetected()}} return \link{tibble} grouped
by cluster.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{color}{Color palette to use for points. Defaults to the viridis color
palette. Use \code{\link[ggplot2:scale_color_gradient]{ggplot2::scale_color_gradient()}} to easily define your own
low/high gradient.}

\item{dark}{Enable dark mode.}

\item{headerLevel}{Include a Markdown header for each gene.}
}
\value{
No value, only graphical output.
}
\description{
Plot Known Markers Detected
}
\examples{
load(system.file(
    file.path("extdata", "knownMarkersDetected.rda"),
    package = "bcbioSingleCell"))
load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

# seurat
# Let's plot the first 2 markers, as a quick example
plotKnownMarkersDetected(
    seurat,
    knownMarkersDetected = knownMarkersDetected[1:2, ])
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotMarkers}}, \code{\link{plotPCElbow}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
