% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deprecated.R,
%   R/plotMarker-methods.R
\docType{methods}
\name{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected}
\alias{plotMarkerTSNE}
\alias{plotMarkerUMAP}
\alias{plotTopMarkers}
\alias{plotMarker}
\alias{plotMarkerTSNE,SingleCellExperiment-method}
\alias{plotMarkerTSNE,seurat-method}
\alias{plotMarkerUMAP,SingleCellExperiment-method}
\alias{plotMarkerUMAP,seurat-method}
\alias{plotTopMarkers,SingleCellExperiment-method}
\alias{plotTopMarkers,seurat-method}
\alias{plotKnownMarkersDetected,SingleCellExperiment-method}
\alias{plotKnownMarkersDetected,seurat-method}
\title{Plot Cell-Type-Specific Gene Markers}
\usage{
plotKnownMarkersDetected(object, ...)

plotMarkerTSNE(object, ...)

plotMarkerUMAP(object, ...)

plotTopMarkers(object, ...)

\S4method{plotMarkerTSNE}{SingleCellExperiment}(object, genes,
  expression = c("mean", "median", "sum"),
  color = getOption("bcbio.discrete.color", NULL),
  pointSize = getOption("bcbio.pointSize", 0.75),
  pointAlpha = getOption("bcbio.pointAlpha", 0.8),
  pointsAsNumbers = FALSE, label = getOption("bcbio.label", TRUE),
  labelSize = getOption("bcbio.labelSize", 6L),
  dark = getOption("bcbio.dark", FALSE), grid = getOption("bcbio.grid",
  FALSE), legend = getOption("bcbio.legend", TRUE),
  aspectRatio = getOption("bcbio.aspectRatio", 1L), title = TRUE)

\S4method{plotMarkerTSNE}{seurat}(object, genes, expression = c("mean",
  "median", "sum"), color = getOption("bcbio.discrete.color", NULL),
  pointSize = getOption("bcbio.pointSize", 0.75),
  pointAlpha = getOption("bcbio.pointAlpha", 0.8),
  pointsAsNumbers = FALSE, label = getOption("bcbio.label", TRUE),
  labelSize = getOption("bcbio.labelSize", 6L),
  dark = getOption("bcbio.dark", FALSE), grid = getOption("bcbio.grid",
  FALSE), legend = getOption("bcbio.legend", TRUE),
  aspectRatio = getOption("bcbio.aspectRatio", 1L), title = TRUE)

\S4method{plotMarkerUMAP}{SingleCellExperiment}(object, genes,
  expression = c("mean", "median", "sum"),
  color = getOption("bcbio.discrete.color", NULL),
  pointSize = getOption("bcbio.pointSize", 0.75),
  pointAlpha = getOption("bcbio.pointAlpha", 0.8),
  pointsAsNumbers = FALSE, label = getOption("bcbio.label", TRUE),
  labelSize = getOption("bcbio.labelSize", 6L),
  dark = getOption("bcbio.dark", FALSE), grid = getOption("bcbio.grid",
  FALSE), legend = getOption("bcbio.legend", TRUE),
  aspectRatio = getOption("bcbio.aspectRatio", 1L), title = TRUE)

\S4method{plotMarkerUMAP}{seurat}(object, genes, expression = c("mean",
  "median", "sum"), color = getOption("bcbio.discrete.color", NULL),
  pointSize = getOption("bcbio.pointSize", 0.75),
  pointAlpha = getOption("bcbio.pointAlpha", 0.8),
  pointsAsNumbers = FALSE, label = getOption("bcbio.label", TRUE),
  labelSize = getOption("bcbio.labelSize", 6L),
  dark = getOption("bcbio.dark", FALSE), grid = getOption("bcbio.grid",
  FALSE), legend = getOption("bcbio.legend", TRUE),
  aspectRatio = getOption("bcbio.aspectRatio", 1L), title = TRUE)

\S4method{plotTopMarkers}{SingleCellExperiment}(object, markers, n = 10L,
  direction = c("positive", "negative", "both"), coding = FALSE,
  reduction = c("TSNE", "UMAP"), headerLevel = 2L, ...)

\S4method{plotTopMarkers}{seurat}(object, markers, n = 10L,
  direction = c("positive", "negative", "both"), coding = FALSE,
  reduction = c("TSNE", "UMAP"), headerLevel = 2L, ...)

\S4method{plotKnownMarkersDetected}{SingleCellExperiment}(object, markers,
  reduction = c("TSNE", "UMAP"), headerLevel = 2L, ...)

\S4method{plotKnownMarkersDetected}{seurat}(object, markers,
  reduction = c("TSNE", "UMAP"), headerLevel = 2L, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{\code{character}. Gene identifiers. Must match the rownames of the
object.}

\item{expression}{\code{string}. Calculation to apply. Uses \code{\link[=match.arg]{match.arg()}} and
defaults to the first argument in the \code{character} vector.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{pointSize}{\code{scalar numeric}. Cell point size.}

\item{pointAlpha}{\code{scalar numeric} (\code{0}-\code{1}). Alpha transparency level.
Useful when there many cells in the dataset, and some cells can be masked.}

\item{pointsAsNumbers}{\code{boolean}. Plot the points as numbers (\code{TRUE}) or
dots (\code{FALSE}).}

\item{label}{\code{boolean}. Overlay a cluster identitiy label on the plot.}

\item{labelSize}{\code{scalar integer}. Size of the text label.}

\item{dark}{\code{boolean}. Plot against a dark background using
\code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{grid}{\code{boolean}. Show major grid lines but hide axis lines.}

\item{legend}{\code{boolean}. Include plot legend.}

\item{aspectRatio}{\code{scalar integer}. Aspect ratio.}

\item{title}{\code{string} or \code{NULL}. Plot title.}

\item{markers}{\code{grouped_df}. Marker genes data.
\itemize{
\item \code{\link[=plotTopMarkers]{plotTopMarkers()}}: must be grouped by "\code{cluster}".
\item \code{\link[=plotKnownMarkersDetected]{plotKnownMarkersDetected()}}: must be grouped by "\code{cellType}".
}}

\item{reduction}{\code{string}. Dimensional reduction method to apply. Defaults to
t-SNE ("\code{TSNE}") but UMAP is also supported ("\code{UMAP}").}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). R Markdown header level.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Plot gene expression per cell in multiple formats:
\enumerate{
\item \code{\link[=plotMarkerTSNE]{plotMarkerTSNE()}}: t-SNE gene expression plot.
\item \code{\link[=plotDot]{plotDot()}}: Dot plot.
\item \code{\link[=plotViolin]{plotViolin()}}: Violin plot.
}
}
\section{Plot top markers}{

The number of markers to plot is determined by the output of the
\code{\link[=topMarkers]{topMarkers()}} function. If you want to reduce the number of genes to plot,
simply reassign first using that function. If necessary, we can add support
for the number of genes to plot here in a future update.
}

\examples{
# SingleCellExperiment ====
object <- cellranger_small

# MHC class II genes
title <- "MHC class II"
genes <- rownames(object)[which(grepl(
    pattern = "^major histocompatibility complex, class II",
    x = rowData(object)$description
))]
print(genes)

# t-SNE
plotMarkerTSNE(object, genes)
plotMarkerTSNE(
    object = object,
    genes = genes,
    expression = "sum",
    pointsAsNumbers = TRUE,
    dark = TRUE,
    label = FALSE,
    title = title
)

# UMAP
plotMarkerUMAP(object, genes)
plotMarkerTSNE(
    object = object,
    genes = genes,
    expression = "mean",
    pointsAsNumbers = TRUE,
    dark = TRUE,
    label = FALSE,
    title = title
)

# seurat ====
object <- seurat_small

# Top markers
markers <- topMarkers(all_markers_small, n = 1)
glimpse(markers)
plotTopMarkers(object, markers = tail(markers, 1))

# Known markers detected
markers <- head(known_markers_small, n = 1)
glimpse(markers)
plotKnownMarkersDetected(object, markers = head(markers, 1))
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}}, \code{\link{plotPCElbow}},
  \code{\link{plotTSNE}},
  \code{\link{sanitizeSeuratMarkers}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Clustering Functions}
