% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deprecated.R,
%   R/methods-plotMarker.R
\docType{methods}
\name{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected}
\alias{plotMarkerTSNE}
\alias{plotMarkerUMAP}
\alias{plotTopMarkers}
\alias{plotMarker}
\alias{plotMarker}
\alias{plotMarkerTSNE,SingleCellExperiment-method}
\alias{plotMarkerTSNE,seurat-method}
\alias{plotMarkerUMAP,SingleCellExperiment-method}
\alias{plotMarkerUMAP,seurat-method}
\alias{plotTopMarkers,SingleCellExperiment-method}
\alias{plotTopMarkers,seurat-method}
\alias{plotKnownMarkersDetected,SingleCellExperiment-method}
\alias{plotKnownMarkersDetected,seurat-method}
\title{Plot Cell-Type-Specific Gene Markers}
\usage{
plotKnownMarkersDetected(object, ...)

plotMarkerTSNE(object, ...)

plotMarkerUMAP(object, ...)

plotTopMarkers(object, ...)

\S4method{plotMarkerTSNE}{SingleCellExperiment}(object, genes,
  expression = c("mean", "median", "sum"), color = NULL,
  pointsAsNumbers = FALSE, pointSize = 0.75, pointAlpha = 0.8,
  label = TRUE, labelSize = 6L, dark = FALSE, grid = FALSE,
  legend = TRUE, aspectRatio = 1L, title = TRUE)

\S4method{plotMarkerTSNE}{seurat}(object, genes, expression = c("mean",
  "median", "sum"), color = NULL, pointsAsNumbers = FALSE,
  pointSize = 0.75, pointAlpha = 0.8, label = TRUE, labelSize = 6L,
  dark = FALSE, grid = FALSE, legend = TRUE, aspectRatio = 1L,
  title = TRUE)

\S4method{plotMarkerUMAP}{SingleCellExperiment}(object, genes,
  expression = c("mean", "median", "sum"), color = NULL,
  pointsAsNumbers = FALSE, pointSize = 0.75, pointAlpha = 0.8,
  label = TRUE, labelSize = 6L, dark = FALSE, grid = FALSE,
  legend = TRUE, aspectRatio = 1L, title = TRUE)

\S4method{plotMarkerUMAP}{seurat}(object, genes, expression = c("mean",
  "median", "sum"), color = NULL, pointsAsNumbers = FALSE,
  pointSize = 0.75, pointAlpha = 0.8, label = TRUE, labelSize = 6L,
  dark = FALSE, grid = FALSE, legend = TRUE, aspectRatio = 1L,
  title = TRUE)

\S4method{plotTopMarkers}{SingleCellExperiment}(object, markers,
  reduction = c("TSNE", "UMAP"), headerLevel = 2L, ...)

\S4method{plotTopMarkers}{seurat}(object, markers, reduction = c("TSNE",
  "UMAP"), headerLevel = 2L, ...)

\S4method{plotKnownMarkersDetected}{SingleCellExperiment}(object, markers,
  reduction = c("TSNE", "UMAP"), headerLevel = 2L, ...)

\S4method{plotKnownMarkersDetected}{seurat}(object, markers,
  reduction = c("TSNE", "UMAP"), headerLevel = 2L, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{Gene identifiers. Must match the rownames of the object.}

\item{expression}{Calculation to apply. Uses \code{\link[=match.arg]{match.arg()}} and defaults to
the first argument in the vector.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{pointSize}{Cell point size.}

\item{pointAlpha}{Alpha transparency level. Useful when there many cells in
the dataset, and some cells can be masked.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{dark}{Plot against a dark background using
\code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{grid}{Show major grid lines but hide axis lines.}

\item{legend}{Include plot legend.}

\item{aspectRatio}{Aspect ratio.}

\item{title}{Plot title.}

\item{markers}{\code{grouped_df} of marker genes.
\itemize{
\item \code{\link[=plotTopMarkers]{plotTopMarkers()}}: must be grouped by "\code{cluster}".
\item \code{\link[=plotKnownMarkersDetected]{plotKnownMarkersDetected()}}: must be grouped by "\code{cellType}".
}}

\item{headerLevel}{R Markdown header level.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Plot gene expression per cell in multiple formats:
\enumerate{
\item \code{\link[=plotMarkerTSNE]{plotMarkerTSNE()}}: t-SNE gene expression plot.
\item \code{\link[=plotDot]{plotDot()}}: Dot plot.
\item \code{\link[=plotViolin]{plotViolin()}}: Violin plot.
}
}
\section{Plot top markers}{

The number of markers to plot is determined by the output of the
\code{\link[=topMarkers]{topMarkers()}} function. If you want to reduce the number of genes to plot,
simply reassign first using that function. If necessary, we can add support
for the number of genes to plot here in a future update.
}

\examples{
object <- seurat_small
title <- "mito genes"
genes <- grep("^MT-", rownames(object), value = TRUE)
print(genes)

# t-SNE
plotMarkerTSNE(
    object = object,
    genes = genes,
    title = title
)

# Dark mode
plotMarkerTSNE(
    object = object,
    genes = genes,
    dark = TRUE,
    title = title
)

# Number cloud
plotMarkerTSNE(
    object = object,
    genes = genes,
    pointsAsNumbers = TRUE,
    title = title
)

# UMAP
plotMarkerUMAP(
    object = object,
    genes = genes,
    title = title
)

# Top markers
markers <- topMarkers(all_markers_small, n = 1)
markers
plotTopMarkers(object, markers = tail(markers, 1))

# Known markers detected
markers <- head(known_markers_small, n = 1)
markers
plotKnownMarkersDetected(object, markers = head(markers, 1))
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}}, \code{\link{plotPCElbow}},
  \code{\link{plotTSNE}}, \code{\link{sanitizeMarkers}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
