% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMarker.R
\docType{methods}
\name{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected}
\alias{plotMarkers}
\alias{plotMarkerTSNE}
\alias{plotTopMarkers}
\alias{plotMarker}
\alias{plotMarkerTSNE,seurat-method}
\alias{plotMarkers,seurat-method}
\alias{plotTopMarkers,seurat-method}
\alias{plotKnownMarkersDetected,seurat-method}
\title{Plot Cell-Type Gene Marker(s)}
\usage{
plotKnownMarkersDetected(object, ...)

plotMarkers(object, ...)

plotMarkerTSNE(object, ...)

plotTopMarkers(object, ...)

\S4method{plotMarkerTSNE}{seurat}(object, genes, expression = c("mean",
  "median", "sum"), color = "auto", pointsAsNumbers = FALSE,
  pointSize = 0.5, pointAlpha = 0.8, label = TRUE, labelSize = 6L,
  dark = TRUE, grid = TRUE, legend = FALSE, aspectRatio = 1L,
  title = TRUE)

\S4method{plotMarkers}{seurat}(object, genes, dark = TRUE, headerLevel = 2L,
  ...)

\S4method{plotTopMarkers}{seurat}(object, markers, headerLevel = 2L, ...)

\S4method{plotKnownMarkersDetected}{seurat}(object, markers, headerLevel = 2L,
  ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{Gene identifiers. Must match the rownames of the object.}

\item{expression}{Calculation to apply on the aggregate marker expression.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{pointSize}{Cell point size.}

\item{pointAlpha}{Alpha transparency level. Useful when there many cells in
the dataset, and some cells can be masked.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{dark}{Plot against a dark background using
\code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{grid}{Show major grid lines but hide axis lines.}

\item{legend}{Include plot legend.}

\item{aspectRatio}{Aspect ratio.}

\item{title}{Plot title.}

\item{headerLevel}{R Markdown header level.}

\item{markers}{\code{grouped_df} of marker genes.
\itemize{
\item \code{\link[=plotTopMarkers]{plotTopMarkers()}}: must be grouped by "\code{cluster}".
\item \code{\link[=plotKnownMarkersDetected]{plotKnownMarkersDetected()}}: must be grouped by "\code{cellType}".
}}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Plot gene expression per cell in multiple formats:
\enumerate{
\item \code{\link[=plotMarkerTSNE]{plotMarkerTSNE()}}: t-SNE gene expression plot.
\item \code{\link[=plotDot]{plotDot()}}: Dot plot.
\item \code{\link[=plotViolin]{plotViolin()}}: Violin plot.
}
}
\section{plotTopMarkers}{

The number of markers to plot is determined by the output of the
\code{\link[=topMarkers]{topMarkers()}} function. If you want to reduce the number of genes to plot,
simply reassign first using that function. If necessary, we can add support
for the number of genes to plot here in a future update.
}

\examples{
object <- seurat_small
genes <- "COL1A1"

# plotMarkerTSNE ====
plotMarkerTSNE(object, genes = genes)
plotMarkerTSNE(object, genes = genes, dark = FALSE, grid = FALSE)

# Mitochondrial genes
mito <- grep("^MT-", rownames(object), value = TRUE)
print(mito)
plotMarkerTSNE(object, genes = mito, title = "mito")

# plotMarkers ====
plotMarkers(object, genes = genes)

# plotTopMarkers ====
markers <- topMarkers(all_markers_small, n = 1)
markers
markers <- head(markers, n = 1)
plotTopMarkers(object, markers = markers)

# plotKnownMarkersDetected ====
markers <- head(known_markers_small, n = 1)
markers
plotKnownMarkersDetected(object, markers = markers)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}}, \code{\link{plotPCElbow}},
  \code{\link{plotTSNE}}, \code{\link{sanitizeMarkers}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
