% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-seurat-plotMarker.R
\docType{methods}
\name{plotMarker}
\alias{plotMarker}
\alias{plotMarkers}
\alias{plotMarker}
\alias{plotMarker,seurat-method}
\alias{plotMarkers,seurat-method}
\title{Plot Cell-Type Gene Marker(s)}
\usage{
plotMarker(object, ...)

plotMarkers(object, ...)

\S4method{plotMarker}{seurat}(object, gene, dark = TRUE, return = c("grid",
  "list"), ...)

\S4method{plotMarkers}{seurat}(object, genes, headerLevel = 2L, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{gene}{Gene identifier. Must intersect with \code{\link[=rownames]{rownames()}}.}

\item{dark}{Plot against a dark background using
\code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{return}{Return type. Uses \code{\link[base:match.arg]{base::match.arg()}} internally and defaults
to the first argument in the vector.}

\item{genes}{Gene identifiers. Must match the rownames of the object.}

\item{headerLevel}{R Markdown header level.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Plot gene expression per cell in multiple formats:
\enumerate{
\item t-SNE plot
\item Violin plot
\item Ridgeline plot
\item Dot plot
}
}
\examples{
# seurat ===
top <- topMarkers(all_markers_small, n = 1L)
genes <- pull(top, "rowname")

# Individual gene
gene <- genes[[1L]]
plotMarker(seurat_small, gene = gene, dark = TRUE)
plotMarker(seurat_small, gene = gene, dark = FALSE)

# Multiple genes
plotMarkers(seurat_small, genes = genes)
}
\seealso{
\itemize{
\item \code{\link[=plotMarkerTSNE]{plotMarkerTSNE()}}.
\item \code{\link[=plotDot]{plotDot()}}.
\item \code{\link[=plotViolin]{plotViolin()}}.
}

Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotPCA}},
  \code{\link{plotPCElbow}}, \code{\link{plotTSNE}},
  \code{\link{plotTopMarkers}},
  \code{\link{sanitizeMarkers}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
