% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMarkers.R
\docType{methods}
\name{plotMarkers}
\alias{plotMarkers}
\alias{plotMarkers,seurat-method}
\title{Plot Cell-Type Gene Markers}
\usage{
plotMarkers(object, ...)

\S4method{plotMarkers}{seurat}(object, genes,
  color = scale_color_viridis(option = "inferno"), dark = TRUE,
  pointsAsNumbers = FALSE, headerLevel = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{genes}{Genes (by symbol name) of which to get expression data.}

\item{color}{Color palette to use for points. Defaults to the inferno
palette from the viridis package. Use \code{\link[ggplot2:scale_color_gradient]{ggplot2::scale_color_gradient()}}
to easily define your own low/high gradient.}

\item{dark}{Enable dark mode.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{headerLevel}{Include a Markdown header for each gene.}
}
\value{
No value, only graphical output.
}
\description{
Plot gene expression per cell in multiple formats:
\enumerate{
\item t-SNE plot
\item Violin plot
\item Ridgeline plot
\item Dot plot
}
}
\examples{
\dontrun{
top <- topMarkers(markers)
genes <- top$symbol[1:4]
plotMarkers(seurat, genes = genes)
}
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{pcCutoff}}, \code{\link{plotKnownMarkers}},
  \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Utilities}
