% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMarkers.R
\docType{methods}
\name{plotMarkers}
\alias{plotMarkers}
\alias{plotMarkers}
\alias{plotMarkers,seurat-method}
\title{Plot Cell-Type Gene Markers}
\usage{
plotMarkers(object, ...)

\S4method{plotMarkers}{seurat}(object, genes, format = "symbol",
  tsneColor = viridis::scale_color_viridis(),
  violinFill = viridis::scale_fill_viridis(discrete = TRUE),
  dotColor = ggplot2::scale_color_gradient(low = "lightgray", high =
  "purple"), dark = TRUE, pointsAsNumbers = FALSE, headerLevel = 2,
  title = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{genes}{Genes (by symbol name) of which to get expression data.}

\item{format}{Gene identifier format. Defaults to \code{symbol} but \code{ensgene}
is also supported, as long as Ensembl gene to symbol identifier mappings
are defined.}

\item{tsneColor}{Color palette to use for tSNE plot.}

\item{violinFill}{Color palette to use for violin plot.}

\item{dotColor}{Color palette to use for dot plot.}

\item{dark}{Plot the TSNE against a dark background.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{headerLevel}{Include a Markdown header for each gene.}

\item{title}{Plot title.}
}
\value{
No value, only graphical output.
}
\description{
Plot gene expression per cell in multiple formats:
\enumerate{
\item t-SNE plot
\item Violin plot
\item Ridgeline plot
\item Dot plot
}
}
\examples{
load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))
load(system.file(
    file.path("extdata", "topMarkers.rda"),
    package = "bcbioSingleCell"))

symbol <- topMarkers$symbol[[1]]
print(symbol)
ensgene <- topMarkers$ensgene[[1]]
print(ensgene)

plotMarkers(seurat, genes = symbol, format = "symbol")
plotMarkers(seurat, genes = ensgene, format = "ensgene")
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotTopMarkers}},
  \code{\link{topBarcodes}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
