% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMarkers.R
\docType{methods}
\name{plotMarkers}
\alias{plotMarkers}
\alias{plotMarkers}
\alias{plotMarkers,seurat-method}
\title{Plot Cell-Type Gene Markers}
\usage{
plotMarkers(object, ...)

\S4method{plotMarkers}{seurat}(object, genes,
  tsneColor = scale_color_viridis(discrete = FALSE),
  violinFill = scale_fill_viridis(discrete = TRUE),
  dotColor = scale_color_gradient(low = "lightgray", high = "purple"),
  dark = TRUE, pointsAsNumbers = FALSE, headerLevel = 2L, title = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{Gene identifiers. Must match the rownames of the object.}

\item{tsneColor}{Color palette to use for tSNE plot.}

\item{violinFill}{Color palette to use for violin plot.}

\item{dotColor}{Color palette to use for dot plot.}

\item{dark}{Plot against a dark background using \code{\link[=midnightTheme]{midnightTheme()}}.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{headerLevel}{R Markdown header level.}

\item{title}{Plot title.}
}
\value{
No value, only graphical output.
}
\description{
Plot gene expression per cell in multiple formats:
\enumerate{
\item t-SNE plot
\item Violin plot
\item Ridgeline plot
\item Dot plot
}
}
\examples{
load(system.file(
    "extdata/all_markers_small.rda",
    package = "bcbioSingleCell"
))
load(system.file("extdata/seurat_small.rda", package = "bcbioSingleCell"))

# seurat ===
top <- topMarkers(all_markers_small, n = 1L)
genes <- pull(top, "geneName")
plotMarkers(seurat_small, genes = genes)
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotTopMarkers}},
  \code{\link{topBarcodes}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
