% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMarkerTSNE.R
\docType{methods}
\name{plotMarkerTSNE}
\alias{plotMarkerTSNE}
\alias{plotMarkerTSNE,grouped_df-method}
\alias{plotMarkerTSNE,seurat-method}
\title{Plot Marker t-SNE}
\usage{
plotMarkerTSNE(object, ...)

\S4method{plotMarkerTSNE}{grouped_df}(object, colorPoints = "geomean",
  color = scale_color_viridis(), dark = TRUE, pointsAsNumbers = FALSE,
  label = TRUE)

\S4method{plotMarkerTSNE}{seurat}(object, genes, colorPoints = "geomean",
  color = scale_color_viridis(), dark = TRUE, pointsAsNumbers = FALSE,
  label = TRUE, title = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{colorPoints}{Color points by geometric mean (\code{geomean}) or expression
of individual gene (\code{expression}).}

\item{color}{Color palette to use for points. Defaults to the viridis color
palette. Use \code{\link[ggplot2:scale_color_gradient]{ggplot2::scale_color_gradient()}} to easily define your own
low/high gradient.}

\item{dark}{Enable dark mode.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{label}{Label the clusters on the plot.}

\item{genes}{Genes (by symbol name) of which to get expression data.}

\item{title}{Plot title.}
}
\value{
\link{ggplot}.
}
\description{
Plot Marker t-SNE
}
\examples{
\dontrun{
data(markers, seurat)
top <- topMarkers(markers)

# Let's take the top markers specific to cluster 0, as an example
genes <- top \%>\%
    dplyr::filter(cluster == 0) \%>\%
    dplyr::pull(symbol)

# Fetch the t-SNE expression data for the desired gene symbols
dat <- fetchTSNEExpressionData(seurat, genes = genes)
print(unique(dat$gene))

# To make t-SNE plot colored by geometric mean of topGenes
plotTSNE(dat, colorPoints = "geomean")

# To make faceted t-SNE plot of each gene (looks good at up to 6 genes)
plotTSNE(dat, colorPoints = "expression") +
    ggplot2::facet_wrap(~gene)
}
}
\seealso{
Other t-SNE Utilities: \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{plotTSNE}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{t-SNE Utilities}
