% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMarkerTSNE.R
\docType{methods}
\name{plotMarkerTSNE}
\alias{plotMarkerTSNE}
\alias{plotMarkerTSNE}
\alias{plotMarkerTSNE,seurat-method}
\title{Plot Marker t-SNE}
\usage{
plotMarkerTSNE(object, ...)

\S4method{plotMarkerTSNE}{seurat}(object, genes, expression = c("mean",
  "median", "sum"), pointsAsNumbers = FALSE, pointSize = 0.5,
  pointAlpha = 0.8, label = TRUE, labelSize = 6L,
  color = scale_color_viridis(), dark = TRUE, legend = TRUE,
  title = NULL, subtitle = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{Genes identifiers (matching the rownames in the object),
of which to get expression data.}

\item{expression}{Calculation to apply on the aggregate marker expression.
Supports \code{mean} (default), \code{median}, or \code{sum}.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{pointSize}{Cell point size.}

\item{pointAlpha}{Alpha transparency level. Useful when there many cells in
the dataset, and some cells can be masked.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{color}{Color palette to use for points.}

\item{dark}{Enable dark mode.}

\item{legend}{Show plot legend.}

\item{title}{Plot title.}

\item{subtitle}{Include gene(s) in the subtitle.}
}
\value{
\code{ggplot}.
}
\description{
Plot Marker t-SNE
}
\examples{
# seurat ====
# Mitochondrial genes
mito <- grep("^MT\\\\.", rownames(counts(seurat_small)), value = TRUE)
print(mito)
plotMarkerTSNE(seurat_small, genes = mito)
}
\seealso{
Other t-SNE Utilities: \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{plotTSNE}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
