% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMarkerTSNE.R
\docType{methods}
\name{plotMarkerTSNE}
\alias{plotMarkerTSNE}
\alias{plotMarkerTSNE,grouped_df-method}
\alias{plotMarkerTSNE,seurat-method}
\title{Plot Marker t-SNE}
\usage{
plotMarkerTSNE(object, ...)

\S4method{plotMarkerTSNE}{grouped_df}(object, colorpoints = "geomean",
  pointsAsNumbers = FALSE, label = TRUE)

\S4method{plotMarkerTSNE}{seurat}(object, genes, colorpoints = "geomean",
  pointsAsNumbers = FALSE, label = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{colorpoints}{Color points by geometric mean or expression of
individual gene.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{label}{Label the clusters on the plot.}

\item{genes}{Genes (by symbol name) of which to get expression data.}
}
\value{
\link{ggplot}.
}
\description{
Plot Marker t-SNE
}
\examples{
\dontrun{
data(markers, seurat)
top <- topMarkers(markers)

# Let's take the top markers specific to cluster 0, as an example
genes <- top \%>\%
    filter(cluster == 0) \%>\%
    pull(symbol)

# Fetch the t-SNE expression data for the desired gene symbols
dat <- fetchTSNEExpressionData(seurat, genes = genes)
print(unique(dat$gene))

# To make t-SNE plot colored by geometric mean of topGenes
plotTSNE(dat, colorpoints = "geomean")

# To make faceted t-SNE plot of each gene (looks good at up to 6 genes)
plotTSNE(dat, colorpoints = "expression") +
    facet_wrap(~gene)
}
}
\seealso{
Other t-SNE Utilities: \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{plotTSNE}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{t-SNE Utilities}
