% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-seurat-plotMarkerTSNE.R
\docType{methods}
\name{plotMarkerTSNE}
\alias{plotMarkerTSNE}
\alias{plotMarkerTSNE}
\alias{plotMarkerTSNE,seurat-method}
\title{Plot Marker t-SNE}
\usage{
plotMarkerTSNE(object, ...)

\S4method{plotMarkerTSNE}{seurat}(object, genes, expression = c("mean",
  "median", "sum"), color, pointsAsNumbers = FALSE, pointSize = 0.5,
  pointAlpha = 0.8, label = TRUE, labelSize = 6L, dark = FALSE,
  legend = TRUE, aspectRatio = 1L, title = NULL, subtitle = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{Gene identifiers. Must match the rownames of the object.}

\item{expression}{Calculation to apply on the aggregate marker expression.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{pointSize}{Cell point size.}

\item{pointAlpha}{Alpha transparency level. Useful when there many cells in
the dataset, and some cells can be masked.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{dark}{Plot against a dark background using \code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{legend}{Include plot legend.}

\item{title}{Plot title.}

\item{subtitle}{Include gene identifiers as subtitles.}
}
\value{
\code{ggplot}.
}
\description{
Plot Marker t-SNE
}
\examples{
# seurat ====
# Mitochondrial genes
mito <- grep("^MT\\\\.", rownames(counts(seurat_small)), value = TRUE)
head(mito)
plotMarkerTSNE(seurat_small, genes = mito)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarker}}, \code{\link{plotPCA}},
  \code{\link{plotPCElbow}}, \code{\link{plotTSNE}},
  \code{\link{plotTopMarkers}},
  \code{\link{sanitizeMarkers}}, \code{\link{topMarkers}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
