% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMarkerTSNE.R
\docType{methods}
\name{plotMarkerTSNE}
\alias{plotMarkerTSNE}
\alias{plotMarkerTSNE}
\alias{plotMarkerTSNE,data.frame-method}
\alias{plotMarkerTSNE,seurat-method}
\title{Plot Marker t-SNE}
\usage{
plotMarkerTSNE(object, ...)

\S4method{plotMarkerTSNE}{data.frame}(object, genes, expression = "mean",
  pointsAsNumbers = FALSE, pointSize = 0.5, pointAlpha = 0.8,
  label = TRUE, labelSize = 6L, color = viridis::scale_color_viridis(),
  dark = TRUE, legend = TRUE, title = NULL, subtitle = TRUE)

\S4method{plotMarkerTSNE}{seurat}(object, genes, expression = "mean",
  pointsAsNumbers = FALSE, pointSize = 0.5, pointAlpha = 0.8,
  label = TRUE, labelSize = 6L, color = viridis::scale_color_viridis(),
  dark = TRUE, title = NULL, subtitle = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{Genes identifiers (matching the rownames in the object),
of which to get expression data.}

\item{expression}{Calculation to apply on the aggregate marker expression.
Supports \code{mean} (default), \code{median}, or \code{sum}.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{pointSize}{Cell point size.}

\item{pointAlpha}{Alpha transparency level. Useful when there many cells in
the dataset, and some cells can be masked.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{color}{Color palette to use for points. Defaults to the viridis color
palette. Use \code{\link[ggplot2:scale_color_gradient]{ggplot2::scale_color_gradient()}} to easily define your own
low/high gradient.}

\item{dark}{Enable dark mode.}

\item{legend}{Show plot legend.}

\item{title}{Plot title.}

\item{subtitle}{Include gene(s) in the subtitle.}
}
\value{
\link{ggplot}.
}
\description{
Plot Marker t-SNE
}
\examples{
load(system.file("extdata/seurat.rda", package = "bcbioSingleCell"))

genes <- counts(seurat) \%>\% rownames() \%>\% head()
print(genes)

# seurat
plotMarkerTSNE(seurat, genes = genes)

# data.frame
df <- fetchTSNEExpressionData(seurat, genes = genes)
plotMarkerTSNE(df, genes = genes)
}
\seealso{
Other t-SNE Utilities: \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{plotTSNE}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
