% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMarkerTSNE.R
\docType{methods}
\name{plotMarkerTSNE}
\alias{plotMarkerTSNE}
\alias{plotMarkerTSNE}
\alias{plotMarkerTSNE,data.frame-method}
\alias{plotMarkerTSNE,seurat-method}
\title{Plot Marker t-SNE}
\usage{
plotMarkerTSNE(object, ...)

\S4method{plotMarkerTSNE}{data.frame}(object, colorPoints = "geomean",
  pointsAsNumbers = FALSE, pointSize = 1L, label = TRUE, labelSize = 6L,
  color = viridis::scale_color_viridis(), dark = TRUE, legend = TRUE,
  title = NULL, subtitle = TRUE)

\S4method{plotMarkerTSNE}{seurat}(object, genes, format = "symbol",
  colorPoints = "geomean", pointsAsNumbers = FALSE, pointSize = 1L,
  label = TRUE, labelSize = 6L, color = viridis::scale_color_viridis(),
  dark = TRUE, title = NULL, subtitle = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{colorPoints}{Color points by geometric mean (\code{geomean}) or expression
of individual gene (\code{expression}).}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{pointSize}{Cell point size.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{color}{Color palette to use for points. Defaults to the viridis color
palette. Use \code{\link[ggplot2:scale_color_gradient]{ggplot2::scale_color_gradient()}} to easily define your own
low/high gradient.}

\item{dark}{Enable dark mode.}

\item{legend}{Show plot legend.}

\item{title}{Plot title.}

\item{subtitle}{Include gene(s) in the subtitle.}

\item{genes}{Genes (by symbol name) of which to get expression data.}

\item{format}{Gene identifier format. Defaults to \code{symbol} but \code{ensgene}
is also supported, as long as Ensembl gene to symbol identifier mappings
are defined.}
}
\value{
\link{ggplot}.
}
\description{
Plot Marker t-SNE
}
\examples{
load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

symbol <- counts(seurat) \%>\% rownames() \%>\% .[[1]]
print(symbol)

ensgene <- bcbio(seurat, "gene2symbol") \%>\%
    .[which(.[["symbol"]] \%in\% symbol), "ensgene", drop = TRUE]
print(ensgene)

# seurat
plotMarkerTSNE(seurat, genes = symbol, format = "symbol")
plotMarkerTSNE(seurat, genes = ensgene, format = "ensgene")

# data.frame
df <- fetchTSNEExpressionData(seurat, genes = symbol)
plotMarkerTSNE(df)
}
\seealso{
Other t-SNE Utilities: \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{plotTSNE}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
