% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMitoRatio-methods.R
\docType{methods}
\name{plotMitoRatio}
\alias{plotMitoRatio}
\alias{plotMitoRatio,SingleCellExperiment-method}
\title{Plot Mitochondrial Transcript Abundance}
\usage{
plotMitoRatio(object, ...)

\S4method{plotMitoRatio}{SingleCellExperiment}(object,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  interestingGroups = NULL, max = 1L,
  fill = getOption("basejump.discrete.fill", NULL), trans = "sqrt",
  title = "mito ratio")
}
\value{
\code{ggplot}.
}
\description{
Plot Mitochondrial Transcript Abundance
}
\examples{
data(indrops_small)
plotMitoRatio(indrops_small)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
