% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMitoRatio.R
\docType{methods}
\name{plotMitoRatio}
\alias{plotMitoRatio}
\alias{plotMitoRatio,bcbioSingleCellANY-method}
\title{Plot Mitochondrial Transcript Abundance}
\usage{
plotMitoRatio(object, ...)

\S4method{plotMitoRatio}{bcbioSingleCellANY}(object, interestingGroups, max,
  filterCells = FALSE, aggregateReplicates = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroups}{Interesting group, to use for colors.}

\item{max}{Recommended maximum value cutoff.}

\item{filterCells}{Show only the cells that have passed filtering cutoffs.}

\item{aggregateReplicates}{Aggregate technical replicates, if specified. This
function uses values assigned in the \code{sampleNameAggregate} column of the
internal sample metadata \link{data.frame}.}
}
\value{
\link{ggplot} grid.
}
\description{
Plot Mitochondrial Transcript Abundance
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Metrics}
