% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMitoRatio.R
\docType{methods}
\name{plotMitoRatio}
\alias{plotMitoRatio}
\alias{plotMitoRatio,bcbioSingleCellANY-method}
\title{Plot Mitochondrial Transcript Abundance}
\usage{
plotMitoRatio(object, ...)

\S4method{plotMitoRatio}{bcbioSingleCellANY}(object, interestingGroup, max,
  filterCells = FALSE, aggregateReplicates = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroup}{Interesting group, to use for colors.}

\item{max}{Recommended maximum value cutoff.}

\item{filterCells}{Show only the cells that have passed filtering cutoffs.}

\item{aggregateReplicates}{Aggregate technical replicates, if present. If
\code{TRUE}, this function uses the values slotted in
\code{sampleMetadata(object)[["sampleNameAggregate"]])}.

[["sampleNameAggregate"]: R:[%22sampleNameAggregate%22}
}
\value{
\link{ggplot} grid.
}
\description{
Plot Mitochondrial Transcript Abundance
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Metrics}
