% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMitoRatio.R
\docType{methods}
\name{plotMitoRatio}
\alias{plotMitoRatio}
\alias{plotMitoRatio,bcbioSCDataSet-method}
\alias{plotMitoRatio,bcbioSCFiltered-method}
\title{Plot Mitochondrial Transcript Abundance}
\usage{
plotMitoRatio(object, ...)

\S4method{plotMitoRatio}{bcbioSCDataSet}(object, max = 0.1)

\S4method{plotMitoRatio}{bcbioSCFiltered}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{max}{Recommended maximum value cutoff.}
}
\value{
\link{ggplot} grid.
}
\description{
Plot Mitochondrial Transcript Abundance
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Metrics}
