% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMitoRatio.R
\docType{methods}
\name{plotMitoRatio}
\alias{plotMitoRatio}
\alias{plotMitoRatio}
\alias{plotMitoRatio,SingleCellExperiment-method}
\alias{plotMitoRatio,seurat-method}
\title{Plot Mitochondrial Transcript Abundance}
\usage{
plotMitoRatio(object, ...)

\S4method{plotMitoRatio}{SingleCellExperiment}(object, geom = c("violin",
  "boxplot", "histogram", "ridgeline"), interestingGroups, max = 1L,
  fill = scale_fill_hue(), title = "mitochondrial abundance")

\S4method{plotMitoRatio}{seurat}(object, geom = c("violin", "boxplot",
  "histogram", "ridgeline"), interestingGroups, max = 1L,
  fill = scale_fill_hue(), title = "mitochondrial abundance")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{geom}{Plot type. Uses \code{\link[=match.arg]{match.arg()}} internally and defaults to the
first argument in the vector.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in camel case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{max}{Recommended maximum value cutoff.}

\item{fill}{Desired ggplot fill scale. Must supply discrete values. When set
to \code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{title}{Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot Mitochondrial Transcript Abundance
}
\examples{
# bcbioSingleCell ====
plotMitoRatio(indrops_small)

# SingleCellExperiment ====
plotMitoRatio(cellranger_small)

# seurat ====
# `object@meta.data` must contain `mitoRatio`
\dontrun{
plotMitoRatio(Seurat::pbmc_small)
}
}
\seealso{
Other Quality Control Functions: \code{\link{barcodeRanksPerSample}},
  \code{\link{barcodeRanks}}, \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
