% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMitoRatio.R
\docType{methods}
\name{plotMitoRatio}
\alias{plotMitoRatio}
\alias{plotMitoRatio}
\alias{plotMitoRatio,bcbioSingleCell-method}
\alias{plotMitoRatio,seurat-method}
\title{Plot Mitochondrial Transcript Abundance}
\usage{
plotMitoRatio(object, ...)

\S4method{plotMitoRatio}{bcbioSingleCell}(object, geom = c("violin",
  "boxplot", "histogram", "ridgeline"), max = Inf, interestingGroups,
  flip = TRUE, fill = scale_fill_viridis(discrete = TRUE))

\S4method{plotMitoRatio}{seurat}(object, geom = c("violin", "boxplot",
  "histogram", "ridgeline"), max = Inf, interestingGroups, flip = TRUE,
  fill = scale_fill_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{geom}{Plot type. Uses \code{\link[=match.arg]{match.arg()}} internally and defaults to the
first argument in the vector.}

\item{max}{Recommended maximum value cutoff.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in \code{\link[=camel]{camel()}} case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{flip}{Plot the samples on the y axis.}

\item{fill}{Desired ggplot fill scale. Must supply discrete values. When set
to \code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}
}
\value{
\code{ggplot}.
}
\description{
Plot Mitochondrial Transcript Abundance
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat_small.rda", package = "bcbioSingleCell"))

# bcbioSingleCell ====
plotMitoRatio(bcb_small)

# seurat ====
plotMitoRatio(seurat_small)
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
