% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMitoRatio.R
\docType{methods}
\name{plotMitoRatio}
\alias{plotMitoRatio}
\alias{plotMitoRatio,bcbioSingleCell-method}
\alias{plotMitoRatio,data.frame-method}
\alias{plotMitoRatio,seurat-method}
\title{Plot Mitochondrial Transcript Abundance}
\usage{
plotMitoRatio(object, ...)

\S4method{plotMitoRatio}{bcbioSingleCell}(object, geom = "violin", max,
  interestingGroups, filterCells = FALSE, samplesOnYAxis = TRUE,
  fill = scale_fill_viridis(discrete = TRUE))

\S4method{plotMitoRatio}{data.frame}(object, geom = "violin", max = Inf,
  interestingGroups, multiplexed = FALSE, samplesOnYAxis = TRUE,
  fill = scale_fill_viridis(discrete = TRUE))

\S4method{plotMitoRatio}{seurat}(object, geom = "violin", max = Inf,
  interestingGroups, multiplexed = FALSE, samplesOnYAxis = TRUE,
  fill = scale_fill_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{geom}{Plot type. Supported formats: \code{boxplot}, \code{histogram},
\code{ridgeline}, and \code{violin}.}

\item{max}{Recommended maximum value cutoff.}

\item{interestingGroups}{Interesting group, to use for colors.}

\item{filterCells}{Show only the cells that have passed filtering cutoffs.}

\item{samplesOnYAxis}{Plot the samples on the y axis. Doesn't apply to
histogram.}

\item{fill}{Desired ggplot fill scale. Defaults to
\code{\link[viridis:scale_fill_viridis]{viridis::scale_fill_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{multiplexed}{Whether the FASTQ files are multiplexed. This is
detected automatically for \link{bcbioSingleCell} objects.}
}
\value{
\link{ggplot}.
}
\description{
Plot Mitochondrial Transcript Abundance
}
\details{
A violin plot is a comact display of a continuous distribution. It
is a blend of \code{\link[=geom_boxplot]{geom_boxplot()}} and \code{\link[=geom_density]{geom_density()}}: a violin plot is a
mirrored density plot displayed in the same way as a boxplot.
}
\examples{
# bcbioSingleCell
\dontrun{
plotMitoRatio(bcb)
}

# seurat
\dontrun{
plotMitoRatio(seurat)
}

# metrics data.frame
\dontrun{
metrics <- metrics(bcb)
plotMitoRatio(metrics)
}
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Metrics}
