% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMitoVsCoding-methods.R
\docType{methods}
\name{plotMitoVsCoding}
\alias{plotMitoVsCoding}
\alias{plotMitoVsCoding,SingleCellExperiment-method}
\title{Plot Mitochondrial vs. Coding Counts}
\usage{
\S4method{plotMitoVsCoding}{SingleCellExperiment}(object,
  interestingGroups = NULL, trendline = FALSE,
  color = getOption("basejump.discrete.color", NULL), trans = "log2",
  title = "mito vs. coding")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character} or \code{NULL}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{trendline}{\code{boolean}. Include trendline on plot.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{ggplot2::scale_color_manual()}. To set the discrete color
palette globally, use
\code{options(basejump.discrete.color = ggplot2::scale_color_viridis_d())}.}

\item{trans}{\code{string}. Name of the axis scale transformation to apply. See
\code{help("scale_x_continuous", "ggplot2")} for more information.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot Mitochondrial vs. Coding Counts
}
\examples{
data(indrops)
plotMitoVsCoding(indrops)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
