% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMitoVsCoding-methods.R
\docType{methods}
\name{plotMitoVsCoding}
\alias{plotMitoVsCoding}
\alias{plotMitoVsCoding,SingleCellExperiment-method}
\title{Plot mitochondrial vs. coding counts}
\usage{
\S4method{plotMitoVsCoding}{SingleCellExperiment}(object,
  interestingGroups = NULL, trendline = FALSE,
  color = getOption("basejump.color.discrete", default = NULL),
  trans = "log2", title = "mito vs. coding")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{trendline}{\code{logical(1)}.
Include trendline on plot.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot mitochondrial vs. coding counts
}
\examples{
data(indrops)
plotMitoVsCoding(indrops)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
