% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMitoVsCoding.R
\docType{methods}
\name{plotMitoVsCoding}
\alias{plotMitoVsCoding}
\alias{plotMitoVsCoding}
\alias{plotMitoVsCoding,SingleCellExperiment-method}
\alias{plotMitoVsCoding,seurat-method}
\title{Plot Mitochondrial vs. Coding Counts}
\usage{
plotMitoVsCoding(object, ...)

\S4method{plotMitoVsCoding}{SingleCellExperiment}(object, interestingGroups,
  trendline = TRUE, color = scale_color_hue(), trans = "log2",
  title = "mito vs. coding")

\S4method{plotMitoVsCoding}{seurat}(object, interestingGroups,
  trendline = TRUE, color = scale_color_hue(), trans = "log2",
  title = "mito vs. coding")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in camel case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{trendline}{Include trendline on plot.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{trans}{Name of the axis scale transformation to apply. See
\code{help("scale_x_continuous", "ggplot2")} for more information.}

\item{title}{Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot Mitochondrial vs. Coding Counts
}
\examples{
# bcbioSingleCell ====
plotMitoVsCoding(indrops_small)

# SingleCellExperiment ====
plotMitoVsCoding(cellranger_small)

# seurat ====
plotMitoVsCoding(seurat_small)
}
\seealso{
Other Quality Control Functions: \code{\link{barcodeRanksPerSample}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotQC}}, \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
