% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMitoVsCoding.R
\docType{methods}
\name{plotMitoVsCoding}
\alias{plotMitoVsCoding}
\alias{plotMitoVsCoding}
\alias{plotMitoVsCoding,bcbioSingleCell-method}
\alias{plotMitoVsCoding,data.frame-method}
\alias{plotMitoVsCoding,seurat-method}
\title{Plot Mitochondrial Counts vs. Coding Counts}
\usage{
plotMitoVsCoding(object, ...)

\S4method{plotMitoVsCoding}{bcbioSingleCell}(object, interestingGroups,
  color = scale_color_viridis(discrete = TRUE))

\S4method{plotMitoVsCoding}{data.frame}(object, interestingGroups,
  color = scale_color_viridis(discrete = TRUE))

\S4method{plotMitoVsCoding}{seurat}(object, interestingGroups,
  color = scale_color_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in \code{\link[=camel]{camel()}} case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}
}
\value{
\code{ggplot}.
}
\description{
Plot Mitochondrial Counts vs. Coding Counts
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat.rda", package = "bcbioSingleCell"))

# bcbioSingleCell
plotMitoVsCoding(bcb)

# seurat
plotMitoVsCoding(seurat)

# data.frame
df <- metrics(bcb)
plotMitoVsCoding(df)
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
