% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMitoVsCoding.R
\docType{methods}
\name{plotMitoVsCoding}
\alias{plotMitoVsCoding}
\alias{plotMitoVsCoding}
\alias{plotMitoVsCoding,bcbioSingleCell-method}
\alias{plotMitoVsCoding,seurat-method}
\title{Plot Mitochondrial Counts vs. Coding Counts}
\usage{
plotMitoVsCoding(object, ...)

\S4method{plotMitoVsCoding}{bcbioSingleCell}(object, interestingGroups,
  color = scale_color_viridis(discrete = TRUE))

\S4method{plotMitoVsCoding}{seurat}(object, interestingGroups,
  color = scale_color_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{Interesting group, to use for colors.}

\item{color}{\emph{Only applies to scatterplot}. Desired ggplot color scale.
Defaults to \code{\link[viridis:scale_color_viridis]{viridis::scale_color_viridis()}}. Must supply discrete values.
When set to \code{NULL}, the default ggplot2 color palette will be used. If
manual color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}
}
\value{
\code{ggplot}.
}
\description{
Plot Mitochondrial Counts vs. Coding Counts
}
\details{
A violin plot is a comact display of a continuous distribution. It
is a blend of \code{\link[=geom_boxplot]{geom_boxplot()}} and \code{\link[=geom_density]{geom_density()}}: a violin plot is a
mirrored density plot displayed in the same way as a boxplot.
}
\examples{
# bcbioSingleCell ====
plotMitoVsCoding(bcb_small)

# seurat ====
plotMitoVsCoding(seurat_small)
}
\seealso{
Other Quality Control Functions: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotQC}}, \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
