% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNovelty-methods.R
\docType{methods}
\name{plotNovelty}
\alias{plotNovelty}
\alias{plotNovelty,SingleCellExperiment-method}
\title{Plot Novelty Score}
\usage{
\S4method{plotNovelty}{SingleCellExperiment}(object,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  interestingGroups = NULL, min = 0L,
  fill = getOption("basejump.fill.discrete", NULL), trans = "identity",
  title = "novelty : genes per UMI")
}
\arguments{
\item{object}{Object.}

\item{geom}{\code{string}. Plot type. Uses \code{base::match.arg()} and defaults to the
first argument in the \code{character} vector.}

\item{interestingGroups}{\code{character} or \code{NULL}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{min}{\code{scalar numeric}. Recommended minimum value cutoff.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{ggplot2::scale_fill_manual()}. To set the discrete fill palette
globally, use \code{options(basejump.fill.discrete = scale_fill_viridis_d())}.}

\item{trans}{\code{string}. Name of the axis scale transformation to apply. See
\code{help("scale_x_continuous", "ggplot2")} for more information.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
"Novelty" refers to log10 genes detected per count.
}
\examples{
data(indrops)
plotNovelty(indrops)
}
\author{
Michael Steinbaugh
}
