% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNovelty-methods.R
\docType{methods}
\name{plotNovelty}
\alias{plotNovelty}
\alias{plotNovelty,bcbioSingleCell-method}
\title{Plot novelty score}
\usage{
plotNovelty(object, ...)

\S4method{plotNovelty}{bcbioSingleCell}(object, geom = c("histogram",
  "ecdf", "violin", "ridgeline", "boxplot"), interestingGroups = NULL,
  min = 0L, fill = getOption(x = "acid.fill.discrete", default =
  acidplots::scale_fill_synesthesia_d()), trans = "identity",
  title = "Novelty")
}
\arguments{
\item{object}{Object.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{min}{\code{numeric(1)}.
Recommended minimum value cutoff.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{title}{\code{character(1)}.
Plot title.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
"Novelty" refers to log10 genes detected per count.
}
\note{
Updated 2019-07-27.
}
\examples{
data(indrops)
plotNovelty(indrops)
}
\author{
Michael Steinbaugh
}
