% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotNovelty.R
\docType{methods}
\name{plotNovelty}
\alias{plotNovelty}
\alias{plotNovelty}
\alias{plotNovelty,bcbioSCDataSet-method}
\alias{plotNovelty,bcbioSCSubset-method}
\title{Plot Novelty Score}
\usage{
plotNovelty(object, ...)

\S4method{plotNovelty}{bcbioSCDataSet}(object, min = 0.8)

\S4method{plotNovelty}{bcbioSCSubset}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{min}{Recommended minimum value cutoff.}
}
\value{
\link{ggplot} grid.
}
\description{
Plot Novelty Score
}
\details{
"Novelty" refers to log10 genes detected per count.
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotCellularBarcodes}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}}
}
