% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotNovelty.R
\docType{methods}
\name{plotNovelty}
\alias{plotNovelty}
\alias{plotNovelty}
\alias{plotNovelty,SingleCellExperiment-method}
\alias{plotNovelty,seurat-method}
\title{Plot Novelty Score}
\usage{
plotNovelty(object, ...)

\S4method{plotNovelty}{SingleCellExperiment}(object, geom = c("ecdf",
  "histogram", "ridgeline", "violin", "boxplot"), interestingGroups, min = 0L,
  fill = scale_fill_hue(), title = "genes per UMI (novelty)")

\S4method{plotNovelty}{seurat}(object, geom = c("ecdf", "histogram",
  "ridgeline", "violin", "boxplot"), interestingGroups, min = 0L,
  fill = scale_fill_hue(), title = "genes per UMI (novelty)")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{geom}{Plot type. Uses \code{\link[=match.arg]{match.arg()}} internally and defaults to the
first argument in the vector.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in camel case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{min}{Recommended minimum value cutoff.}

\item{fill}{Desired ggplot fill scale. Must supply discrete values. When set
to \code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{title}{Plot title.}
}
\value{
\code{ggplot}.
}
\description{
"Novelty" refers to log10 genes detected per count.
}
\examples{
# bcbioSingleCell ====
plotNovelty(indrops_small)

# SingleCellExperiment ====
plotNovelty(cellranger_small)

# seurat ====
# `object@meta.data` must contain `log10GenesPerUMI`
\dontrun{
plotNovelty(Seurat::pbmc_small)
}
}
\seealso{
Other Quality Control Functions: \code{\link{barcodeRanksPerSample}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
