% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotPCA.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,seurat-method}
\title{Plot Principal Component Analysis (PCA)}
\usage{
\S4method{plotPCA}{seurat}(object, interestingGroups = "ident",
  pointsAsNumbers = FALSE, pointSize = 0.5, pointAlpha = 0.8,
  label = TRUE, labelSize = 6L, color = scale_color_hue(), dark = TRUE,
  title = NULL)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Interesting group to use for plot colors.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{pointSize}{Cell point size.}

\item{pointAlpha}{Alpha transparency level. Useful when there many cells in
the dataset, and some cells can be masked.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{color}{Color palette to use for points.}

\item{dark}{Enable dark mode.}

\item{title}{Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot Principal Component Analysis (PCA)
}
\examples{
# seurat ====
plotPCA(pbmc_small)
}
\seealso{
\code{\link[=plotTSNE]{plotTSNE()}}.

Other PCA Utilities: \code{\link{fetchPCAData}}
}
\author{
Michael Steinbaugh
}
