% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotPCA.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,seurat-method}
\title{Plot Principal Component Analysis (PCA)}
\usage{
\S4method{plotPCA}{seurat}(object, interestingGroup = "ident", label = TRUE)
}
\arguments{
\item{object}{Object.}

\item{interestingGroup}{Interesting group to use for plot colors.}

\item{label}{Label the clusters on the plot.}
}
\value{
\link{ggplot}.
}
\description{
Generate a t-distributed stochastic neighbor embedding (t-SNE) plot.
}
\examples{
\dontrun{
plotTSNE(seurat)
}
}
\seealso{
\code{\link[=plotTSNE]{plotTSNE()}}.

Other PCA Utilities: \code{\link{fetchPCAData}}
}
\author{
Michael Steinbaugh
}
\concept{PCA Utilities}
