% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-seurat-plotPCA.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,seurat-method}
\title{Plot Principal Component Analysis (PCA)}
\usage{
\S4method{plotPCA}{seurat}(object, interestingGroups = "ident",
  color = scale_color_hue(), pointsAsNumbers = FALSE, pointSize = 0.5,
  pointAlpha = 0.8, label = TRUE, labelSize = 6L, dark = TRUE,
  aspectRatio = 1L, title = NULL)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in camel case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{pointSize}{Cell point size.}

\item{pointAlpha}{Alpha transparency level. Useful when there many cells in
the dataset, and some cells can be masked.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{dark}{Plot against a dark background using
\code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{aspectRatio}{Aspect ratio.}

\item{title}{Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot Principal Component Analysis (PCA)
}
\examples{
# seurat ====
plotPCA(Seurat::pbmc_small)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotMarker}},
  \code{\link{plotPCElbow}}, \code{\link{plotTSNE}},
  \code{\link{plotTopMarkers}},
  \code{\link{sanitizeMarkers}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
