% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotPCA.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,seurat-method}
\title{Plot Principal Component Analysis (PCA)}
\usage{
\S4method{plotPCA}{seurat}(object, interestingGroups = "ident",
  pointsAsNumbers = FALSE, pointSize = 1, label = TRUE, labelSize = 6,
  color = ggplot2::scale_color_hue(), dark = TRUE, title = NULL)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Interesting group to use for plot colors.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{pointSize}{Cell point size.}

\item{label}{Overlay a cluster identitiy label on the plot.}

\item{labelSize}{Size of the text label.}

\item{color}{Color palette to use for points. Defaults to the viridis color
palette. Use \code{\link[ggplot2:scale_color_gradient]{ggplot2::scale_color_gradient()}} to easily define your own
low/high gradient.}

\item{dark}{Enable dark mode.}

\item{title}{Plot title.}
}
\value{
\link{ggplot}.
}
\description{
Plot Principal Component Analysis (PCA)
}
\seealso{
\code{\link[=plotTSNE]{plotTSNE()}}.

Other PCA Utilities: \code{\link{fetchPCAData}}
}
\author{
Michael Steinbaugh
}
