% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotPCElbow.R
\docType{methods}
\name{plotPCElbow}
\alias{plotPCElbow}
\alias{plotPCElbow}
\alias{plotPCElbow,seurat-method}
\title{Plot PC Elbow}
\usage{
plotPCElbow(object, ...)

\S4method{plotPCElbow}{seurat}(object, maxPct = 0.05, minCumPct = 0.9,
  plot = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{maxPct}{Maximum percent standard deviation.}

\item{minCumPct}{Minimum cumulative percent standard deviation.}

\item{plot}{Plot the PC standard deviations.}
}
\value{
\itemize{
\item Show graphical output of elbow plots.
\item Invisibly return numeric sequence vector of PCs to include for
dimensionality reduction analysis.
}
}
\description{
Calculate the principal component (PC) with either a maximum SD percentage or
minimum % SD cumulative sum cutoff value. This defaults to keeping the larger
PC cutoff for either 5% SD (\code{maxPct}) or 90% cumulative (\code{minCumsum}).
}
\examples{
seurat <- examples[["seurat"]]
pcUse <- plotPCElbow(seurat)
pcUse
}
\seealso{
\link[Seurat:PCElbowPlot]{Seurat::PCElbowPlot}.

Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotKnownMarkers}},
  \code{\link{plotMarkers}}, \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
