% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotPCElbow.R
\docType{methods}
\name{plotPCElbow}
\alias{plotPCElbow}
\alias{plotPCElbow}
\alias{plotPCElbow,seurat-method}
\title{Plot PC Elbow}
\usage{
plotPCElbow(object, ...)

\S4method{plotPCElbow}{seurat}(object, minSD = 1L, minPct = 0.025,
  maxCumPct = 0.9, trans = c("identity", "sqrt"), plot = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{minSD}{Minimum standard deviation.}

\item{minPct}{Minimum percent standard deviation.}

\item{maxCumPct}{Maximum cumulative percent standard deviation.}

\item{plot}{Include plot.}
}
\value{
\itemize{
\item Show graphical output of elbow plots.
\item Invisibly return numeric sequence vector of PCs to include for
dimensionality reduction analysis.
}
}
\description{
Calculate the principal component (PC) cutoff using a heuristic approach.
}
\details{
Automatically return the smallest number of PCs that match the \code{minSD},
\code{minPct}, and \code{maxCumPct} cutoffs.
}
\examples{
# seurat ====
plotPCElbow(pbmc_small)
}
\seealso{
\code{\link[Seurat:PCElbowPlot]{Seurat::PCElbowPlot()}}.

Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotMarkers}},
  \code{\link{plotPCA}}, \code{\link{plotTSNE}},
  \code{\link{plotTopMarkers}},
  \code{\link{sanitizeMarkers}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
