% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotPCElbow-methods.R
\docType{methods}
\name{plotPCElbow}
\alias{plotPCElbow}
\alias{plotPCElbow,seurat-method}
\title{Plot PC Elbow}
\usage{
plotPCElbow(object, ...)

\S4method{plotPCElbow}{seurat}(object, minSD = 1L, minPct = 0.01,
  maxCumPct = 0.9, trans = c("identity", "sqrt"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{minSD}{\code{scalar numeric}. Minimum standard deviation.}

\item{minPct}{\code{scalar numeric} (\code{0}-\code{1}). Minimum percent standard deviation.}

\item{maxCumPct}{\code{scalar numeric} (\code{0}-\code{1}).Maximum cumulative percent
standard deviation.}

\item{trans}{\code{string}. Name of the axis scale transformation to apply. See
\code{help("scale_x_continuous", "ggplot2")} for more information.}
}
\value{
\itemize{
\item Show graphical output of elbow plots.
\item Invisibly return numeric sequence vector of PCs to include for
dimensionality reduction analysis.
}
}
\description{
Calculate the principal component (PC) cutoff using a heuristic approach.
}
\details{
Automatically return the smallest number of PCs that match the \code{minSD},
\code{minPct}, and \code{maxCumPct} cutoffs.
}
\examples{
# seurat ====
plotPCElbow(seurat_small)
}
\seealso{
\itemize{
\item \code{\link[Seurat:PCElbowPlot]{Seurat::PCElbowPlot()}}.
}

Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeature}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotReducedDim}},
  \code{\link{sanitizeSeuratMarkers}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Functions}
