% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC-methods.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\alias{plotQC,bcbioSingleCell-method}
\title{Quality control}
\usage{
plotQC(object, ...)

\S4method{plotQC}{bcbioSingleCell}(object, interestingGroups = NULL,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  headerLevel = 2L, legend = getOption(x = "acid.legend", default =
  TRUE), return = c("grid", "list", "markdown"))
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
R Markdown header level.}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{grid}: Graphical output.
\item \code{list}: \code{list} containing \code{ggplot} objects.
\item \code{markdown}: R Markdown report, with reports separated by headers.
}
}
\description{
Quality control
}
\note{
Updated 2019-07-29.
}
\examples{
data(indrops)
plotQC(indrops)
}
\author{
Michael Steinbaugh
}
