% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC-methods.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\alias{plotQC,SingleCellExperiment-method}
\title{Quality control}
\usage{
\S4method{plotQC}{SingleCellExperiment}(object, interestingGroups = NULL,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  headerLevel = 2L, legend = getOption("basejump.legend", default =
  TRUE), return = c("grid", "list", "markdown"))
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
R Markdown header level.}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}
}
\value{
\itemize{
\item \code{grid}: Graphical output.
\item \code{list}: \code{list} containing \code{ggplot} objects.
\item \code{markdown}: R Markdown report, with reports separated by headers.
}
}
\description{
Quality control
}
\examples{
data(indrops)
plotQC(indrops)
}
\author{
Michael Steinbaugh
}
