% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotQC.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\alias{plotQC,bcbioSingleCell-method}
\alias{plotQC,seurat-method}
\title{Quality Control Plots}
\usage{
\S4method{plotQC}{bcbioSingleCell}(object, interestingGroups,
  geom = c("violin", "boxplot", "histogram", "ridgeline"), headerLevel = 2L,
  legend = FALSE, return = c("grid", "list", "markdown"))

\S4method{plotQC}{seurat}(object, interestingGroups, geom = c("violin",
  "boxplot", "histogram", "ridgeline"), headerLevel = 2L, legend = FALSE,
  return = c("grid", "list", "markdown"))
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in \code{\link[=camel]{camel()}} case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{geom}{Plot type.}

\item{headerLevel}{R Markdown header level.}

\item{legend}{Include plot legend.}

\item{return}{Return type. Uses \code{\link[=match.arg]{match.arg()}} internally and defaults to the
first item in the arguments vector.}
}
\value{
\itemize{
\item \code{grid}: \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}} graphical output.
\item \code{list}: \code{list} containing \code{ggplot} objects.
\item \code{markdown}: R Markdown report, with reports separated by headers.
}
}
\description{
Utility function that loops our standard quality control plots, for easy
visualization.
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat.rda", package = "bcbioSingleCell"))

# bcbioSingleCell
plotQC(bcb)

# seurat
plotQC(seurat)
}
\author{
Michael Steinbaugh
}
