% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC-methods.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\alias{plotQC,SingleCellExperiment-method}
\alias{plotQC,seurat-method}
\title{Quality Control Plots}
\usage{
\S4method{plotQC}{SingleCellExperiment}(object, interestingGroups,
  geom = c("violin", "ridgeline", "ecdf", "histogram", "boxplot"),
  headerLevel = 2L, legend = getOption("bcbio.legend", FALSE),
  return = c("grid", "list", "markdown"))

\S4method{plotQC}{seurat}(object, interestingGroups, geom = c("violin",
  "ridgeline", "ecdf", "histogram", "boxplot"), headerLevel = 2L,
  legend = getOption("bcbio.legend", FALSE), return = c("grid", "list",
  "markdown"))
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}

\item{geom}{\code{string}. Plot type. Uses \code{\link[=match.arg]{match.arg()}} and defaults to the first
argument in the \code{character} vector.}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). R Markdown header level.}

\item{legend}{\code{boolean}. Include plot legend.}

\item{return}{\code{string}. Return type. Uses \code{\link[base:match.arg]{base::match.arg()}} internally and
defaults to the first argument in the \code{character} vector.}
}
\value{
\itemize{
\item \code{grid}: \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}} graphical output.
\item \code{list}: \code{list} containing \code{ggplot} objects.
\item \code{markdown}: R Markdown report, with reports separated by headers.
}
}
\description{
Utility function that loops our standard quality control plots, for easy
visualization.
}
\examples{
# bcbioSingleCell ====
plotQC(indrops_small)

# SingleCellExperiment ====
plotQC(cellranger_small)

# seurat ====
plotQC(seurat_small)
}
\seealso{
Other Quality Control Functions: \code{\link{barcodeRanksPerSample}},
  \code{\link{filterCells}}, \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Functions}
