% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotQC.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\alias{plotQC,bcbioSingleCell-method}
\alias{plotQC,seurat-method}
\title{Quality Control Plots}
\usage{
\S4method{plotQC}{bcbioSingleCell}(object, interestingGroups, geom = "violin",
  headerLevel = 2L, legend = FALSE, return = "grid")

\S4method{plotQC}{seurat}(object, interestingGroups, geom = "violin",
  headerLevel = 2L, legend = FALSE, return = "grid")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Interesting group, to use for colors.}

\item{geom}{Plot type. Supported formats: \code{boxplot}, \code{histogram},
\code{ridgeline}, and \code{violin}. Applies to \code{\link[=plotUMIsPerCell]{plotUMIsPerCell()}},
\code{\link[=plotGenesPerCell]{plotGenesPerCell()}}, \code{\link[=plotMitoRatio]{plotMitoRatio()}}, and \code{\link[=plotNovelty]{plotNovelty()}} output.}

\item{headerLevel}{R Markdown header level.}

\item{legend}{Include plot legend.}

\item{return}{\itemize{
\item \code{grid}: \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}} graphical output.
\itemize{
\item \code{list}: \link{ggplot} \link{list}.
\item \code{markdown}: R Markdown report, with reports separated by headers.
}
}}
}
\value{
R Markdown template code for quality control analysis.
}
\description{
Utility function that loops our standard quality control plots, for easy
visualization.
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat.rda", package = "bcbioSingleCell"))

# bcbioSingleCell
plotQC(bcb)

# seurat
plotQC(seurat)
}
\author{
Michael Steinbaugh
}
