% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotQuantileHeatmap.R
\docType{methods}
\name{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap,bcbioSingleCell-method}
\alias{plotQuantileHeatmap,seurat-method}
\title{Plot Heatmap with Quantile Breaks}
\usage{
\S4method{plotQuantileHeatmap}{bcbioSingleCell}(object, n = 10,
  annotationCol = NA, clusterCols = FALSE, clusterRows = FALSE,
  color = viridis::viridis, legendColor = viridis::viridis, title = NULL)

\S4method{plotQuantileHeatmap}{seurat}(object, n = 10, annotationCol = NA,
  clusterCols = FALSE, clusterRows = FALSE, color = viridis::viridis,
  legendColor = viridis::viridis, title = NULL)
}
\arguments{
\item{object}{Object.}

\item{n}{The number of breaks to create.}

\item{annotationCol}{\emph{Optional}. \link{data.frame} that defines annotation
mappings for the columns.}

\item{color}{Colors to use for plot. Defaults to the \code{\link[viridis:viridis]{viridis::viridis()}}
palette.}

\item{legendColor}{Colors to use for legend labels. Defaults to the
\code{\link[viridis:viridis]{viridis::viridis()}} palette.}

\item{title}{\emph{Optional}. Plot title.}
}
\value{
Show heatmap. Invisibly return \link{list} containing breaks and
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}} \code{gtable}.
}
\description{
Plot Heatmap with Quantile Breaks
}
\details{
This is helpful for more usefully visualizing single cell data.
Ideas and code from: http://slowkow.com/notes/heatmap-tutorial/
}
\examples{
load(system.file(
    file.path("extdata", "filtered.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
plotQuantileHeatmap(filtered)
}
\author{
Rory Kirchner, Michael Steinbaugh
}
