% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotQuantileHeatmap.R
\docType{methods}
\name{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap,bcbioSingleCell-method}
\alias{plotQuantileHeatmap,seurat-method}
\title{Plot Heatmap with Quantile Breaks}
\usage{
\S4method{plotQuantileHeatmap}{bcbioSingleCell}(object, n = 10,
  annotationCol = NA, clusterRows = FALSE, clusterCols = FALSE,
  color = viridis::viridis)

\S4method{plotQuantileHeatmap}{seurat}(object, n = 10, annotationCol = NA,
  clusterRows = FALSE, clusterCols = FALSE, color = viridis::viridis)
}
\arguments{
\item{object}{Matrix of data.}

\item{n}{The number of breaks to create.}

\item{annotationCol}{Column annotations \link{data.frame}.}

\item{clusterRows}{Perform row clustering.}

\item{clusterCols}{Perform column clustering.}

\item{color}{Color palette function.}
}
\value{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
}
\description{
Plot Heatmap with Quantile Breaks
}
\details{
This is helpful for more usefully visualizing single cell data.
Ideas and code from: http://slowkow.com/notes/heatmap-tutorial/
}
\examples{
load(system.file(
    file.path("extdata", "filtered.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
plotQuantileHeatmap(filtered)
}
\author{
Rory Kirchner, Michael Steinbaugh
}
