% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotQuantileHeatmap.R
\docType{methods}
\name{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap,seurat-method}
\title{Plot Heatmap with Quantile Breaks}
\usage{
\S4method{plotQuantileHeatmap}{seurat}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments must be formatted in camel case, not snake case.}
}
\value{
Show heatmap and invisibly return a \code{list} of the components.
}
\description{
Plot Heatmap with Quantile Breaks
}
\details{
This is helpful for visualizing single cell data.
Ideas and code from: http://slowkow.com/notes/heatmap-tutorial/
}
\examples{
# bcbioSingleCell ====
plotQuantileHeatmap(
    object = bcb_small,
    showColnames = FALSE,
    showRownames = FALSE
)

# seurat ====
plotQuantileHeatmap(
    object = seurat_small,
    showColnames = FALSE,
    showRownames = FALSE
)
}
\seealso{
Other Quality Control Functions: \code{\link{inflectionPoint}},
  \code{\link{plotCellCounts}},
  \code{\link{plotCumulativeUMIsPerCell}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
