% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotQuantileHeatmap.R
\docType{methods}
\name{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap,dgCMatrix-method}
\alias{plotQuantileHeatmap,matrix-method}
\alias{plotQuantileHeatmap,seurat-method}
\title{Plot Heatmap with Quantile Breaks}
\usage{
plotQuantileHeatmap(object, ...)

\S4method{plotQuantileHeatmap}{dgCMatrix}(object, ...)

\S4method{plotQuantileHeatmap}{matrix}(object, annotation = NA,
  clusterRows = TRUE, clusterCols = TRUE)

\S4method{plotQuantileHeatmap}{seurat}(object, annotation = NA,
  clusterRows = TRUE, clusterCols = TRUE)
}
\arguments{
\item{object}{Matrix of data.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{annotation}{Column annotations.}

\item{clusterRows}{Perform row clustering.}

\item{clusterCols}{Perform column clustering.}
}
\value{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
}
\description{
Plot Heatmap with Quantile Breaks
}
\details{
This is helpful for more usefully visualizing single cell data.
Ideas and code from: http://slowkow.com/notes/heatmap-tutorial/
}
\author{
Rory Kirchner
}
