% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotQuantileHeatmap.R
\docType{methods}
\name{plotQuantileHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotQuantileHeatmap,seurat-method}
\title{Plot Heatmap with Quantile Breaks}
\usage{
\S4method{plotQuantileHeatmap}{seurat}(object, n = 5L, annotationCol = NA,
  clusterCols = FALSE, clusterRows = FALSE, color = viridis,
  legendColor = viridis, title = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{n}{The number of breaks to create.}

\item{annotationCol}{\emph{Optional.} \code{data.frame} that defines annotation
mappings for the columns.}

\item{clusterCols}{Logical determining if columns should be arranged with
hierarchical clustering. Alternatively, can define an \code{hclust} object.}

\item{clusterRows}{Logical determining if rows should be arranged with
hierarchical clustering. Alternatively, can define an \code{hclust} object.}

\item{color}{Colors to use for plot. Defaults to the \code{\link[=viridis]{viridis()}}
palette.}

\item{legendColor}{Colors to use for legend labels. Defaults to the
\code{\link[=viridis]{viridis()}} palette.}

\item{title}{\emph{Optional.} Plot title.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.}
}
\value{
Show heatmap. Invisibly return \code{list} containing breaks and \code{gtable}.
}
\description{
Plot Heatmap with Quantile Breaks
}
\details{
This is helpful for visualizing single cell data.
Ideas and code from: http://slowkow.com/notes/heatmap-tutorial/
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat_small.rda", package = "bcbioSingleCell"))

# bcbioSingleCell ====
plotQuantileHeatmap(bcb_small)

# seurat ====
plotQuantileHeatmap(seurat_small)
}
\seealso{
Other Quality Control Functions: \code{\link{calculateMetrics}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
