% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotReadsPerCell-methods.R
\docType{methods}
\name{plotReadsPerCell}
\alias{plotReadsPerCell}
\alias{plotReadsPerCell,bcbioSingleCell-method}
\title{Plot Read Counts per Cell}
\usage{
plotReadsPerCell(object, ...)

\S4method{plotReadsPerCell}{bcbioSingleCell}(object,
  interestingGroups = NULL, geom = c("histogram", "ecdf", "violin",
  "ridgeline", "boxplot"), cutoffLine = FALSE,
  color = getOption("basejump.discrete.color", NULL),
  fill = getOption("basejump.discrete.fill", NULL),
  title = "reads per cell")
}
\arguments{
\item{cutoffLine}{\code{boolean}. Include a line marking the cutoff.}
}
\value{
\code{ggplot}.
}
\description{
Plot the distribution of read counts for all unfiltered cellular barcodes.
}
\examples{
data(indrops_small)
plotReadsPerCell(indrops_small, geom = "histogram")
plotReadsPerCell(indrops_small, geom = "ecdf")
}
\author{
Michael Steinbaugh, Rory Kirchner
}
