% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReadsPerCell-methods.R
\docType{methods}
\name{plotReadsPerCell}
\alias{plotReadsPerCell}
\alias{plotReadsPerCell,bcbioSingleCell-method}
\title{Plot read counts per cell}
\usage{
\S4method{plotReadsPerCell}{bcbioSingleCell}(object,
  interestingGroups = NULL, geom = c("histogram", "ecdf", "violin",
  "ridgeline", "boxplot"), cutoffLine = FALSE,
  color = getOption("basejump.color.discrete", NULL),
  fill = getOption("basejump.fill.discrete", NULL),
  title = "reads per cell")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{cutoffLine}{\code{logical(1)}.
Include a line marking the cutoff.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(basejump.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot the distribution of read counts for all unfiltered cellular barcodes.
}
\examples{
data(indrops)
plotReadsPerCell(indrops, geom = "histogram")
plotReadsPerCell(indrops, geom = "ecdf")
}
\author{
Michael Steinbaugh, Rory Kirchner
}
