% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotReadsPerCell.R
\docType{methods}
\name{plotReadsPerCell}
\alias{plotReadsPerCell}
\alias{plotReadsPerCell,bcbioSingleCellANY-method}
\title{Plot Read Counts per Cell}
\usage{
plotReadsPerCell(object, ...)

\S4method{plotReadsPerCell}{bcbioSingleCellANY}(object, interestingGroup,
  filterCells = FALSE, aggregateReplicates = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroup}{Interesting group, to use for colors.}

\item{filterCells}{Show only the cells that have passed filtering cutoffs.}

\item{aggregateReplicates}{Aggregate technical replicates, if present. If
\code{TRUE}, this function uses the values slotted in
\code{sampleMetadata(object)[["sampleNameAggregate"]])}.

[["sampleNameAggregate"]: R:[%22sampleNameAggregate%22}
}
\value{
\link{ggplot}.
}
\description{
Plot Read Counts per Cell
}
\details{
A violin plot is a comact display of a continuous distribution. It
is a blend of \code{\link[=geom_boxplot]{geom_boxplot()}} and \code{\link[=geom_density]{geom_density()}}: a violin plot is a
mirrored density plot displayed in the same way as a boxplot.
}
\note{
Here by "cell" we mean "cellular barcode".
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Metrics}
