% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotReadsPerCell.R
\docType{methods}
\name{plotReadsPerCell}
\alias{plotReadsPerCell}
\alias{plotReadsPerCell}
\alias{plotReadsPerCell,bcbioSingleCell-method}
\alias{plotReadsPerCell,seurat-method}
\title{Plot Read Counts per Cell}
\usage{
plotReadsPerCell(object, ...)

\S4method{plotReadsPerCell}{bcbioSingleCell}(object, geom = "histogram",
  interestingGroups)

\S4method{plotReadsPerCell}{seurat}(object, geom = "histogram",
  interestingGroups)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{geom}{Plot type. Supported formats: proportional \code{histogram}
(\emph{recommended}), raw \code{ridgeline}, and raw \code{violin}.}

\item{interestingGroups}{Interesting group, to use for colors.}
}
\value{
\code{ggplot}.
}
\description{
Plot Read Counts per Cell
}
\note{
Here by "cell" we mean "cellular barcode".
}
\examples{
# bcbioSingleCell ====
plotReadsPerCell(bcb_small)

# seurat ====
plotReadsPerCell(seurat_small)
}
\seealso{
Other Quality Control Functions: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
