% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotTopMarkers-methods.R
\docType{methods}
\name{plotTopMarkers}
\alias{plotTopMarkers}
\alias{plotTopMarkers,seurat,grouped_df-method}
\title{Plot Top Markers}
\usage{
plotTopMarkers(x, y, ...)

\S4method{plotTopMarkers}{seurat,grouped_df}(x, y, markdown = TRUE)
}
\arguments{
\item{x}{Primary object.}

\item{y}{Top markers grouped \link{tibble} returned by \code{\link[=topMarkers]{topMarkers()}}.}

\item{...}{Additional arguments.}

\item{markdown}{Print Markdown headers.}
}
\value{
\link{ggplot}.
}
\description{
Plot Top Markers
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotClusters}},
  \code{\link{plotKnownMarkers}},
  \code{\link{plotStressGenes}}, \code{\link{topBarcodes}},
  \code{\link{topMarkers}}
}
