% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotTopMarkers.R
\docType{methods}
\name{plotTopMarkers}
\alias{plotTopMarkers}
\alias{plotTopMarkers}
\alias{plotTopMarkers,seurat-method}
\title{Plot Top Markers}
\usage{
plotTopMarkers(object, ...)

\S4method{plotTopMarkers}{seurat}(object, topMarkers, headerLevel = 2L, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{topMarkers}{\code{grouped_df} grouped by "\code{cluster}", returned by
\code{\link[=topMarkers]{topMarkers()}}.}

\item{headerLevel}{R Markdown header level.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Plot Top Markers
}
\note{
The number of markers to plot is determined by the output of the
\code{\link[=topMarkers]{topMarkers()}} function. If you want to reduce the number of genes to plot,
simply reassign first using that function. If necessary, we can add support
for the number of genes to plot here in a future update.
}
\examples{
# seurat ====
plotTopMarkers(
    object = seurat_small,
    topMarkers = topMarkers(all_markers_small, n = 1L)
)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotMarker}},
  \code{\link{plotPCElbow}}, \code{\link{plotTSNE}},
  \code{\link{sanitizeMarkers}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
