% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotTopMarkers.R
\docType{methods}
\name{plotTopMarkers}
\alias{plotTopMarkers}
\alias{plotTopMarkers,seurat,grouped_df-method}
\title{Plot Top Markers}
\usage{
plotTopMarkers(object, topMarkers, ...)

\S4method{plotTopMarkers}{seurat,grouped_df}(object, topMarkers,
  headerLevel = 2)
}
\arguments{
\item{object}{Object.}

\item{topMarkers}{Top markers grouped \link{tibble} returned by \code{\link[=topMarkers]{topMarkers()}}.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{headerLevel}{Include a Markdown header for each gene.}
}
\value{
No value, only graphical output.
}
\description{
Plot Top Markers
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{pcCutoff}}, \code{\link{plotFeatures}},
  \code{\link{plotKnownMarkers}},
  \code{\link{plotMarkers}}, \code{\link{plotStressGenes}},
  \code{\link{topBarcodes}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Utilities}
