% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotTopMarkers.R
\docType{methods}
\name{plotTopMarkers}
\alias{plotTopMarkers}
\alias{plotTopMarkers}
\alias{plotTopMarkers,seurat,grouped_df-method}
\title{Plot Top Markers}
\usage{
plotTopMarkers(object, topMarkers, ...)

\S4method{plotTopMarkers}{seurat,grouped_df}(object, topMarkers,
  tsneColor = scale_color_viridis(discrete = FALSE),
  violinFill = scale_fill_viridis(discrete = TRUE),
  dotColor = scale_color_gradient(low = "lightgray", high = "purple"),
  dark = TRUE, pointsAsNumbers = FALSE, headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{topMarkers}{\code{grouped_df} grouped by "\code{cluster}", returned by
\code{\link[=topMarkers]{topMarkers()}}.}

\item{...}{Additional arguments.}

\item{tsneColor}{Color palette to use for tSNE plot.}

\item{violinFill}{Color palette to use for violin plot.}

\item{dotColor}{Color palette to use for dot plot.}

\item{dark}{Plot against a dark background using \code{\link[basejump:midnightTheme]{basejump::midnightTheme()}}.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{headerLevel}{R Markdown header level.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Plot Top Markers
}
\note{
The number of markers to plot is determined by the output of the
\code{\link[=topMarkers]{topMarkers()}} function. If you want to reduce the number of genes to plot,
simply reassign first using that function. If necessary, we can add support
for the number of genes to plot here in a future update.
}
\examples{
# seurat, grouped_df ====
plotTopMarkers(
    object = seurat_small,
    topMarkers = topMarkers(all_markers_small, n = 1L)
)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotMarkers}},
  \code{\link{plotPCA}}, \code{\link{plotPCElbow}},
  \code{\link{plotTSNE}}, \code{\link{sanitizeMarkers}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
