% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotTopMarkers.R
\docType{methods}
\name{plotTopMarkers}
\alias{plotTopMarkers}
\alias{plotTopMarkers}
\alias{plotTopMarkers,seurat,grouped_df-method}
\title{Plot Top Markers}
\usage{
plotTopMarkers(object, topMarkers, ...)

\S4method{plotTopMarkers}{seurat,grouped_df}(object, topMarkers,
  tsneColor = viridis::scale_color_viridis(),
  violinFill = viridis::scale_fill_viridis(discrete = TRUE),
  dotColor = ggplot2::scale_color_gradient(low = "lightgray", high =
  "purple"), dark = TRUE, pointsAsNumbers = FALSE, headerLevel = 2)
}
\arguments{
\item{object}{Object.}

\item{topMarkers}{Top markers \link{tibble} grouped by cluster, returned by
\code{\link[=topMarkers]{topMarkers()}}.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{tsneColor}{Color palette to use for tSNE plot.}

\item{violinFill}{Color palette to use for violin plot.}

\item{dotColor}{Color palette to use for dot plot.}

\item{dark}{Plot the TSNE against a dark background.}

\item{pointsAsNumbers}{Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{headerLevel}{Include a Markdown header for each gene.}
}
\value{
No value, only graphical output.
}
\description{
Plot Top Markers
}
\note{
The number of markers to plot is determined by the output of the
\code{\link[=topMarkers]{topMarkers()}} function. If you want to reduce the number of genes to plot,
simply reassign first using that function. If necessary, we can add support
for the number of genes to plot here in a future update.
}
\examples{
load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))
load(system.file(
    file.path("extdata", "topMarkers.rda"),
    package = "bcbioSingleCell"))

# seurat, grouped_df
# Let's plot the top 2 markers from cluster 0, as a quick example
plotTopMarkers(seurat, topMarkers[1:2, ])
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkers}}, \code{\link{plotPCElbow}},
  \code{\link{topBarcodes}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
