% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotTSNE.R
\docType{methods}
\name{plotTSNE}
\alias{plotTSNE}
\alias{plotTSNE,seurat-method}
\title{Plot t-SNE}
\usage{
plotTSNE(object, ...)

\S4method{plotTSNE}{seurat}(object, interestingGroups = "ident",
  label = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroups}{Interesting group to use for plot colors.}

\item{label}{Label the clusters on the plot.}
}
\value{
\link{ggplot}.
}
\description{
Generate a t-distributed stochastic neighbor embedding (t-SNE) plot.
}
\examples{
\dontrun{
plotTSNE(seurat)
}
}
\seealso{
Other t-SNE Utilities: \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{plotMarkerTSNE}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{t-SNE Utilities}
