% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotTSNEExpressionData.R
\docType{methods}
\name{plotTSNEExpressionData}
\alias{plotTSNEExpressionData}
\title{Plot Fetched t-SNE Expression Data}
\usage{
plotTSNEExpressionData(object, ...)

\S4method{plotTSNEExpressionData}{grouped_df}(object, colorpoints = "geomean")
}
\arguments{
\item{object}{\link{data.frame} returned from \code{\link[=fetchTSNEExpressionData]{fetchTSNEExpressionData()}}.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{colorpoints}{Color points by geometric mean or expression of
individual gene.}
}
\value{
t-SNE \link{ggplot}.
}
\description{
Plot Fetched t-SNE Expression Data
}
\examples{
\dontrun{
data(markers, seurat)
top <- topMarkers(markers)

# Let's take the top markers specific to cluster 0, as an example
cluster0 <- top \%>\%
    filter(cluster == 0) \%>\%
    pull(symbol)

# Fetch the t-SNE expression data for the desired gene symbols
dat <- fetchTSNEExpressionData(seurat, cluster0)
print(unique(dat$gene))

# To make t-SNE plot colored by geometric mean of topGenes
plotTSNEExpressionData(dat, colorpoints = "geomean")

# To make faceted t-SNE plot of each gene (looks good at up to 6 genes)
plotTSNEExpressionData(dat, colorpoints = "expression") +
    facet_wrap(~gene)
}
}
\seealso{
Other t-SNE Utilities: \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{t-SNE Utilities}
