% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMIsPerCell-methods.R
\docType{methods}
\name{plotUMIsPerCell}
\alias{plotUMIsPerCell}
\alias{plotUMIsPerCell,SingleCellExperiment-method}
\title{Plot UMIs per Cell}
\usage{
\S4method{plotUMIsPerCell}{SingleCellExperiment}(object,
  geom = c("histogram", "ecdf", "violin", "ridgeline", "boxplot"),
  interestingGroups = NULL, min = 0L, max = Inf, point = c("none",
  "inflection", "knee"), trans = "log10",
  color = getOption("basejump.color.discrete", NULL),
  fill = getOption("basejump.fill.discrete", NULL),
  title = "UMIs per cell")
}
\arguments{
\item{object}{Object.}

\item{geom}{\code{string}. Plot type. Uses \code{base::match.arg()} and defaults to the
first argument in the \code{character} vector.}

\item{interestingGroups}{\code{character}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{min}{\code{scalar numeric}. Recommended minimum value cutoff.}

\item{max}{\code{scalar numeric}. Recommended maximum value cutoff.}

\item{point}{\code{string}. Label either the "\code{knee}" or "\code{inflection}" points per
sample. To disable, use "\code{none}". Requires \code{geom = "ecdf"}.}

\item{trans}{\code{string}. Name of the axis scale transformation to apply. See
\code{help("scale_x_continuous", "ggplot2")} for more information.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{ggplot2::scale_color_manual()}. To set the discrete color
palette globally, use
\code{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())}.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}. Desired ggplot2 fill scale. Must
supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{ggplot2::scale_fill_manual()}. To set the discrete fill palette
globally, use \code{options(basejump.fill.discrete = scale_fill_viridis_d())}.}

\item{title}{\code{string}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot the universal molecular identifiers (UMIs) per cell.
}
\examples{
data(indrops)
plotUMIsPerCell(indrops, geom = "violin")
plotUMIsPerCell(indrops, geom = "ridgeline")
plotUMIsPerCell(indrops, geom = "ecdf")
plotUMIsPerCell(indrops, geom = "histogram")
plotUMIsPerCell(indrops, geom = "boxplot")
}
\author{
Michael Steinbaugh, Rory Kirchner
}
