% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMIsPerCell-methods.R
\docType{methods}
\name{plotUMIsPerCell}
\alias{plotUMIsPerCell}
\alias{plotUMIsPerCell,bcbioSingleCell-method}
\title{Plot UMIs per cell}
\usage{
plotUMIsPerCell(object, ...)

\S4method{plotUMIsPerCell}{bcbioSingleCell}(object, geom = c("histogram",
  "ecdf", "violin", "ridgeline", "boxplot"), interestingGroups = NULL,
  min = 0L, max = Inf, point = c("none", "inflection", "knee"),
  trans = "log10", color = getOption(x = "acid.color.discrete", default
  = acidplots::scale_colour_synesthesia_d()), fill = getOption(x =
  "acid.fill.discrete", default = acidplots::scale_fill_synesthesia_d()),
  title = "UMIs per cell")
}
\arguments{
\item{object}{Object.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{min}{\code{numeric(1)}.
Recommended minimum value cutoff.}

\item{max}{\code{numeric(1)}.
Recommended maximum value cutoff.}

\item{point}{\code{character(1)}.
Label either the "\code{knee}" or "\code{inflection}" points per sample. To disable,
use "\code{none}". Requires \code{geom = "ecdf"}.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{title}{\code{character(1)}.
Plot title.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot the universal molecular identifiers (UMIs) per cell.
}
\note{
Updated 2019-07-24.
}
\examples{
data(indrops)
plotUMIsPerCell(indrops, geom = "violin")
plotUMIsPerCell(indrops, geom = "ridgeline")
plotUMIsPerCell(indrops, geom = "ecdf")
plotUMIsPerCell(indrops, geom = "histogram")
plotUMIsPerCell(indrops, geom = "boxplot")
}
\author{
Michael Steinbaugh, Rory Kirchner
}
