% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotUMIsPerCell.R
\docType{methods}
\name{plotUMIsPerCell}
\alias{plotUMIsPerCell}
\alias{plotUMIsPerCell,bcbioSingleCell-method}
\title{Plot UMIs per Cell}
\usage{
plotUMIsPerCell(object, ...)

\S4method{plotUMIsPerCell}{bcbioSingleCell}(object, interestingGroup, min,
  filterCells = FALSE, aggregateReplicates = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroup}{Interesting group, to use for colors.}

\item{min}{Recommended minimum value cutoff.}

\item{filterCells}{Show only the cells that have passed filtering cutoffs.}

\item{aggregateReplicates}{Aggregate technical replicates, if specified. This
function uses values assigned in the \code{sampleNameAggregate} column of the
internal sample metadata \link{data.frame}.}
}
\value{
\link{ggplot} grid.
}
\description{
Plot the universal molecular identifiers (UMIs) per cell.
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Metrics}
