% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotUMIsPerCell.R
\docType{methods}
\name{plotUMIsPerCell}
\alias{plotUMIsPerCell}
\alias{plotUMIsPerCell}
\alias{plotUMIsPerCell,bcbioSCDataSet-method}
\alias{plotUMIsPerCell,bcbioSCFiltered-method}
\title{Plot UMIs per Cell}
\usage{
plotUMIsPerCell(object, ...)

\S4method{plotUMIsPerCell}{bcbioSCDataSet}(object, min = 1000L)

\S4method{plotUMIsPerCell}{bcbioSCFiltered}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{min}{Recommended minimum value cutoff.}
}
\value{
\link{ggplot} grid.
}
\description{
Plot the universal molecular identifiers (UMIs) per cell.
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
