% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotUMIsPerCell.R
\docType{methods}
\name{plotUMIsPerCell}
\alias{plotUMIsPerCell}
\alias{plotUMIsPerCell}
\alias{plotUMIsPerCell,SingleCellExperiment-method}
\alias{plotUMIsPerCell,seurat-method}
\title{Plot UMIs per Cell}
\usage{
plotUMIsPerCell(object, ...)

\S4method{plotUMIsPerCell}{SingleCellExperiment}(object, geom = c("ecdf",
  "histogram", "ridgeline", "boxplot", "violin"), interestingGroups, min = 0L,
  barcodeRanks = FALSE, ranksPerSample = TRUE, labelPoint = c("knee",
  "inflection"), trans = "log10", color = scale_color_viridis(discrete =
  TRUE), fill = scale_fill_viridis(discrete = TRUE))

\S4method{plotUMIsPerCell}{seurat}(object, geom = c("ecdf", "histogram",
  "ridgeline", "boxplot", "violin"), interestingGroups, min = 0L,
  barcodeRanks = FALSE, ranksPerSample = TRUE, labelPoint = c("knee",
  "inflection"), trans = "log10", color = scale_color_viridis(discrete =
  TRUE), fill = scale_fill_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{geom}{Plot type. Uses \code{\link[=match.arg]{match.arg()}} internally and defaults to the
first argument in the vector.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in camel case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{min}{Recommended minimum value cutoff.}

\item{barcodeRanks}{Calculate barcode ranks to label knee or inflection
points.}

\item{ranksPerSample}{Calculate the barcode ranks individually per sample.
Generally recommended unless there's a reason to use a single cutoff point
across all samples in the analysis.}

\item{labelPoint}{Label either the "\code{knee}" or "\code{inflection}" point.}

\item{trans}{Name of the axis scale transformation to apply. See
\code{help("scale_x_continuous", "ggplot2")} for more information.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{fill}{Desired ggplot fill scale. Must supply discrete values. When set
to \code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}
}
\value{
\code{ggplot}.
}
\description{
Plot the universal molecular identifiers (UMIs) per cell.
}
\examples{
# bcbioSingleCell ====
plotUMIsPerCell(bcb_small)

# Label the barcode ranks
plotUMIsPerCell(bcb_small, barcodeRanks = TRUE)

# SingleCellExperiment ====
plotUMIsPerCell(cellranger_small)

# seurat ====
plotUMIsPerCell(Seurat::pbmc_small)
}
\seealso{
Other Quality Control Functions: \code{\link{barcodeRanksPerSample}},
  \code{\link{barcodeRanks}}, \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
