% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotUMIsPerCell.R
\docType{methods}
\name{plotUMIsPerCell}
\alias{plotUMIsPerCell}
\alias{plotUMIsPerCell}
\alias{plotUMIsPerCell,bcbioSingleCell-method}
\alias{plotUMIsPerCell,seurat-method}
\title{Plot UMIs per Cell}
\usage{
plotUMIsPerCell(object, ...)

\S4method{plotUMIsPerCell}{bcbioSingleCell}(object, geom = c("violin",
  "boxplot", "histogram", "ridgeline"), interestingGroups, min,
  fill = scale_fill_viridis(discrete = TRUE))

\S4method{plotUMIsPerCell}{seurat}(object, geom = c("violin", "boxplot",
  "histogram", "ridgeline"), interestingGroups, min,
  fill = scale_fill_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{geom}{Plot type. Uses \code{\link[=match.arg]{match.arg()}} internally and defaults to the
first argument in the vector.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in camel case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{min}{Recommended minimum value cutoff.}

\item{fill}{Desired ggplot fill scale. Must supply discrete values. When set
to \code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}
}
\value{
\code{ggplot}.
}
\description{
Plot the universal molecular identifiers (UMIs) per cell.
}
\examples{
# bcbioSingleCell ====
plotUMIsPerCell(bcb_small)

# seurat ====
plotUMIsPerCell(seurat_small)
plotUMIsPerCell(pbmc_small)
}
\seealso{
Other Quality Control Functions: \code{\link{inflectionPoint}},
  \code{\link{plotCellCounts}},
  \code{\link{plotCumulativeUMIsPerCell}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotQuantileHeatmap}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
