% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotUMIsPerCell.R
\docType{methods}
\name{plotUMIsPerCell}
\alias{plotUMIsPerCell}
\alias{plotUMIsPerCell,bcbioSingleCell-method}
\alias{plotUMIsPerCell,data.frame-method}
\alias{plotUMIsPerCell,seurat-method}
\title{Plot UMIs per Cell}
\usage{
plotUMIsPerCell(object, ...)

\S4method{plotUMIsPerCell}{bcbioSingleCell}(object, geom = "boxplot", min,
  interestingGroups, filterCells = TRUE, aggregateReplicates = TRUE,
  samplesOnYAxis = TRUE, fill = scale_fill_viridis(discrete = TRUE))

\S4method{plotUMIsPerCell}{data.frame}(object, geom = "boxplot", min = 0,
  interestingGroups, multiplexed = FALSE, samplesOnYAxis = TRUE,
  fill = scale_fill_viridis(discrete = TRUE))

\S4method{plotUMIsPerCell}{seurat}(object, geom = "boxplot", min,
  interestingGroups, multiplexed = FALSE, samplesOnYAxis = TRUE,
  fill = scale_fill_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{geom}{Plot type. Supported formats: \code{boxplot}, \code{histogram},
\code{ridgeline}, and \code{violin}.}

\item{min}{Recommended minimum value cutoff.}

\item{interestingGroups}{Interesting group, to use for colors.}

\item{filterCells}{Show only the cells that have passed filtering cutoffs.}

\item{aggregateReplicates}{Aggregate technical replicates, if specified. This
function uses values assigned in the \code{sampleNameAggregate} column of the
internal sample metadata \link{data.frame}.}

\item{samplesOnYAxis}{Plot the samples on the y axis. Doesn't apply to
histogram.}

\item{fill}{Desired ggplot fill scale. Defaults to
\code{\link[viridis:scale_fill_viridis]{viridis::scale_fill_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{multiplexed}{Whether the FASTQ files are multiplexed. This is
detected automatically for \link{bcbioSingleCell} objects.}
}
\value{
\link{ggplot}.
}
\description{
Plot the universal molecular identifiers (UMIs) per cell.
}
\details{
A violin plot is a comact display of a continuous distribution. It
is a blend of \code{\link[=geom_boxplot]{geom_boxplot()}} and \code{\link[=geom_density]{geom_density()}}: a violin plot is a
mirrored density plot displayed in the same way as a boxplot.
}
\examples{
# bcbioSingleCell
\dontrun{
plotUMIsPerCell(bcb)
}

# seurat
\dontrun{
plotUMIsPerCell(seurat)}

# data.frame
\dontrun{
metrics <- metrics(bcb)
plotUMIsPerCell(metrics)
}
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsVsGenes}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Metrics}
