% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotUMIsVsGenes.R
\docType{methods}
\name{plotUMIsVsGenes}
\alias{plotUMIsVsGenes}
\alias{plotUMIsVsGenes}
\alias{plotUMIsVsGenes,bcbioSCDataSet-method}
\alias{plotUMIsVsGenes,bcbioSCSubset-method}
\title{Plot UMI and Gene Correlation}
\usage{
plotUMIsVsGenes(object, ...)

\S4method{plotUMIsVsGenes}{bcbioSCDataSet}(object)

\S4method{plotUMIsVsGenes}{bcbioSCSubset}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{ggplot}.
}
\description{
Plot UMI and Gene Correlation
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotCellularBarcodes}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotUMIsPerCell}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
