% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotUMIsVsGenes.R
\docType{methods}
\name{plotUMIsVsGenes}
\alias{plotUMIsVsGenes}
\alias{plotUMIsVsGenes}
\alias{plotUMIsVsGenes,bcbioSingleCell-method}
\alias{plotUMIsVsGenes,seurat-method}
\title{Plot UMI and Gene Correlation}
\usage{
plotUMIsVsGenes(object, ...)

\S4method{plotUMIsVsGenes}{bcbioSingleCell}(object, interestingGroups,
  color = scale_color_viridis(discrete = TRUE))

\S4method{plotUMIsVsGenes}{seurat}(object, interestingGroups,
  color = scale_color_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in \code{\link[=camel]{camel()}} case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}
}
\description{
Plot UMI and Gene Correlation
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat_small.rda", package = "bcbioSingleCell"))

# bcbioSingleCell ====
plotUMIsVsGenes(bcb_small)

# seurat ====
plotUMIsVsGenes(pbmc_small)
plotUMIsVsGenes(seurat_small)

# data.frame
df <- metrics(bcb)
plotUMIsPerCell(df)
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
