% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotUMIsVsGenes.R
\docType{methods}
\name{plotUMIsVsGenes}
\alias{plotUMIsVsGenes}
\alias{plotUMIsVsGenes,bcbioSingleCell-method}
\alias{plotUMIsVsGenes,data.frame-method}
\alias{plotUMIsVsGenes,seurat-method}
\title{Plot UMI and Gene Correlation}
\usage{
plotUMIsVsGenes(object, ...)

\S4method{plotUMIsVsGenes}{bcbioSingleCell}(object, interestingGroups,
  filterCells = TRUE, aggregateReplicates = TRUE, samplesOnYAxis = TRUE,
  color = scale_color_viridis(discrete = TRUE))

\S4method{plotUMIsVsGenes}{data.frame}(object, interestingGroups,
  multiplexed = FALSE, color = scale_color_viridis(discrete = TRUE))

\S4method{plotUMIsVsGenes}{seurat}(object, interestingGroups,
  multiplexed = FALSE, color = scale_color_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroups}{Interesting group, to use for colors.}

\item{filterCells}{Show only the cells that have passed filtering cutoffs.}

\item{aggregateReplicates}{Aggregate technical replicates, if specified. This
function uses values assigned in the \code{sampleNameAggregate} column of the
internal sample metadata \link{data.frame}.}

\item{samplesOnYAxis}{Plot the samples on the y axis. Doesn't apply to
histogram.}

\item{color}{\emph{Only applies to scatterplot}. Desired ggplot color scale.
Defaults to \code{\link[viridis:scale_color_viridis]{viridis::scale_color_viridis()}}. Must supply discrete values.
When set to \code{NULL}, the default ggplot2 color palette will be used. If
manual color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{multiplexed}{Whether the FASTQ files are multiplexed. This is
detected automatically for \link{bcbioSingleCell} objects.}
}
\value{
\link{ggplot}.
}
\description{
Plot UMI and Gene Correlation
}
\examples{
# bcbioSingleCell
\dontrun{
plotUMIsVsGenes(bcb)
}

# seurat
\dontrun{
plotUMIsVsGenes(seurat)
}

# data.frame
\dontrun{
metrics <- metrics(bcb)
plotUMIsPerCell(metrics)
}
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Quality Control Metrics}
