% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotViolin.R
\docType{methods}
\name{plotViolin}
\alias{plotViolin}
\alias{plotViolin,seurat-method}
\title{Plot Violin}
\usage{
\S4method{plotViolin}{seurat}(object, genes,
  fill = viridis::scale_fill_viridis(discrete = TRUE), return = "grid",
  headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{genes}{Gene identifiers to plot.}

\item{fill}{Fill color palette. Defaults to viridis.}

\item{return}{Return type. "grid", "list", and "markdown" are supported.}

\item{headerLevel}{Markdown header level. Only applicable when
\code{return = "markdown"}.}
}
\description{
Plot Violin
}
\examples{
load(system.file("extdata/seurat.rda", package = "bcbioSingleCell"))

genes <- slot(seurat, "raw.data") \%>\% rownames() \%>\% .[1:2]
print(genes)

# grid
plotViolin(seurat, genes = genes, return = "grid")

# list
list <- plotViolin(seurat, genes = genes, return = "list")
glimpse(list)

# markdown
plotViolin(seurat, genes = genes, return = "markdown")
}
\author{
Michael Steinbaugh
}
