% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotViolin.R
\docType{methods}
\name{plotViolin}
\alias{plotViolin}
\alias{plotViolin,seurat-method}
\title{Plot Violin}
\usage{
\S4method{plotViolin}{seurat}(object, genes, format = "symbol",
  fill = viridis::scale_fill_viridis(discrete = TRUE), return = "grid",
  headerLevel = 2)
}
\arguments{
\item{object}{Object.}

\item{genes}{Gene identifiers to plot.}

\item{format}{Gene identifier format. Supports \code{ensgene} or \code{symbol}.}
}
\description{
Plot Violin
}
\examples{

genes <- slot(seurat, "raw.data") \%>\% rownames() \%>\% .[1:2]
print(genes)

# grid
plotViolin(seurat, genes = genes, return = "grid")

# list
list <- plotViolin(seurat, genes = genes, return = "list")
glimpse(list)

# markdown
plotViolin(seurat, genes = genes, return = "markdown")
}
\author{
Michael Steinbaugh
}
