% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotViolin.R
\docType{methods}
\name{plotViolin}
\alias{plotViolin}
\alias{plotViolin,seurat-method}
\title{Plot Violin}
\usage{
\S4method{plotViolin}{seurat}(object, genes,
  fill = viridis::scale_fill_viridis(discrete = TRUE), return = "grid",
  headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{genes}{Gene identifiers to plot.}

\item{fill}{Fill color palette. Defaults to viridis.}

\item{return}{Return type. "grid", "list", and "markdown" are supported.}

\item{headerLevel}{Markdown header level. Only applicable when
\code{return = "markdown"}.}
}
\description{
Plot Violin
}
\examples{
# seurat ====
genes <- slot(pbmc_small, "raw.data") \%>\% rownames() \%>\% head(2L)

# grid
plotViolin(pbmc_small, genes = genes, return = "grid")

# list
list <- plotViolin(pbmc_small, genes = genes, return = "list")
names(list)

# markdown
plotViolin(pbmc_small, genes = genes, return = "markdown")
}
\author{
Michael Steinbaugh
}
