% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotViolin.R
\docType{methods}
\name{plotViolin}
\alias{plotViolin}
\alias{plotViolin,seurat-method}
\title{Plot Violin}
\usage{
\S4method{plotViolin}{seurat}(object, genes,
  fill = scale_fill_viridis(discrete = TRUE), headerLevel = 2L,
  return = c("grid", "list", "markdown"))
}
\arguments{
\item{object}{Object.}

\item{genes}{Gene identifiers to plot.}

\item{fill}{Desired ggplot fill scale. Must supply discrete values. When set
to \code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{headerLevel}{R Markdown header level.}

\item{return}{Return type. Uses \code{\link[=match.arg]{match.arg()}} internally and defaults to the
first argument in the vector.}
}
\description{
Plot Violin
}
\examples{
# seurat ====
genes <- slot(pbmc_small, "raw.data") \%>\% rownames() \%>\% head(2L)

# grid
plotViolin(pbmc_small, genes = genes, return = "grid")

# list
list <- plotViolin(pbmc_small, genes = genes, return = "list")
names(list)

# markdown
plotViolin(pbmc_small, genes = genes, return = "markdown")
}
\author{
Michael Steinbaugh
}
