% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotViolin.R
\docType{methods}
\name{plotViolin}
\alias{plotViolin}
\alias{plotViolin,seurat-method}
\title{Plot Violin}
\usage{
\S4method{plotViolin}{seurat}(object, genes, scale = c("count", "width",
  "area"), fill = scale_fill_hue(), dark = FALSE, headerLevel = 2L,
  return = c("grid", "list", "markdown"))
}
\arguments{
\item{object}{Object.}

\item{genes}{Gene identifiers. Must match the rownames of the object.}

\item{scale}{if "area" (default), all violins have the same area (before trimming
the tails). If "count", areas are scaled proportionally to the number of
observations. If "width", all violins have the same maximum width.}

\item{fill}{Desired ggplot fill scale. Must supply discrete values. When set
to \code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{dark}{Plot against a dark background using
\code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{headerLevel}{R Markdown header level.}

\item{return}{Return type. Uses \code{\link[base:match.arg]{base::match.arg()}} internally and defaults
to the first argument in the vector.}
}
\value{
\code{ggplot}.
}
\description{
Plot Violin
}
\examples{
# seurat ====
genes <- head(rownames(Seurat::pbmc_small), 2L)
print(genes)

# grid
plotViolin(Seurat::pbmc_small, genes = genes, return = "grid")

# list
list <- plotViolin(Seurat::pbmc_small, genes = genes, return = "list")
names(list)

# markdown
plotViolin(Seurat::pbmc_small, genes = genes, return = "markdown")
}
\seealso{
Other Gene Expression Functions: \code{\link{plotDot}}
}
\author{
Michael Steinbaugh
}
