% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotViolin.R
\docType{methods}
\name{plotViolin}
\alias{plotViolin}
\alias{plotViolin,seurat-method}
\title{Plot Violin}
\usage{
\S4method{plotViolin}{seurat}(object, genes, scale = c("count", "width",
  "area"), fill = NULL, legend = FALSE)
}
\arguments{
\item{object}{Object.}

\item{genes}{Gene identifiers. Must match the rownames of the object.}

\item{scale}{if "area" (default), all violins have the same area (before trimming
the tails). If "count", areas are scaled proportionally to the number of
observations. If "width", all violins have the same maximum width.}

\item{fill}{Desired ggplot fill scale. Must supply discrete values. When set
to \code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{legend}{Include plot legend.}
}
\value{
\code{ggplot}.
}
\description{
Plot Violin
}
\examples{
# seurat ====
object <- Seurat::pbmc_small
genes <- head(rownames(object), 4L)
plotViolin(object, genes = genes)
}
\seealso{
\code{\link[Seurat:VlnPlot]{Seurat::VlnPlot()}}.

Other Gene Expression Functions: \code{\link{plotDot}},
  \code{\link{plotGene}}
}
\author{
Michael Steinbaugh
}
