% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotZeroesVsDepth.R
\docType{methods}
\name{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\alias{plotZerosVsDepth,bcbioSingleCell-method}
\title{Plot Percentage of Zeros vs. Library Depth}
\usage{
plotZerosVsDepth(object, ...)

\S4method{plotZerosVsDepth}{bcbioSingleCell}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{ggplot}.
}
\description{
Plot Percentage of Zeros vs. Library Depth
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}}, \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\concept{Quality Control Metrics}
