% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotZerosVsDepth.R
\docType{methods}
\name{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\alias{plotZerosVsDepth,bcbioSingleCell-method}
\alias{plotZerosVsDepth,seurat-method}
\title{Plot Percentage of Zeros vs. Library Depth}
\usage{
plotZerosVsDepth(object, ...)

\S4method{plotZerosVsDepth}{bcbioSingleCell}(object)

\S4method{plotZerosVsDepth}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{metrics}{Metrics \code{data.frame}.}
}
\value{
\code{ggplot}.
}
\description{
This function helps us visualize the dropout rate.
}
\examples{
# bcbioSingleCell ====
plotZerosVsDepth(bcb_small)

# seurat ====
plotZerosVsDepth(pbmc_small)
plotZerosVsDepth(seurat_small)
}
\seealso{
Other Quality Control Functions: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
