% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotZerosVsDepth.R
\docType{methods}
\name{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\alias{plotZerosVsDepth,bcbioSingleCell-method}
\alias{plotZerosVsDepth,dgCMatrix-method}
\alias{plotZerosVsDepth,matrix-method}
\title{Plot Percentage of Zeros vs. Library Depth}
\usage{
plotZerosVsDepth(object, ...)

\S4method{plotZerosVsDepth}{bcbioSingleCell}(object)

\S4method{plotZerosVsDepth}{dgCMatrix}(object, metrics)

\S4method{plotZerosVsDepth}{matrix}(object, metrics)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{metrics}{Metrics \link{data.frame}.}
}
\value{
\link{ggplot}.
}
\description{
This function helps us visualize the dropout rate.
}
\examples{
load(system.file(
    file.path("inst", "extdata", "bcb.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
plotZerosVsDepth(bcb)

# dgCMatrix
counts <- counts(bcb)
metrics <- metrics(bcb)
plotZerosVsDepth(counts, metrics = metrics)
}
\seealso{
Other Quality Control Metrics: \code{\link{metrics}},
  \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}},
  \code{\link{plotUMIsVsGenes}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
