% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotZerosVsDepth.R
\docType{methods}
\name{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\alias{plotZerosVsDepth}
\alias{plotZerosVsDepth,SingleCellExperiment-method}
\alias{plotZerosVsDepth,seurat-method}
\title{Plot Percentage of Zeros vs. Library Depth}
\usage{
plotZerosVsDepth(object, ...)

\S4method{plotZerosVsDepth}{SingleCellExperiment}(object,
  color = scale_color_hue())

\S4method{plotZerosVsDepth}{seurat}(object, color = scale_color_hue())
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}
}
\value{
\code{ggplot}.
}
\description{
This function helps us visualize the dropout rate.
}
\examples{
# bcbioSingleCell ====
plotZerosVsDepth(indrops_small)

# SingleCellExperiment ====
plotZerosVsDepth(cellranger_small)

# seurat ====
plotZerosVsDepth(Seurat::pbmc_small)
}
\seealso{
Other Quality Control Functions: \code{\link{barcodeRanksPerSample}},
  \code{\link{barcodeRanks}}, \code{\link{plotCellCounts}},
  \code{\link{plotGenesPerCell}},
  \code{\link{plotMitoRatio}},
  \code{\link{plotMitoVsCoding}},
  \code{\link{plotNovelty}}, \code{\link{plotQC}},
  \code{\link{plotReadsPerCell}},
  \code{\link{plotUMIsPerCell}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
