% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-prepareSingleCellTemplate.R
\docType{methods}
\name{prepareSingleCellTemplate}
\alias{prepareSingleCellTemplate}
\alias{prepareSingleCellTemplate}
\alias{prepareSingleCellTemplate,missing-method}
\title{Prepare Single-Cell RNA-Seq RMarkdown Template}
\usage{
prepareSingleCellTemplate(object, ...)

\S4method{prepareSingleCellTemplate}{missing}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
No value.
}
\description{
Prepare Single-Cell RNA-Seq RMarkdown Template
}
\examples{
\dontrun{
prepareSingleCellTemplate()
}
}
\seealso{
Other RMarkdown Utilities: \code{\link{metadataTable}}
}
\author{
Michael Steinbaugh
}
