% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSingleCellTemplate.R
\name{prepareSingleCellTemplate}
\alias{prepareSingleCellTemplate}
\title{Prepare Single-Cell RNA-Seq R Markdown Template}
\usage{
prepareSingleCellTemplate(overwrite = FALSE)
}
\arguments{
\item{overwrite}{\code{logical}. Should existing destination files be overwritten?}
}
\value{
Invisible \code{logical} vector indicating which files were copied.
}
\description{
Prepare Single-Cell RNA-Seq R Markdown Template
}
\details{
By default, this function will create local copies of these files:
\itemize{
\item \code{_footer.Rmd}
\item \code{_header.Rmd}
\item \code{_output.yaml}
\item \code{_setup.R}
\item \code{bibliography.bib}
}

This code is used internally by:
\itemize{
\item \code{bcbioRNASeq::prepareRNASeqTemplate()}.
\item \code{bcbioSingleCell::prepareSingleCellTemplate()}.
}
}
\examples{
x <- prepareSingleCellTemplate()
x
}
\seealso{
Other Prepare Functions: \code{\link{prepareSummarizedExperiment}}
}
\author{
Michael Steinbaugh
}
