% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSingleCellTemplate.R
\name{prepareSingleCellTemplate}
\alias{prepareSingleCellTemplate}
\title{Prepare Single-Cell RNA-Seq R Markdown Template}
\usage{
prepareSingleCellTemplate(overwrite = FALSE)
}
\arguments{
\item{overwrite}{\code{boolean}. Should existing destination files be overwritten?}
}
\value{
Invisible \code{logical}. Was the file copied?.
}
\description{
If the required template dependency files aren't present, copy them from the
requested package. Existing files are not overwritten by default. This
function will copy dependency files from a requested package inside the
\code{rmarkdown/shared} directory. If a package doesn't contain this
subdirectory, the function will return an error.
}
\details{
This code is used internally by:
\itemize{
\item \code{bcbioRNASeq::prepareRNASeqTemplate()}.
\item \code{bcbioSingleCell::prepareSingleCellTemplate()}.
}
}
\examples{
x <- prepareSingleCellTemplate()
x

# Clean up
unlink(c(
    "_footer.Rmd",
    "_header.Rmd",
    "_output.yaml",
    "_setup.R",
    "bibliography.bib"
))
}
\seealso{
Other R Markdown Functions: \code{\link{markdownHeader}},
  \code{\link{markdownLink}}, \code{\link{markdownList}},
  \code{\link{markdownPlotlist}},
  \code{\link{markdownTables}}
}
\author{
Michael Steinbaugh
}
\concept{R Markdown Functions}
