% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellularBarcodes-internal.R
\name{.proportionalCB}
\alias{.proportionalCB}
\title{Proportional Cellular Barcodes}
\usage{
.proportionalCB(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{tibble}.
}
\description{
Proportional Cellular Barcodes
}
\details{
Modified version of Klein Lab MATLAB code.
}
\author{
Rory Kirchner, Michael Steinbaugh
}
\keyword{internal}
