% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metrics.R
\name{plot_cell_counts_barplot}
\alias{plot_cell_counts_barplot}
\alias{plot_cell_counts}
\alias{plot_read_counts}
\alias{plot_genes_detected}
\alias{plot_reads_vs_genes}
\alias{plot_mito_ratio}
\alias{plot_novelty}
\title{Cell metrics quality control plots}
\usage{
plot_cell_counts_barplot(run)

plot_cell_counts(run)

plot_read_counts(run, min = 1000)

plot_genes_detected(run, min = 500, max = NULL)

plot_reads_vs_genes(run, intgroup = NULL)

plot_mito_ratio(run, max = 0.2)

plot_novelty(run, min = 0.8)
}
\arguments{
\item{run}{\link{bcbioSCDataSet}.}

\item{min}{Recommended minimum value cutoff.}

\item{max}{Recommended maximum value cutoff.}

\item{intgroup}{Interesting group used to define colors.}
}
\value{
\link{ggplot}.
}
\description{
Novelty score means log10 genes detected per count.
}
\author{
Michael Steinbaugh

Rory Kirchner
}
