% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-quantileHeatmap.R
\docType{methods}
\name{quantileHeatmap}
\alias{quantileHeatmap}
\alias{quantileHeatmap}
\alias{quantileHeatmap,dgCMatrix-method}
\alias{quantileHeatmap,matrix-method}
\alias{quantileHeatmap,seurat-method}
\title{Heatmap with Quantile Breaks}
\usage{
quantileHeatmap(object, ...)

\S4method{quantileHeatmap}{dgCMatrix}(object, ...)

\S4method{quantileHeatmap}{matrix}(object, annotation = NA,
  clusterRows = TRUE, clusterCols = TRUE)

\S4method{quantileHeatmap}{seurat}(object, annotation = NA,
  clusterRows = TRUE, clusterCols = TRUE)
}
\arguments{
\item{object}{Matrix of data.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{annotation}{Column annotations.}

\item{clusterRows}{Perform row clustering.}

\item{clusterCols}{Perform column clustering.}
}
\value{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
}
\description{
Heatmap with Quantile Breaks
}
\details{
This is helpful for more usefully visualizing single cell data.
Ideas and code from: http://slowkow.com/notes/heatmap-tutorial/
}
\author{
Rory Kirchner
}
