% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10x.R
\name{read_10x_counts}
\alias{read_10x_counts}
\title{Read 10x Genomics Cell Ranger output}
\usage{
read_10x_counts(data_dir)
}
\arguments{
\item{data_dir}{Data directory cont}
}
\value{
Sparse counts matrix (dgCMatrix)
}
\description{
Generate a sparse matrix from barcodes.tsv, genes.tsv, and matrix.mtx files
}
\author{
Michael Steinbaugh
}
