% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_10x}
\alias{read_10x}
\title{Read \href{https://www.10xgenomics.com/software/}{10x Genomics Chromium} output}
\usage{
read_10x(data_dir)
}
\arguments{
\item{data_dir}{Data directory}
}
\value{
Sparse counts matrix (dgCMatrix).
}
\description{
Read into sparse matrix from barcodes.tsv, genes.tsv, and matrix.mtx files.
}
\author{
Michael Steinbaugh
}
