% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_barcode_file.R
\name{read_barcode_file}
\alias{read_barcode_file}
\title{Load a barcode histogram file generated from bcbio-nextgen}
\usage{
read_barcode_file(filename)
}
\arguments{
\item{filename}{path to a barcode histogram file}
}
\value{
dataframe of reads per barcode
}
\description{
Load a barcode histogram file generated from bcbio-nextgen
}
\author{
Rory Kirchner
}
\keyword{internal}
