% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_read_barcode_file.R
\name{read_barcode_file}
\alias{read_barcode_file}
\title{Load a barcode histogram file}
\usage{
read_barcode_file(file_name)
}
\arguments{
\item{file_name}{path to a barcode histogram file.}
}
\value{
Data frame of reads per barcode.
}
\description{
Load a barcode histogram file
}
\author{
Rory Kirchner
}
\keyword{internal}
