% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bcbio.R
\name{read_bcbio_metadata}
\alias{read_bcbio_metadata}
\title{Read data from bcbio-nextgen run}
\usage{
read_bcbio_metadata(run, save = FALSE)
}
\arguments{
\item{run}{\code{bcbio-nextgen}}

\item{save}{Save data}
}
\value{
Metadata data frame
}
\description{
Import metadata
}
\author{
Michael Steinbaugh

Rory Kirchner
}
