% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_bcbio_barcodes}
\alias{read_bcbio_barcodes}
\alias{read_bcbio_counts}
\alias{read_bcbio_programs}
\title{Read \href{https://github.com/chapmanb/bcbio-nextgen}{bcbio-nextgen} output}
\usage{
read_bcbio_barcodes(run)

read_bcbio_counts(run, strip_version = TRUE)

read_bcbio_programs(run)
}
\arguments{
\item{run}{\link{bcbioSCDataSet}.}

\item{strip_version}{Strip transcript version from identifier.}
}
\value{
Sparse counts matrix.
}
\description{
Read \href{https://github.com/chapmanb/bcbio-nextgen}{bcbio-nextgen} output

Import transcript-level count data from a bcbio run into a
sparse matrix.
}
\author{
Michael Steinbaugh

Rory Kirchner
}
