% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bcbio.R
\name{read_bcbio_sparsecounts}
\alias{read_bcbio_sparsecounts}
\title{Read data from bcbio-nextgen run}
\usage{
read_bcbio_sparsecounts(run, strip_version = TRUE)
}
\arguments{
\item{run}{\code{bcbio-nextgen}}

\item{strip_version}{Strip transcript version from identifier}

\item{save}{Save data
Import single-cell RNA-seq count data

Import transcript-level count data from a bcbio run into a sparse matrix}

\item{run}{\code{bcbio-nextgen} run}
}
\value{
Sparse counts matrix
}
\description{
Read data from bcbio-nextgen run
}
\author{
Michael Steinbaugh

Rory Kirchner

Rory Kirchner

Michael Steinbaugh
}
