% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bcbio.R
\name{read_bcbio_counts}
\alias{read_bcbio_counts}
\title{Read \href{https://github.com/chapmanb/bcbio-nextgen}{bcbio-nextgen} run
output}
\usage{
read_bcbio_counts(run, strip_version = TRUE)
}
\arguments{
\item{run}{bcbio-nextgen}

\item{strip_version}{Strip transcript version from identifier}
}
\value{
Sparse counts matrix
}
\description{
Import transcript-level count data from a bcbio run into a
  sparse matrix
}
\author{
Michael Steinbaugh

Rory Kirchner
}
