% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bcbio.R
\name{read_sparsecounts}
\alias{read_sparsecounts}
\title{Import scRNA-Seq count data}
\usage{
read_sparsecounts(run)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run}
}
\value{
Sparse counts matrix
}
\description{
Import transcript-level count data from a bcbio run into a sparse matrix
}
\author{
Rory Kirchner

Michael Steinbaugh
}
