% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/readCellRanger.R
\name{loadCellRanger}
\alias{loadCellRanger}
\alias{readCellRanger}
\title{Read 10X Genomics Cell Ranger Data}
\usage{
readCellRanger(uploadDir, refdataDir, sampleMetadataFile,
  interestingGroups = "sampleName", transgeneNames = NULL,
  spikeNames = NULL, prefilter = TRUE, ...)
}
\arguments{
\item{...}{Additional arguments, to be stashed in the \code{\link[=metadata]{metadata()}} slot.}

\item{uploadDir}{Path to Cell Ranger output directory. This directory path
must contain \code{filtered_gene_bc_matrices*} as a child directory.}

\item{refdataDir}{Directory path to Cell Ranger reference annotation data.}

\item{sampleMetadataFile}{Sample barcode metadata file. Optional for runs
with demultiplixed index barcodes (e.g. SureCell), but otherwise required
for runs with multipliexed FASTQs containing multiple index barcodes (e.g.
inDrop).}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in camel case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{transgeneNames}{\code{character} vector indicating which \code{\link[=assay]{assay()}} rows
denote transgenes (e.g. EGFP, TDTOMATO).}

\item{spikeNames}{\code{character} vector indicating which \code{\link[=assay]{assay()}} rows denote
spike-in sequences (e.g. ERCCs).}

\item{prefilter}{Prefilter counts prior to quality control analysis. This
applies very minimal filtering to the dataset, dropping only cells that
don't contain any genes, and is generally recommended.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Read \href{https://www.10xgenomics.com/software/}{10x Genomics Chromium} cell
counts from \code{barcodes.tsv}, \code{genes.tsv}, and \code{matrix.mtx} files.
}
\details{
This function is a variant of the main \code{\link[=bcbioSingleCell]{bcbioSingleCell()}}
constructor, but optimized for handling Cell Ranger output.
}
\note{
Unlike the \code{\link[=bcbioSingleCell]{bcbioSingleCell()}} function, here the \code{organism},
\code{ensemblRelease}, and \code{genomeBuild} are always detected automatically,
based on the 10X \code{refdataDir} YAML metadata. Therefore, these parameters
cannot be set by the user.
}
\examples{
uploadDir <- system.file("extdata/cellranger", package = "bcbioSingleCell")
readCellRanger(
    uploadDir = uploadDir,
    refdataDir = file.path(uploadDir, "refdata-cellranger-hg19-1.2.0"),
    sampleMetadataFile = file.path(uploadDir, "metadata.csv")
)
}
\seealso{
Other Read Functions: \code{\link{readCellTypeMarkers}}
}
\author{
Michael Steinbaugh
}
