% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCellRanger.R
\name{readCellRanger}
\alias{readCellRanger}
\title{Read 10X Genomics Cell Ranger Data}
\usage{
readCellRanger(uploadDir, format = c("mtx", "hdf5"), filtered = TRUE,
  organism = NULL, sampleMetadataFile = NULL, refdataDir = NULL,
  interestingGroups = "sampleName", transgeneNames = NULL,
  spikeNames = NULL, ...)
}
\arguments{
\item{uploadDir}{\code{string}. Path to Cell Ranger output directory. This
directory path must contain \code{filtered_gene_bc_matrices*} as a child
directory.}

\item{format}{\code{string}. Output format, either MatrixMarket ("\code{mtx}") or HDF5
("\code{hdf5}").}

\item{filtered}{\code{boolean}. Use filtered (recommended) or raw counts. Note
that raw counts still contain only whitelisted cellular barcodes.}

\item{organism}{\code{string}. Organism name. Use the full latin name (e.g.
"Homo sapiens"), since this will be input downstream to
AnnotationHub and ensembldb, unless \code{gffFile} is set.}

\item{sampleMetadataFile}{\code{string}. Sample barcode metadata file. Optional
for runs with demultiplixed index barcodes (e.g. SureCell), but otherwise
required for runs with multipliexed FASTQs containing multiple index
barcodes (e.g. inDrop).}

\item{refdataDir}{\code{string} or \code{NULL}. Directory path to Cell Ranger reference
annotation data.}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}

\item{transgeneNames}{\code{character} vector indicating which \code{\link[=assay]{assay()}} rows
denote transgenes (e.g. EGFP, TDTOMATO).}

\item{spikeNames}{\code{character} vector indicating which \code{\link[=assay]{assay()}} rows denote
spike-in sequences (e.g. ERCCs).}

\item{...}{Additional arguments, to be stashed in the \code{\link[=metadata]{metadata()}} slot.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Read \href{https://www.10xgenomics.com/software/}{10x Genomics Chromium} cell
counts from \code{barcodes.tsv}, \code{genes.tsv}, and \code{matrix.mtx} files.
}
\section{Directory structure}{

Cell Ranger can vary in its output directory structure, but we're requiring a
single, consistent directory structure for all datasets, even those that only
contain a single sample:

\preformatted{
file.path(
    "<uploadDir>",
    "<sampleName>",
    "outs",
    "filtered_gene_bc_matrices*",
    "outs",
    "<genomeBuild>",
    "matrix.mtx"
)
}
}

\section{Sample metadata}{

A user-supplied sample metadata file defined by \code{sampleMetadataFile} is
required for multiplexed datasets. Otherwise this can be left \code{NULL}, and
minimal sample data will be used, based on the directory names.
}

\section{Reference data}{

We strongly recommend supplying the corresponding reference data required for
Cell Ranger with the \code{refdataDir} argument. It will convert the gene
annotations defined in the GTF file into a \code{GRanges} object, which get
slotted in \code{\link[=rowRanges]{rowRanges()}}. Otherwise, the function will attempt to use the
most current annotations available from Ensembl, and some gene IDs may not
match, due to deprecation in the current Ensembl release.
}

\examples{
uploadDir <- system.file("extdata/cellranger", package = "bcbioSingleCell")
x <- readCellRanger(
    uploadDir = uploadDir,
    organism = "Homo sapiens"
)
show(x)
}
\seealso{
Other Read Functions: \code{\link{readCellTypeMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Read Functions}
