% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCellTypeMarkers.R
\name{readCellTypeMarkers}
\alias{readCellTypeMarkers}
\title{Read Cell Type Markers File}
\usage{
readCellTypeMarkers(file, gene2symbol)
}
\arguments{
\item{file}{Gene markers file (CSV or Excel).}

\item{gene2symbol}{\code{data.frame} containing gene-to-symbol mappings. Columns
must contain \code{geneID} and \code{geneName}.}
}
\value{
\code{grouped_df}, grouped by \code{cellType}.
}
\description{
Read Cell Type Markers File
}
\examples{
# Homo sapiens
file <- system.file(
    file.path("extdata", "cell_type_markers.csv"),
    package = "bcbioSingleCell"
)
gene2symbol <- makeGene2symbolFromEnsembl("Homo sapiens")
readCellTypeMarkers(file, gene2symbol = gene2symbol)
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{colData}},
  \code{\link{counts}}, \code{\link{fetchGeneData}},
  \code{\link{fetchPCAData}}, \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}}, \code{\link{gene2symbol}},
  \code{\link{interestingGroups}}, \code{\link{rowData}},
  \code{\link{sampleData}}, \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
