% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCellTypeMarkersFile.R
\name{readCellTypeMarkersFile}
\alias{readCellTypeMarkersFile}
\title{Read Cell Type Markers File}
\usage{
readCellTypeMarkersFile(file, gene2symbol)
}
\arguments{
\item{file}{Gene markers file (CSV or Excel).}

\item{gene2symbol}{\code{data.frame} containing gene-to-symbol mappings. Columns
must contain \code{geneID} and \code{geneName}.}
}
\value{
\code{grouped_df}, grouped by \code{cellType}.
}
\description{
Read Cell Type Markers File
}
\examples{
file <- system.file(
    file.path("extdata", "cell_type_markers.csv"),
    package = "bcbioSingleCell"
)
gene2symbol <- gene2symbol("Mus musculus")
readCellTypeMarkersFile(file, gene2symbol = gene2symbol)
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateReplicates}},
  \code{\link{filterCells}}, \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
