% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-readMarkers.R
\docType{methods}
\name{readMarkers}
\alias{readMarkers}
\alias{readMarkers}
\alias{readMarkers,character-method}
\title{Read Known Markers}
\usage{
readMarkers(object, ...)

\S4method{readMarkers}{character}(object, gene2symbol, show = FALSE)
}
\arguments{
\item{object}{Gene markers file (CSV or Excel).}

\item{...}{Additional arguments.}

\item{gene2symbol}{Gene-to-symbol annotation \link{data.frame}.}

\item{show}{Show \link{kable}.}
}
\value{
\link{tibble}.
}
\description{
Read Known Markers
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}},
  \code{\link{filterCells}},
  \code{\link{interestingGroups}},
  \code{\link{sampleMetadata}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
