% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{assignAndSaveData}
\alias{basenameSansExt}
\alias{export}
\alias{markdownHeader}
\alias{saveData}
\alias{set_names}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{assignAndSaveData}}, \code{\link[basejump]{basenameSansExt}}, \code{\link[basejump]{export}}, \code{\link[basejump]{markdownHeader}}, \code{\link[basejump]{saveData}}}

  \item{magrittr}{\code{\link[magrittr]{set_names}}}
}}

