% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plotCountsPerCell}
\alias{plotCountsVsFeatures}
\alias{plotFeaturesPerCell}
\alias{plotMitoRatio}
\alias{plotNovelty}
\alias{plotQC}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{acidplots}{\code{\link[acidplots]{plotCountsPerCell}}, \code{\link[acidplots]{plotCountsVsFeatures}}, \code{\link[acidplots]{plotFeaturesPerCell}}, \code{\link[acidplots]{plotMitoRatio}}, \code{\link[acidplots]{plotNovelty}}, \code{\link[acidplots]{plotQC}}}
}}

