% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeRanksPerSample-methods.R,
%   R/deprecated.R, R/filterCells-methods.R, R/plotBarcodeRanks-methods.R,
%   R/plotCellCounts-methods.R, R/plotGenesPerCell-methods.R,
%   R/plotMitoRatio-methods.R, R/plotMitoVsCoding-methods.R,
%   R/plotNovelty-methods.R, R/plotQC-methods.R, R/plotReadsPerCell-methods.R,
%   R/plotUMIsPerCell-methods.R, R/plotUMIsVsGenes-methods.R, R/reexports.R,
%   R/show-methods.R, R/updateObject-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plotGene}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{barcodeRanksPerSample}}, \code{\link[basejump]{plotGene}}, \code{\link[basejump]{filterCells}}, \code{\link[basejump]{plotBarcodeRanks}}, \code{\link[basejump]{plotCellCounts}}, \code{\link[basejump]{plotGenesPerCell}}, \code{\link[basejump]{plotMitoRatio}}, \code{\link[basejump]{plotMitoVsCoding}}, \code{\link[basejump]{plotNovelty}}, \code{\link[basejump]{plotQC}}, \code{\link[basejump]{plotReadsPerCell}}, \code{\link[basejump]{plotUMIsPerCell}}, \code{\link[basejump]{plotUMIsVsGenes}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{updateObject}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{methods}{\code{\link[methods]{show}}}
}}

