% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{barcodeRanksPerSample}
\alias{plotCellCounts}
\alias{plotCountsPerCell}
\alias{plotCountsVsFeatures}
\alias{plotFeaturesPerCell}
\alias{plotMitoRatio}
\alias{plotNovelty}
\alias{plotBarcodeRanks}
\alias{plotQC}
\alias{calculateMetrics}
\alias{filterCells}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidPlots}{\code{\link[AcidPlots]{barcodeRanksPerSample}}, \code{\link[AcidPlots]{plotBarcodeRanks}}, \code{\link[AcidPlots]{plotCellCounts}}, \code{\link[AcidPlots]{plotCountsPerCell}}, \code{\link[AcidPlots]{plotCountsVsFeatures}}, \code{\link[AcidPlots]{plotFeaturesPerCell}}, \code{\link[AcidPlots]{plotMitoRatio}}, \code{\link[AcidPlots]{plotNovelty}}, \code{\link[AcidPlots]{plotQC}}}

  \item{basejump}{\code{\link[basejump:reexports]{calculateMetrics}}, \code{\link[basejump:reexports]{filterCells}}}
}}

