% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{calculateMetrics}
\alias{filterCells}
\alias{plotBarcodeRanks}
\alias{plotCellCounts}
\alias{plotCountsPerCell}
\alias{plotCountsVsFeatures}
\alias{plotFeaturesPerCell}
\alias{plotMitoRatio}
\alias{plotNovelty}
\alias{plotQc}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidGenerics}{\code{\link[AcidGenerics]{calculateMetrics}}, \code{\link[AcidGenerics]{filterCells}}, \code{\link[AcidGenerics]{plotBarcodeRanks}}, \code{\link[AcidGenerics]{plotCellCounts}}, \code{\link[AcidGenerics]{plotCountsPerCell}}, \code{\link[AcidGenerics]{plotCountsVsFeatures}}, \code{\link[AcidGenerics]{plotFeaturesPerCell}}, \code{\link[AcidGenerics]{plotMitoRatio}}, \code{\link[AcidGenerics]{plotNovelty}}, \code{\link[AcidGenerics]{plotQc}}}
}}

