% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeRanksPerSample-methods.R,
%   R/deprecated.R, R/filterCells-methods.R, R/plotBarcodeRanks-methods.R,
%   R/plotCellCounts-methods.R, R/plotGenesPerCell-methods.R,
%   R/plotMitoRatio-methods.R, R/plotMitoVsCoding-methods.R,
%   R/plotNovelty-methods.R, R/plotQC-methods.R, R/plotReadsPerCell-methods.R,
%   R/plotUMIsPerCell-methods.R, R/plotUMIsVsGenes-methods.R, R/show-methods.R,
%   R/updateObject-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plotGene}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{plotGene}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{updateObject}}}

  \item{bioverbs}{\code{\link[bioverbs]{barcodeRanksPerSample}}, \code{\link[bioverbs]{filterCells}}, \code{\link[bioverbs]{plotBarcodeRanks}}, \code{\link[bioverbs]{plotCellCounts}}, \code{\link[bioverbs]{plotGenesPerCell}}, \code{\link[bioverbs]{plotMitoRatio}}, \code{\link[bioverbs]{plotMitoVsCoding}}, \code{\link[bioverbs]{plotNovelty}}, \code{\link[bioverbs]{plotQC}}, \code{\link[bioverbs]{plotReadsPerCell}}, \code{\link[bioverbs]{plotUMIsPerCell}}, \code{\link[bioverbs]{plotUMIsVsGenes}}}

  \item{methods}{\code{\link[methods]{show}}}
}}

