% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioSinglecell.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{write_csv}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{readr}{\code{\link[readr]{write_csv}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}}
}}

