% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-methods.R, R/metrics-methods.R,
%   R/plotGene-methods.R, R/plotQC-methods.R, R/sampleData-methods.R,
%   R/selectSamples-methods.R, R/updateObject-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{combine}
\alias{reexports}
\alias{metrics}
\alias{reexports}
\alias{plotGene}
\alias{reexports}
\alias{plotQC}
\alias{reexports}
\alias{sampleData}
\alias{reexports}
\alias{sampleData<-}
\alias{reexports}
\alias{selectSamples}
\alias{reexports}
\alias{updateObject}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{sampleData}}, \code{\link[basejump]{sampleData<-}}, \code{\link[basejump]{selectSamples}}}

  \item{bcbioBase}{\code{\link[bcbioBase]{metrics}}, \code{\link[bcbioBase]{plotGene}}, \code{\link[bcbioBase]{plotQC}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{combine}}, \code{\link[BiocGenerics]{updateObject}}}
}}

