% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{doc_date}
\alias{reexports}
\alias{dot_notation}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{top_n}
\alias{reexports}
\alias{save_data}
\alias{reexports}
\alias{colSums}
\alias{reexports}
\alias{rowSums}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{\%>\%}}, \code{\link[basejump]{doc_date}}, \code{\link[basejump]{dot_notation}}, \code{\link[basejump]{group_by}}, \code{\link[basejump]{top_n}}, \code{\link[basejump]{save_data}}}

  \item{Matrix}{\code{\link[Matrix]{colSums}}, \code{\link[Matrix]{rowSums}}}
}}

