% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pull}
\alias{reexports}
\alias{ggplot}
\alias{reexports}
\alias{scale_color_gradient}
\alias{reexports}
\alias{scale_color_hue}
\alias{reexports}
\alias{scale_colour_gradient}
\alias{reexports}
\alias{scale_colour_gradientn}
\alias{reexports}
\alias{scale_colour_hue}
\alias{reexports}
\alias{scale_fill_hue}
\alias{reexports}
\alias{gzip}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{inferno}
\alias{reexports}
\alias{scale_color_viridis}
\alias{reexports}
\alias{scale_colour_viridis}
\alias{reexports}
\alias{scale_fill_viridis}
\alias{reexports}
\alias{viridis}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{pull}}}

  \item{ggplot2}{\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{scale_color_gradient}}, \code{\link[ggplot2]{scale_color_hue}}, \code{\link[ggplot2]{scale_colour_gradient}}, \code{\link[ggplot2]{scale_colour_gradientn}}, \code{\link[ggplot2]{scale_colour_hue}}, \code{\link[ggplot2]{scale_fill_hue}}}

  \item{R.utils}{\code{\link[R.utils]{gzip}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}}

  \item{viridis}{\code{\link[viridis]{inferno}}, \code{\link[viridis]{scale_color_viridis}}, \code{\link[viridis]{scale_colour_viridis}}, \code{\link[viridis]{scale_fill_viridis}}, \code{\link[viridis]{viridis}}}
}}

