% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{assay}
\alias{reexports}
\alias{assays}
\alias{reexports}
\alias{colData}
\alias{reexports}
\alias{detect_hpc}
\alias{reexports}
\alias{dotted}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{metadata}
\alias{reexports}
\alias{pct}
\alias{reexports}
\alias{rowData}
\alias{reexports}
\alias{save_data}
\alias{reexports}
\alias{tibble}
\alias{reexports}
\alias{top_n}
\alias{reexports}
\alias{write_counts}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{\%>\%}}, \code{\link[basejump]{assay}}, \code{\link[basejump]{assays}}, \code{\link[basejump]{colData}}, \code{\link[basejump]{detect_hpc}}, \code{\link[basejump]{dotted}}, \code{\link[basejump]{group_by}}, \code{\link[basejump]{metadata}}, \code{\link[basejump]{pct}}, \code{\link[basejump]{rowData}}, \code{\link[basejump]{save_data}}, \code{\link[basejump]{tibble}}, \code{\link[basejump]{top_n}}, \code{\link[basejump]{write_counts}}}
}}

