% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioSinglecell.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{arrange}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{left_join}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{read_excel}
\alias{reexports}
\alias{remove_rownames}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{write_csv}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{readr}{\code{\link[readr]{write_csv}}}

  \item{readxl}{\code{\link[readxl]{read_excel}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}, \code{\link[tibble]{remove_rownames}}}
}}

