% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-rowData.R
\docType{methods}
\name{rowData}
\alias{rowData}
\alias{rowData,bcbioSingleCell-method}
\alias{rowData,seurat-method}
\title{Row Data}
\usage{
\S4method{rowData}{bcbioSingleCell}(x, return = c("DataFrame", "data.frame",
  "AsIs"))

\S4method{rowData}{seurat}(x, return = c("DataFrame", "data.frame", "AsIs"))
}
\arguments{
\item{x}{Primary object.}
}
\value{
\code{DataFrame}, \code{data.frame}, or \code{GRanges}.
}
\description{
Row Data
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat.rda", package = "bcbioSingleCell"))

# bcbioSingleCell
rowData(bcb) \%>\% glimpse()

# seurat
rowData(seurat) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
