% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-rowData.R
\docType{methods}
\name{rowData}
\alias{rowData}
\alias{rowData,bcbioSingleCell-method}
\alias{rowData,seurat-method}
\title{Row Data}
\usage{
\S4method{rowData}{bcbioSingleCell}(x, return = c("data.frame", "DataFrame"))

\S4method{rowData}{seurat}(x, return = c("data.frame", "DataFrame"))
}
\arguments{
\item{x}{Primary object.}

\item{return}{Return type. Uses \code{\link[=match.arg]{match.arg()}} internally and defaults to the
first argument in the vector.}
}
\value{
Return data as \code{data.frame}, or \code{DataFrame}.
}
\description{
Metadata describing the genes. These annotations are obtained from
AnnotationHub and ensembldb using \code{\link[basejump:ensembl]{basejump::ensembl()}}.
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat_small.rda", package = "bcbioSingleCell"))

# bcbioSingleCell ====
rowData(bcb_small) \%>\% glimpse()

# seurat ====
rowData(pbmc_small) \%>\% glimpse()
rowData(seurat_small) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
