% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-rowData.R
\docType{methods}
\name{rowData}
\alias{rowData}
\alias{rowData,bcbioSingleCell-method}
\alias{rowData,seurat-method}
\title{Row Data}
\usage{
\S4method{rowData}{bcbioSingleCell}(x, return = c("data.frame", "DataFrame"))

\S4method{rowData}{seurat}(x, return = c("data.frame", "DataFrame"))
}
\arguments{
\item{x}{Primary object.}

\item{return}{Return type. Uses \code{\link[base:match.arg]{base::match.arg()}} internally and defaults
to the first argument in the vector.}
}
\value{
Return data as \code{data.frame}, or \code{DataFrame}.
}
\description{
Metadata describing the genes.
}
\examples{
# bcbioSingleCell ====
rowData(bcb_small) \%>\% glimpse()

# seurat ====
rowData(pbmc_small) \%>\% glimpse()
rowData(seurat_small) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{calculateMetrics}},
  \code{\link{cell2sample}}, \code{\link{colData}},
  \code{\link{counts}}, \code{\link{fetchGeneData}},
  \code{\link{fetchPCAData}}, \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{sampleData}}, \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
