% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-sample_dirs.R
\name{.sample_dirs}
\alias{.sample_dirs}
\title{Detect sample directories}
\usage{
.sample_dirs(upload_dir, nested_file = NULL)
}
\arguments{
\item{upload_dir}{Upload directory.}

\item{nested_file}{Match sample directories by the presence of a nested file.}
}
\value{
Named character vector containing sample directory paths. Function
will \link{stop} if no sample directories are found.
}
\description{
Detect sample directories
}
\author{
Michael Steinbaugh
}
\keyword{internal}
