% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-sample_dirs.R
\name{.sample_dirs}
\alias{.sample_dirs}
\title{Detect sample directories}
\usage{
.sample_dirs(upload_dir, mtx_mode = FALSE)
}
\arguments{
\item{upload_dir}{Upload directory.}

\item{mtx_mode}{Locate nested MatrixMarket (\code{.mtx}) files to define sample
directories.}
}
\value{
Named character vector containing sample directory paths. Function
will \link{stop} if no sample directories are found.
}
\description{
Detect sample directories
}
\author{
Michael Steinbaugh
}
\keyword{internal}
