% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sample_metadata.R
\docType{methods}
\name{.sample_metadata}
\alias{.sample_metadata}
\alias{sample_metadata,bcbioSCDataSet-method}
\alias{sample_metadata,SummarizedExperiment-method}
\title{Sample metadata}
\usage{
\S4method{sample_metadata}{bcbioSCDataSet}(object, unique_names = FALSE)

\S4method{sample_metadata}{SummarizedExperiment}(object, unique_names = FALSE)
}
\arguments{
\item{object}{\link{bcbioSCDataSet}.}

\item{unique_names}{Unique sample names.}
}
\value{
\link{data.frame}.
}
\description{
Sample metadata
}
\author{
Michael Steinbaugh
}
