% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-sample_metadata.R
\docType{methods}
\name{sample_metadata}
\alias{sample_metadata}
\alias{.sample_metadata}
\alias{sample_metadata,bcbioSCDataSet-method}
\alias{sample_metadata,SCSubset-method}
\title{Sample metadata}
\usage{
\S4method{sample_metadata}{bcbioSCDataSet}(object)

\S4method{sample_metadata}{SCSubset}(object)
}
\arguments{
\item{object}{Primary object.}
}
\value{
\link{tibble}.
}
\description{
Sample metadata
}
\author{
Michael Steinbaugh
}
