% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-sample_metadata.R,
%   R/methods-sample_metadata.R
\docType{methods}
\name{.sample_metadata}
\alias{.sample_metadata}
\alias{.sample_metadata}
\alias{sample_metadata,bcbioSCDataSet-method}
\alias{sample_metadata,SummarizedExperiment-method}
\title{Sample metadata}
\usage{
.sample_metadata(object, unique_names = FALSE)

\S4method{sample_metadata}{bcbioSCDataSet}(object, unique_names = FALSE)

\S4method{sample_metadata}{SummarizedExperiment}(object, unique_names = FALSE)
}
\arguments{
\item{object}{\link{bcbioSCDataSet}.}

\item{unique_names}{Unique sample names.}

\item{metadata}{Metadata data frame.}

\item{sample_dirs}{(\emph{Optional}). Named character vector of sample directory
paths.}
}
\value{
\link{DataFrame}.

\link{data.frame}.
}
\description{
Appends reverse complement sequences by matching against the \code{sequence}
column.

Sample metadata
}
\author{
Michael Steinbaugh

Michael Steinbaugh
}
\keyword{internal}
