% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/internal-sample_metadata.R,
%   R/methods-sample_metadata.R
\docType{methods}
\name{sample_metadata}
\alias{sample_metadata}
\alias{.sample_metadata}
\alias{sample_metadata,bcbioSCDataSet-method}
\title{Sample metadata}
\usage{
sample_metadata(object)

.sample_metadata(file, sample_dirs)

\S4method{sample_metadata}{bcbioSCDataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{sample_dirs}{(\emph{Optional}). Named character vector of sample directory
paths.}

\item{metadata}{Metadata data frame.}
}
\value{
\link{DataFrame}.
}
\description{
Appends reverse complement sequences by matching against the \code{sequence}
column.

Sample metadata
}
\author{
Michael Steinbaugh

Michael Steinbaugh
}
\keyword{internal}
