% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-sample_metadata.R
\docType{methods}
\name{sample_metadata}
\alias{sample_metadata}
\alias{sample_metadata,bcbioSCDataSet-method}
\title{Sample metadata}
\usage{
sample_metadata(object)

\S4method{sample_metadata}{bcbioSCDataSet}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Sample metadata
}
\author{
Michael Steinbaugh
}
