% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-sample_metadata_file.R
\name{.sample_metadata_file}
\alias{.sample_metadata_file}
\title{Prepare sample metadata from file}
\usage{
.sample_metadata_file(file, sample_dirs)
}
\arguments{
\item{sample_dirs}{(\emph{Optional}). Named character vector of sample directory
paths.}

\item{metadata}{Metadata data frame.}
}
\value{
\link{tibble}.
}
\description{
Appends reverse complement sequences by matching against the \code{sequence}
column.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
