% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleData.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData,SingleCellExperiment-method}
\alias{sampleData,seurat-method}
\title{Sample Data}
\usage{
\S4method{sampleData}{SingleCellExperiment}(object, return = c("DataFrame",
  "data.frame", "kable"))

\S4method{sampleData}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{return}{Return type. Uses \code{\link[base:match.arg]{base::match.arg()}} internally and defaults
to the first argument in the vector.}
}
\value{
\code{data.frame}.
}
\description{
Metadata in columns describing the samples, which are defined in the
rownames. Similar to \code{\link[=colData]{colData()}}, which for \code{bcbioSingleCell} and
\code{SingleCellExperiment} objects describes cells in the columns, rather than
the samples.
}
\examples{
# bcbioSingleCell ====
sampleData(bcb_small) \%>\% glimpse()

# seurat ====
sampleData(pbmc_small) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{counts}},
  \code{\link{fetchGeneData}}, \code{\link{fetchPCAData}},
  \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
