% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleData.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData,SingleCellExperiment-method}
\alias{sampleData,seurat-method}
\alias{sampleData<-,SingleCellExperiment,DataFrame-method}
\alias{sampleData<-,seurat,DataFrame-method}
\title{Sample Data}
\usage{
\S4method{sampleData}{SingleCellExperiment}(object, interestingGroups,
  return = c("DataFrame", "data.frame", "kable"))

\S4method{sampleData}{seurat}(object, interestingGroups,
  return = c("DataFrame", "data.frame"))

\S4method{sampleData}{SingleCellExperiment,DataFrame}(object) <- value

\S4method{sampleData}{seurat,DataFrame}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in camel case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{return}{Return type. Uses \code{\link[base:match.arg]{base::match.arg()}} internally and defaults
to the first argument in the vector.}

\item{value}{Object to assign.}
}
\value{
\code{data.frame}.
}
\description{
Metadata in columns describing the samples, which are defined in the
rownames. Similar to \code{\link[=colData]{colData()}}, which for \code{bcbioSingleCell} and
\code{SingleCellExperiment} objects describes cells in the columns, rather than
the samples.
}
\examples{
# bcbioSingleCell ====
x <- bcb_small
sampleData(x) \%>\% glimpse()
sampleData(x)[["batch"]] <- 1L
sampleData(x) \%>\% glimpse()

# seurat ====
x <- seurat_small
sampleData(x) \%>\% glimpse()
sampleData(x)[["batch"]] <- 1L
sampleData(x) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
