% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData-methods.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData,SingleCellExperiment-method}
\alias{sampleData,seurat-method}
\alias{sampleData<-,SingleCellExperiment,DataFrame-method}
\alias{sampleData<-,SingleCellExperiment,DataFrame-method}
\title{Sample Data}
\usage{
\S4method{sampleData}{SingleCellExperiment}(object, interestingGroups,
  clean = FALSE, return = c("DataFrame", "data.frame", "kable"))

\S4method{sampleData}{seurat}(object, interestingGroups, clean = FALSE,
  return = c("DataFrame", "data.frame", "kable"))

\S4method{sampleData}{SingleCellExperiment,DataFrame}(object) <- value

\S4method{sampleData}{SingleCellExperiment,DataFrame}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in camel case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{clean}{Only return factor columns.}

\item{return}{Return type. Uses \code{\link[base:match.arg]{base::match.arg()}} internally and defaults
to the first argument in the vector.}

\item{value}{Object to assign.}
}
\value{
Return either a DataFrame, data.frame, or kable.
}
\description{
Metadata in columns describing the samples, which are defined in the
rownames. Similar to \code{\link[=colData]{colData()}}, which for \code{bcbioSingleCell} and
\code{SingleCellExperiment} objects describes cells in the columns, rather than
the samples.
}
\examples{
# SingleCellExperiment ====
x <- cellranger_small
sampleData(x, clean = FALSE) \%>\% glimpse()
sampleData(x, clean = TRUE) \%>\% glimpse()

# Assignment support
sampleData(x)[["batch"]] <- 1L
sampleData(x) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{fetchGeneData}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
