% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleData.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData,bcbioSingleCell-method}
\alias{sampleData,seurat-method}
\title{Sample Data}
\usage{
\S4method{sampleData}{bcbioSingleCell}(object, interestingGroups)

\S4method{sampleData}{seurat}(object, interestingGroups)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in \code{\link[=camel]{camel()}} case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}
}
\value{
\code{data.frame}.
}
\description{
Sample Data
}
\examples{
# bcbioSingleCell ====
sampleData(bcb_small) \%>\% glimpse()

# seurat ====
sampleData(pbmc_small) \%>\% glimpse()
sampleData(seurat_small) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
