% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleData.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData,bcbioSingleCell-method}
\alias{sampleData,seurat-method}
\title{Sample Data}
\usage{
\S4method{sampleData}{bcbioSingleCell}(object, interestingGroups,
  aggregateReplicates = FALSE)

\S4method{sampleData}{seurat}(object, interestingGroups)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Interesting group, to use for colors.}

\item{aggregateReplicates}{Aggregate technical replicates, if specified. This
function uses values assigned in the \code{sampleNameAggregate} column of the
internal sample metadata \link{data.frame}.}
}
\value{
\link{data.frame}.
}
\description{
Sample Data
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioSingleCell"))
load(system.file("extdata/seurat.rda", package = "bcbioSingleCell"))

# bcbioSingleCell
sampleData(bcb) \%>\% glimpse()

# seurat
sampleData(seurat) \%>\% glimpse()
}
