% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData-methods.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData<-}
\alias{sampleData,SingleCellExperiment-method}
\alias{sampleData,seurat-method}
\alias{sampleData<-,SingleCellExperiment,DataFrame-method}
\alias{sampleData<-,seurat,DataFrame-method}
\title{Sample Data}
\usage{
\S4method{sampleData}{SingleCellExperiment}(object, interestingGroups)

\S4method{sampleData}{seurat}(object, interestingGroups)

\S4method{sampleData}{SingleCellExperiment,DataFrame}(object) <- value

\S4method{sampleData}{seurat,DataFrame}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}

\item{value}{Object to assign.}
}
\value{
\code{DataFrame}.
}
\description{
Metadata in columns describing the samples, which are defined in the
rownames. Similar to \code{\link[=colData]{colData()}}, which for \code{bcbioSingleCell} and
\code{SingleCellExperiment} objects describes cells in the columns, rather than
the samples.
}
\examples{
# SingleCellExperiment ====
x <- cellranger_small
sampleData(x) \%>\% glimpse()
sampleData(x)[["batch"]] <- 1L
sampleData(x) \%>\% glimpse()
"batch" \%in\% colnames(colData(x))

# seurat ====
x <- seurat_small
sampleData(x) \%>\% glimpse()
sampleData(x)[["batch"]] <- 1L
sampleData(x) \%>\% glimpse()
"batch" \%in\% colnames(colData(x))
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{fetchGeneData}},
  \code{\link{metricsPerSample}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
