% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleData.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData,SingleCellExperiment-method}
\alias{sampleData,seurat-method}
\alias{sampleData<-,SingleCellExperiment,DataFrame-method}
\alias{sampleData<-,seurat,DataFrame-method}
\title{Sample Data}
\usage{
\S4method{sampleData}{SingleCellExperiment}(object, clean = TRUE,
  interestingGroups, return = c("DataFrame", "data.frame", "kable"))

\S4method{sampleData}{seurat}(object, clean = TRUE, interestingGroups,
  return = c("DataFrame", "data.frame", "kable"))

\S4method{sampleData}{SingleCellExperiment,DataFrame}(object) <- value

\S4method{sampleData}{seurat,DataFrame}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{clean}{\code{logical}. Only return \code{factor} columns not defined in
\link{metadataBlacklist}.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}

\item{value}{Value to assign.}
}
\value{
Data describing the samples.
}
\description{
Metadata in columns describing the samples, which are defined in the
rownames. Similar to \code{\link[=colData]{colData()}}, which for \code{bcbioSingleCell} and
\code{SingleCellExperiment} objects describes cells in the columns, rather than
the samples.
}
\examples{
# SingleCellExperiment ====
x <- cellranger_small
sampleData(x, clean = FALSE) \%>\% glimpse()
sampleData(x, clean = TRUE) \%>\% glimpse()

# Assignment support
sampleData(x)[["batch"]] <- 1L
sampleData(x) \%>\% glimpse()
}
\seealso{
\link{metadataBlacklist}.

Other Data Functions: \code{\link{convertGenesToSymbols}},
  \code{\link{counts}}, \code{\link{flatFiles}},
  \code{\link{gene2symbol}}, \code{\link{sampleDirs}},
  \code{\link{sampleNames}}, \code{\link{selectSamples}},
  \code{\link{uniteInterestingGroups}}
}
\author{
Michael Steinbaugh
}
