% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleMetadata.R
\docType{methods}
\name{sampleMetadata}
\alias{sampleMetadata}
\alias{sampleMetadata,bcbioSingleCellANY-method}
\alias{sampleMetadata,seurat-method}
\title{Sample Metadata}
\usage{
\S4method{sampleMetadata}{bcbioSingleCellANY}(object,
  aggregateReplicates = FALSE)

\S4method{sampleMetadata}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{aggregateReplicates}{Aggregate technical replicates, if present. If
\code{TRUE}, this function uses the values slotted in
\code{sampleMetadata(object)[["sampleNameAggregate"]])}.

[["sampleNameAggregate"]: R:[%22sampleNameAggregate%22}
}
\value{
\link{data.frame}.
}
\description{
Sample Metadata
}
