% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sampleMetadata.R
\docType{methods}
\name{sampleMetadata}
\alias{sampleMetadata}
\alias{sampleMetadata}
\alias{sampleMetadata,bcbioSCDataSet-method}
\alias{sampleMetadata,bcbioSCFiltered-method}
\title{Sample Metadata}
\usage{
sampleMetadata(object, ...)

\S4method{sampleMetadata}{bcbioSCDataSet}(object)

\S4method{sampleMetadata}{bcbioSCFiltered}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{data.frame}.
}
\description{
Sample Metadata
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}},
  \code{\link{filterCells}},
  \code{\link{interestingGroups}},
  \code{\link{readMarkers}}, \code{\link{selectSamples}}
}
\author{
Michael Steinbaugh
}
