% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sampleMetadata.R
\docType{methods}
\name{sampleMetadata}
\alias{sampleMetadata}
\alias{sampleMetadata,bcbioSCDataSet-method}
\alias{sampleMetadata,bcbioSCFiltered-method}
\alias{sampleMetadata,seurat-method}
\title{Sample Metadata}
\usage{
sampleMetadata(object, ...)

\S4method{sampleMetadata}{bcbioSCDataSet}(object, kable = FALSE)

\S4method{sampleMetadata}{bcbioSCFiltered}(object, kable = FALSE)

\S4method{sampleMetadata}{seurat}(object, kable = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{kable}{Return the output as \link{kable} instead of \link{data.frame}.}
}
\value{
\link{data.frame}, unless \code{kable = TRUE}, which will return a \link{kable}.
}
\description{
Sample Metadata
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}},
  \code{\link{filterCells}},
  \code{\link{interestingGroups}},
  \code{\link{readMarkers}}, \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
\concept{Data Management Utilities}
