% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleMetadata.R
\docType{methods}
\name{sampleMetadata}
\alias{sampleMetadata}
\alias{sampleMetadata,bcbioSingleCell-method}
\alias{sampleMetadata,seurat-method}
\title{Sample Metadata}
\usage{
\S4method{sampleMetadata}{bcbioSingleCell}(object, interestingGroups,
  aggregateReplicates = FALSE)

\S4method{sampleMetadata}{seurat}(object, interestingGroups)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Interesting group, to use for colors.}

\item{aggregateReplicates}{Aggregate technical replicates, if specified. This
function uses values assigned in the \code{sampleNameAggregate} column of the
internal sample metadata \link{data.frame}.}
}
\value{
\link{data.frame}.
}
\description{
Sample Metadata
}
