% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleMetadata.R
\docType{methods}
\name{sampleMetadata}
\alias{sampleMetadata}
\alias{sampleMetadata,bcbioSingleCellANY-method}
\alias{sampleMetadata,seurat-method}
\title{Sample Metadata}
\usage{
\S4method{sampleMetadata}{bcbioSingleCellANY}(object, kable = FALSE)

\S4method{sampleMetadata}{seurat}(object, kable = FALSE)
}
\arguments{
\item{object}{Object.}

\item{kable}{Return the output as \link{kable} instead of \link{data.frame}.}
}
\value{
\link{data.frame}, unless \code{kable = TRUE}, which will return a \link{kable}.
}
\description{
Sample Metadata
}
