% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleMetadata.R
\docType{methods}
\name{sampleMetadata}
\alias{sampleMetadata}
\alias{sampleMetadata,bcbioSingleCellANY-method}
\alias{sampleMetadata,seurat-method}
\title{Sample Metadata}
\usage{
\S4method{sampleMetadata}{bcbioSingleCellANY}(object)

\S4method{sampleMetadata}{seurat}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{data.frame}.
}
\description{
Sample Metadata
}
