% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeMarkers.R
\name{sanitizeMarkers}
\alias{sanitizeMarkers}
\title{Sanitize Markers Output}
\usage{
sanitizeMarkers(data, rowRanges)
}
\arguments{
\item{data}{\code{data.frame}. \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} or
\code{\link[Seurat:FindMarkers]{Seurat::FindMarkers()}} return.}

\item{rowRanges}{\code{GRanges}. Gene annotations. Must contain \code{geneName} column
that corresponds to the gene names in the \code{seurat} object used to generate
the markers \code{data.frame}.}
}
\value{
\code{data.frame}, arranged by adjusted P value.
}
\description{
Currently only Seurat markers are supported.
}
\note{
\code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} maps the counts matrix rownames correctly in
the \code{gene} column, whereas \code{\link[Seurat:FindMarkers]{Seurat::FindMarkers()}} maps them correctly in
the rownames of the returned marker \code{data.frame}.
}
\examples{
# seurat ====
object <- seurat_small

# `FindAllMarkers()` return
invisible(capture.output(
    all_markers <- Seurat::FindAllMarkers(object)
))
all_sanitized <- sanitizeMarkers(
    data = all_markers,
    rowRanges = rowRanges(object)
)
glimpse(all_sanitized)

# `FindMarkers()` return
invisible(capture.output(
    ident_3_markers <- Seurat::FindMarkers(
        object = object,
        ident.1 = "3",
        ident.2 = NULL
    )
))
ident_3_sanitized <- sanitizeMarkers(
    data = ident_3_markers,
    rowRanges = rowRanges(object)
)
glimpse(ident_3_sanitized)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotTSNE}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
