% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sanitizeMarkers.R
\docType{methods}
\name{sanitizeMarkers}
\alias{sanitizeMarkers}
\alias{sanitizeMarkers}
\alias{sanitizeMarkers,seurat-method}
\title{Sanitize Markers Output}
\usage{
sanitizeMarkers(object, ...)

\S4method{sanitizeMarkers}{seurat}(object, markers)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{markers}{Original \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} \code{data.frame}.}
}
\value{
\code{data.frame}, arranged by adjusted P value.
}
\description{
Sanitize Markers Output
}
\note{
\code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} maps the counts matrix rownames correctly in
the \code{gene} column, whereas \code{\link[Seurat:FindMarkers]{Seurat::FindMarkers()}} maps them correctly in
the rownames of the returned marker \code{data.frame}.
}
\examples{
# seurat ====
invisible(capture.output(
    all_markers <- Seurat::FindAllMarkers(seurat_small)
))
all_sanitized <- sanitizeMarkers(
    object = seurat_small,
    markers = all_markers
)
glimpse(all_sanitized)

invisible(capture.output(
    ident_3_markers <- Seurat::FindMarkers(
        object = seurat_small,
        ident.1 = "3",
        ident.2 = NULL
    )
))
ident_3_sanitized <- sanitizeMarkers(
    object = seurat_small,
    markers = ident_3_markers
)
glimpse(ident_3_sanitized)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotMarker}},
  \code{\link{plotPCElbow}}, \code{\link{plotTSNE}},
  \code{\link{plotTopMarkers}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
