% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sanitizeMarkers.R
\docType{methods}
\name{sanitizeMarkers}
\alias{sanitizeMarkers}
\alias{sanitizeMarkers}
\alias{sanitizeMarkers,seurat,data.frame-method}
\title{Sanitize Markers Output}
\usage{
sanitizeMarkers(object, markers, ...)

\S4method{sanitizeMarkers}{seurat,data.frame}(object, markers)
}
\arguments{
\item{object}{Object.}

\item{markers}{Original \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} \code{data.frame}.}

\item{...}{Additional arguments.}
}
\value{
Tibble arranged by adjusted P value.
\itemize{
\item 
\item \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}}: \code{grouped_df}, grouped by cluster ident.
}
}
\description{
Sanitize Markers Output
}
\note{
\code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} maps the counts matrix rownames correctly in
the \code{gene} column, whereas \code{\link[Seurat:FindMarkers]{Seurat::FindMarkers()}} maps them correctly in
the rownames of the returned marker \code{data.frame}.
}
\examples{
# seurat ====
all_markers <- FindAllMarkers(seurat_small)
all_sanitized <- sanitizeMarkers(
    object = seurat_small,
    markers = all_markers
)
glimpse(all_sanitized)

ident_3_markers <- FindMarkers(seurat_small, ident.1 = "3", ident.2 = NULL)
ident_3_sanitized <- sanitizeMarkers(
    object = seurat_small,
    markers = ident_3_markers
)
glimpse(ident_3_sanitized)
}
\seealso{
Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotMarkers}},
  \code{\link{plotPCA}}, \code{\link{plotPCElbow}},
  \code{\link{plotTSNE}}, \code{\link{plotTopMarkers}},
  \code{\link{topBarcodes}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
