% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-sanitizeMarkers.R
\docType{methods}
\name{sanitizeMarkers}
\alias{sanitizeMarkers}
\alias{sanitizeMarkers,seurat,data.frame-method}
\title{Sanitize Markers Output}
\usage{
sanitizeMarkers(object, markers, ...)

\S4method{sanitizeMarkers}{seurat,data.frame}(object, markers)
}
\arguments{
\item{object}{Object.}

\item{markers}{\code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} return.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{grouped_df}, grouped by cluster ident.
}
\description{
Sanitize Markers Output
}
