% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-select_samples.R
\docType{methods}
\name{select_samples}
\alias{select_samples}
\alias{.select_samples}
\alias{select_samples,bcbioSCDataSet-method}
\alias{select_samples,bcbioSCSubset-method}
\title{Select Samples}
\usage{
select_samples(object, ...)

\S4method{select_samples}{bcbioSCDataSet}(object)

\S4method{select_samples}{bcbioSCSubset}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Columns to use for grep pattern matching. Supply a named character
vector containing the column name and the grep pattern.}
}
\value{
\link{bcbioSCSubset}.
}
\description{
Select Samples
}
\details{
Internally, pattern matching against sample and file names is
applied using \code{\link[=str_detect]{str_detect()}}.
}
\note{
Supports an object containing filtered counts generated by \code{\link[=filter]{filter()}}
on a \link{bcbioSCDataSet}.
}
\author{
Michael Steinbaugh
}
