% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-select_samples.R
\docType{methods}
\name{select_samples}
\alias{select_samples}
\alias{select_samples,bcbioSCDataSet-method}
\alias{select_samples,SummarizedExperiment-method}
\title{Select samples}
\usage{
\S4method{select_samples}{SummarizedExperiment}(object, sample_name = NULL,
  file_name = NULL)
}
\arguments{
\item{object}{Primary object.}

\item{sample_name}{Character vector of sample names.}

\item{file_name}{Character vector of file names.}
}
\value{
\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment}.
}
\description{
Select samples
}
\details{
Internally, pattern matching against sample and file names is
applied using \code{\link[=str_detect]{str_detect()}}.
}
\note{
Supports an object containing filtered counts generated by
\code{\link[=filter_barcodes]{filter_barcodes()}} on a \link{bcbioSCDataSet}.
}
