% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-select_samples.R
\docType{methods}
\name{select_samples}
\alias{select_samples}
\alias{select_samples,bcbioSCDataSet-method}
\title{Select samples}
\usage{
select_samples(object, ...)

\S4method{select_samples}{bcbioSCDataSet}(object, samples)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters.}

\item{samples}{Character vector of desired samples.}
}
\value{
\link{bcbioSCDataSet}.
}
\description{
Select samples
}
