% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-selectSamples.R
\docType{methods}
\name{selectSamples}
\alias{selectSamples}
\alias{selectSamples,bcbioSCFiltered-method}
\title{Select Samples}
\usage{
\S4method{selectSamples}{bcbioSCFiltered}(object, ...)
}
\arguments{
\item{...}{Columns to use for grep pattern matching. Supply a named character
vector containing the column name and the grep pattern.}
}
\value{
\link{bcbioSCFiltered}.
}
\description{
Select Samples
}
\details{
Internally, pattern matching against sample and file names is
applied using \code{\link[=str_detect]{str_detect()}}.
}
\examples{
\dontrun{
data(bcbFiltered)
# grep pattern matching with string
selectSamples(bcbFiltered, sampleName = "wt")

# Exact name matching with character vector
selectSamples(bcbFiltered, sampleName = c("wt1", "wt2"))
}
}
