% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-selectSamples.R
\docType{methods}
\name{selectSamples}
\alias{selectSamples}
\alias{selectSamples}
\alias{selectSamples,bcbioSCDataSet-method}
\alias{selectSamples,bcbioSCSubset-method}
\title{Select Samples}
\usage{
selectSamples(object, ...)

\S4method{selectSamples}{bcbioSCDataSet}(object, ...)

\S4method{selectSamples}{bcbioSCSubset}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Columns to use for grep pattern matching. Supply a named character
vector containing the column name and the grep pattern.}
}
\value{
\link{bcbioSCSubset}.
}
\description{
Select Samples
}
\details{
Internally, pattern matching against sample and file names is
applied using \code{\link[=str_detect]{str_detect()}}.
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateFeatures}},
  \code{\link{aggregateReplicates}},
  \code{\link{filterCells}},
  \code{\link{interestingGroups}},
  \code{\link{readKnownMarkers}},
  \code{\link{sampleMetadata}}
}
\author{
Michael Steinbaugh
}
