% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSamples-methods.R
\docType{methods}
\name{selectSamples}
\alias{selectSamples}
\alias{selectSamples,SingleCellExperiment-method}
\title{Select Samples}
\usage{
\S4method{selectSamples}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Columns to use for grep pattern matching. Supply a named character
vector containing the column name and the grep pattern.}
}
\value{
\code{bcbioSingleCell}.
}
\description{
Select Samples
}
\details{
Internally, pattern matching against sample and file names is
applied using logical grep matching.
}
\note{
Bracket based subsetting with \code{[} also works on \code{bcbioSingleCell}
objects. In this case, provide cellular barcode identifiers for columns
and Ensembl gene identifiers for rows.
}
\examples{
# bcbioSingleCell ====
object <- indrops_small
sample <- sampleNames(object) \%>\% head(1L)
print(sample)
selectSamples(object, sampleName = sample)
}
\seealso{
\code{\link[=sampleData]{sampleData()}}.

Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{fetchGeneData}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{sampleData}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
