% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSamples-methods.R
\docType{methods}
\name{selectSamples}
\alias{selectSamples}
\alias{selectSamples,SingleCellExperiment-method}
\title{Select Samples}
\usage{
\S4method{selectSamples}{SingleCellExperiment}(object, ...,
  prefilter = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Selection arguments that map to the column names of
\code{\link[=sampleData]{sampleData()}}. \code{character} vectors are supported. Avoid using \code{logical} or
\code{numeric} indices (e.g. \code{\link[base:which]{base::which()}} calls) here, for improved code
legibility and reproducibility.}

\item{prefilter}{\code{boolean}. Apply \code{filterCells()} with default minimal
cutoffs to the sample subset. This is recommended to automatically remove
genes with all zero counts.}
}
\value{
\code{bcbioSingleCell}.
}
\description{
Sample selection utility function designed for interactive use.
}
\details{
Internally, pattern matching against sample and file names is
applied using logical grep matching.
}
\note{
Bracket based subsetting with \code{[} also works on \code{bcbioSingleCell}
objects. In this case, provide cellular barcode identifiers for columns
and Ensembl gene identifiers for rows.
}
\examples{
object <- indrops_small
sample <- sampleNames(object) \%>\% head(1L)
print(sample)
selectSamples(object, sampleName = sample)
}
\seealso{
\code{\link[=sampleData]{sampleData()}}.

Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{cell2sample}},
  \code{\link{metricsPerSample}}, \code{\link{sampleData}},
  \code{\link{subsetPerSample}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
