% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-selectSamples.R
\docType{methods}
\name{selectSamples}
\alias{selectSamples}
\alias{selectSamples,bcbioSingleCell-method}
\title{Select Samples}
\usage{
\S4method{selectSamples}{bcbioSingleCell}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Columns to use for grep pattern matching. Supply a named character
vector containing the column name and the grep pattern.}
}
\value{
\link{bcbioSingleCell}.
}
\description{
Select Samples
}
\details{
Internally, pattern matching against sample and file names is
applied using logical grep matching.
}
\note{
Bracket based subsetting with \code{[} also works on [bcbioSingleCell]
objects. In this case, provide cellular barcode identifiers for columns
and Ensembl gene identifiers for rows.

[` also works on [bcbioSingleCell]: R:%60%20also%20works%20on%20[bcbioSingleCell
}
\examples{
load(system.file("extdata/filtered.rda", package = "bcbioSingleCell"))

# bcbioSingleCell
# Quality control filtering must be applied first!
selectSamples(filtered, sampleName = "M1")
}
