% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-seurat-SingleCellExperiment.R
\docType{methods}
\name{seurat-SingleCellExperiment}
\alias{seurat-SingleCellExperiment}
\alias{assay,seurat,ANY-method}
\alias{colData,seurat-method}
\alias{colnames,seurat-method}
\alias{gene2symbol,seurat-method}
\alias{interestingGroups,seurat-method}
\alias{interestingGroups<-,seurat,character-method}
\alias{metadata,seurat-method}
\alias{metadata<-,seurat-method}
\alias{rowData,seurat-method}
\alias{rowRanges,seurat-method}
\alias{rownames,seurat-method}
\title{Extend S4 Methods for \code{seurat} Class}
\usage{
\S4method{assay}{seurat,ANY}(x)

\S4method{colData}{seurat}(x)

\S4method{colnames}{seurat}(x)

\S4method{gene2symbol}{seurat}(object)

\S4method{interestingGroups}{seurat}(object)

\S4method{interestingGroups}{seurat,character}(object) <- value

\S4method{metadata}{seurat}(x)

\S4method{metadata}{seurat}(x) <- value

\S4method{rowData}{seurat}(x)

\S4method{rowRanges}{seurat}(x)

\S4method{rownames}{seurat}(x)
}
\arguments{
\item{x}{Primary object.}

\item{object}{Object.}

\item{value}{Object to assign.}
}
\value{
Match \code{SummarizedExperiment} method return.
}
\description{
Provide limited \code{SingleCellExperiment}-like method support.
}
\examples{
# assay ====
assay(pbmc_small) \%>\% glimpse()
# colData ====
colData(seurat_small) \%>\% glimpse()
colData(pbmc_small) \%>\% glimpse()
# colnames ====
colnames(pbmc_small) \%>\% head()
# gene2symbol ====
gene2symbol(seurat_small) \%>\% glimpse()
# interestingGroups ====
interestingGroups(seurat_small)
interestingGroups(seurat_small) <- "sampleID"
interestingGroups(seurat_small)
# metadata ====
names(metadata(seurat_small))
metadata(seurat_small)[["stash"]] <- "XXX"
metadata(seurat_small)[["stash"]]
# rowData ====
rowData(seurat_small) \%>\% glimpse()
rowData(pbmc_small) \%>\% glimpse()
# rowRanges ====
rowRanges(seurat_small) \%>\% glimpse()
rowRanges(pbmc_small) \%>\% glimpse()
# rownames ====
rownames(pbmc_small) \%>\% head()
}
\seealso{
\code{getMethod("metadata<-", "Annotated")}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
