% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2sample-methods.R,
%   R/plotCellCounts-methods.R, R/seurat-SingleCellExperiment-methods.R
\docType{methods}
\name{cell2sample,seurat-method}
\alias{cell2sample,seurat-method}
\alias{plotCellCounts,seurat-method}
\alias{seurat-SingleCellExperiment}
\alias{assay,seurat,ANY-method}
\alias{colData,seurat-method}
\alias{colnames,seurat-method}
\alias{counts,seurat-method}
\alias{metadata,seurat-method}
\alias{metadata<-,seurat-method}
\alias{rowData,seurat-method}
\alias{rowRanges,seurat-method}
\alias{rownames,seurat-method}
\alias{gene2symbol,seurat-method}
\alias{interestingGroups,seurat-method}
\alias{interestingGroups<-,seurat,character-method}
\alias{sampleNames,seurat-method}
\title{Extend S4 Methods for \code{seurat} Class}
\usage{
\S4method{cell2sample}{seurat}(object)

\S4method{plotCellCounts}{seurat}(object, interestingGroups, fill = NULL,
  title = "cell counts")

\S4method{assay}{seurat,ANY}(x)

\S4method{colData}{seurat}(x)

\S4method{colnames}{seurat}(x)

\S4method{counts}{seurat}(object)

\S4method{metadata}{seurat}(x)

\S4method{metadata}{seurat}(x) <- value

\S4method{rowData}{seurat}(x)

\S4method{rowRanges}{seurat}(x)

\S4method{rownames}{seurat}(x)

\S4method{gene2symbol}{seurat}(object)

\S4method{interestingGroups}{seurat}(object)

\S4method{interestingGroups}{seurat,character}(object) <- value

\S4method{sampleNames}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Character vector of interesting groups. Must be
formatted in camel case and intersect with \code{\link[=sampleData]{sampleData()}} colnames.}

\item{fill}{Desired ggplot fill scale. Must supply discrete values. When set
to \code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{title}{Plot title.}

\item{x}{Primary object.}

\item{value}{Object to assign.}

\item{normalized}{Normalized (\code{TRUE}) or raw (\code{FALSE}) counts.}
}
\value{
Match \code{SummarizedExperiment} method return.
}
\description{
Provide limited \code{SingleCellExperiment}-like method support.
}
\seealso{
\code{getMethod("metadata<-", "Annotated")}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
