% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-seurat.R
\docType{methods}
\name{seurat}
\alias{seurat}
\alias{assay,seurat,ANY-method}
\alias{colnames,seurat-method}
\alias{metadata,seurat-method}
\alias{rownames,seurat-method}
\alias{metadata<-,seurat-method}
\title{Extend S4 Methods for \code{seurat} Objects}
\usage{
\S4method{assay}{seurat,ANY}(x)

\S4method{colnames}{seurat}(x)

\S4method{metadata}{seurat}(x)

\S4method{rownames}{seurat}(x)

\S4method{metadata}{seurat}(x) <- value
}
\description{
Extend S4 Methods for \code{seurat} Objects
}
\examples{
load(system.file("extdata/seurat_small.rda", package = "bcbioSingleCell"))

# assay ====
assay(pbmc_small) \%>\% glimpse()

# dimensions ====
rownames(pbmc_small) \%>\% head()
colnames(pbmc_small) \%>\% head()

# metadata ====
# Only works for objects created with bcbioSingleCell
names(metadata(seurat_small))

# Assignment method support
metadata(seurat_small)[["stash"]] <- "XXX"
metadata(seurat_small)[["stash"]]
}
\seealso{
\code{getMethod("metadata<-", "Annotated")}
}
\author{
Michael Steinbaugh
}
