% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{seurat_small}
\alias{seurat_small}
\title{Minimal \code{seurat} Example}
\format{An object of class \code{seurat} of length 1.}
\usage{
seurat_small
}
\description{
Minimal \code{seurat} Example
}
\author{
Michael Steinbaugh
}
\keyword{internal}
