% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-tx2gene.R,
%   R/methods-aggregate_replicates.R
\name{.strip_transcript_versions}
\alias{.strip_transcript_versions}
\alias{.aggregate_sparse_features}
\alias{.aggregate_sparse_replicates}
\title{Sparse matrix operations}
\usage{
.strip_transcript_versions(sparse)

.aggregate_sparse_features(sparse, featureids)

.aggregate_sparse_replicates(sparse, cellids)
}
\arguments{
\item{sparse}{Sparse counts matrix.}

\item{featureids}{New feature IDs for the collapsed features.}

\item{cellids}{New cellular ids of the collapsed cells, one for each cell.}
}
\value{
Sparse counts matrix with the transcript versions stripped.

Modified sparse matrix.
}
\description{
Strip transcript versions

Collapse features with the same feature id by summing them.

Pool cellular technical replicate counts.
}
\author{
Rory Kirchner

Michael Steinbaugh
}
