% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{aggregate_sparse_features}
\alias{aggregate_sparse_features}
\alias{aggregate_sparse_replicates}
\alias{strip_transcript_versions}
\title{Sparse matrix operations}
\usage{
aggregate_sparse_features(sparse, featureids)

aggregate_sparse_replicates(sparse, cellids)

strip_transcript_versions(sparse)
}
\arguments{
\item{sparse}{Sparse counts matrix.}

\item{featureids}{New feature IDs for the collapsed features.}

\item{cellids}{New cellular ids of the collapsed cells, one for each cell.}
}
\value{
Modified sparse matrix.

Sparse counts matrix with the transcript versions stripped.
}
\description{
Collapse features with the same feature id by summing them.

Pool cellular technical replicate counts.

Strip transcript versions
}
\author{
Rory Kirchner

Michael Steinbaugh
}
