% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_transcript_versions.R
\name{strip_transcript_versions}
\alias{strip_transcript_versions}
\title{Strip transcript versions off of the rownames of a dataframe or matrix}
\usage{
strip_transcript_versions(matrix)
}
\arguments{
\item{matrix}{a matrix or dataframe}
}
\value{
a matrix or dataframe with the transcript versions stripped off
}
\description{
Strip transcript versions off of the rownames of a dataframe or matrix
}
\author{
Rory Kirchner
}
