% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/subsetPerSample-methods.R
\docType{methods}
\name{subsetPerSample}
\alias{subsetPerSample}
\alias{subsetPerSample,bcbioSingleCell-method}
\title{Subset Per Sample}
\usage{
subsetPerSample(object, ...)

\S4method{subsetPerSample}{bcbioSingleCell}(object, minCells = 200L,
  prefilter = TRUE, assignAndSave = FALSE, envir = parent.frame(),
  dir = ".")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{minCells}{\code{scalar integer}. Minimum number of cells required per
sample.}

\item{prefilter}{\code{boolean}. Apply prefiltering to remove zero count genes.}

\item{assignAndSave}{\code{boolean}. Assign and save the individual datasets.}

\item{envir}{\code{environment}. Where to assign the subsets. Only applicable when
\code{assignAndSave = TRUE}.}

\item{dir}{\code{string}. Output directory. Only applicable when
\code{assignAndSave = TRUE}.}
}
\value{
\itemize{
\item \code{assignAndSave = FALSE}: Per sample objects in a \code{list}.
\item \code{assignAndSave = TRUE}: Subset file paths.
}
}
\description{
Subset Per Sample
}
\examples{
# List mode (default)
list <- subsetPerSample(indrops_small, assignAndSave = FALSE)
names(list)

# Assign and save mode (useful for large datasets)
subsetPerSample(
    object = indrops_small,
    assignAndSave = TRUE,
    envir = parent.frame(),
    dir = "subsetPerSample"
)
list.files("subsetPerSample")

# Clean up
unlink("subsetPerSample", recursive = TRUE)
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}},
  \code{\link{metricsPerSample}}, \code{\link{sampleData}},
  \code{\link{selectSamples}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
