% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-subsetPerSample.R
\docType{methods}
\name{subsetPerSample}
\alias{subsetPerSample}
\alias{subsetPerSample}
\alias{subsetPerSample,bcbioSingleCell-method}
\title{Subset Per Sample}
\usage{
subsetPerSample(object, ...)

\S4method{subsetPerSample}{bcbioSingleCell}(object, minCells = 200L,
  envir = parent.frame(), dir = ".")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{minCells}{Minimum number of cells required per sample.}

\item{envir}{Environment where to assign the subsets.}

\item{dir}{Output directory.}
}
\value{
Named vector of saved \code{bcbioSingleCell} subset file paths.
}
\description{
Subset Per Sample
}
\examples{
# bcbioRNASeq ====
# This will save the subsets per sample to disk
subsetPerSample(bcb_small, dir = "subsetPerSample")
list.files("subsetPerSample")

# Clean up
unlink("subsetPerSample", recursive = TRUE)
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{calculateMetrics}},
  \code{\link{cell2sample}}, \code{\link{colData}},
  \code{\link{counts}}, \code{\link{fetchGeneData}},
  \code{\link{fetchPCAData}}, \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}},
  \code{\link{interestingGroups}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{rowData}}, \code{\link{sampleData}},
  \code{\link{selectSamples}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
