% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-subsetPerSample.R
\docType{methods}
\name{subsetPerSample}
\alias{subsetPerSample}
\alias{subsetPerSample}
\alias{subsetPerSample,bcbioSingleCell-method}
\title{Subset Per Sample}
\usage{
subsetPerSample(object, ...)

\S4method{subsetPerSample}{bcbioSingleCell}(object, minCells = 200L,
  envir = parent.frame(), dir = ".")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{minCells}{Minimum number of cells required per sample.}

\item{envir}{Environment where to assign the subsets.}

\item{dir}{Output directory.}
}
\value{
Named vector of saved \code{bcbioSingleCell} subset file paths.
}
\description{
Subset Per Sample
}
\examples{
load(system.file("extdata/filtered.rda", package = "bcbioSingleCell"))

# This will save the subsets per sample to disk
subsetPerSample(filtered, dir = "subsetPerSample")
dir("subsetPerSample")

# Clean up
unlink("subsetPerSample", recursive = TRUE)
}
\seealso{
Other Data Management Utilities: \code{\link{aggregateReplicates}},
  \code{\link{filterCells}},
  \code{\link{readCellTypeMarkersFile}}
}
\author{
Michael Steinbaugh
}
