% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-tenx.R
\name{.cellranger}
\alias{.cellranger}
\title{Read \href{https://www.10xgenomics.com/software/}{10x Genomics Chromium} output}
\usage{
.cellranger(upload_dir, aggregate = TRUE)
}
\arguments{
\item{upload_dir}{Final upload directory.}

\item{aggregate}{Aggregate samples into single sparse matrix.}
}
\value{
Sparse counts matrix (\code{dgCMatrix}).
}
\description{
Read into sparse matrix from barcodes.tsv, genes.tsv, and matrix.mtx files.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
