% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-top_barcodes.R
\docType{methods}
\name{top_barcodes}
\alias{top_barcodes}
\alias{top_barcodes,bcbioSCDataSet-method}
\title{Top barcodes}
\usage{
top_barcodes(object, ...)

\S4method{top_barcodes}{bcbioSCDataSet}(object, n = 2)
}
\arguments{
\item{object}{\link{bcbioSCDataSet}.}

\item{n}{Number of barcodes to return per sample.}
}
\value{
\link{data.frame}.
}
\description{
Top barcodes
}
