% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-top_barcodes.R
\docType{methods}
\name{top_barcodes}
\alias{top_barcodes}
\alias{.top_barcodes}
\alias{top_barcodes,bcbioSCDataSet-method}
\alias{top_barcodes,SCSubset-method}
\title{Top barcodes}
\usage{
\S4method{top_barcodes}{bcbioSCDataSet}(object, n = 2L)

\S4method{top_barcodes}{SCSubset}(object, n = 2L)
}
\arguments{
\item{object}{Primary object.}

\item{n}{Number of barcodes to return per sample.}
}
\value{
\link{tibble}.
}
\description{
Top barcodes
}
