% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{top_barcodes}
\alias{top_barcodes}
\title{Top barcodes}
\usage{
top_barcodes(run, n = 2)
}
\arguments{
\item{run}{\link{bcbioSCDataSet}.}

\item{n}{Number of barcodes to return per sample.}
}
\description{
Top barcodes
}
