% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-top_markers.R
\docType{methods}
\name{top_markers}
\alias{top_markers}
\alias{top_markers,data.frame-method}
\title{Top markers}
\usage{
\S4method{top_markers}{data.frame}(object, n = 4L)
}
\arguments{
\item{object}{Primary object.}

\item{n}{Number of genes per cluster.}
}
\value{
\link{tibble}.
}
\description{
Top markers
}
