% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{top_markers}
\alias{top_markers}
\title{Top marker genes}
\usage{
top_markers(markers, n = 4L)
}
\arguments{
\item{markers}{Markers.}

\item{n}{Number of genes per cluster.}
}
\value{
\link{data.frame}
}
\description{
Top marker genes
}
