% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-topBarcodes.R
\docType{methods}
\name{topBarcodes}
\alias{topBarcodes}
\alias{topBarcodes}
\alias{topBarcodes,bcbioSingleCell-method}
\alias{topBarcodes,seurat-method}
\title{Top Barcodes}
\usage{
topBarcodes(object, ...)

\S4method{topBarcodes}{bcbioSingleCell}(object, n = 10L)

\S4method{topBarcodes}{seurat}(object, n = 10L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{n}{Number of barcodes to return per sample.}
}
\value{
\code{data.frame}.
}
\description{
Top Barcodes
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioSingleCell"))

# bcbioSingleCell ====
topBarcodes(bcb_small) \%>\% glimpse()

# seurat ====
topBarcodes(pbmc_small) \%>\% glimpse()
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkers}}, \code{\link{plotPCElbow}},
  \code{\link{plotTopMarkers}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
