% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-topBarcodes.R
\docType{methods}
\name{topBarcodes}
\alias{topBarcodes}
\alias{topBarcodes}
\alias{topBarcodes,bcbioSingleCell-method}
\alias{topBarcodes,seurat-method}
\title{Top Barcodes}
\usage{
topBarcodes(object, ...)

\S4method{topBarcodes}{bcbioSingleCell}(object, n = 10L)

\S4method{topBarcodes}{seurat}(object, n = 10L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{n}{Number of barcodes to return per sample.}
}
\value{
\code{data.frame}.
}
\description{
Top Barcodes
}
\examples{
# bcbioSingleCell ====
topBarcodes(bcb_small) \%>\% glimpse()

# seurat ====
topBarcodes(pbmc_small) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{calculateMetrics}},
  \code{\link{cell2sample}}, \code{\link{colData}},
  \code{\link{counts}}, \code{\link{fetchGeneData}},
  \code{\link{fetchPCAData}}, \code{\link{fetchTSNEData}},
  \code{\link{fetchTSNEExpressionData}},
  \code{\link{filterCells}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{rowData}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
