% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/topBarcodes-methods.R
\docType{methods}
\name{topBarcodes}
\alias{topBarcodes}
\alias{topBarcodes,SingleCellExperiment-method}
\title{Top Barcodes}
\usage{
topBarcodes(object, ...)

\S4method{topBarcodes}{SingleCellExperiment}(object, n = 1000L,
  return = c("data.frame", "list"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{n}{\code{scalar integer}. Number of barcodes to return per sample.}

\item{return}{\code{string}. Return type. Uses \code{\link[base:match.arg]{base::match.arg()}} internally and
defaults to the first argument in the \code{character} vector.}
}
\value{
\itemize{
\item "\code{list}": \code{list} containing top barcodes as a \code{character} vector, split by
\code{sampleID}.
\item "\code{data.frame}": \code{tibble} grouped by \code{sampleID} and arranged by \code{nUMI}
column in descending order. Cellular barcodes are in the \code{cellID} column.
}
}
\description{
Obtain the top cellular barcodes based on UMI counts.
}
\examples{
# Return list
x <- topBarcodes(indrops_small, return = "list")
lapply(x, class)

# Return data.frame
x <- topBarcodes(indrops_small, return = "data.frame")
glimpse(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateCols}},
  \code{\link{cell2sample}},
  \code{\link{metricsPerSample}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
