% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/topBarcodes-methods.R
\docType{methods}
\name{topBarcodes}
\alias{topBarcodes}
\alias{topBarcodes}
\alias{topBarcodes,SingleCellExperiment-method}
\alias{topBarcodes,seurat-method}
\title{Top Barcodes}
\usage{
topBarcodes(object, ...)

\S4method{topBarcodes}{SingleCellExperiment}(object, n = 1000L,
  return = c("data.frame", "list"))

\S4method{topBarcodes}{seurat}(object, n = 1000L, return = c("data.frame",
  "list"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{n}{Number of barcodes to return per sample.}

\item{return}{Return type. Uses \code{\link[base:match.arg]{base::match.arg()}} internally and defaults
to the first argument in the vector.}
}
\value{
\itemize{
\item "\code{list}": \code{list} containing top barcodes as a \code{character} vector, split by
\code{sampleID}.
\item "\code{data.frame}": \code{tibble} grouped by \code{sampleID} and arranged by \code{nUMI}
column in descending order. Cellular barcodes are in the \code{cellID} column.
}
}
\description{
Obtain the top cellular barcodes based on UMI counts.
}
\examples{
# SingleCellExperiment ====
# list
x <- topBarcodes(cellranger_small, return = "list")
lapply(x, class)

# data.frame
x <- topBarcodes(cellranger_small, return = "data.frame")
glimpse(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{fetchGeneData}},
  \code{\link{metricsPerSample}}, \code{\link{sampleData}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
