% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-topBarcodes.R
\docType{methods}
\name{topBarcodes}
\alias{topBarcodes}
\alias{topBarcodes}
\alias{topBarcodes,bcbioSingleCell-method}
\alias{topBarcodes,seurat-method}
\title{Top Barcodes}
\usage{
topBarcodes(object, ...)

\S4method{topBarcodes}{bcbioSingleCell}(object, n = 10)

\S4method{topBarcodes}{seurat}(object, n = 10)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{n}{Number of barcodes to return per sample.}
}
\value{
\link{data.frame}
}
\description{
Top Barcodes
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioSingleCell"))
load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
topBarcodes(bcb) \%>\% glimpse()

# seurat
topBarcodes(seurat) \%>\% glimpse()
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotKnownMarkers}},
  \code{\link{plotMarkers}}, \code{\link{plotPCElbow}},
  \code{\link{plotTopMarkers}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
