% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-topBarcodes.R
\docType{methods}
\name{topBarcodes}
\alias{topBarcodes}
\alias{topBarcodes}
\alias{topBarcodes,bcbioSCDataSet-method}
\alias{topBarcodes,bcbioSCSubset-method}
\title{Top Barcodes}
\usage{
topBarcodes(object, ...)

\S4method{topBarcodes}{bcbioSCDataSet}(object, n = 2L)

\S4method{topBarcodes}{bcbioSCSubset}(object, n = 2L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{n}{Number of barcodes to return per sample.}
}
\value{
\link{tibble}.
}
\description{
Top Barcodes
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{plotClusters}},
  \code{\link{plotKnownMarkers}},
  \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
