% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-topBarcodes.R
\docType{methods}
\name{topBarcodes}
\alias{topBarcodes}
\alias{topBarcodes}
\alias{topBarcodes,SingleCellExperiment-method}
\alias{topBarcodes,seurat-method}
\title{Top Barcodes}
\usage{
topBarcodes(object, ...)

\S4method{topBarcodes}{SingleCellExperiment}(object, n = 10L)

\S4method{topBarcodes}{seurat}(object, n = 10L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{n}{Number of barcodes to return per sample.}
}
\value{
\code{grouped_df} containing \code{\link[=metrics]{metrics()}} return. Grouped by \code{sampleID}
and arranged by \code{nUMI} column in descending order.
}
\description{
Top Barcodes
}
\examples{
# SingleCellExperiment ====
x <- topBarcodes(cellranger_small)
glimpse(x)
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{cell2sample}}, \code{\link{fetchGeneData}},
  \code{\link{filterCells}},
  \code{\link{mapCellsToSamples}},
  \code{\link{metricsPerSample}}, \code{\link{metrics}},
  \code{\link{selectSamples}},
  \code{\link{subsetPerSample}}
}
\author{
Michael Steinbaugh
}
