% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/topBarcodes-methods.R
\docType{methods}
\name{topBarcodes}
\alias{topBarcodes}
\alias{topBarcodes,SingleCellExperiment-method}
\title{Top Barcodes}
\usage{
topBarcodes(object, ...)

\S4method{topBarcodes}{SingleCellExperiment}(object, n = 100L,
  return = c("tibble", "list"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{n}{\code{scalar integer}. Number of barcodes to return per sample.}

\item{return}{\code{string}. Return type. Uses \code{\link[base:match.arg]{base::match.arg()}} internally and
defaults to the first argument in the \code{character} vector.}
}
\value{
\itemize{
\item "\code{tibble}": \code{grouped_df}. Grouped by \code{sampleID} and arranged by \code{nUMI}
column in descending order. Cellular barcodes are in the \code{cellID} column.
\item "\code{list}": \code{list}. Top barcodes as \code{character}, split by \code{sampleID}.
}
}
\description{
Obtain the top cellular barcodes based on UMI counts.
}
\examples{
object <- indrops_small

# tibble
x <- topBarcodes(object), return = "tibble")
glimpse(x)

# list
x <- topBarcodes(object, return = "list")
lapply(x, class)
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
