% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-topMarkers.R
\docType{methods}
\name{topMarkers}
\alias{topMarkers}
\alias{topMarkers,grouped_df-method}
\title{Top Markers}
\usage{
topMarkers(object, ...)

\S4method{topMarkers}{grouped_df}(object, n = 10, coding = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{n}{Number of genes per cluster.}

\item{coding}{Only include protein coding genes.}
}
\value{
\link{tibble}.
}
\description{
Top Markers
}
\seealso{
\itemize{
\item \code{\link[dplyr:slice]{dplyr::slice()}}
\item \code{\link[dplyr:top_n]{dplyr::top_n()}}.
}

Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{pcCutoff}}, \code{\link{plotKnownMarkers}},
  \code{\link{plotMarkers}}, \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Utilities}
