% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-seurat-topMarkers.R
\docType{methods}
\name{topMarkers}
\alias{topMarkers}
\alias{topMarkers}
\alias{topMarkers,grouped_df-method}
\title{Top Markers}
\usage{
topMarkers(object, ...)

\S4method{topMarkers}{grouped_df}(object, n = 10L, direction = c("positive",
  "negative", "both"), coding = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{n}{Number of genes per cluster.}

\item{direction}{Whether to include "\code{positive}", "\code{negative}", or "\code{both}"
directions of association per cluster.}

\item{coding}{Only include protein coding genes.}
}
\value{
\code{grouped_df}.
}
\description{
Top Markers
}
\examples{
# grouped_df ====
topMarkers(
    object = all_markers_small,
    n = 2L,
    direction = "positive"
)
}
\seealso{
\itemize{
\item \code{\link[dplyr:slice]{dplyr::slice()}}.
\item \code{\link[dplyr:top_n]{dplyr::top_n()}}.
}

Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarkerTSNE}}, \code{\link{plotMarkers}},
  \code{\link{plotPCA}}, \code{\link{plotPCElbow}},
  \code{\link{plotTSNE}}, \code{\link{plotTopMarkers}},
  \code{\link{sanitizeMarkers}}
}
\author{
Michael Steinbaugh
}
