% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topMarkers.R
\name{topMarkers}
\alias{topMarkers}
\title{Top Markers}
\usage{
topMarkers(data, n = 10L, direction = c("positive", "negative",
  "both"), coding = FALSE)
}
\arguments{
\item{data}{\code{grouped_df}. \code{Seurat::FindAllMarkers()} return, sanitized with #
our \code{sanitizeMarkers()} function.}

\item{n}{\code{scalar integer}. Number of genes per cluster.}

\item{direction}{\code{string}. Whether to include "\code{positive}", "\code{negative}", or
"\code{both}" directions of association per cluster.}

\item{coding}{\code{boolean}. Include only protein coding genes.}
}
\value{
\code{grouped_df}.
}
\description{
Top Markers
}
\examples{
# grouped_df ====
x <- topMarkers(
    data = all_markers_small,
    n = 2L,
    direction = "positive"
)
head(x)
}
\seealso{
\itemize{
\item \code{\link[dplyr:slice]{dplyr::slice()}}.
\item \code{\link[dplyr:top_n]{dplyr::top_n()}}.
}

Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeatureTSNE}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotTSNE}},
  \code{\link{sanitizeSeuratMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Functions}
