% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-topMarkers.R
\docType{methods}
\name{topMarkers}
\alias{topMarkers}
\alias{topMarkers,data.frame-method}
\title{Top Markers}
\usage{
topMarkers(object, ...)

\S4method{topMarkers}{data.frame}(object, n = 4L, show = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{n}{Number of genes per cluster.}

\item{show}{Show \link{kable}.}
}
\value{
\link{tibble}.
}
\description{
Top Markers
}
\seealso{
Other Clustering Utilities: \code{\link{knownMarkersDetected}},
  \code{\link{pcCutoff}}, \code{\link{plotClusters}},
  \code{\link{plotKnownMarkers}},
  \code{\link{plotStressGenes}},
  \code{\link{plotTopMarkers}}, \code{\link{topBarcodes}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Utilities}
