% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-updateObject.R
\docType{methods}
\name{updateObject}
\alias{updateObject}
\alias{updateObject,bcbioSingleCell-method}
\title{Update Object}
\usage{
\S4method{updateObject}{bcbioSingleCell}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Update Object
}
\note{
SingleCellExperiment migration started in v0.0.31.
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell up to v0.0.30
# Extended from SummarizedExperiment instead of SingleCellExperiment
error <- tryCatch(
    validObject(bcb),
    error = function(e) e
)
print(error)

# Update object to SingleCellExperiment class extension
bcbSCE <- updateObject(bcb)
validObject(bcbSCE)
is(bcbSCE, "SingleCellExperiment")
bcbSCE
}
\author{
Michael Steinbaugh
}
