% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-updateObject.R
\docType{methods}
\name{updateObject}
\alias{updateObject}
\alias{updateObject,bcbioSingleCell-method}
\title{Update Object}
\usage{
\S4method{updateObject}{bcbioSingleCell}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Update Object
}
\note{
SingleCellExperiment migration started in v0.0.31.
}
\examples{
loadRemoteData("http://bcbiosinglecell.seq.cloud/v0.0.26/bcb.rda")

# bcbioSingleCell prior to v0.0.31 extended from SummarizedExperiment instead
of SingleCellExperiment
metadata(bcb)[["version"]]
error <- tryCatch(
    validObject(bcb),
    error = function(e) e
)
print(error)
bcbSCE <- updateObject(bcb)
validObject(bcbSCE)
is(bcbSCE, "SingleCellExperiment")
bcbSCE
}
\author{
Michael Steinbaugh
}
