% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-updateObject.R
\docType{methods}
\name{updateObject}
\alias{updateObject}
\alias{updateObject,bcbioSingleCell-method}
\title{Update Object}
\usage{
\S4method{updateObject}{bcbioSingleCell}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{bcbioRNASeq}.
}
\description{
Update old objects created by the bcbioSingleCell package. The session
information metadata is preserved from the time when the bcbio data was
originally loaded into R.
}
\examples{
updateObject(indrops_small)
}
\seealso{
Other S4 Class Definition: \code{\link{bcbioSingleCell}},
  \code{\link{coerce}}, \code{\link{show}},
  \code{\link{subset}}
}
\author{
Michael Steinbaugh
}
