% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_counts.R
\name{write_counts}
\alias{write_counts}
\title{Write MatrixMart sparse matrix and supporting files}
\usage{
write_counts(counts, gzip = TRUE)
}
\arguments{
\item{counts}{Sparse counts matrix.}

\item{gzip}{Compress matrix file with gzip.}
}
\value{
Path to matrix file.
}
\description{
Outputs matrix (.mtx), sample barcodes (.colnames), and gene names
(.rownames) as separate files.
}
\author{
Rory Kirchner

Michael Steinbaugh
}
